### 2025-04-11 #8043 @editorialbot
[PRE REVIEW]: StateSpaceDynamics.jl: A Julia package for probabilistic state space models (SSMs)
**Submitting author:** @rsenne (<a href="http://orcid.org/0000-0003-3776-4576">Ryan Aaron Senne</a>)
**Repository:** https://github.com/depasquale-lab/StateSpaceDynamics.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
