### 2025-04-11 #8043 @espottesmith
@rsenne - Thank you for sending for work to JOSS, I'll be the handling editor for this submission.

    Do you have suggestions for suitable reviewers? You can browse the list of people for suggestions, otherwise I will try to find reviewers but it can take a longer time span.
    You can also investigate the "AI"-suggested papers and their authors (not reviewers) located in the previous posts from our editorial bot.

Thank you in advance.
