### 2025-04-15 #8043 @gdalle
Since I reviewed `dynamax`, it's only fair I give this one a shot! Count me in
