+++ docs/src/HiddenMarkovModels.md
@@ -3,17 +3,21 @@
 A **Hidden Markov Model (HMM)** is a graphical model that describes how systems change over time. When modeling a time series with $T$ observations using an HMM, we assume that the observed data $y_{1:T}$ depends on hidden states $x_{1:T}$ that are not observed. Specifically, an HMM is a type of **state-space model** in which the hidden states are discrete.
 
 The three components of an HMM are as follows:
+
 - **An initial state distribution ($\pi$):** which hidden states we are likely to start in.
 - **A transition matrix ($A$):** how the hidden states evolve over time.
 - **An emission model:** how the hidden states generate the observed data.
 
 The generative model is given by:
 
-$$
-x_1 \sim \text{Cat}(\pi) \\
-x_t \mid x_{t-1} \sim \text{Cat}(A_{x_{t-1}, :}) \\
-y_t \mid x_t \sim p(y_t \mid \theta_{x_t})
-$$
+```math
+\begin{align}
+    x_1 &\sim \text{Cat}(\pi) \\
+    x_t &\mid x_{t-1} \sim \text{Cat}(A_{x_{t-1}, :}) \\
+    y_t &\mid x_t \sim p(y_t \mid \theta_{x_t})
+\end{align}
+```
+
 
 Where:
 
@@ -25,9 +29,9 @@ Where:
 
 The emission model can take many forms: Gaussian, Poisson, Bernoulli, categorical, etc... In the case of a Gaussian emission distribution, this becomes:
 
-$$
+```math
 y_t \mid (x_t = k) \sim \mathcal{N}(\mu_k, \Sigma_k)
-$$
+```
 
 Where:
 
@@ -35,15 +39,18 @@ Where:
 - $\Sigma_k$ is the covariance of the emission distribution for state $k$
 
 # What is a Generalized Linear Model - Hidden Markov Model
+
 A **Hidden Markov Model - Generalized Linear Model (GLM-HMM)** - also known as **Switching Regression Model** - is an extension to classic HMMs where the emission models are state-dependent GLMs that link an observed input to an observed output. This formulation allows each hidden state to define its own regression relationship between inputs and outputs, enabling the model to capture complex, state-dependent dynamics in the data. Currently, StateSpaceDynamics.jl support Gaussian, Bernoulli, Poisson, and Autoregressive GLMs as emission models.
 
 The generative model is as follows:
 
-$$
-x_1 \sim \text{Cat}(\pi) \\
-x_t \mid x_{t-1} \sim \text{Cat}(A_{x_{t-1}, :}) \\
-y_t \mid x_t, u_t \sim p(y_t \mid \theta_{x_t}, u_t)
-$$
+```math
+\begin{align}
+    x_1 &\sim \text{Cat}(\pi) \\
+    x_t &\mid x_{t-1} \sim \text{Cat}(A_{x_{t-1}, :}) \\
+    y_t &\mid x_t, u_t \sim p(y_t \mid \theta_{x_t}, u_t)
+\end{align}
+```
 
 Where:
 
@@ -56,20 +63,21 @@ Where:
 
 For example, if the emission is a Gaussian GLM:
 
-$$
+```math
 y_t \mid (x_t = k), u_t \sim \mathcal{N}(\mu_k + \beta_k^\top u_t, \sigma_k^2)
-$$
+```
 
 Where:
+
 - $\beta_k$ are the regression weights for state $k$
 - $\sigma_k^2$ is the state-dependent variance
 - $\mu_k$ is the state-dependent bias
 
 If the emission is Bernoulli (for binary outputs):
 
-$$
+```math
[…]
