+++ docs/src/LinearDynamicalSystems.md
@@ -14,48 +14,49 @@ The **Gaussian Linear Dynamical System** — typically just referred to as an LD
 The generative model is given by:
 
 ```math
-\begin{align}
+\begin{align*}
     x_t &\sim \mathcal{N}(A x_{t-1}, Q) \\
     y_t &\sim \mathcal{N}(C x_t, R)
-\end{align}
+\end{align*}
 ```
 
 Where:
 
-- $x_t$ is the hidden state at time $t$
-- $y_t$ is the observed data at time $t$
-- $A$ is the state transition matrix
-- $C$ is the observation matrix
-- $Q$ is the process noise covariance
-- $R$ is the observation noise covariance
+- ``x_t`` is the hidden state at time ``t``
+- ``y_t`` is the observed data at time ``t``
+- ``\mathbf{A}`` is the state transition matrix
+- ``\mathbf{C}`` is the observation matrix
+- ``\mathbf{Q}`` is the process noise covariance
+- ``\mathbf{R}`` is the observation noise covariance
 
 This can equivalently be written in equation form:
 
 ```math
-\begin{align}
+\begin{align*}
     x_t &= A x_{t-1} + \epsilon_t \\
     y_t &= C x_t + \eta_t
-\end{align}
+\end{align*}
 ```
 
 Where:
 
-- $\epsilon_t \sim \mathcal{N}(0, Q)$ is the process noise
-- $\eta_t \sim \mathcal{N}(0, R)$ is the observation noise
+- ``\epsilon_t \sim \mathcal{N}(0, Q)`` is the process noise
+- ``\eta_t \sim \mathcal{N}(0, R)`` is the observation noise
 
 # The Poisson Linear Dynamical System
 
 The **Poisson Linear Dynamical System** is a variant of the LDS where the observations are modeled as counts. This is useful in useful in fields like neuroscience where we are often interested in modeling spike count data. To relate the spiking data to the Gaussian latent variable, we use a nonlinear link function, specifically the exponential function. Thus our generative model is given by: 
 
 ```math
-x_t \sim \mathcal{N}(A x_{t-1}, Q) \\
-y_t \sim \text{Poisson}(\text{exp}(Cx_t + b))
-
+\begin{align*}
+    x_t &\sim \mathcal{N}(A x_{t-1}, Q) \\
+    y_t &\sim \text{Poisson}(\text{exp}(Cx_t + b))
+\end{align*}
 ```
 
 Where:
 
-- $b$ is a bias term
+- ``\mathbf{b}`` is a bias term
 
 # Inference in Linear Dynamical Systems
 In StateSpaceDynamics.jl, we directly maximize the complete-data log-likelihood function with respect to the latent states given the data and the parameters of the model. In other words, the **maximum a priori** (MAP) estimate of the latent state path is:
@@ -76,32 +77,32 @@ x^{(i+1)} = x^{(i)} - \left[ \nabla^2 \mathcal{L}(x^{(i)}) \right]^{-1} \nabla \
 
 Where:
 
-- $\mathcal{L}(x)$ is the complete-data log-likelihood:
+- ``\mathcal{L}(x)`` is the complete-data log-likelihood:
 
 ```math
 \mathcal{L}(x) = \log p(x_0) + \sum_{t=2}^T \log p(x_t \mid x_{t-1}) + \sum_{t=1}^T \log p(y_t \mid x_t)
 ```
 
-- $\nabla \mathcal{L}(x)$ is the gradient of the full log-likelihood with respect to all latent states,
-- $\nabla^2 \mathcal{L}(x)$ is the Hessian of the full log-likelihood.
+- ``\nabla \mathcal{L}(x)`` is the gradient of the full log-likelihood with respect to all latent states,
[…]
