### 2025-04-17 d4c7943
Re-pushing fixes to LL and A update for trailized HMM GLM
