### 2025-04-18 #8077 @danielskatz
@gdalle - try again, but without the please. Editorialbot is quite rigid in what it acts on 🙂
