### 2025-04-18 #8077 @gdalle
I thought after having known each other for so long we could start getting a bit warmer, but it seems the bot is cold-hearted as ever
