### 2025-04-24 3dcacd1
fixed logic for multioutput intercepts in poisson and bernoulli regressions
