### 2025-05-03 depasquale-lab/StateSpaceDynamics.jl#50 @gdalle
JOSS review - remarks about paper
Hi there!
I'm one of the reviewers for your JOSS submissions, so I'll be opening a series of issues on this repo to improve the paper, code and docs :)
Related:

- https://github.com/openjournals/joss-reviews/issues/8077

Main remarks about the paper:

- L3: Has James Fourie contributed in meaningful ways to the code? He doesn't appear on the GitHub list of contributors.
- L78: Is the code for the experiments available somewhere?
- L82: Was `dynamax` properly JIT-compiled for the experiments? I had [issues with that](https://github.com/probml/dynamax/issues/359) myself.
- L90: Is it fair to compare these packages if you only perform approximate (Laplace) EM and they perform exact EM?
- L91: Add citation for BenchmarkTools.
- L91: Benchmarking through PythonCall implies the assumption that the Julia-to-Python overhead is negligible, this assumption must be made explicit.
- L94 and L103: The benchmark plots are not readable. One option would be splitting each single plot into three plots, for instance based on the value of $n$.  Using thicker lines with different linestyles (in addition to the different markers) would also be helpful. Moreover, the value of drawing a line when there are only 3 distinct measurements is debatable: perhaps a bar plot would be easier to read?
- L94: For the first benchmark plot, I recommend replacing `julia` with `SSD` in the legend.
- L109: With proper storage, the Hessian should only require linear memory. What representation do you use?

Minor details and typos:

- L10: "is _that_ they assume"
- L16: "open-source" $\implies$ When used as adjectives, compound modifiers like these [get a hyphen](https://www.grammarly.com/blog/punctuation-capitalization/hyphen/#4). This mistake happens at lots of other places in the article, as a result I haven't listed them.
- L28: "modern neuroscientists" is a bit weird, "modern neuroscience" might be better
- L31: "precludes model _inferenece_"
- L35: "computationally _inefficeint_"
- L38: "discrete and continuous _variableS_"
- L143: There is a missing DOI in your bibliography, as spotted by the editorial bot
