### 2025-05-03 depasquale-lab/StateSpaceDynamics.jl#51 @gdalle
JOSS review - remarks about code
Some remarks on the code as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/8077) for your package:

- The code example in the README fails with a `SingularException` when I run it on SSD v0.1.1 and the latest Julia 1.11.

```julia
julia> fit!(plds, observations)
Fitting Poisson LDS via LaPlaceEM...   0%|▏                                                 |  ETA: 0:40:45 ( 2.45  s/it)ERROR: TaskFailedException

    nested task error: PosDefException: matrix is not positive definite; Factorization failed.
    Stacktrace:
      [1] cholesky!(F::SparseArrays.CHOLMOD.Factor{…}, A::SparseArrays.CHOLMOD.Sparse{…}; shift::Float64, check::Bool)
        @ SparseArrays.CHOLMOD ~/.julia/juliaup/julia-1.11.5+0.aarch64.apple.darwin14/share/julia/stdlib/v1.11/SparseArrays/src/solvers/cholmod.jl:1462
      [2] cholesky!
        @ ~/.julia/juliaup/julia-1.11.5+0.aarch64.apple.darwin14/share/julia/stdlib/v1.11/SparseArrays/src/solvers/cholmod.jl:1455 [inlined]
      [3] #cholesky#14
        @ ~/.julia/juliaup/julia-1.11.5+0.aarch64.apple.darwin14/share/julia/stdlib/v1.11/SparseArrays/src/solvers/cholmod.jl:1497 [inlined]
      [4] cholesky
        @ ~/.julia/juliaup/julia-1.11.5+0.aarch64.apple.darwin14/share/julia/stdlib/v1.11/SparseArrays/src/solvers/cholmod.jl:1490 [inlined]
      [5] cholesky
        @ ~/.julia/juliaup/julia-1.11.5+0.aarch64.apple.darwin14/share/julia/stdlib/v1.11/SparseArrays/src/solvers/cholmod.jl:1609 [inlined]
      [6] gaussian_entropy(H::Symmetric{Float64, SparseArrays.SparseMatrixCSC{Float64, Int64}})
        @ StateSpaceDynamics ~/.julia/packages/StateSpaceDynamics/zrsML/src/Utilities.jl:513
      [7] smooth(lds::StateSpaceDynamics.LinearDynamicalSystem{…}, y::Matrix{…}, w::Vector{…})
        @ StateSpaceDynamics ~/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:599
      [8] smooth
        @ ~/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:545 [inlined]
      [9] macro expansion
        @ ~/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:656 [inlined]
     [10] (::StateSpaceDynamics.var"#84#threadsfor_fun#26"{…})(tid::Int64; onethread::Bool)
        @ StateSpaceDynamics ./threadingconstructs.jl:253
     [11] #84#threadsfor_fun
        @ ./threadingconstructs.jl:220 [inlined]
     [12] (::Base.Threads.var"#1#2"{StateSpaceDynamics.var"#84#threadsfor_fun#26"{…}, Int64})()
        @ Base.Threads ./threadingconstructs.jl:154

...and 4 more exceptions.

Stacktrace:
 [1] threading_run(fun::StateSpaceDynamics.var"#84#threadsfor_fun#26"{…}, static::Bool)
   @ Base.Threads ./threadingconstructs.jl:173
 [2] macro expansion
   @ ./threadingconstructs.jl:190 [inlined]
 [3] smooth(lds::StateSpaceDynamics.LinearDynamicalSystem{…}, y::Array{…})
   @ StateSpaceDynamics ~/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:655
 [4] estep(lds::StateSpaceDynamics.LinearDynamicalSystem{…}, y::Array{…})
   @ StateSpaceDynamics ~/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:934
 [5] fit!(lds::StateSpaceDynamics.LinearDynamicalSystem{…}, y::Array{…}; max_iter::Int64, tol::Float64)
   @ StateSpaceDynamics ~/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:1333
 [6] fit!(lds::StateSpaceDynamics.LinearDynamicalSystem{…}, y::Array{…})
   @ StateSpaceDynamics ~/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:1306
 [7] top-level scope
   @ ~/Documents/GitHub/Julia/_Scratchpad/mwe.jl:27
Some type information was truncated. Use `show(err)` to see complete types.
```

- It is usually a good idea to also run tests on the latest stable Julia version (in addition to 1.10).
- The random sampling functions should always accept and pass along an `rng` argument for reproducibility.
- Many structs have fields with abstract element types (like `x::Vector{<:Real}`), which is a [performance pitfall](https://docs.julialang.org/en/v1/manual/performance-tips/#Avoid-fields-with-abstract-containers).
- Correcting these structs and generalizing others may enable type-genericity, e.g. using `Float32` instead of `Float64`, or differentiating through an inference step with ForwardDiff.
- There are many uses of `@inbounds`, is it possible to prove that even for accidentally wrong user inputs the indexing will never be out of bounds?
- Many array slices (`x[i, :]`) could be views instead (`@view(x[i, :])`) to get rid of allocations.
- Instead of implementing your block tridiagonal matrices manually, you could use [BlockBandedMatrices.jl](https://github.com/JuliaLinearAlgebra/BlockBandedMatrices.jl), and maybe contribute implementations upstream if they are missing.
- [JET](https://github.com/aviatesk/JET.jl) reports some errors inside your package. Some can be ignored (typically the union split), some are probably limitations of JET itself (like the ones with `T` at the start), but some are actual issues or typos.

```julia
julia> using JET; JET.test_package(StateSpaceDynamics; target_defined_modules=true)
JET-test failed at /Users/guillaumedalle/.julia/packages/JET/yNWjn/src/JETBase.jl:1046
  Expression: (JET.report_package)(StateSpaceDynamics; toplevel_logger = nothing, target_defined_modules = true)
  ═════ 17 possible errors found ═════
  ┌ StateSpaceDynamics.GaussianStateModel() @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:33
  │ `StateSpaceDynamics.T` is not defined: StateSpaceDynamics.T
  └────────────────────
  ┌ StateSpaceDynamics.GaussianObservationModel() @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:77
  │ `StateSpaceDynamics.T` is not defined: StateSpaceDynamics.T
  └────────────────────
  ┌ StateSpaceDynamics.PoissonObservationModel() @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:121
  │ `StateSpaceDynamics.T` is not defined: StateSpaceDynamics.T
  └────────────────────
  ┌ PoissonLDS(; A::Matrix{…} where T<:Real, C::Matrix{…} where T<:Real, Q::Matrix{…} where T<:Real, log_d::Vector{…} where T<:Real, x0::Vector{…} where T<:Real, P0::Matrix{…} where T<:Real, fit_bool::Vector{…}, obs_dim::Int64, latent_dim::Int64) @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:1394
  │ `StateSpaceDynamics.ethrow` is not defined: StateSpaceDynamics.ethrow
  └────────────────────
  ┌ objective(opt::Union{…}, β_vec::Vector{…}) where T<:Real @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/EmissionModels.jl:513
  │ no matching method found `calc_regularization(::Base.ReshapedArray{T, 2, Vector{T1}, Tuple{}} where {T<:Real, T1<:Real}, ::Float64, ::Bool)` (1/2 union split): regularization = StateSpaceDynamics.calc_regularization(β_mat, ((opt::Union{…}).model::Union{…}).λ::Float64, ((opt::Union{…}).model::Union{…}).include_intercept::Bool)
  └────────────────────
  ┌ objective_gradient!(G::Vector{…}, opt::Union{…}, β_vec::Vector{…}) where T<:Real @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/EmissionModels.jl:529
  │ no matching method found `calc_regularization_gradient(::Base.ReshapedArray{T, 2, Vector{T1}, Tuple{}} where {T<:Real, T1<:Real}, ::Float64, ::Bool)` (1/2 union split): regularization = StateSpaceDynamics.calc_regularization_gradient(β_mat, ((opt::Union{…}).model::Union{…}).λ::Float64, ((opt::Union{…}).model::Union{…}).include_intercept::Bool)
  └────────────────────
  ┌ objective(opt::StateSpaceDynamics.RegressionOptimization{…}, β_vec::Vector{…}) where T<:Real @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/EmissionModels.jl:660
  │ no matching method found `calc_regularization(::Base.ReshapedArray{T, 2, Vector{T1}, Tuple{}} where {T<:Real, T1<:Real}, ::Float64, ::Bool)` (1/2 union split): regularization = StateSpaceDynamics.calc_regularization(β_mat, ((opt::StateSpaceDynamics.RegressionOptimization{…}).model::StateSpaceDynamics.BernoulliRegressionEmission).λ::Float64, ((opt::StateSpaceDynamics.RegressionOptimization{…}).model::StateSpaceDynamics.BernoulliRegressionEmission).include_intercept::Bool)
  └────────────────────
  ┌ objective_gradient!(G::Vector{…}, opt::StateSpaceDynamics.RegressionOptimization{…}, β_vec::Vector{…}) where T<:Real @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/EmissionModels.jl:676
  │ no matching method found `calc_regularization_gradient(::Base.ReshapedArray{T, 2, Vector{T1}, Tuple{}} where {T<:Real, T1<:Real}, ::Float64, ::Bool)` (1/2 union split): regularization = StateSpaceDynamics.calc_regularization_gradient(β_mat, ((opt::StateSpaceDynamics.RegressionOptimization{…}).model::StateSpaceDynamics.BernoulliRegressionEmission).λ::Float64, ((opt::StateSpaceDynamics.RegressionOptimization{…}).model::StateSpaceDynamics.BernoulliRegressionEmission).include_intercept::Bool)
  └────────────────────
  ┌ objective(opt::StateSpaceDynamics.RegressionOptimization{…}, β_vec::Vector{…}) where T<:Real @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/EmissionModels.jl:797
  │ no matching method found `calc_regularization(::Base.ReshapedArray{T, 2, Vector{T1}, Tuple{}} where {T<:Real, T1<:Real}, ::Float64, ::Bool)` (1/2 union split): regularization = StateSpaceDynamics.calc_regularization(β_mat, ((opt::StateSpaceDynamics.RegressionOptimization{…}).model::StateSpaceDynamics.PoissonRegressionEmission).λ::Float64, ((opt::StateSpaceDynamics.RegressionOptimization{…}).model::StateSpaceDynamics.PoissonRegressionEmission).include_intercept::Bool)
  └────────────────────
  ┌ objective_gradient!(G::Vector{…}, opt::StateSpaceDynamics.RegressionOptimization{…}, β_vec::Vector{…}) where T<:Real @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/EmissionModels.jl:815
  │ no matching method found `calc_regularization_gradient(::Base.ReshapedArray{T, 2, Vector{T1}, Tuple{}} where {T<:Real, T1<:Real}, ::Float64, ::Bool)` (1/2 union split): regularization = StateSpaceDynamics.calc_regularization_gradient(β_mat, ((opt::StateSpaceDynamics.RegressionOptimization{…}).model::StateSpaceDynamics.PoissonRegressionEmission).λ::Float64, ((opt::StateSpaceDynamics.RegressionOptimization{…}).model::StateSpaceDynamics.PoissonRegressionEmission).include_intercept::Bool)
  └────────────────────
  ┌  @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/HiddenMarkovModels.jl:131
  │┌  @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/HiddenMarkovModels.jl:131
  ││┌  @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/HiddenMarkovModels.jl:164
  │││ no matching method found `sample(::StateSpaceDynamics.EmissionModel, ::Nothing)`: sample(((model::StateSpaceDynamics.HiddenMarkovModel).B::Vector{StateSpaceDynamics.EmissionModel})[state_sequence[1]::Int64]::StateSpaceDynamics.EmissionModel, X::Nothing)
  ││└────────────────────
  ┌ variational_expectation!(model::StateSpaceDynamics.SwitchingLinearDynamicalSystem, y::Any, FB::StateSpaceDynamics.ForwardBackward, FS::Array{…}) where T<:Real @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/SLDS.jl:279
  │ no matching method found `sufficient_statistics(::Base.ReshapedArray{T, 3, Matrix{T1}, Tuple{}} where {T<:Real, T1<:Real}, ::Base.ReshapedArray{T, 4, Array{T1, 3}, Tuple{}} where {T<:Real, T1<:Real}, ::Any)`, `sufficient_statistics(::Array{T, 3} where T<:Real, ::Base.ReshapedArray{T, 4, Array{T1, 3}, Tuple{}} where {T<:Real, T1<:Real}, ::Any)`, `sufficient_statistics(::Base.ReshapedArray{T, 3, Matrix{T1}, Tuple{}} where {T<:Real, T1<:Real}, ::Array{T, 4} where T<:Real, ::Any)` (3/4 union split): StateSpaceDynamics.sufficient_statistics(StateSpaceDynamics.reshape(tuple(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k]::StateSpaceDynamics.FilterSmooth).x_smooth::Matrix{T} where T<:Real)::Tuple{Matrix{T} where T<:Real}, StateSpaceDynamics.size(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k]::StateSpaceDynamics.FilterSmooth).x_smooth::Matrix{T} where T<:Real)::Tuple{Int64, Int64}, tuple(1)::Tuple{Int64}...)::Union{Base.ReshapedArray{T, 3, Matrix{T1}, Tuple{}} where {T<:Real, T1<:Real}, Array{T, 3} where T<:Real}, StateSpaceDynamics.reshape(tuple(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k]::StateSpaceDynamics.FilterSmooth).p_smooth::Array{T, 3} where T<:Real)::Tuple{Array{T, 3} where T<:Real}, StateSpaceDynamics.size(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k]::StateSpaceDynamics.FilterSmooth).p_smooth::Array{T, 3} where T<:Real)::Tuple{Int64, Int64, Int64}, tuple(1)::Tuple{Int64}...)::Union{Base.ReshapedArray{T, 4, Array{T1, 3}, Tuple{}} where {T<:Real, T1<:Real}, Array{T, 4} where T<:Real}, StateSpaceDynamics.reshape(tuple(inverse_offdiag)::Tuple{Any}, StateSpaceDynamics.size(inverse_offdiag)::Any, tuple(1)::Tuple{Int64}...)::Any)
  └────────────────────
  ┌ variational_expectation!(model::StateSpaceDynamics.SwitchingLinearDynamicalSystem, y::Any, FB::StateSpaceDynamics.ForwardBackward, FS::Array{…}) where T<:Real @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/SLDS.jl:285
  │ no matching method found `calculate_elbo(::StateSpaceDynamics.LinearDynamicalSystem, ::Array{T, 3} where T<:Real, ::Array{T, 4} where T<:Real, ::Array{T, 4} where T<:Real, ::Base.ReshapedArray{T, 4, Array{T1, 3}, Tuple{}} where {T<:Real, T1<:Real}, ::Any, ::Any, ::Any)` (1/2 union split): elbo = StateSpaceDynamics.calculate_elbo(((model::StateSpaceDynamics.SwitchingLinearDynamicalSystem).B::Vector{StateSpaceDynamics.LinearDynamicalSystem})[k]::StateSpaceDynamics.LinearDynamicalSystem, ((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k]::StateSpaceDynamics.FilterSmooth).E_z::Array{T, 3} where T<:Real, ((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k]::StateSpaceDynamics.FilterSmooth).E_zz::Array{T, 4} where T<:Real, ((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k]::StateSpaceDynamics.FilterSmooth).E_zz_prev::Array{T, 4} where T<:Real, StateSpaceDynamics.reshape(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k]::StateSpaceDynamics.FilterSmooth).p_smooth::Array{T, 3} where T<:Real, tuple(StateSpaceDynamics.size(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k]::StateSpaceDynamics.FilterSmooth).p_smooth::Array{T, 3} where T<:Real)::Tuple{Int64, Int64, Int64}, tuple(1)::Tuple{Int64}...)::NTuple{4, Int64})::Union{Base.ReshapedArray{T, 4, Array{T1, 3}, Tuple{}} where {T<:Real, T1<:Real}, Array{T, 4} where T<:Real}, StateSpaceDynamics.reshape(y::Any, tuple(StateSpaceDynamics.size(y::Any)::Any, tuple(1)::Tuple{Int64}...)::Tuple)::Any, total_entropy, StateSpaceDynamics.vec((hs::Any)[k, StateSpaceDynamics.:(:)]::Any)::Any)
  └────────────────────
  ┌ variational_expectation!(model::StateSpaceDynamics.SwitchingLinearDynamicalSystem, y::Any, FB::StateSpaceDynamics.ForwardBackward, FS::Array{…}) where T<:Real @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/SLDS.jl:279
  │ no matching method found `sufficient_statistics(::Base.ReshapedArray{T, 3, Matrix{T1}, Tuple{}} where {T<:Real, T1<:Real}, ::Base.ReshapedArray{T, 4, Array{T1, 3}, Tuple{}} where {T<:Real, T1<:Real}, ::Any)`, `sufficient_statistics(::Array{T, 3} where T<:Real, ::Base.ReshapedArray{T, 4, Array{T1, 3}, Tuple{}} where {T<:Real, T1<:Real}, ::Any)`, `sufficient_statistics(::Base.ReshapedArray{T, 3, Matrix{T1}, Tuple{}} where {T<:Real, T1<:Real}, ::Array{T, 4} where T<:Real, ::Any)` (3/4 union split): StateSpaceDynamics.sufficient_statistics(StateSpaceDynamics.reshape(tuple(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k::Int64]::StateSpaceDynamics.FilterSmooth).x_smooth::Matrix{T} where T<:Real)::Tuple{Matrix{T} where T<:Real}, StateSpaceDynamics.size(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k::Int64]::StateSpaceDynamics.FilterSmooth).x_smooth::Matrix{T} where T<:Real)::Tuple{Int64, Int64}, tuple(1)::Tuple{Int64}...)::Union{Base.ReshapedArray{T, 3, Matrix{T1}, Tuple{}} where {T<:Real, T1<:Real}, Array{T, 3} where T<:Real}, StateSpaceDynamics.reshape(tuple(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k::Int64]::StateSpaceDynamics.FilterSmooth).p_smooth::Array{T, 3} where T<:Real)::Tuple{Array{T, 3} where T<:Real}, StateSpaceDynamics.size(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k::Int64]::StateSpaceDynamics.FilterSmooth).p_smooth::Array{T, 3} where T<:Real)::Tuple{Int64, Int64, Int64}, tuple(1)::Tuple{Int64}...)::Union{Base.ReshapedArray{T, 4, Array{T1, 3}, Tuple{}} where {T<:Real, T1<:Real}, Array{T, 4} where T<:Real}, StateSpaceDynamics.reshape(tuple(inverse_offdiag::Any)::Tuple{Any}, StateSpaceDynamics.size(inverse_offdiag::Any)::Any, tuple(1)::Tuple{Int64}...)::Any)::Tuple{Array{T, 3} where T<:Real, Any, Any}
  └────────────────────
  ┌ variational_expectation!(model::StateSpaceDynamics.SwitchingLinearDynamicalSystem, y::Any, FB::StateSpaceDynamics.ForwardBackward, FS::Array{…}) where T<:Real @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/SLDS.jl:285
  │ no matching method found `calculate_elbo(::StateSpaceDynamics.LinearDynamicalSystem, ::Array{T, 3} where T<:Real, ::Array{T, 4} where T<:Real, ::Array{T, 4} where T<:Real, ::Base.ReshapedArray{T, 4, Array{T1, 3}, Tuple{}} where {T<:Real, T1<:Real}, ::Any, ::Any, ::Any)` (1/2 union split): elbo = StateSpaceDynamics.calculate_elbo(((model::StateSpaceDynamics.SwitchingLinearDynamicalSystem).B::Vector{StateSpaceDynamics.LinearDynamicalSystem})[k::Int64]::StateSpaceDynamics.LinearDynamicalSystem, ((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k::Int64]::StateSpaceDynamics.FilterSmooth).E_z::Array{T, 3} where T<:Real, ((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k::Int64]::StateSpaceDynamics.FilterSmooth).E_zz::Array{T, 4} where T<:Real, ((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k::Int64]::StateSpaceDynamics.FilterSmooth).E_zz_prev::Array{T, 4} where T<:Real, StateSpaceDynamics.reshape(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k::Int64]::StateSpaceDynamics.FilterSmooth).p_smooth::Array{T, 3} where T<:Real, tuple(StateSpaceDynamics.size(((FS::Array{StateSpaceDynamics.FilterSmooth{T}, 1} where T<:Real)[k::Int64]::StateSpaceDynamics.FilterSmooth).p_smooth::Array{T, 3} where T<:Real)::Tuple{Int64, Int64, Int64}, tuple(1)::Tuple{Int64}...)::NTuple{4, Int64})::Union{Base.ReshapedArray{T, 4, Array{T1, 3}, Tuple{}} where {T<:Real, T1<:Real}, Array{T, 4} where T<:Real}, StateSpaceDynamics.reshape(y::Any, tuple(StateSpaceDynamics.size(y::Any)::Any, tuple(1)::Tuple{Int64}...)::Tuple)::Any, total_entropy::Any, StateSpaceDynamics.vec((hs::Any)[k::Int64, StateSpaceDynamics.:(:)]::Any)::Any)::Any
  └────────────────────
  ┌ mstep!(slds::StateSpaceDynamics.SwitchingLinearDynamicalSystem, FS::Array{…}, y::Matrix{…}, FB::StateSpaceDynamics.ForwardBackward) where T<:Real @ StateSpaceDynamics /Users/guillaumedalle/.julia/packages/StateSpaceDynamics/zrsML/src/SLDS.jl:428
[…]
