### 2025-05-03 depasquale-lab/StateSpaceDynamics.jl#52 @gdalle
JOSS review - remarks about docs
Some remarks on the (stable) documentation as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/8077) of your paper:

## General

- Please make sure that all examples included in the documentation are included as [doctests](https://documenter.juliadocs.org/stable/man/doctests/).
- It seems you document every single object and put it on the corresponding page. There needs to be a clearer separation between public and private API. By default, in Julia, what is documented is public unless stated otherwise. Starting with Julia 1.11, the `public` keyword enables a clearer distinction.

## Home

- The first example on the home page fails with SSD v0.1.1 and the latest Julia 1.11.

```julia
julia> x, y = sample(lds, 1000)  # 1000 timepoints
ERROR: MethodError: no method matching sample(::StateSpaceDynamics.LinearDynamicalSystem{…}, ::Int64)
The function `sample` exists, but no method is defined for this combination of argument types.

Closest candidates are:
  sample(::StateSpaceDynamics.LinearDynamicalSystem{S, O}, ::Int64, ::Int64) where {T<:Real, S<:StateSpaceDynamics.GaussianStateModel{T}, O<:StateSpaceDynamics.GaussianObservationModel{T}}
   @ StateSpaceDynamics ~/.julia/packages/StateSpaceDynamics/zrsML/src/LinearDynamicalSystems.jl:299
  sample(::SwitchingLinearDynamicalSystem, ::Int64)
   @ StateSpaceDynamics ~/.julia/packages/StateSpaceDynamics/zrsML/src/SLDS.jl:16
  sample(::GaussianMixtureModel, ::Int64)
   @ StateSpaceDynamics ~/.julia/packages/StateSpaceDynamics/zrsML/src/MixtureModels.jl:46
  ...

Stacktrace:
 [1] top-level scope
   @ ~/Documents/GitHub/Julia/_Scratchpad/mwe.jl:10
Some type information was truncated. Use `show(err)` to see complete types.
```

- The link to the repo is outdated.
- Contribution guidelines should be expanded a little.
- For citation, you could add a [Zenodo](https://zenodo.org/) DOI.

## Models

- It might help to intersperse the docstrings with the text? You can force them to be collapsed by default to avoid taking too much space.
- The math doesn't display properly. For the list you might need the backtick syntax (instead of dollar signs) to avoid these alignment issues.
- Why does it make sense to construct the default `GaussianStateModel` with matrices of size 0?

---

I think it would be easier for me to review the docs once the major issues pointed out above have been fixed
