### 2025-05-20 03ccbce
refactored function signatures for HMM.jl
