+++ src/HiddenMarkovModels.jl
@@ -4,7 +4,6 @@ export kmeans_init!
 # for unit tests
 export estep
 export class_probabilities
-
 """
     HiddenMarkovModel
 
@@ -17,11 +16,15 @@ A Hidden Markov Model (HMM) with custom emissions.
 - `A::Matrix{<:Real}`: Transition matrix.
 - `πₖ::Vector{Float64}`: Initial state distribution.
 """
-mutable struct HiddenMarkovModel <: AbstractHMM
-    A::Matrix{<:Real} # transition matrix
+mutable struct HiddenMarkovModel{T<:Real} <: AbstractHMM
+    A:: Matrix{T} # transition matrix
     B::Vector{EmissionModel} # Vector of emission Models
     πₖ::Vector{Float64} # initial state distribution
     K::Int # number of states
+
+    function HiddenMarkovModel(A::Matrix{Float64}, B::Vector{EmissionModel}, πₖ::Vector{Float64}, K::Int)
+        new{Float64}(A, B, πₖ, K)
+    end
 end
 
 function initialize_forward_backward(model::AbstractHMM, num_obs::Int)
@@ -66,7 +69,7 @@ function HiddenMarkovModel(;
     K::Int,
     B::Vector=Vector(),
     emission=nothing,
-    A::Matrix{<:Real}=initialize_transition_matrix(K),
+    A::AbstractMatrix{<:Real}=initialize_transition_matrix(K),
     πₖ::Vector{Float64}=initialize_state_distribution(K),
 )
 
@@ -81,12 +84,14 @@ function HiddenMarkovModel(;
         for i in (length(B) + 1):K
             push!(B, deepcopy(emission))
         end
+        
     end
 
     emission_models = B
     #emission_models = Emission.(B)
 
-    model = HiddenMarkovModel(A, emission_models, πₖ, K)
+    A64  = to_f64(A)
+    model = HiddenMarkovModel(A64, emission_models, πₖ, K)
 
     # check that the transition matrix is the proper shape
     @assert size(model.A) == (model.K, model.K)
@@ -128,7 +133,15 @@ Generate `n` samples from a Hidden Markov Model. Returns a tuple of the state se
 - `state_sequence::Vector{Int}`: The state sequence, where each element is an integer 1:K.
 - `observation_sequence::Matrix{Float64}`: The observation sequence. This takes the form of the emission model's output.
 """
-function sample(model::HiddenMarkovModel, X::Union{Matrix{<:Real},Nothing}=nothing; n::Int, autoregressive::Bool=false)
+function sample(model::HiddenMarkovModel, X::AbstractMatrix{<:Real}; n::Int, autoregressive::Bool=false)
+    sample(model, to_f64(X); n=n, autoregressive=autoregressive)
+end
+
+function sample(model::HiddenMarkovModel; n::Int, autoregressive::Bool=false)
+    sample(model, nothing; n=n, autoregressive=autoregressive)
+end
+
+function sample(model::HiddenMarkovModel, X::Union{Matrix{<:Float64},Nothing}=nothing; n::Int, autoregressive::Bool=false)
     
     if autoregressive ==false
         state_sequence = Vector{Int}(undef, n)
@@ -368,10 +381,18 @@ Fit the Hidden Markov Model using the EM algorithm.
 - `max_iters::Int=100`: The maximum number of iterations to run the EM algorithm.
 - `tol::Float64=1e-6`: When the log likelihood is improving by less than this value, the algorithm will stop.
 """
[…]
