### 2025-05-21 ace51dc
Update HiddenMarkovModels.jl
