### 2025-05-21 f093e48
refactored HMM constructor types
