### 2025-05-21 4688f16
refactoring preprocessing types, HMMConstructors
