### 2025-05-21 c7afde4
refactor emission model function signatures
