+++ src/EmissionModels.jl
@@ -22,8 +22,8 @@ GaussianEmission model with mean and covariance.
 """
 mutable struct GaussianEmission <: EmissionModel
     output_dim::Int # dimension of the data
-    μ::Vector{<:Real}  # mean 
-    Σ::Matrix{<:Real}  # covariance matrix
+    μ::Vector{Float64}  # mean 
+    Σ::Matrix{Float64}  # covariance matrix
 end
 
 """
@@ -33,21 +33,21 @@ Functon to create a GaussianEmission with given output dimension, mean, and cova
 
 # Arguments
 - `output_dim::Int`: The output dimension of the emission
-- `μ::Vector{<:Real}=zeros(output_dim)`: The mean of the Gaussian
-- `Σ::Matrix{<:Real}=Matrix{Float64}(I, output_dim, output_dim))`: The covariance matrix of the Gaussian
+- `μ::Vector{Float64}=zeros(output_dim)`: The mean of the Gaussian
+- `Σ::Matrix{Float64}=Matrix{Float64}(I, output_dim, output_dim))`: The covariance matrix of the Gaussian
 
 # Returns
 """
 function GaussianEmission(;
     output_dim::Int,
-    μ::Vector{<:Real}=zeros(output_dim),
-    Σ::Matrix{<:Real}=Matrix{Float64}(I, output_dim, output_dim),
+    μ::Vector{Float64}=zeros(output_dim),
+    Σ::Matrix{Float64}=Matrix{Float64}(I, output_dim, output_dim),
 )
     return GaussianEmission(output_dim, μ, Σ)
 end
 
 """
-    loglikelihood(model::GaussianEmission, Y::Matrix{<:Real})
+    loglikelihood(model::GaussianEmission, Y::Matrix{Float64})
 
 Calculate the log likelihood of the data `Y` given the Gaussian emission model.
 
@@ -59,6 +59,10 @@ Calculate the log likelihood of the data `Y` given the Gaussian emission model.
 - `Vector{Float64}`: A vector of log likelihoods, one for each observation in the data.
 """
 function loglikelihood(model::GaussianEmission, Y::Matrix{<:Real})
+    loglikelihood(model, to_f64(Y))
+end
+
+function loglikelihood(model::GaussianEmission, Y::Matrix{Float64})
     # Create MvNormal distribution with the model parameters
     dist = MvNormal(model.μ, model.Σ)
 
@@ -86,8 +90,12 @@ Fit a GaussianEmission model to the data `Y`.
 - `Y::Matrix{<:Real}`: Data to fit the model to. Should be a matrix of size `(n, output_dim)`.
 - `w::Vector{Float64}=ones(size(Y, 1))`: Weights for the data. Should be a vector of size `n`.
 """
+function fit!(model::GaussianEmission, Y::Matrix{<:Real}, w::Vector{Float64})
+    fit!(model, to_f64(Y), w)
+end 
+
 function fit!(
-    model::GaussianEmission, Y::Matrix{<:Real}, w::Vector{Float64}=ones(size(Y, 1))
+    model::GaussianEmission, Y::Matrix{Float64}, w::Vector{Float64}=ones(size(Y, 1))
 )
     weighted_sum = sum(Y .* w; dims=1)
     new_mean = weighted_sum[:] ./ sum(w)
@@ -112,8 +120,8 @@ Holds the optimization problem data for regression emissions.
 """
 struct RegressionOptimization{T<:RegressionEmission}
     model::T
-    X::Matrix{<:Real}
-    y::Matrix{<:Real}
+    X::Matrix{Float64}
+    y::Matrix{Float64}
     w::Vector{Float64}
     β_shape::Tuple{Int,Int}  # Added to track original shape
 end
@@ -130,7 +138,7 @@ function create_optimization(
     end
 
[…]
