### 2025-05-22 14340b6
corrected signatures in HMMconstructors
