+++ src/EmissionModels.jl
@@ -161,7 +161,7 @@ Calculate L2 regularization term for regression coefficients.
 # Returns
 - `Float64`: The regularization term value
 """
-function calc_regularization(β::Matrix{Float64}, λ::Float64, include_intercept::Bool=true)
+function calc_regularization(β::AbstractMatrix{T}, λ::Float64, include_intercept::Bool=true) where {T<:Number}
     # calculate L2 penalty
     if include_intercept
         regularization = 0.5 * λ * sum(abs2, β[2:end, :])
@@ -183,8 +183,8 @@ Calculate gradient of L2 regularization term for regression coefficients.
 - `include_intercept::Bool`: Whether to exclude the intercept term from regularization
 """
 function calc_regularization_gradient(
-    β::Matrix{Float64}, λ::Float64, include_intercept::Bool=true
-)
+    β::AbstractMatrix{T}, λ::Float64, include_intercept::Bool=true
+) where {T<:Number}
     # calculate the gradient of the regularization component
     regularization = zeros(size(β))
 
@@ -301,6 +301,15 @@ function loglikelihood(
 )
     w = ones(size(Y,1))                   
     return loglikelihood(model, Φ, Y, w)  
+end
+
+function loglikelihood(
+    model::GaussianRegressionEmission,
+    Φ::Matrix{<:Real},
+    Y::Matrix{<:Real},
+    w::AbstractVector{<:Real},
+)
+    loglikelihood(model, to_f64(Φ), to_f64(Y), to_f64(w), )
 end
 
 function loglikelihood(
@@ -480,7 +489,7 @@ function loglikelihood(
     Y::Matrix{<:Real},
     w::Vector{Float64}=ones(size(Y, 1)),
 )
-    return loglikelihood(model.innerGaussianRegression, X, Y, w)
+    return loglikelihood(model.innerGaussianRegression, to_f64(X), to_f64(Y), to_f64(w))
 end
 
 """
@@ -534,8 +543,8 @@ end
 
 
 function objective(
-    opt::Union{RegressionOptimization{GaussianRegressionEmission}, RegressionOptimization{AutoRegressionEmission}}, β_vec::Vector{T}
-) where {T<:AbstractFloat}
+    opt::Union{RegressionOptimization{GaussianRegressionEmission}, RegressionOptimization{AutoRegressionEmission}}, β_vec::AbstractVector{T}
+) where {T<:Number}
     β_mat = vec_to_matrix(β_vec, opt.β_shape)
     residuals = opt.y - opt.X * β_mat
     w_reshaped = reshape(opt.w, :, 1)
@@ -551,8 +560,8 @@ end
 function objective_gradient!(
     G::Vector{Float64},
     opt::Union{RegressionOptimization{GaussianRegressionEmission}, RegressionOptimization{AutoRegressionEmission}},
-    β_vec::Vector{T},
-) where {T<:AbstractFloat}
+    β_vec::AbstractVector{T},
+) where {T<:Number}
     β_mat = vec_to_matrix(β_vec, opt.β_shape)
     residuals = opt.y - opt.X * β_mat
 
@@ -679,6 +688,15 @@ end
 
 function loglikelihood(
     model::BernoulliRegressionEmission,
+    Φ::Matrix{<:Real},
+    Y::Matrix{<:Real},
+    w::AbstractVector{<:Real},
+)
+    loglikelihood(model, to_f64(Φ), to_f64(Y), to_f64(w), )
+end
+
+function loglikelihood(
+    model::BernoulliRegressionEmission,
     Φ::Matrix{Float64},
[…]
