### 2025-05-28 f9de59c
corrected 6 JET errors across package
