### 2025-05-29 b00f3f2
LDS/SLDS abstract types, moving to_f64 methods to Utilities
