+++ src/LinearDynamicalSystems.jl
@@ -1,4 +1,5 @@
 export GaussianLDS, PoissonLDS, sample, smooth, fit!
+import ForwardDiff
 
 """
     GaussianStateModel{T<:Real} <: AbstractStateModel
@@ -11,7 +12,7 @@ Represents the state model of a Linear Dynamical System with Gaussian noise.
 - `x0::Vector{T}`: Initial state
 - `P0::Matrix{T}`: Initial state covariance
 """
-mutable struct GaussianStateModel{T<:Real} <: AbstractStateModel{T}
+mutable struct GaussianStateModel{T<:Float64} <: AbstractStateModel{T}
     A::Matrix{T}
     Q::Matrix{T}
     x0::Vector{T}
@@ -36,7 +37,7 @@ function GaussianStateModel(;
     x0::Vector{T}=Vector{T}(undef, 0),
     P0::Matrix{T}=Matrix{T}(undef, 0, 0),
     latent_dim::Int=0,
-) where {T<:Real}
+) where {T<:Float64}
     if latent_dim == 0 && (isempty(A) || isempty(Q) || isempty(x0) || isempty(P0))
         throw(ArgumentError("Must provide latent_dim if any matrix is not provided."))
     end
@@ -49,6 +50,19 @@ function GaussianStateModel(;
     return GaussianStateModel{T}(A, Q, x0, P0)
 end
 
+function GaussianStateModel(
+    A::AbstractMatrix{T},
+    Q::AbstractMatrix{T},
+    x0::AbstractVector{T},
+    P0::AbstractMatrix{T},
+) where {T<:Real}
+    A64 = to_f64(A)
+    Q64 = to_f64(Q)
+    x064 = to_f64(x0)
+    P064 = to_f64(P0)
+    GaussianStateModel(A64, Q64, x064, P064)
+end
+
 """
     GaussianObservationModel{T<:Real} <: AbstractObservationModel
 
@@ -58,7 +72,7 @@ Represents the observation model of a Linear Dynamical System with Gaussian nois
 - `C::Matrix{T}`: Observation matrix
 - `R::Matrix{T}`: Observation noise covariance
 """
-mutable struct GaussianObservationModel{T<:Real} <: AbstractObservationModel{T}
+mutable struct GaussianObservationModel{T<:Float64} <: AbstractObservationModel{T}
     C::Matrix{T}
     R::Matrix{T}
 end
@@ -79,7 +93,7 @@ function GaussianObservationModel(;
     R::Matrix{T}=Matrix{T}(undef, 0, 0),
     obs_dim::Int=0,
     latent_dim::Int=0,
-) where {T<:Real}
+) where {T<:Float64}
     if obs_dim == 0 && (isempty(C) || isempty(R))
         throw(ArgumentError("Must provide obs_dim if C or R is not provided."))
     end
@@ -93,6 +107,17 @@ function GaussianObservationModel(;
     return GaussianObservationModel{T}(C, R)
 end
 
+function GaussianObservationModel(
+    C::AbstractMatrix{T}, 
+    R::AbstractMatrix{T}, 
+    obs_dim::Int,
+    latent_dim::Int, 
+) where {T<:Real}
+    c64 = to_f64(C)
+    R64 = to_f64(R)
+    GaussianObservationModel(c64, R64, obs_dim, latent_dim)
+end
+
 """
     PoissonObservationModel{T<:Real} <: AbstractObservationModel
 
@@ -102,7 +127,7 @@ Represents the observation model of a Linear Dynamical System with Poisson obser
 - `C::Matrix{T}`: Observation matrix
[…]
