+++ src/SLDS.jl
@@ -3,7 +3,7 @@ export SwitchingLinearDynamicalSystem, fit!, sample, initialize_slds, variationa
 Switching Linear Dynamical System
 """
 mutable struct SwitchingLinearDynamicalSystem <: AbstractHMM
-    A::Matrix{<:Real}                 # Transition matrix for mode switching
+    A::Matrix{Float64}                 # Transition matrix for mode switching
     B::Vector{LinearDynamicalSystem}  # Vector of Linear Dynamical System models
     πₖ::Vector{Float64}               # Initial state distribution
     K::Int                            # Number of modes
@@ -125,8 +125,11 @@ Fit a Switching Linear Dynamical System using the variational Expectation-Maximi
 - `mls::Vector{T}`: Vector of log-likelihood values for each iteration.
 """
 function fit!(
-    slds::SwitchingLinearDynamicalSystem, y::Matrix{T}; 
-    max_iter::Int=1000, tol::Real=1e-3) where {T<:Real}
+    slds::SwitchingLinearDynamicalSystem, 
+    y::AbstractMatrix{Float64}; 
+    max_iter::Int=1000, 
+    tol::Float64=1e-3
+  )
 
     # Initialize log-likelihood
     prev_ml = -T(Inf)
@@ -191,6 +194,16 @@ function fit!(
     return mls, param_diff, FB, FS
 end
 
+function fit!(
+    slds::SwitchingLinearDynamicalSystem,
+    y::AbstractMatrix{T};
+    max_iter::Int = 1000,
+    tol::Float64 = 1e-3,
+) where {T<:Real}
+    y64 = to_f64(y)
+    fit!(slds, y64; max_iter=max_iter, tol=tol)
+end
+
 
 """
     variational_expectation!(model::SwitchingLinearDynamicalSystem, y, FB, FS) -> Float64
@@ -244,7 +257,7 @@ This function performs the variational expectation step for a Switching Linear D
 elbo = variational_expectation!(model, y, FB, FS)
 println("Computed ELBO: ", elbo)
 """
-function variational_expectation!(model::SwitchingLinearDynamicalSystem, y, FB::ForwardBackward, FS::Vector{FilterSmooth{T}}) where {T<:Real}
+function variational_expectation!(model::SwitchingLinearDynamicalSystem, y, FB::ForwardBackward, FS::Vector{FilterSmooth{T}}) where {T<:Float64}
   # For now a hardcoded tolerance
   tol = 1e-6
   # Get starting point for iterative E-step
@@ -382,7 +395,7 @@ variational_qs!(model, FB, y, FS)
 println(FB.loglikelihoods)
 """
 function variational_qs!(model::Vector{GaussianObservationModel{T}}, FB::ForwardBackward, 
-  y, FS::Vector{FilterSmooth{T}}) where {T<:Real}
+  y, FS::Vector{FilterSmooth{T}}) where {T<:Float64}
 
   T_steps = size(y, 2)
   K = length(model)
@@ -403,7 +416,7 @@ end
 """
 """
 function mstep!(slds::SwitchingLinearDynamicalSystem,
-    FS::Vector{FilterSmooth{T}}, y::Matrix{T}, FB::ForwardBackward) where {T<:Real}
+    FS::Vector{FilterSmooth{T}}, y::Matrix{T}, FB::ForwardBackward) where {T<:Float64}
 
     K = slds.K
 
