+++ test/HiddenMarkovModels/SwitchingPoissonRegression.jl
@@ -130,3 +130,117 @@ function sample_states(num_samples, initial_probs, transition_matrix)
     end
     return states
 end
+
+
+function test_sample_non_f64_SwitchingPoissonRegression()
+  
+    emission_1 = PoissonRegressionEmission(;
+        input_dim=3, output_dim=1, include_intercept=true, β=reshape([4, 3, 2, 4], :, 1)
+    )
+    emission_2 = PoissonRegressionEmission(;
+        input_dim=3, output_dim=1, include_intercept=true, β=reshape([-4, -2, 1, 3], :, 1)
+    )
+
+    # Initialize the SwitchingPoissonRegression
+    true_model = SwitchingPoissonRegression(; K=2, input_dim=3, output_dim=1)
+
+    # Plug in the emission models
+    true_model.B[1] = emission_1
+    true_model.B[2] = emission_2
+
+    n = 1000
+    Φ_int = randn(3, n)
+    Φ_f32  = Float32.(randn(3, n))
+
+    # int features
+    labels_int, Y_int = StateSpaceDynamics.sample(true_model, Φ_int; n=n)
+    @test length(labels_int) == n
+    @test eltype(Y_int) == Float64
+    @test all(Y_int .>= 0)                 
+    @test all(round.(Y_int) .== Y_int)    
+
+    # Float32 features
+    labels_f32, Y_f32 = StateSpaceDynamics.sample(true_model, Φ_f32; n=n)
+    @test length(labels_f32) == n
+    @test eltype(Y_f32) == Float64
+    @test all(Y_f32 .>= 0)
+    @test all(round.(Y_f32) .== Y_f32)
+end
+
+function test_fit_non_f64_SwitchingPoissonRegression()
+  
+    num_trials = 50  # Number of trials
+    trial_length = 1000  # Number of time steps per trial
+
+    # Create the emission models
+    emission_1 = PoissonRegressionEmission(;
+        input_dim=3, output_dim=1, include_intercept=true, β=reshape([4, 3, 2, 4], :, 1)
+    )
+    emission_2 = PoissonRegressionEmission(;
+        input_dim=3, output_dim=1, include_intercept=true, β=reshape([-4, -2, 1, 3], :, 1)
+    )
+
+    # Initialize the SwitchingPoissonRegression model
+    true_model = SwitchingPoissonRegression(; K=2, input_dim=3, output_dim=1)
+    true_model.A = [0.9 0.1; 0.2 0.8]
+
+    # Plug in the emission models
+    true_model.B[1] = emission_1
+    true_model.B[2] = emission_2
+
+    # Create lists to hold data and labels for each trial
+    Φ_trials      = [randn(3, trial_length) for _ in 1:num_trials]
+    data_trials   = Vector{Matrix{Float64}}(undef, num_trials)
+    labels_trials = Vector{Vector{Int}}(undef, num_trials)
+
+    # Sample data for each trial
+    for t in 1:num_trials
+        labels_t, data_t = StateSpaceDynamics.sample(true_model, Φ_trials[t]; n=trial_length)
+        labels_trials[t] = labels_t
+        data_trials[t]  = data_t
+    end 
+   
+    all_Y_int = [Int.(round.(Y)) for Y in data_trials]
+    all_Y_f32 = [Float32.(Y)       for Y in data_trials]
+
+    all_Φ_int    = [Int.(round.(Φ))   for Φ in Φ_trials]
+    all_Φ_f32    = [Float32.(Φ)       for Φ in Φ_trials]
+
+    #int model 
[…]
