+++ test/HiddenMarkovModels/GaussianHMM.jl
@@ -185,3 +185,66 @@ function test_incomplete_initialization()
     @test model.B[1].μ == model.B[2].μ
     @test model.B[1].Σ == model.B[2].Σ
 end
+
+function test_sample_non_f64()
+    
+    model = GaussianHMM(K=1, output_dim=3)
+    μ = [0.0,  1.0,  -1.0]
+    Σ = 0.5 * I(3)
+    model.B[1] = GaussianEmission(output_dim=3, μ=μ, Σ=Σ)
+
+    n = 100
+    true_labels, Y = StateSpaceDynamics.sample(model; n=n)
+
+    @test length(true_labels) == n
+    @test size(Y) == (3, n)
+    @test eltype(Y) == Float64
+end
+
+function test_fit_non_f64()
+
+    output_dim = 2
+    μ = [-5.0, -4.0]
+    Σ = 0.1 * Matrix{Float64}(I, output_dim, output_dim)
+    emission_1 = GaussianEmission(; output_dim=output_dim, μ=μ, Σ=Σ)
+
+    μ = [2.0, 1.0]
+    Σ = 0.1 * Matrix{Float64}(I, output_dim, output_dim)
+    emission_2 = GaussianEmission(; output_dim=output_dim, μ=μ, Σ=Σ)
+
+    true_model = GaussianHMM(; K=2, output_dim=2)
+    true_model.B[1] = emission_1
+    true_model.B[2] = emission_2
+    true_model.A = [0.9 0.1; 0.8 0.2]
+
+    n = 1000
+    num_trials = 50
+    Y = Vector{Matrix{Float64}}(undef, num_trials)
+    trial_labels = Vector{Vector{Int}}(undef, num_trials)
+
+    for i in 1:num_trials
+        true_labels, data = StateSpaceDynamics.sample(true_model; n=n) 
+        Y[i] = data  
+    end
+
+    all_data = hcat(Y...)     
+
+    Y_int = Int.(round.(all_data))
+    Y_f32 = Float32.(all_data) 
+
+    #create two models with different types 
+    model_int = GaussianHMM(; K=2, output_dim=2)
+    ll_int = StateSpaceDynamics.fit!(model_int, Y_int; max_iters=100)
+
+    @test all(eltype(ll_int) .== Float64)
+    @test eltype(model_int.B[1].μ) == Float64
+    @test eltype(model_int.B[1].Σ) == Float64
+
+    model_f32 = GaussianHMM(; K=2, output_dim=2)
+    ll_f32 = StateSpaceDynamics.fit!(model_f32, Y_f32)
+
+    @test all(eltype(ll_f32) .== Float64)
+    @test eltype(model_f32.B[1].μ) == Float64
+    @test eltype(model_f32.B[1].Σ) == Float64
+
+end

[…]
