### 2025-06-02 967ea11
Paper edits
