+++ paper/paper.md
@@ -30,45 +30,44 @@ bibliography: paper.bib
 
 # Summary
 
-State-space models (SSMs) are powerful tools for modeling time series data that naturally arise in a variety of domains, including neuroscience, finance, and engineering. The unifying principle of these models is they assume an observation sequence, $Y_1, Y_2,...,Y_T$, is generated through an underlying latent sequence, $X_1, X_2,...,X_T$. This framework encompasses two popular models for time series analysis: the hidden Markov model (HMM) and the (Gaussian) linear dynamical system (LDS, i.e., the Kalman filter). Thus, SSMs provide a probabilistic framework for describing the temporal evolution of many phenomena, and their generality naturally leads to a variety of use cases. We introduce `StateSpaceDynamics.jl` [@SSDjl2024], an open source, modular package designed to be fast, readable, and self contained for the purpose of easily fitting a plurality of SSMs in the Julia language.
+State-space models (SSMs) are powerful tools for modeling time series data that naturally arise in a variety of domains, including neuroscience, finance, and engineering. The unifying principle of these models is that they assume an observation sequence, $Y_1, Y_2,...,Y_T$, is generated through an underlying latent sequence, $X_1, X_2,...,X_T$. This framework encompasses two popular models for time series analysis: the hidden Markov model (HMM) and the (Gaussian) linear dynamical system (LDS, i.e., the Kalman filter). Thus, SSMs provide a probabilistic framework for describing the temporal evolution of many phenomena, and their generality naturally leads to a variety of use cases. We introduce `StateSpaceDynamics.jl` [@SSDjl2024], an open-source, modular package designed to be fast, readable, and self-contained for the purpose of easily fitting a plurality of SSMs in the Julia language.
 
 # Statement of need
 
 Advances in neuroscience have enabled the collection of massive, multivariate, and complex time series datasets, where simultaneous observations from hundreds to thousands of neurons are increasingly common. Interpreting these high-dimensional datasets presents significant challenges. Recent modeling approaches suggest that neural activity can be characterized by a set of latent factors evolving within a low-dimensional manifold. Consequently, there is a growing need for models that combine dimensionality reduction with temporal dynamics, for which state-space models provide a natural framework.
 
-While state space model implementations exist in Python, such as the SSM package [@PySSM2022] and Dynamax [@ChangUnknown-wn], the Julia programming language lacks an equivalent that meets the needs of modern neuroscientists. Existing Julia offerings, like `StateSpaceModels.jl` [@SSMjl2024], can accomdodate continuous-state SSMs (e.g., LDS) but are limited to Gaussian observation models and rely on analytical calculation of the marginal log-likelihood. This latter limitation precludes model inferenece and parameter learning for non-conjugate observations which are common in neuroscience, where neural activity follow Poisson or other discrete distributions. Packages for performing inference and learning using sampling based methods exist in Julia (such as `Turing.jl` [@Turing2018]) but are computationally inefficeint compared to tailored approaches based on Expectation-Maximization (EM). For discrete SSMs, an existing Julia offering, `HiddenMarkovModels.jl` [@HMMjl2024], is efficient and scalable but not intentionally designed with the functionality for mixing models that contain both discrete and continuous latent variable, such as the switching linear dynamical system model (SLDS) [@Linderman2016-xe; @slds] increasingly used in neuroscience.
[…]
