+++ paper/paper.bib
@@ -88,6 +88,7 @@
   author       = {Paninski, Liam and Ahmadian, Yashar and Ferreira, Daniel Gil
                   and Koyama, Shinsuke and Rahnama Rad, Kamiar and Vidne,
                   Michael and Vogelstein, Joshua and Wu, Wei},
+  doi          = {10.1007/s10827-009-0179-x},
   journaltitle = {J. Comput. Neurosci.},
   publisher    = {Springer Science and Business Media LLC},
   volume       = {29},
@@ -137,6 +138,7 @@ state space models},
   number={1},
   pages={65--98},
   year={2017},
+  doi={10.48550/arXiv.1411.1607},
   publisher={SIAM}
 }
 
@@ -147,6 +149,7 @@ state space models},
   journaltitle = {arXiv [stat.ML]},
   date         = {2016-10-26},
   eprintclass  = {stat.ML},
+  doi          = {10.48550/arXiv.1610.08466},
   abstract     = {Many natural systems, such as neurons firing in the brain or
                   basketball teams traversing a court, give rise to time series
                   data with complex, nonlinear dynamics. We can gain insight
@@ -228,3 +231,18 @@ state space models},
   url       = {https://www.nature.com/articles/s41593-021-01007-z}
 }
 
+@software{pythoncall,
+  title    = {PythonCall.jl},
+  author   = {Christopher Doris},
+  date     = {2021},
+  howpublished = {\url{https://github.com/JuliaPy/PythonCall.jl}},
+  url = {https://github.com/JuliaPy/PythonCall.jl},
+}
+
+@software{bench,
+  title    = {BenchmarkTools.jl},
+  author   = {Jarrett Revels},
+  date     = {2017},
+  howpublished = {\url{https://github.com/JuliaCI/BenchmarkTools.jl}},
+  url = {https://github.com/JuliaCI/BenchmarkTools.jl},
+}
