### 2025-06-02 e7edd93
refactoring matrix slices to @view
