### 2025-06-04 c238c62
Switches problematic LineSearch method "MoreThuente" to "HagerZhang"
