+++ src/LinearDynamicalSystems.jl
@@ -1,4 +1,4 @@
-export GaussianLDS, PoissonLDS, sample, smooth, fit!
+export GaussianLDS, PoissonLDS, rand, smooth, fit!
 
 """
     GaussianStateModel{T<:Real} <: AbstractStateModel
@@ -277,47 +277,71 @@ end
 
 
 """
-    sample(lds::LinearDynamicalSystem{S,O}, T_steps::Int, n_trials::Int) where {T<:Real, S<:GaussianStateModel{T}, O<:GaussianObservationModel{T}}
+    Random.rand(rng::AbstractRNG, lds::LinearDynamicalSystem; tsteps::Int, ntrials::Int)
 
-Sample from a Linear Dynamical System (LDS) model for multiple trials.
+Sample directly from a Linear Dynamical System.
 
 # Arguments
-- `lds::LinearDynamicalSystem{S,O}`: The Linear Dynamical System model.
-- `T_steps::Int`: The number of time steps to sample for each trial.
-- `n_trials::Int`: The number of trials to sample.=
+- `rng::AbstractRNG`: Random number generator
+- `lds::LinearDynamicalSystem`: The LDS model
+- `tsteps::Int`: Number of time steps to sample
+- `ntrials::Int`: Number of trials to sample
 
 # Returns
-- `x::Array{T, 3}`: The latent state variables. Dimensions: (latent_dim, T_Steps, n_trials)
-- `y::Array{T, 3}`: The observed data. Dimensions: (obs_dim, T_steps, n_trials)
-
-# Examples
-```julia
-lds = GaussianLDS(obs_dim=4, latent_dim=3)
-x, y = sample(lds, 10, 100)  # 10 trials, 100 time steps each
-```
+- `Tuple{Array,Array}`: Latent states and observations
 """
-function sample(
-    lds::LinearDynamicalSystem{S,O}, T_steps::Int, n_trials::Int
-) where {T<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
-    A, Q, x0, P0 = lds.state_model.A,
-    lds.state_model.Q, lds.state_model.x0,
-    lds.state_model.P0
+function Random.rand(rng::AbstractRNG, lds::LinearDynamicalSystem{S,O}; tsteps::Int, ntrials::Int) where {T<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
+    A, Q, x0, P0 = lds.state_model.A, lds.state_model.Q, lds.state_model.x0, lds.state_model.P0
     C, R = lds.obs_model.C, lds.obs_model.R
 
-    x = Array{T,3}(undef, lds.latent_dim, T_steps, n_trials)
-    y = Array{T,3}(undef, lds.obs_dim, T_steps, n_trials)
+    x = Array{T,3}(undef, lds.latent_dim, tsteps, ntrials)
+    y = Array{T,3}(undef, lds.obs_dim, tsteps, ntrials)
 
-    for trial in 1:n_trials
-        x[:, 1, trial] = rand(MvNormal(x0, P0))
-        y[:, 1, trial] = rand(MvNormal(C * x[:, 1, trial], R))
+    for trial in 1:ntrials
+        x[:, 1, trial] = rand(rng, MvNormal(x0, P0))
+        y[:, 1, trial] = rand(rng, MvNormal(C * x[:, 1, trial], R))
 
-        for t in 2:T_steps
-            x[:, t, trial] = rand(MvNormal(A * x[:, t - 1, trial], Q))
-            y[:, t, trial] = rand(MvNormal(C * x[:, t, trial], R))
+        for t in 2:tsteps
+            x[:, t, trial] = rand(rng, MvNormal(A * x[:, t - 1, trial], Q))
[…]
