### 2025-06-04 aee1ac6
Removed old sample function.
