### 2025-06-04 c460c08
configured Poisson LDS tests
