+++ src/LinearDynamicalSystems.jl
@@ -1,5 +1,4 @@
 export GaussianLDS, PoissonLDS, sample, smooth, fit!
-import ForwardDiff
 
 """
     GaussianStateModel{T<:Real} <: AbstractStateModel
@@ -380,7 +379,7 @@ function loglikelihood(
     lds::LinearDynamicalSystem{S,O}, 
     y::AbstractMatrix{U},
     w::Vector{Float64}=ones(size(y, 2))
-) where {T<:Real,U<:Real,W<:Float64,S<:GaussianStateModel{W},O<:GaussianObservationModel{W}}
+) where {T<:Real,U<:Float64,S<:GaussianStateModel{U},O<:GaussianObservationModel{U}}
     T_steps = size(y, 2)
     A, Q, x0, P0 = lds.state_model.A,
     lds.state_model.Q, lds.state_model.x0,
@@ -590,7 +589,7 @@ x, p_smooth, inverse_offdiag, Q_val = DirectSmoother(lds, y)
 """
 function smooth(
     lds::LinearDynamicalSystem{S,O}, y::Matrix{T}, w::Vector{Float64}=ones(size(y, 2))
-) where {T<:Real,S<:GaussianStateModel{<:Float64},O<:AbstractObservationModel{<:Float64}}
+) where {T<:Float64,S<:GaussianStateModel{<:Float64},O<:AbstractObservationModel{<:Float64}}
     T_steps, D = size(y, 2), lds.latent_dim
 
     # set initial "solution" and preallocate x_reshape
@@ -681,7 +680,7 @@ x, p_smooth, inverse_offdiag = smooth(lds, y)
 """
 function smooth(
     lds::LinearDynamicalSystem{S,O}, y::Array{T,3}
-) where {T<:Real,S<:GaussianStateModel{<:Float64},O<:AbstractObservationModel{<:Float64}}
+) where {T<:Float64,S<:GaussianStateModel{<:Float64},O<:AbstractObservationModel{<:Float64}}
     obs_dim, T_steps, n_trials = size(y)
     latent_dim = lds.latent_dim
 
@@ -734,7 +733,7 @@ function Q_state(
     E_z::AbstractMatrix{T},
     E_zz::AbstractArray{T,3},
     E_zz_prev::AbstractArray{T,3},
-) where {T<:Real}
+) where {T<:Float64}
     T_step = size(E_z, 2)
     state_dim = size(A, 1)
     
@@ -800,7 +799,7 @@ function Q_obs(
     E_z::AbstractVector{T},
     E_zz::AbstractMatrix{T},
     y::AbstractVector{T},
-) where {T<:Real}
+) where {T<:Float64}
 
     obs_dim = size(H, 1)
 
@@ -846,7 +845,7 @@ function Q_obs(
     E_zz::AbstractArray{T,3},
     y::AbstractMatrix{T},
     weights::Vector{Float64}=ones(size(y, 2))
-) where {T<:Real}
+) where {T<:Float64}
     obs_dim = size(H, 1)
     T_step = size(E_z, 2)
     
@@ -904,7 +903,7 @@ function Q_function(
     E_zz_prev::AbstractArray{T,3},
     y::AbstractMatrix{T},
     weights::Vector{Float64}=ones(size(y, 2))
-) where {T<:Real}
+) where {T<:Float64}
     Q_val_state = Q_state(A, Q, P0, x0, E_z, E_zz, E_zz_prev)
     Q_val_obs = Q_obs(C, R, E_z, E_zz, y, weights)
     return Q_val_state + Q_val_obs
@@ -931,7 +930,7 @@ Compute sufficient statistics for the EM algorithm in a Linear Dynamical System.
 """
[…]
