### 2025-06-05 b7a99f8
refactored structs using T notation
