+++ src/EmissionModels.jl
@@ -20,10 +20,10 @@ Gaussian Emission Models
 
 GaussianEmission model with mean and covariance.
 """
-mutable struct GaussianEmission <: EmissionModel
+mutable struct GaussianEmission{T<:Float64} <: EmissionModel
     output_dim::Int # dimension of the data
-    μ::Vector{Float64}  # mean 
-    Σ::AbstractMatrix{Float64}  # covariance matrix
+    μ::AbstractVector{T}  # mean 
+    Σ::AbstractMatrix{T}  # covariance matrix
 end
 
 """
@@ -40,10 +40,10 @@ Functon to create a GaussianEmission with given output dimension, mean, and cova
 """
 function GaussianEmission(;
     output_dim::Int,
-    μ::Vector{Float64}=zeros(output_dim),
-    Σ::AbstractMatrix{Float64}=Matrix{Float64}(I, output_dim, output_dim),
-)
-    return GaussianEmission(output_dim, μ, Σ)
+    μ::Vector{T}=zeros(output_dim),
+    Σ::AbstractMatrix{T}=Matrix{Float64}(I, output_dim, output_dim),
+) where {T<:Float64}
+    return GaussianEmission{T}(output_dim, μ, Σ)
 end
 
 """
@@ -217,31 +217,31 @@ A Gaussian regression Emission model.
 - `Σ::Matrix{<:Real} = Matrix{Float64}(I, output_dim, output_dim)`: Covariance matrix of the model.
 - `λ::Float64 = 0.0`: Regularization parameter.
 """
-mutable struct GaussianRegressionEmission <: RegressionEmission
+mutable struct GaussianRegressionEmission{T<:Float64} <: RegressionEmission
     input_dim::Int
     output_dim::Int
-    β::Matrix{Float64} # coefficient matrix of the model. Shape input_dim by output_dim. Column one is coefficients for target one, etc. The first row are the intercept terms, if included. 
-    Σ::AbstractMatrix{Float64} # covariance matrix of the model 
+    β::AbstractMatrix{T} # coefficient matrix of the model. Shape input_dim by output_dim. Column one is coefficients for target one, etc. The first row are the intercept terms, if included. 
+    Σ::AbstractMatrix{T} # covariance matrix of the model 
     include_intercept::Bool # whether to include an intercept term; if true, the first column of β is assumed to be the intercept/bias
-    λ::Float64 # regularization parameter
+    λ::T # regularization parameter
 end
 
 function GaussianRegressionEmission(;
     input_dim::Int,
     output_dim::Int,
     include_intercept::Bool=true,
-    β::AbstractMatrix{<:Real}=if include_intercept
+    β::AbstractMatrix{T}=if include_intercept
         zeros(input_dim + 1, output_dim)
     else
         zeros(input_dim, output_dim)
     end,
-    Σ::AbstractMatrix{<:Real}=Matrix{Float64}(I, output_dim, output_dim),
+    Σ::AbstractMatrix{U}=Matrix{Float64}(I, output_dim, output_dim),
     λ::Float64=0.0,
-)
+) where {T<:Real, U<:Real}
     βf = to_f64(β)
     Σf = to_f64(Σ)
 
-    GaussianRegressionEmission(input_dim, output_dim, βf, Σf, include_intercept, λ)
+    GaussianRegressionEmission{Float64}(input_dim, output_dim, βf, Σf, include_intercept, λ)
 end
 
 """
@@ -257,11 +257,11 @@ Generate `n` samples from a Gaussian regression model. Returns a matrix of size 
 # Returns
[…]
