### 2025-06-06 ff0ea04
Increase type genericity
