+++ src/LinearDynamicalSystems.jl
@@ -6,10 +6,10 @@ export GaussianLDS, PoissonLDS, sample, smooth, fit!
 Represents the state model of a Linear Dynamical System with Gaussian noise.
 
 # Fields
-- `A::Matrix{T}`: Transition matrix
-- `Q::Matrix{T}`: Process noise covariance
-- `x0::Vector{T}`: Initial state
-- `P0::Matrix{T}`: Initial state covariance
+- `A::AbstractMatrix{T}`: Transition matrix
+- `Q::AbstractMatrix{T}`: Process noise covariance
+- `x0::AbstractVector{T}`: Initial state
+- `P0::AbstractMatrix{T}`: Initial state covariance
 """
 mutable struct GaussianStateModel{T<:Real, M<:AbstractMatrix{T}, V<:AbstractVector{T}} <: AbstractStateModel{T}
     A::M
@@ -24,17 +24,17 @@ end
 Construct a GaussianStateModel with the given parameters or random initializations.
 
 # Arguments
-- `A::Matrix{T}=Matrix{T}(undef, 0, 0)`: Transition matrix
-- `Q::Matrix{T}=Matrix{T}(undef, 0, 0)`: Process noise covariance
-- `x0::Vector{T}=Vector{T}(undef, 0)`: Initial state
-- `P0::Matrix{T}=Matrix{T}(undef, 0, 0)`: Initial state covariance
+- `A::AbstractMatrix{T}=Matrix{T}(undef, 0, 0)`: Transition matrix
+- `Q::AbstractMatrix{T}=Matrix{T}(undef, 0, 0)`: Process noise covariance
+- `x0::AbstractVector{T}=Vector{T}(undef, 0)`: Initial state
+- `P0::AbstractMatrix{T}=Matrix{T}(undef, 0, 0)`: Initial state covariance
 - `latent_dim::Int`: Dimension of the latent state (required if any matrix is not provided.)
 """
 function GaussianStateModel(;
-    A::Matrix{T}=Matrix{T}(undef, 0, 0),
-    Q::Matrix{T}=Matrix{T}(undef, 0, 0),
-    x0::Vector{T}=Vector{T}(undef, 0),
-    P0::Matrix{T}=Matrix{T}(undef, 0, 0),
+    A::AbstractMatrix{T}=Matrix{T}(undef, 0, 0),
+    Q::AbstractMatrix{T}=Matrix{T}(undef, 0, 0),
+    x0::AbstractVector{T}=Vector{T}(undef, 0),
+    P0::AbstractMatrix{T}=Matrix{T}(undef, 0, 0),
     latent_dim::Int=0,
 ) where {T<:Real}
     if latent_dim == 0 && (isempty(A) || isempty(Q) || isempty(x0) || isempty(P0))
@@ -55,10 +55,10 @@ end
 Represents the observation model of a Linear Dynamical System with Gaussian noise.
 
 # Fields
-- `C::Matrix{T}`: Observation matrix
-- `R::Matrix{T}`: Observation noise covariance
+- `C::AbstractMatrix{T}`: Observation matrix
+- `R::AbstractMatrix{T}`: Observation noise covariance
 """
-mutable struct GaussianObservationModel{T<:Float64, M<:AbstractMatrix{T}} <: AbstractObservationModel{T}
+mutable struct GaussianObservationModel{T<:Real, M<:AbstractMatrix{T}} <: AbstractObservationModel{T}
     C::M
     R::M
 end
@@ -69,14 +69,14 @@ end
 Construct a GaussianObservationModel with the given parameters or random initializations.
 
 # Arguments
-- `C::Matrix{T}=Matrix{T}(undef, 0, 0)`: Observation matrix
-- `R::Matrix{T}=Matrix{T}(undef, 0, 0)`: Observation noise covariance
[…]
