### 2025-06-06 f38a648
Makes PPCA adhere to type general standards
