+++ src/Preprocessing.jl
@@ -13,13 +13,13 @@ Probabilistic PCA model from Bishop's Pattern Recognition and Machine Learning.
     D: Number of features
     z: Latent variables
 """
-mutable struct ProbabilisticPCA
-    W::Matrix{<:AbstractFloat} # weight matrix
-    σ²::AbstractFloat # noise variance
-    μ::Matrix{<:AbstractFloat} # mean of the data
+mutable struct ProbabilisticPCA{T<:Real, M<:AbstractMatrix{T}, V<:AbstractVector{T}}
+    W::M # weight matrix
+    σ²::T # noise variance
+    μ::V # mean of the data
     k::Int # number of latent dimensions
     D::Int # dimension of the data
-    z::Matrix{<:AbstractFloat} # latent variables
+    z::M # latent variables
 end
 
 """
@@ -42,20 +42,47 @@ end
 # ppca = ProbabilisticPCA(W=rand(2, 1), σ²=0.1, μ=rand(2), k=1, D=2)
 # ```
 # """
-function ProbabilisticPCA(;
-    W::Matrix{<:AbstractFloat}=Matrix{Float64}(undef, 0, 0),
-    μ::Matrix{<:AbstractFloat}=Matrix{Float64}(undef, 0, 0),
-    σ²::AbstractFloat=0.0,
+function ProbabilisticPCA(
+    ::Type{T} = Float64;
+    W::Union{AbstractMatrix{T}, Nothing} = nothing,
+    μ::Union{AbstractVector{T}, Nothing} = nothing, 
+    σ²::Union{T, Nothing} = nothing,
     k::Int,
-    D::Int,
-)
-    # if W is not provided, initialize it randomly
-    W = isempty(W) ? rand(D, k) / sqrt(k) : W
-    # if σ² is not provided, initialize it randomly
-    σ² = σ² === 0.0 ? abs(rand()) : σ²
-    # add empty z
-    z = Matrix{Float64}(undef, 0, 0)
-    return ProbabilisticPCA(W, σ², μ, k, D, z)
+    D::Int
+) where {T<:Real}
+    
+    # Initialize W with proper type
+    if W === nothing
+        W_matrix = randn(T, D, k) / sqrt(T(k))
+    else
+        @assert size(W) == (D, k) "W must have size ($D, $k)"
+        W_matrix = W
+    end
+    
+    # Initialize μ with proper type
+    if μ === nothing
+        μ_vector = zeros(T, D)  # Will be set during fitting
+    else
+        @assert length(μ) == D "μ must have length $D"
+        μ_vector = μ
+    end
+    
+    # Initialize σ²
+    if σ² === nothing
+        σ²_val = one(T)  # Default to 1.0 in type T
+    else
+        @assert σ² > zero(T) "σ² must be positive"
+        σ²_val = σ²
+    end
+    
+    # Empty latent variables - will be allocated during fitting
+    z_matrix = Matrix{T}(undef, 0, k)
+    
+    # Infer matrix and vector types from the actual objects
+    M = typeof(W_matrix)
+    V = typeof(μ_vector)
+    
+    return ProbabilisticPCA{T, M, V}(W_matrix, σ²_val, μ_vector, k, D, z_matrix)
 end
 
 """
@@ -73,20 +100,31 @@ ppca = ProbabilisticPCA(K=1, D=2)
 E_Step(ppca, rand(10, 2))
 ```
 """
-function E_Step(ppca::ProbabilisticPCA, X::Matrix{<:Real})
[…]
