+++ test/HiddenMarkovModels/SwitchingBernoulliRegression.jl
@@ -110,91 +110,3 @@ function test_trialized_SwitchingBernoulliRegression()
     # Test that the ll is always increasing (accept some numerical instability)
     @test any(diff(ll) .< -1) == false
 end
-
-function test_sample_non_float64()
-
-    emission_1 = StateSpaceDynamics.BernoulliRegressionEmission(;
-        input_dim=2, output_dim=1, include_intercept=true, β=reshape([3.0, 1.0, 2.0], :, 1)
-    )
-    emission_2 = StateSpaceDynamics.BernoulliRegressionEmission(;
-        input_dim=2,
-        output_dim=1,
-        include_intercept=true,
-        β=reshape([-3.0, -2.0, 1.0], :, 1),
-    )
-
-    # Create Switching Bernoulli Regression and add the emissions
-    true_model = StateSpaceDynamics.SwitchingBernoulliRegression(; K=2, input_dim=2)
-    true_model.A = [0.9 0.1; 0.2 0.8]
-    true_model.B[1] = emission_1
-    true_model.B[2] = emission_2
-
-    # Sample from the model
-    n = 30000
-    Φ_int = randn(2, n)
-    Φ_f32 = Float32.(randn(2, n))
-
-    # int features
-    labels_int, Y_int = StateSpaceDynamics.sample(true_model, Φ_int; n=n)
-    @test length(labels_int) == n
-    @test eltype(Y_int) == Float64
-
-    # Float32 features
-    labels_f32, Y_f32 = StateSpaceDynamics.sample(true_model, Φ_f32; n=n)
-    @test length(labels_f32) == n
-    @test eltype(Y_f32) == Float64
-end
-
-function test_fit_non_float_SwitchingBernoulliRegression()
-
-    num_trials = 50  # Number of trials
-    trial_length = 1000 # Number of time steps per trial
-
-    # Create the emission models
-    emission_1 = StateSpaceDynamics.BernoulliRegressionEmission(;
-        input_dim=2, output_dim=1, include_intercept=true, β=reshape([3, 1, 2], :, 1)
-    )
-    emission_2 = StateSpaceDynamics.BernoulliRegressionEmission(;
-        input_dim=2, output_dim=1, include_intercept=true, β=reshape([-3, -2, 0.1], :, 1)
-    )
-
-    # Initialize the SwitchingPoissonRegression model
-    true_model = SwitchingBernoulliRegression(; K=2, input_dim=2, output_dim=1)
-    true_model.A = [0.9 0.1; 0.2 0.8]
-
-    # Plug in the emission models
-    true_model.B[1] = emission_1
-    true_model.B[2] = emission_2
-
-    # Create lists to hold data and labels for each trial
-    Φ_trials = [randn(2, trial_length) for _ in 1:num_trials]  # Input features for each trial
-    true_labels_trials = Vector{Vector{Int}}(undef, num_trials)
-    data_trials = Vector{Matrix{Float64}}(undef, num_trials)
-
-    # Sample data for each trial
-    for i in 1:num_trials
-        true_labels_trials[i], data_trials[i] = StateSpaceDynamics.sample(
-            true_model, Φ_trials[i]; n=trial_length
-        )
-    end
-
-    all_Y_int = [Int.(round.(Y)) for Y in data_trials]
-    all_Y_f32 = [Float32.(Y)       for Y in data_trials]
-
-    all_Φ_int    = [Int.(round.(Φ))   for Φ in Φ_trials]
-    all_Φ_f32    = [Float32.(Φ)       for Φ in Φ_trials]
-
-    # Int model
[…]
