### 2025-06-06 c705a96
Update HiddenMarkovModels.jl
