### 2025-06-06 e89bf66
Make FB Constructor compatible with new wrapper, HMM struct wrapper formatted
