### 2025-06-06 4a700c0
Increase genericity to LinearDynamicalSystems.jl
