+++ src/LinearDynamicalSystems.jl
@@ -6,16 +6,16 @@ export GaussianLDS, PoissonLDS, sample, smooth, fit!
 Represents the state model of a Linear Dynamical System with Gaussian noise.
 
 # Fields
-- `A::Matrix{T}`: Transition matrix
-- `Q::Matrix{T}`: Process noise covariance
+- `A::AbstractMatrix{T}`: Transition matrix
+- `Q::AbstractMatrix{T}`: Process noise covariance
 - `x0::Vector{T}`: Initial state
-- `P0::Matrix{T}`: Initial state covariance
+- `P0::AbstractMatrix{T}`: Initial state covariance
 """
-mutable struct GaussianStateModel{T<:Real} <: AbstractStateModel{T}
-    A::Matrix{T}
-    Q::Matrix{T}
-    x0::Vector{T}
-    P0::Matrix{T}
+mutable struct GaussianStateModel{T<:Real, M<:AbstractMatrix{T}, V<:AbstractVector{T}} <: AbstractStateModel{T}
+    A::M
+    Q::M
+    x0::V
+    P0::M 
 end
 
 """
@@ -31,10 +31,10 @@ Construct a GaussianStateModel with the given parameters or random initializatio
 - `latent_dim::Int`: Dimension of the latent state (required if any matrix is not provided.)
 """
 function GaussianStateModel(;
-    A::Matrix{T}=Matrix{T}(undef, 0, 0),
-    Q::Matrix{T}=Matrix{T}(undef, 0, 0),
-    x0::Vector{T}=Vector{T}(undef, 0),
-    P0::Matrix{T}=Matrix{T}(undef, 0, 0),
+    A::AbstractMatrix{T}=Matrix{T}(undef, 0, 0),
+    Q::AbstractMatrix{T}=Matrix{T}(undef, 0, 0),
+    x0::AbstractVector{T}=Vector{T}(undef, 0),
+    P0::AbstractMatrix{T}=Matrix{T}(undef, 0, 0),
     latent_dim::Int=0,
 ) where {T<:Real}
     if latent_dim == 0 && (isempty(A) || isempty(Q) || isempty(x0) || isempty(P0))
@@ -46,7 +46,7 @@ function GaussianStateModel(;
     x0 = isempty(x0) ? randn(T, latent_dim) : x0
     P0 = isempty(P0) ? Matrix{T}(I, latent_dim, latent_dim) : P0
 
-    return GaussianStateModel{T}(A, Q, x0, P0)
+    return GaussianStateModel{T, typeof(A), typeof(x0)}(A, Q, x0, P0)
 end
 
 """
@@ -55,12 +55,12 @@ end
 Represents the observation model of a Linear Dynamical System with Gaussian noise.
 
 # Fields
-- `C::Matrix{T}`: Observation matrix
-- `R::Matrix{T}`: Observation noise covariance
+- `C::AbstractMatrix{T}`: Observation matrix
+- `R::AbstractMatrix{T}`: Observation noise covariance
 """
-mutable struct GaussianObservationModel{T<:Real} <: AbstractObservationModel{T}
-    C::Matrix{T}
-    R::Matrix{T}
+mutable struct GaussianObservationModel{T<:Real, M<:AbstractMatrix{T}} <: AbstractObservationModel{T}
+    C::M
+    R::M
 end
 
 """
@@ -75,8 +75,8 @@ Construct a GaussianObservationModel with the given parameters or random initial
 - `latent_dim::Int`: Dimension of the latent state (required if C is not provided.)
 """
 function GaussianObservationModel(;
-    C::Matrix{T}=Matrix{T}(undef, 0, 0),
-    R::Matrix{T}=Matrix{T}(undef, 0, 0),
+    C::AbstractMatrix{T}=Matrix{T}(undef, 0, 0),
[…]
