### 2025-06-06 71ab04f
augmenting @views usage to avoid unnecessary allocations
