+++ src/LinearDynamicalSystems.jl
@@ -311,15 +311,12 @@ function sample(
     for trial in 1:n_trials
         x[:, 1, trial] = rand(MvNormal(x0, P0))
 
-        x1_view = @view x[:, 1, trial]
-        y[:, 1, trial] = rand(MvNormal(C * x1_view, R))
+        y[:, 1, trial] = rand(MvNormal(C * @view(x[:, 1, trial]), R))
 
         for t in 2:T_steps
-            x_prev = @view x[:, t-1, trial]
-            x[:, t, trial] = rand(MvNormal(A * x_prev, Q))
+            x[:, t, trial] = rand(MvNormal(A * @view(x[:, t-1, trial]), Q))
 
-            x_current = @view x[:, t, trial]
-            y[:, t, trial] = rand(MvNormal(C * x_current, R))
+            y[:, t, trial] = rand(MvNormal(C * @view(x[:, t, trial]), R))
         end
     end
 
@@ -429,18 +426,8 @@ function Gradient(
     grad[:, 1] .= A_inv_Q * dx2 + w[1] * C_inv_R * dy1 - (P0_chol \ dx1)
 
     # Middle time steps
-    @inbounds for t in 2:(T_steps - 1)
-
-        x_prev = @view x[:, t - 1]
-        x_curr = @view x[:, t]
-        x_next = @view x[:, t + 1]
-        y_curr = @view y[:, t]
-
-        dxt = x_curr .- A * x_prev
-        dxt_next = x_next .- A * x_curr
-        dyt = y_curr .- C * x_curr
-
-        grad[:, t] .= w[t] * C_inv_R * dyt - (Q_chol \ dxt) + (A_inv_Q * dxt_next)
+    @views @inbounds for t in 2:(T_steps - 1)
+        grad[:, t] .= w[t] * C_inv_R * (y[:, t] .- C * x[:, t]) - (Q_chol \ (x[:, t] .- A * x[:, t - 1])) + (A_inv_Q * (x[:, t + 1] .- A * x[:, t]))
     end
 
     # Last time step
@@ -1177,14 +1164,9 @@ function update_C!(
         sum_zz = zeros(T, size(E_zz)[1:2])
 
         for trial in 1:n_trials
-            @inbounds for t in 1:T_steps
-
-                y_view   = @view y[:, t, trial]  
-                Ez_view  = @view E_z[:, t, trial]    
-                Ezz_view = @view E_zz[:, :, t, trial]
-                
-                sum_yz .+= w[t] * (y_view * Ez_view')
-                sum_zz .+= w[t] * Ezz_view
+            @views @inbounds for t in 1:T_steps
+                sum_yz .+= w[t] * (y[:, t, trial]* E_z[:, t, trial]')
+                sum_zz .+= w[t] * E_zz[:, :, t, trial]
             end
         end
 
@@ -1224,12 +1206,10 @@ function update_R!(
         
         # Reorganize as sum of outer products
         for trial in 1:n_trials
-            @inbounds for t in 1:T_steps
+            @views @inbounds for t in 1:T_steps
                 # Compute innovation using pre-allocated arrays
-                yt = @view y[:, t, trial]
-                zt = @view E_z[:, t, trial]
-                mul!(Czt, C, zt)
-                @. innovation = (yt - Czt)
+                mul!(Czt, C, E_z[:, t, trial])
+                @. innovation = (y[:, t, trial] - Czt)
                 
                 # Add innovation outer product
                 #BLAS.ger!(one(T), innovation, innovation, R_new)
@@ -1237,8 +1217,7 @@ function update_R!(
                 
                 # Add correction term efficiently:
                 # First compute state_uncertainty = Σ_t - z_t*z_t'
[…]
