+++ src/EmissionModels.jl
@@ -40,30 +40,30 @@ Functon to create a GaussianEmission with given output dimension, mean, and cova
 """
 function GaussianEmission(;
     output_dim::Int,
-    μ::Vector{<:Real}=zeros(output_dim),
-    Σ::Matrix{<:Real}=Matrix{Float64}(I, output_dim, output_dim),
+    μ::AbstractVector{<:Real}=zeros(output_dim),
+    Σ::AbstractMatrix{<:Real}=Matrix{Float64}(I, output_dim, output_dim),
 )
     return GaussianEmission(output_dim, μ, Σ)
 end
 
 """
-    loglikelihood(model::GaussianEmission, Y::Matrix{<:Real})
+    loglikelihood(model::GaussianEmission, Y::AbstractMatrix{<:Real})
 
 Calculate the log likelihood of the data `Y` given the Gaussian emission model.
 
 # Arguments
 - `model::GaussianEmission`: The Gaussian emission model for which to calculate the log likelihood.
-- `Y::Matrix{<:Real}`: The data matrix, where each row represents an observation.
+- `Y::AbstractMatrix{<:Real}`: The data matrix, where each row represents an observation.
 
 # Returns
 - `Vector{Float64}`: A vector of log likelihoods, one for each observation in the data.
 """
-function loglikelihood(model::GaussianEmission, Y::Matrix{<:Real})
+function loglikelihood(model::GaussianEmission, Y::AbstractMatrix{<:Real})
     # Create MvNormal distribution with the model parameters
     dist = MvNormal(model.μ, model.Σ)
 
     # Calculate log likelihood for each observation
-    return [logpdf(dist, Y[i, :]) for i in axes(Y, 1)]
+    return [logpdf(dist, @view(Y[i, :])) for i in axes(Y, 1)]
 end
 
 """
@@ -77,17 +77,17 @@ function sample(model::GaussianEmission; n::Int=1)
 end
 
 """
-    fit!(model::GaussianEmission, Y::Matrix{<:Real}, w::Vector{Float64}=ones(size(Y, 1)))
+    fit!(model::GaussianEmission, Y::AbstractMatrix{<:Real}, w::AbstractVector{Float64}=ones(size(Y, 1)))
 
 Fit a GaussianEmission model to the data `Y`. 
 
 # Arguments
 - `model::GaussianEmission`: Gaussian model to fit.
-- `Y::Matrix{<:Real}`: Data to fit the model to. Should be a matrix of size `(n, output_dim)`.
-- `w::Vector{Float64}=ones(size(Y, 1))`: Weights for the data. Should be a vector of size `n`.
+- `Y::AbstractMatrix{<:Real}`: Data to fit the model to. Should be a matrix of size `(n, output_dim)`.
+- `w::AbstractVector{Float64}=ones(size(Y, 1))`: Weights for the data. Should be a vector of size `n`.
 """
 function fit!(
-    model::GaussianEmission, Y::Matrix{<:Real}, w::Vector{Float64}=ones(size(Y, 1))
+    model::GaussianEmission, Y::AbstractMatrix{<:Real}, w::AbstractVector{Float64}=ones(size(Y, 1))
 )
     weighted_sum = sum(Y .* w; dims=1)
     new_mean = weighted_sum[:] ./ sum(w)
@@ -112,18 +112,18 @@ Holds the optimization problem data for regression emissions.
 """
 struct RegressionOptimization{T<:RegressionEmission}
     model::T
-    X::Matrix{<:Real}
-    y::Matrix{<:Real}
-    w::Vector{Float64}
+    X::AbstractMatrix{<:Real}
+    y::AbstractMatrix{<:Real}
+    w::AbstractVector{Float64}
[…]
