+++ src/LinearDynamicalSystems.jl
@@ -393,7 +393,7 @@ Compute the gradient of the log-likelihood with respect to the latent states for
 - `grad::Matrix{T}`: Gradient of the log-likelihood with respect to the latent states.
 """
 function Gradient(
-    lds::LinearDynamicalSystem{S,O}, y::Matrix{T}, x::Matrix{T},
+    lds::LinearDynamicalSystem{S,O}, y::AbstractMatrix{T}, x::AbstractMatrix{T},
     w::Vector{Float64}=ones(size(y, 2))
 ) where {T<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
     # Dims etc.
@@ -541,7 +541,7 @@ x, p_smooth, inverse_offdiag, Q_val = DirectSmoother(lds, y)
 ```
 """
 function smooth(
-    lds::LinearDynamicalSystem{S,O}, y::Matrix{T}, w::Vector{Float64}=ones(size(y, 2))
+    lds::LinearDynamicalSystem{S,O}, y::AbstractMatrix{T}, w::Vector{Float64}=ones(size(y, 2))
 ) where {T<:Real,S<:GaussianStateModel{T},O<:AbstractObservationModel{T}}
     T_steps, D = size(y, 2), lds.latent_dim
 
@@ -600,7 +600,7 @@ function smooth(
     gauss_entropy = gaussian_entropy(Symmetric(H))
 
     # Symmetrize the covariance matrices
-    @inbounds for i in 1:T_steps
+    @inbounds @views for i in 1:T_steps
         p_smooth[:, :, i] .= 0.5 .* (p_smooth[:, :, i] .+ p_smooth[:, :, i]')
     end
 
@@ -653,13 +653,14 @@ function smooth(
     inverse_offdiag = Array{T,4}(undef, latent_dim, latent_dim, T_steps, n_trials)
     total_entropy = 0.0
 
-    @threads for trial in 1:n_trials
+    @views @threads for trial in 1:n_trials
         x_sm, p_sm, p_prev, ent = smooth(lds, y[:, :, trial])
         total_entropy += ent
         x_smooth[:, :, trial] .= x_sm
         p_smooth[:, :, :, trial] .= p_sm
         inverse_offdiag[:, :, :, trial] .= p_prev
     end
+    
 
     return x_smooth, p_smooth, inverse_offdiag, total_entropy
 end
@@ -889,17 +890,18 @@ function sufficient_statistics(
     E_zz = similar(p_smooth)
     E_zz_prev = similar(p_smooth)
 
-    for trial in 1:n_trials
-        @inbounds for t in 1:T_steps
-            E_zz[:, :, t, trial] .=
-                p_smooth[:, :, t, trial] + x_smooth[:, t, trial] * x_smooth[:, t, trial]'
+    @inbounds for trial in 1:n_trials
+        @views for t in 1:T_steps
+            xt = view(x_smooth, :, t, trial)
+            pt = view(p_smooth, :, :, t, trial)
+            E_zz[:, :, t, trial] .= pt .+ xt * xt'
             if t > 1
-                E_zz_prev[:, :, t, trial] .=
-                    p_smooth_t1[:, :, t, trial] +
-                    x_smooth[:, t, trial] * x_smooth[:, t - 1, trial]'
+                xtm1 = view(x_smooth, :, t - 1, trial)
+                pt1 = view(p_smooth_t1, :, :, t, trial)
+                E_zz_prev[:, :, t, trial] .= pt1 .+ xt * xtm1'
             end
         end
-        E_zz_prev[:, :, 1, trial] .= 0
+        @views E_zz_prev[:, :, 1, trial] .= 0
     end
 
[…]

[…]
