### 2025-06-07 9935daa
Replaces all instances of "sample" with Random.rand
