+++ test/HiddenMarkovModels/AutoRegressionHMM.jl
@@ -12,7 +12,7 @@ function test_ARHMM_sampling()
     # Sample from the model
     n = 2
     Φ = randn(2, n)
-    true_labels, data = StateSpaceDynamics.sample(true_model, Φ, n=n)
+    true_labels, data = rand(true_model, Φ, n=n)
     
     order=1
     output_dim = 2
@@ -32,7 +32,7 @@ function test_ARHMM_sampling()
     Y = StateSpaceDynamics.construct_AR_feature_matrix(data, order, false)
     X = copy(Y[:, end])
     X = reshape(Y[:, end], 2, 1)
-    AR_labels, AR_data = StateSpaceDynamics.sample(true_model, X, n=num_points, autoregressive=true)
+    AR_labels, AR_data = rand(true_model, X, n=num_points, autoregressive=true)
 
     @test size(AR_data, 2) == num_points
     @test size(AR_data, 1) == output_dim
@@ -52,7 +52,7 @@ function test_ARHMM_fit()
     # Sample from the model
     n = 2
     Φ = randn(2, n)
-    true_labels, data = StateSpaceDynamics.sample(true_model, Φ, n=n)
+    true_labels, data = rand(true_model, Φ, n=n)
     
     """
     Construct true AR HMM
@@ -71,7 +71,7 @@ function test_ARHMM_fit()
     Y = StateSpaceDynamics.construct_AR_feature_matrix(data, 1, false)
     X = copy(Y[:, end])
     X = reshape(Y[:, end], 2, 1)
-    AR_labels, AR_data = StateSpaceDynamics.sample(true_model, X, n=2000, autoregressive=true)
+    AR_labels, AR_data = rand(true_model, X, n=2000, autoregressive=true)
     
     
     """
@@ -126,7 +126,7 @@ function test_timeseries_to_AR_feature_matrix()
     # Sample from the model
     n = 100
     Φ = randn(2, n)
-    true_labels, data = StateSpaceDynamics.sample(true_model, Φ, n=n)
+    true_labels, data = rand(true_model, Φ, n=n)
 
     # Sample from the AR HMM using its own emission models and a starting point
     order = 1
@@ -154,7 +154,7 @@ function test_trialized_timeseries_to_AR_feature_matrix()
     # Sample from the model
     for i in 1:num_trials
         Φ = randn(2, 100)
-        true_labels, data = StateSpaceDynamics.sample(true_model, Φ, n=100)
+        true_labels, data = rand(true_model, Φ, n=100)
         push!(all_data, data)
         push!(Φ_total, Φ)
     end
