### 2025-06-07 59b424b
Removes cases of @inbounds. Lack of safety doesn't justify performance gain
