+++ src/LinearDynamicalSystems.jl
@@ -382,7 +382,7 @@ function loglikelihood(
     temp_dx = zeros(T, size(x, 1))
     temp_dy = zeros(promote_type(T, U), size(y, 1))
 
-    @inbounds for t in 1:tsteps
+    for t in 1:tsteps
         if t > 1
             mul!(temp_dx, A, view(x, :, t-1), -1.0, false)
             temp_dx .+= view(x, :, t)
@@ -445,7 +445,7 @@ function Gradient(
     grad[:, 1] .= A_inv_Q * dx2 + w[1] * C_inv_R * dy1 - (P0_chol \ dx1)
 
     # Middle time steps
-    @inbounds for t in 2:(tsteps - 1)
+    for t in 2:(tsteps - 1)
         dxt = x[:, t] - A * x[:, t - 1]
         dxt_next = x[:, t + 1] - A * x[:, t]
         dyt = y[:, t] - C * x[:, t]
@@ -523,14 +523,14 @@ function Hessian(
     x_t = -inv_P0
 
     # Build off-diagonals
-    @inbounds for i in 1:(tsteps - 1)
+    for i in 1:(tsteps - 1)
         H_sub[i] = H_sub_entry
         H_super[i] = H_super_entry
     end
 
     # Build main diagonal
     H_diag[1] = w[1] * yt_given_xt + xt1_given_xt + x_t
-    @inbounds for i in 2:(tsteps - 1)
+    for i in 2:(tsteps - 1)
         H_diag[i] = w[i] * yt_given_xt + xt_given_xt_1 + xt1_given_xt
     end
     H_diag[tsteps] = w[tsteps] * (yt_given_xt) + xt_given_xt_1
@@ -623,7 +623,7 @@ function smooth(
     gauss_entropy = gaussian_entropy(Symmetric(H))
 
     # Symmetrize the covariance matrices
-    @inbounds for i in 1:tsteps
+    for i in 1:tsteps
         p_smooth[:, :, i] .= 0.5 .* (p_smooth[:, :, i] .+ p_smooth[:, :, i]')
     end
 
@@ -735,7 +735,7 @@ function Q_state(
     sum_E_zz_prev_time = zeros(state_dim, state_dim)
     
     # Compute sums with views
-    @inbounds for t in 2:T_step
+    for t in 2:T_step
         sum_E_zz_current .+= view(E_zz, :, :, t)
         sum_E_zz_prev_cross .+= view(E_zz_prev, :, :, t)
         sum_E_zz_prev_time .+= view(E_zz, :, :, t-1)
@@ -914,7 +914,7 @@ function sufficient_statistics(
     E_zz_prev = similar(p_smooth)
 
     for trial in 1:ntrials
-        @inbounds for t in 1:tsteps
+        for t in 1:tsteps
             E_zz[:, :, t, trial] .=
                 p_smooth[:, :, t, trial] + x_smooth[:, t, trial] * x_smooth[:, t, trial]'
             if t > 1
@@ -1138,7 +1138,7 @@ function update_Q!(
         A = lds.state_model.A
 
         for trial in 1:ntrials
-            @inbounds for t in 2:tsteps
+            for t in 2:tsteps
                 # Get current state covariance and previous-current cross covariance
                 Σt = E_zz[:, :, t, trial]          # E[z_t z_t']
                 Σt_prev = E_zz[:, :, t - 1, trial]   # E[z_{t-1} z_{t-1}']
@@ -1188,7 +1188,7 @@ function update_C!(
         sum_zz = zeros(T, size(E_zz)[1:2])
 
         for trial in 1:ntrials
-            @inbounds for t in 1:tsteps
+            for t in 1:tsteps
[…]
