### 2025-06-09 7f90e9d
refactored SLDS, Mixture Models structs
