### 2025-06-09 bc716f2
LDS constructor initialization types tests
