### 2025-06-09 6350ec4
Updated tests to use floats only
