### 2025-06-09 5f4b461
Add PPCA sampling test.
