### 2025-06-09 depasquale-lab/StateSpaceDynamics.jl#51 @rsenne
Hi @gdalle. As we have made some good progress, and pull requests are starting to be integrated into the `joss_` branch, I will now start give our progress on these points.

> The code example in the README fails with a SingularException when I run it on SSD v0.1.1 and the latest Julia 1.11.

This has been more of a pernicious issue than I initially expected. I think i know what the issue is, but I want to give an update on where I am with it. One issue is that as we did not previously supply RNGs as you pointed out down below--we were at the whim of randomness for this example to work. For example, my desktop at home almost _never_ has the issues yo found (which is where I made this example), but my laptop has errors at a much higher rate. Overall here are the issues as I see them:

1. Our default initializations, (especially for the PoissonLDS model), are not particularly good. They rely on randomly generating matrices and in reality for these models, domain specific knowledge, or problem-specific routines should be used (i.e., one could use [exponential family pca](https://github.com/sisl/ExpFamilyPCA.jl) to initialize or [subspace identification methods](https://sites.stat.columbia.edu/liam/research/pubs/pfau-subspace-id.pdf).)

I'm of two minds on this. On one hand, its convenient to not have to pass in parameters when making one of these models. However, since these are latent variable models, and the loglikelihood surface is non-convex, I'm thinking it would be better to require the user to specify their own starting points so that the initializations can be based off something reasonable, and if there is demand--to maybe have some convenience functions that could use reasonable heuristics to generate problem-specific starting points.

2.) When implementing multiple trials, I'm finding that the gradients of the observation model, for poorly behaved trials can explode. Some naive experiments with gradient clipping seem to promise promising results, and will allow the model to learn very reasonable dynamics.

3.) In synthetic datasets where this is all of this is an issue, I am finding that the ELBO calculation starts to become unstable, (despite learning good latent dynamics). Why this happens eludes me, but, This is certainly odd behavior and needs to be fixed somehow. I have half a mind to sample the estimated posterior to get an [unbiased estimate of the ELBO](https://chrisorm.github.io/VI-ELBO-MC-approx.html). I can always use the norm of the parameter change as the convergence criteria, and also the analytical ELBO calculations can be a pain to implement and have a non-trivial performance hit.

> It is usually a good idea to also run tests on the latest stable Julia version (in addition to 1.10).

Our code is now tested against the release version as of #55

> The random sampling functions should always accept and pass along an rng argument for reproducibility.

All `sample` functions are now instances of `Random.rand` and as such accept an RNG. If a user fails to provide one, it will default to `default_rng()`. This was fixed in #55

> Many structs have fields with abstract element types (like x::Vector{<:Real}), which is a [performance pitfall](https://docs.julialang.org/en/v1/manual/performance-tips/#Avoid-fields-with-abstract-containers).

> Correcting these structs and generalizing others may enable type-genericity, e.g. using Float32 instead of Float64, or differentiating through an inference step with ForwardDiff.

This is an ongoing fix--but we are making good progress. I expect it will be finished this week. The branch `JOSS_Review_Abstract_Types_Struct_Fixes` should take care of this.

> Many array slices (x[i, :]) could be views instead (@view(x[i, :])) to get rid of allocations.

We have added these as suggested in #54. I *think* we got all of them.

> [JET](https://github.com/aviatesk/JET.jl) reports some errors inside your package. Some can be ignored (typically the union split), some are probably limitations of JET itself (like the ones with T at the start), but some are actual issues or typos.

This is also ongoing and we are in process of implementing this in the `add_JET` branch. Also, hope to be done by this week.
