### 2025-06-09 9982229
Edit paper and bibliography
