+++ paper/paper.bib
@@ -235,14 +235,12 @@ state space models},
   title    = {PythonCall.jl},
   author   = {Christopher Doris},
   date     = {2021},
-  howpublished = {\url{https://github.com/JuliaPy/PythonCall.jl}},
-  url = {https://github.com/JuliaPy/PythonCall.jl},
+  howpublished = {\url{https://github.com/JuliaPy/PythonCall.jl}}
 }
 
 @software{bench,
   title    = {BenchmarkTools.jl},
   author   = {Jarrett Revels},
   date     = {2017},
-  howpublished = {\url{https://github.com/JuliaCI/BenchmarkTools.jl}},
-  url = {https://github.com/JuliaCI/BenchmarkTools.jl},
+  howpublished = {\url{https://github.com/JuliaCI/BenchmarkTools.jl}}
 }
