### 2025-06-10 d43879a
Updated GaussianHMM, Gaussian Regression, and Autoregression to be type general and not have initializations
