+++ src/EmissionModels.jl
@@ -113,21 +113,23 @@ Regression Emission Models
     RegressionOptimization{T<:RegressionEmission}
 
 Holds the optimization problem data for regression emissions.
+
+This handle is inclusive to X and y being different types for Poisson and Bernoulli models
 """
-struct RegressionOptimization{R<:RegressionEmission, T<:Real, V<:AbstractVector{T}, MX<:AbstractMatrix{<:Real}, MY<:AbstractMatrix{<:Real}}
+struct RegressionOptimization{R<:RegressionEmission, V<:AbstractVector{<:AbstractFloat}, MX<:AbstractMatrix{<:AbstractFloat}, MY<:AbstractMatrix{<:AbstractFloat}}
     model::R
     X::MX
     y::MY
     w::V
-    β_shape::Tuple{Int,Int}  # Added to track original shape
+    β_shape::Tuple{Int, Int}  # Added to track original shape
 end
 
 # Unified interface for creating optimization problems
 function create_optimization(
     model::RegressionEmission,
-    X::AbstractMatrix{<:Real},
-    y::AbstractMatrix{<:Real},
-    w::Vector{<:Real}=ones(size(y, 1)),
+    X::AbstractMatrix{<:AbstractFloat},
+    y::AbstractMatrix{<:AbstractFloat},
+    w::Vector{<:AbstractFloat}=ones(size(y, 1)),
 )
     if model.include_intercept
         X = hcat(ones(size(X, 1)), X)
@@ -157,7 +159,7 @@ Calculate L2 regularization term for regression coefficients.
 # Returns
 - `Float64`: The regularization term value
 """
-function calc_regularization(β::AbstractMatrix{T}, λ::Float64, include_intercept::Bool=true) where {T<:Real}
+function calc_regularization(β::AbstractMatrix{T}, λ::T, include_intercept::Bool=true) where {T<:AbstractFloat}
     # calculate L2 penalty
     if include_intercept
         regularization = 0.5 * λ * sum(abs2, β[2:end, :])
@@ -180,7 +182,7 @@ Calculate gradient of L2 regularization term for regression coefficients.
 """
 function calc_regularization_gradient(
     β::AbstractMatrix{T}, λ::T, include_intercept::Bool=true
-) where {T<:Real}
+) where {T<:AbstractFloat}
     # calculate the gradient of the regularization component
     regularization = zeros(size(β))
 
@@ -209,7 +211,7 @@ A Gaussian regression Emission model.
 - `Σ::Matrix{<:Real} = Matrix{Float64}(I, output_dim, output_dim)`: Covariance matrix of the model.
 - `λ::Float64 = 0.0`: Regularization parameter.
 """
-mutable struct GaussianRegressionEmission{T<:Real, M<:AbstractMatrix{T}} <: RegressionEmission
+mutable struct GaussianRegressionEmission{T<:AbstractFloat, M<:AbstractMatrix{T}} <: RegressionEmission
     input_dim::Int
     output_dim::Int
     β::M # coefficient matrix of the model. Shape input_dim by output_dim. Column one is coefficients for target one, etc. The first row are the intercept terms, if included. 
@@ -218,32 +220,21 @@ mutable struct GaussianRegressionEmission{T<:Real, M<:AbstractMatrix{T}} <: Regr
     λ::T # regularization parameter
 end
 
-function GaussianRegressionEmission(; 
+# This may look redundant, but this is to enforce types to be how we like
+function GaussianRegressionEmission(;
[…]
