### 2025-06-10 c89a8b2
Update gaussian HMM tests
