+++ test/HiddenMarkovModels/GaussianHMM.jl
@@ -1,40 +1,41 @@
 function test_SwitchingGaussian_fit()
-    Random.seed!(1234)
-    # Create Guassian Emission Models
+    # Define the output dimensionality of the HMM
     output_dim = 2
-    μ = [0.0, 0.0]
-    Σ = 0.1 * Matrix{Float64}(I, output_dim, output_dim)
-    emission_1 = GaussianEmission(output_dim, μ, Σ)
 
-    μ = [2.0, 1.0]
-    Σ = 0.1 * Matrix{Float64}(I, output_dim, output_dim)
-    emission_2 = GaussianEmission(output_dim, μ, Σ)
+    # Define the transition matrix and the initial state distribution
+    A = [0.99 0.01; 0.05 0.95];
+    πₖ = [0.5; 0.5]
 
-    # Create GaussianHMM
-    true_model = StateSpaceDynamics.GaussianHMM(; K=2, output_dim=2)
-    true_model.B[1] = emission_1
-    true_model.B[2] = emission_2
-    true_model.A = [0.9 0.1; 0.2 0.8]
+    # Initialize the emission models
+    μ_1 = [0.0, 0.0]
+    Σ_1 = 0.1 * Matrix{Float64}(I, output_dim, output_dim)
+    emission_1 = GaussianEmission(output_dim=output_dim, μ=μ_1, Σ=Σ_1)
+
+    μ_2 = [2.0, 1.0]
+    Σ_2 = 0.2 * Matrix{Float64}(I, output_dim, output_dim)
+    emission_2 = GaussianEmission(output_dim=output_dim, μ=μ_2, Σ=Σ_2)
+
+    # The general HMM constructor is used as follows
+    true_model = HiddenMarkovModel(K=2, B=[emission_1, emission_2], A=A, πₖ=πₖ)
 
     # Sample from the model
     n = 50000
     true_labels, data = StateSpaceDynamics.sample(true_model; n=n)
 
-    # Fit a gaussian hmm to the data
-    test_model = StateSpaceDynamics.GaussianHMM(; K=2, output_dim=2)
+    # Initialize a new GaussianHMM
+    μ_1 = rand(output_dim)
+    Σ_1 = 0.3 * Matrix{Float64}(I, output_dim, output_dim)
+    emission_1 = GaussianEmission(output_dim=output_dim, μ=μ_1, Σ=Σ_1)
 
-    μ = [0.5, 0.1]
-    Σ = 0.1 * Matrix{Float64}(I, output_dim, output_dim)
-    emission_1 = GaussianEmission(output_dim, μ, Σ)
+    μ_2 = rand(output_dim)
+    Σ_2 = 0.5 * Matrix{Float64}(I, output_dim, output_dim)
+    emission_2 = GaussianEmission(output_dim=output_dim, μ=μ_1, Σ=Σ_1)
 
-    μ = [1.8, 1.2]
-    Σ = 0.1 * Matrix{Float64}(I, output_dim, output_dim)
-    emission_2 = GaussianEmission(output_dim, μ, Σ)
+    A = [0.8 0.2; 0.05 0.95]
+    πₖ = [0.6,0.4]
+    test_model = HiddenMarkovModel(K=2, B=[emission_1, emission_2], A=A, πₖ=πₖ)
 
-    test_model.B[1] = emission_1
-    test_model.B[2] = emission_2
-
-    test_model.A = [0.8 0.2; 0.05 0.95]
+    # Fit the model to recover the original parameters
     ll = StateSpaceDynamics.fit!(test_model, data)
 
     # Test that the transition matrix is correct
@@ -56,22 +57,29 @@ function test_SwitchingGaussian_fit()
 end
 
 function test_SwitchingGaussian_SingleState_fit()
-    Random.seed!(1234)
[…]
