### 2025-06-10 9bf0504
Remove unneeded test
