+++ docs/make.jl
@@ -1,8 +1,9 @@
 using Documenter
 using StateSpaceDynamics
+using Random
 
-
-DocMeta.setdocmeta!(StateSpaceDynamics, :DocTestSetup, :(using StateSpaceDynamics); recursive=true)
+# Set up the documentation environment
+DocMeta.setdocmeta!(StateSpaceDynamics, :DocTestSetup, :(using StateSpaceDynamics, Random); recursive=true)
 
 makedocs(;
     modules=[StateSpaceDynamics],
@@ -22,7 +23,9 @@ makedocs(;
             "Mixture Models" => "MixtureModels.md",
         ],
         "Miscellaneous" => "Misc.md",
-    ]
+    ],
+    checkdocs = :exports,  # Only check exported functions
+    warnonly = true # Do not fail on warnings
 )
 
 deploydocs(; repo="github.com/depasquale-lab/StateSpaceDynamics.jl", devbranch="docs_dev_")
