### 2025-06-10 7b45d83
Fixed Poisson + Bernoulli struct handles. Made functions type general.
