+++ src/EmissionModels.jl
@@ -116,7 +116,7 @@ Holds the optimization problem data for regression emissions.
 
 This handle is inclusive to X and y being different types for Poisson and Bernoulli models
 """
-struct RegressionOptimization{R<:RegressionEmission, V<:AbstractVector{<:AbstractFloat}, MX<:AbstractMatrix{<:AbstractFloat}, MY<:AbstractMatrix{<:AbstractFloat}}
+struct RegressionOptimization{R<:RegressionEmission, V<:AbstractVector{<:Real}, MX<:AbstractMatrix{<:Real}, MY<:AbstractMatrix{<:Real}}
     model::R
     X::MX
     y::MY
@@ -127,9 +127,9 @@ end
 # Unified interface for creating optimization problems
 function create_optimization(
     model::RegressionEmission,
-    X::AbstractMatrix{<:AbstractFloat},
-    y::AbstractMatrix{<:AbstractFloat},
-    w::Vector{<:AbstractFloat}=ones(size(y, 1)),
+    X::AbstractMatrix{<:Real},
+    y::AbstractMatrix{<:Real},
+    w::Vector{<:Real}=ones(size(y, 1)),
 )
     if model.include_intercept
         X = hcat(ones(size(X, 1)), X)
@@ -159,7 +159,8 @@ Calculate L2 regularization term for regression coefficients.
 # Returns
 - `Float64`: The regularization term value
 """
-function calc_regularization(β::AbstractMatrix{T}, λ::T, include_intercept::Bool=true) where {T<:AbstractFloat}
+function calc_regularization(β::AbstractMatrix{T1}, λ::T2, include_intercept::Bool=true) where {T1<:Real, T2<:Real}
+    # Includes  T1 and T2 since autodiff passes in DualNumber which is not subtype float for β
     # calculate L2 penalty
     if include_intercept
         regularization = 0.5 * λ * sum(abs2, β[2:end, :])
@@ -181,8 +182,8 @@ Calculate gradient of L2 regularization term for regression coefficients.
 - `include_intercept::Bool`: Whether to exclude the intercept term from regularization
 """
 function calc_regularization_gradient(
-    β::AbstractMatrix{T}, λ::T, include_intercept::Bool=true
-) where {T<:AbstractFloat}
+    β::AbstractMatrix{T1}, λ::T2, include_intercept::Bool=true
+) where {T1<:Real, T2<:Real}
     # calculate the gradient of the regularization component
     regularization = zeros(size(β))
 
@@ -211,7 +212,7 @@ A Gaussian regression Emission model.
 - `Σ::Matrix{<:Real} = Matrix{Float64}(I, output_dim, output_dim)`: Covariance matrix of the model.
 - `λ::Float64 = 0.0`: Regularization parameter.
 """
-mutable struct GaussianRegressionEmission{T<:AbstractFloat, M<:AbstractMatrix{T}} <: RegressionEmission
+mutable struct GaussianRegressionEmission{T<:Real, M<:AbstractMatrix{T}} <: RegressionEmission
     input_dim::Int
     output_dim::Int
     β::M # coefficient matrix of the model. Shape input_dim by output_dim. Column one is coefficients for target one, etc. The first row are the intercept terms, if included. 
@@ -227,7 +228,7 @@ function GaussianRegressionEmission(;
     include_intercept::Bool=true,
     β::AbstractMatrix,
     Σ::AbstractMatrix,
-    λ::AbstractFloat
+    λ::Real
 )
 
     if !check_same_type(β[1], Σ[1], λ)
[…]
