+++ test/HiddenMarkovModels/AutoRegressionHMM.jl
@@ -1,34 +1,33 @@
 function test_ARHMM_sampling()
-    """
-    Use Gaussian Regression to get first point for sampling from AR
-    """
-    emission_1 = GaussianRegressionEmission(input_dim=2, output_dim=1, include_intercept=false, β=reshape([2.0 1.0 ], :, 1))
-    emission_2 = GaussianRegressionEmission(input_dim=2, output_dim=1, include_intercept=false, β=reshape([-2.0 -1.0], :, 1))
+    # Create Emission Models
+    emission_1 = GaussianRegressionEmission(input_dim=2, output_dim=1, include_intercept=false, β=reshape([3.0, 2.0], :, 1), Σ=[1.0;;], λ=0.0)
+
     # Create Switching Regression Model
-    true_model = SwitchingGaussianRegression(K=2, input_dim=2, output_dim=1, include_intercept=false)
-    # Plug in the emission models
-    true_model.B[1] = emission_1
-    true_model.B[2] = emission_2
+    A = [1.0;;]
+    πₖ = [1.0]
+
+    true_model = HiddenMarkovModel(K=1, A=A, πₖ=πₖ, B=[emission_1])
+
     # Sample from the model
     n = 2
     Φ = randn(2, n)
     true_labels, data = StateSpaceDynamics.sample(true_model, Φ, n=n)
-    
-    order=1
+
+    # Create autoregressive HMM and sample from it -> one growing, one shrinking
+    order = 1
     output_dim = 2
-    # Create autoregressive emission models
-    AR_emission_1 = AutoRegressionEmission(output_dim=output_dim, order=order, include_intercept=false, β=[0.8 -0.6; 0.6 0.8])
-    AR_emission_2 = AutoRegressionEmission(output_dim=output_dim, order=order, include_intercept=false, β=[0.0 -1.0; 1.0 0.0])
+    AR_emission_1 = AutoRegressionEmission(output_dim=output_dim, order=1, include_intercept=false, β=[0.8 -0.6; 0.6 0.8], Σ=[0.001 0.0; 0.0 0.001], λ=0.0);
+    AR_emission_2 = AutoRegressionEmission(output_dim=output_dim, order=1, include_intercept=false, β=[0.0 -1.0; 1.0 0.0], Σ=[0.001 0.0; 0.0 0.001], λ=0.0);
+
+    A = [0.9 0.1; 0.2 0.8]
+    πₖ = [0.9; 0.1]
 
     # Create the autoregressive model and plug in emission models
-    true_model = StateSpaceDynamics.SwitchingAutoRegression(K=2, output_dim=output_dim, order=order)
-    true_model.A = [0.8 0.2; 0.2 0.8]
-    true_model.B[1] = AR_emission_1
-    true_model.B[2] = AR_emission_2
+    true_model = HiddenMarkovModel(K=2, A=A, πₖ=πₖ, B=[AR_emission_1, AR_emission_2]);
 
     # Sample from the AR HMM using its own emission models and a starting point
     num_points = 1000
-    
+
     Y = StateSpaceDynamics.construct_AR_feature_matrix(data, order, false)
     X = copy(Y[:, end])
     X = reshape(Y[:, end], 2, 1)
@@ -39,64 +38,56 @@ function test_ARHMM_sampling()
 end
 
 function test_ARHMM_fit()
-    """
-    Get initial sample from GaussianRegression
-    """
-    emission_1 = GaussianRegressionEmission(input_dim=2, output_dim=1, include_intercept=false, β=reshape([2.0 1.0 ], :, 1))
[…]
