### 2025-06-10 dbb4030
Updated constructor handles. They are redundant know and we could consider dropping them altogether
