### 2025-06-10 c1f19f3
refactored weight initializations in functions
