+++ src/LinearDynamicalSystems.jl
@@ -93,7 +93,11 @@ function GaussianObservationModel(;
     C = isempty(C) ? randn(T, obs_dim, latent_dim) : C
     R = isempty(R) ? Matrix{T}(I, obs_dim, obs_dim) : R
 
-    return GaussianObservationModel{T, typeof(C)}(C, R)
+    if !check_same_type(C[1], R[1])
+        error("C and R must be of the same element type. Got $(eltype(C[1])) and $(eltype(R[1]))")
+    end
+
+    return GaussianObservationModel(C, R)
 end
 
 """
@@ -137,7 +141,11 @@ function PoissonObservationModel(;
     C = isempty(C) ? randn(T, obs_dim, latent_dim) : C
     log_d = isempty(log_d) ? randn(T, obs_dim) : log_d
 
-    return PoissonObservationModel{T, typeof(C), typeof(log_d)}(C, log_d)
+    if !check_same_type(C[1], log_d[1])
+        error("C and log_d must be of the same element type. Got $(eltype(C[1])) and $(eltype(log_d[1]))")
+    end
+
+    return PoissonObservationModel(C, log_d)
 end
 
 """
@@ -268,7 +276,10 @@ filter_smooth = initialize_FilterSmooth(model, num_observations)
 
 # `filter_smooth` now contains zero-initialized arrays for smoothing operations
 """
-function initialize_FilterSmooth(model::LinearDynamicalSystem{T, S, O}, num_obs::Int) where {T<:Real, S<:GaussianStateModel{T}, O<:GaussianObservationModel{T}}
+function initialize_FilterSmooth(
+    model::LinearDynamicalSystem{T, S, O}, 
+    num_obs::Int
+) where {T<:Real, S<:GaussianStateModel{T}, O<:GaussianObservationModel{T}}
     num_states = model.latent_dim
     FilterSmooth(
         zeros(T, num_states, num_obs),
@@ -325,9 +336,15 @@ function sample(
 end
 
 function loglikelihood(
-    x::AbstractMatrix{U}, lds::LinearDynamicalSystem{T,S,O}, y::AbstractMatrix{T},
-    w::Vector{T}=ones(T, size(y, 2))
-) where {T<:Real,U<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
+    x::AbstractMatrix{T}, lds::LinearDynamicalSystem{T,S,O}, y::AbstractMatrix{T},
+    w::Union{Nothing,AbstractVector{T}} = nothing
+) where {T<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
+    if w === nothing
+        w = ones(T, size(y,2)) 
+    elseif eltype(w) !== T
+        error("weights must be Vector{$(T)}; Got Vector{$(eltype(w))}")
+    end
+
     T_steps = size(y, 2)
     A, Q, x0, P0 = lds.state_model.A, lds.state_model.Q, lds.state_model.x0, lds.state_model.P0
     C, R = lds.obs_model.C, lds.obs_model.R 
@@ -342,21 +359,20 @@ function loglikelihood(
     ll = sum(abs2, P0_chol \ dx0)
 
     # Create temporaries with compatible element types
-    CommonType = promote_type(T, U)
-    temp_dx = zeros(CommonType, size(x, 1)) 
-    temp_dy = zeros(CommonType, size(y, 1))  
+    temp_dx = zeros(T, size(x, 1)) 
+    temp_dy = zeros(T, size(y, 1))  
 
     @inbounds for t in 1:T_steps
         if t > 1
[…]
