### 2025-06-10 1490e95
fix some tests
