+++ src/LinearDynamicalSystems.jl
@@ -293,7 +293,7 @@ Sample directly from a Linear Dynamical System.
 # Returns
 - `Tuple{Array,Array}`: Latent states and observations
 """
-function Random.rand(rng::AbstractRNG, lds::LinearDynamicalSystem{S,O}; tsteps::Int, ntrials::Int) where {T<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
+function Random.rand(rng::AbstractRNG, lds::LinearDynamicalSystem{T,S,O}; tsteps::Int, ntrials::Int) where {T<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
     A, Q, x0, P0 = lds.state_model.A, lds.state_model.Q, lds.state_model.x0, lds.state_model.P0
     C, R = lds.obs_model.C, lds.obs_model.R
 
@@ -314,7 +314,7 @@ function Random.rand(rng::AbstractRNG, lds::LinearDynamicalSystem{S,O}; tsteps::
 end
 
 # For Poisson LDS
-function Random.rand(rng::AbstractRNG, lds::LinearDynamicalSystem{S,O}; tsteps::Int, ntrials::Int) where {T<:Real,S<:GaussianStateModel{T},O<:PoissonObservationModel{T}}
+function Random.rand(rng::AbstractRNG, lds::LinearDynamicalSystem{T,S,O}; tsteps::Int, ntrials::Int) where {T<:Real,S<:GaussianStateModel{T},O<:PoissonObservationModel{T}}
     # Extract model components
     A, Q = lds.state_model.A, lds.state_model.Q
     C, log_d = lds.obs_model.C, lds.obs_model.log_d
@@ -363,7 +363,7 @@ Calculate the complete-data log-likelihood of a linear dynamical system (LDS) gi
 """
 function loglikelihood(
     x::AbstractMatrix{T}, 
-    lds::LinearDynamicalSystem{S,O}, 
+    lds::LinearDynamicalSystem{U,S,O}, 
     y::AbstractMatrix{U},
     w::Vector{Float64}=ones(size(y, 2))
 ) where {T<:Real,U<:Real,S<:GaussianStateModel{<:Real},O<:GaussianObservationModel{<:Real}}
@@ -644,12 +644,12 @@ This function performs direct smoothing for a linear dynamical system (LDS) give
 
 # Arguments
 - `lds::LinearDynamicalSystem{T,S,O}`: The LDS object representing the system parameters.
-- `y::AbstractArray{T,3}`: The observed data array with dimensions (obs_dim, tiem_steps, n_trials).
+- `y::AbstractArray{T,3}`: The observed data array with dimensions (obs_dim, tsteps, ntrials).
 
 # Returns
-- `x::AbstractArray{T,3}`: The optimal state estimates with dimensions (n_trials, time_steps, latent_dim).
-- `p_smooth::AbstractArray{T,4}`: The posterior covariance matrices with dimensions (latent_dim, latent_dim, time_steps, n_trials).
-- `inverse_offdiag::AbstractArray{T,4}`: The inverse off-diagonal matrices with dimensions (latent_dim, latent_dim, time_steps, n_trials).
+- `x::AbstractArray{T,3}`: The optimal state estimates with dimensions (ntrials, tsteps, latent_dim).
+- `p_smooth::AbstractArray{T,4}`: The posterior covariance matrices with dimensions (latent_dim, latent_dim, tsteps, ntrials).
+- `inverse_offdiag::AbstractArray{T,4}`: The inverse off-diagonal matrices with dimensions (latent_dim, latent_dim, tsteps, ntrials).
 
 # Example
[…]
