+++ src/LinearDynamicalSystems.jl
@@ -373,44 +373,46 @@ Calculate the complete-data log-likelihood of a linear dynamical system (LDS) gi
 - `ll::T`: The complete-data log-likelihood of the LDS.
 """
 function loglikelihood(
-    x::AbstractMatrix{T}, lds::LinearDynamicalSystem{T,S,O}, y::AbstractMatrix{T},
+    x::AbstractMatrix{U}, lds::LinearDynamicalSystem{T,S,O}, y::AbstractMatrix{T},
     w::Union{Nothing,AbstractVector{T}} = nothing
-) where {T<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
+) where {U<:Real, T<:Real, S<:GaussianStateModel{T}, O<:GaussianObservationModel{T}}
+
     if w === nothing
-        w = ones(T, size(y,2)) 
-    elseif eltype(w) !== T
-        error("weights must be Vector{$(T)}; Got Vector{$(eltype(w))}")
+        w = ones(eltype(y), size(y, 2))
+    elseif eltype(w) !== eltype(y)
+        error("weights must be Vector{$(eltype(y))}; Got Vector{$(eltype(w))}")
     end
 
-    T_steps = size(y, 2)
+    tsteps = size(y, 2)
     A, Q, x0, P0 = lds.state_model.A, lds.state_model.Q, lds.state_model.x0, lds.state_model.P0
-    C, R = lds.obs_model.C, lds.obs_model.R 
+    C, R = lds.obs_model.C, lds.obs_model.R
 
-    # Pre-compute Cholesky factors
+    # Use symmetric Cholesky to allow AD with dual types
     R_chol = cholesky(Symmetric(R)).U
     Q_chol = cholesky(Symmetric(Q)).U
     P0_chol = cholesky(Symmetric(P0)).U
 
-    # Initial state contribution
-    dx0 = view(x, :, 1) - x0
+    dx0 = view(x, :, 1) .- x0
     ll = sum(abs2, P0_chol \ dx0)
 
-    # Create temporaries with compatible element types
-    temp_dx = zeros(T, size(x, 1)) 
-    temp_dy = zeros(T, size(y, 1))  
+    # Allocate temporary arrays with correct type
+    temp_dx = zeros(eltype(x), size(x, 1))
+    temp_dy = zeros(eltype(x), size(y, 1))
 
     for t in 1:tsteps
         if t > 1
-            mul!(temp_dx, A, view(x, :, t-1), -one(T), false)  
+            mul!(temp_dx, A, view(x, :, t-1), -one(eltype(x)), false)
             temp_dx .+= view(x, :, t)
             ll += sum(abs2, Q_chol \ temp_dx)
         end
-        mul!(temp_dy, C, view(x, :, t), -one(T), false)  
+        mul!(temp_dy, C, view(x, :, t), -one(eltype(x)), false)
         temp_dy .+= view(y, :, t)
         ll += w[t] * sum(abs2, R_chol \ temp_dy)
     end
-    return -T(0.5) * ll 
+
+    return -eltype(x)(0.5) * ll
 end
+
 
 """
     Gradient(lds::LinearDynamicalSystem{T,S,O}, y::AbstractMatrix{T}, x::AbstractMatrix{T}) where {T<:Real, S<:GaussianStateModel{T}, O<:GaussianObservationModel{T}}
@@ -613,9 +615,11 @@ function smooth(
         return g .= vec(-grad)
     end
 
-    function h!(h::AbstractSparseMatrix, vec_x::Vector{T})
+    function h!(h::AbstractMatrix{T}, vec_x::Vector{T}) where T
         x = reshape(vec_x, D, tsteps)
[…]
