### 2025-06-11 bb26b53
Fix Float32/BigFloat use. Numps min julia version to 1.11
