### 2025-06-11 ca26481
Remove GaussianLDS and PoissonLDS default constructors
