+++ test/LinearDynamicalSystems/PoissonLDS.jl
@@ -1,6 +1,8 @@
+const CHECKED_TYPES = [Float32, Float64, BigFloat]
+
 # define parameters for a PoissonLDS
 x0 = [1.0, -1.0]
-p0 = Matrix(Diagonal([0.1, 0.1]))
+P0 = Matrix(Diagonal([0.1, 0.1]))  # Fixed: was p0, now P0
 A = [cos(0.1) -sin(0.1); sin(0.1) cos(0.1)]
 Q = Matrix(Diagonal([0.1, 0.1]))
 C = [0.6 0.6; 0.6 0.6; 0.6 0.6] .* 2
@@ -9,13 +11,19 @@ log_d = log.([0.1, 0.1, 0.1])
 function toy_PoissonLDS(
     ntrials::Int=1, fit_bool::Vector{Bool}=[true, true, true, true, true, true]
 )
-    poisson_lds = PoissonLDS(;
-        A=A, C=C, Q=Q, x0=x0, P0=p0, log_d=log_d, obs_dim=3, latent_dim=2, fit_bool=fit_bool
+    gaussian_sm = GaussianStateModel(A=A, Q=Q, x0=x0, P0=P0)
+    poisson_om = PoissonObservationModel(C=C, log_d=log_d)  # Fixed: was poisson_sm
+    poisson_lds = LinearDynamicalSystem(
+        state_model=gaussian_sm, 
+        obs_model=poisson_om,  
+        latent_dim=2, 
+        obs_dim=3, 
+        fit_bool=fill(true, 6)
     )
 
     # sample data
     T = 100
-    x, y = StateSpaceDynamics.rand(poisson_lds; tsteps=T, ntrials=ntrials) # 100 timepoints, ntrials
+    x, y = rand(poisson_lds; tsteps=T, ntrials=ntrials) # 100 timepoints, ntrials
 
     return poisson_lds, x, y
 end
@@ -42,10 +50,8 @@ function test_pobs_constructor_type_preservation()
     log_d_int = [5, 6]
 
     pom_int = PoissonObservationModel(
-      C = C_int,
-      log_d = log_d_int,
-      obs_dim = 2,
-      latent_dim = 2
+        C = C_int,
+        log_d = log_d_int
     )
 
     @test eltype(pom_int.C) === Int
@@ -53,16 +59,13 @@ function test_pobs_constructor_type_preservation()
     @test size(pom_int.C) == (2, 2)
     @test length(pom_int.log_d) == 2
 
-
     # Float32
     C_f32     = Float32[1 2; 3 4]
     log_d_f32 = Float32[0.5, 0.6]
 
     pom_f32 = PoissonObservationModel(
-      C = C_f32,
-      log_d = log_d_f32,
-      obs_dim = 2,
-      latent_dim = 2
+        C = C_f32,
+        log_d = log_d_f32
     )
 
     @test eltype(pom_f32.C) === Float32
@@ -70,16 +73,13 @@ function test_pobs_constructor_type_preservation()
     @test size(pom_f32.C) == (2, 2)
     @test length(pom_f32.log_d) == 2
 
-
     # BigFloat
     C_bf = BigFloat[1 2; 3 4]
     log_d_bf = BigFloat[0.1, 0.2]
 
     pom_bf = PoissonObservationModel(
-      C = C_bf,
-      log_d = log_d_bf,
-      obs_dim = 2,
-      latent_dim = 2
+        C = C_bf,
+        log_d = log_d_bf
     )
 
     @test eltype(pom_bf.C) === BigFloat
@@ -97,11 +97,14 @@ function test_plds_constructor_type_preservation()
     x0_int = [5, 6]
     P0_int = [4 0; 0 4]
 
[…]
