+++ test/LinearDynamicalSystems/GaussianLDS.jl
@@ -1,6 +1,6 @@
 const CHECKED_TYPES = [Float32, Float64, BigFloat]
 
-# Define the parameters of a penndulum
+# Define the parameters of a pendulum
 g = 9.81 # gravity
 l = 1.0 # length of pendulum
 dt = 0.01 # time step
@@ -21,8 +21,14 @@ R = Matrix{Float64}((observation_noise_std^2) * I(2))  # Observation noise covar
 function toy_lds(
     ntrials::Int=1, fit_bool::Vector{Bool}=[true, true, true, true, true, true]
 )
-    lds = GaussianLDS(;
-        A=A, C=C, Q=Q, R=R, x0=x0, P0=P0, obs_dim=2, latent_dim=2, fit_bool=fit_bool
+    gaussian_sm = GaussianStateModel(A=A, Q=Q, x0=x0, P0=P0)
+    gaussian_om = GaussianObservationModel(C=C, R=R)
+    lds = LinearDynamicalSystem(
+        state_model=gaussian_sm,
+        obs_model=gaussian_om,
+        latent_dim=2,
+        obs_dim=2,
+        fit_bool=fill(true, 6)
     )
 
     # sample data
@@ -59,14 +65,12 @@ function test_gaussian_obs_constructor_type_preservation()
         Q = Q_int,
         x0 = x0_int,
         P0 = P0_int,
-        latent_dim = 2
     )
 
     @test eltype(gsm_int.A) === Int
     @test eltype(gsm_int.Q) === Int
     @test eltype(gsm_int.x0) === Int
     @test eltype(gsm_int.P0) === Int
-
 
     # Float32
     A_f32 = Float32[1 2; 3 4]
@@ -78,15 +82,13 @@ function test_gaussian_obs_constructor_type_preservation()
         A = A_f32,
         Q = Q_f32,
         x0 = x0_f32,
-        P0 = P0_f32,
-        latent_dim = 2
+        P0 = P0_f32
     )
 
     @test eltype(gsm_f32.A)   === Float32
     @test eltype(gsm_f32.Q)   === Float32
     @test eltype(gsm_f32.x0)  === Float32
     @test eltype(gsm_f32.P0)  === Float32
-
 
     # BigFloat
     A_bf = BigFloat[1 2; 3 4]
@@ -98,8 +100,7 @@ function test_gaussian_obs_constructor_type_preservation()
         A = A_bf,
         Q = Q_bf,
         x0 = x0_bf,
-        P0 = P0_bf,
-        latent_dim = 2
+        P0 = P0_bf
     )
 
     @test eltype(gsm_bf.A) === BigFloat
@@ -117,10 +118,14 @@ function test_gaussian_lds_constructor_type_preservation()
     x0_int = [5, 6]
     P0_int = [4 0; 0 4]
 
-    gls_int = GaussianLDS(Int;
-        A = A_int, C = C_int, Q = Q_int, R = R_int,
-        x0 = x0_int, P0 = P0_int,
-        obs_dim = 2, latent_dim = 2
+    gsm_int = GaussianStateModel(A=A_int, Q=Q_int, x0=x0_int, P0=P0_int)
+    gom_int = GaussianObservationModel(C=C_int, R=R_int)
+    gls_int = LinearDynamicalSystem(
+        state_model=gsm_int,
+        obs_model=gom_int,
+        latent_dim=2,
+        obs_dim=2,
+        fit_bool=fill(true, 6)
     )
 
     @test eltype(gls_int.state_model.A) === Int
@@ -140,10 +145,14 @@ function test_gaussian_lds_constructor_type_preservation()
     x0_f32 = Float32[0.5, 1.5]
     P0_f32 = Float32[4 0; 0 4]
 
[…]

[…]
