+++ src/LinearDynamicalSystems.jl
@@ -4,15 +4,15 @@ export GaussianLDS, PoissonLDS, rand, smooth, fit!
 export GaussianStateModel, PoissonObservationModel 
 
 """
-    GaussianStateModel{T<:Real} <: AbstractStateModel
+    GaussianStateModel{T<:Real. M<:AbstractMatrix{T}, V<:AbstractVector{T}}}
 
 Represents the state model of a Linear Dynamical System with Gaussian noise.
 
-# Fields
-- `A::AbstractMatrix{T}`: Transition matrix
-- `Q::AbstractMatrix{T}`: Process noise covariance
-- `x0::Vector{T}`: Initial state
-- `P0::AbstractMatrix{T}`: Initial state covariance
+# Fields 
+- `A::M`: Transition matrix (size `latent_dim×latent_dim`). 
+- `Q::M`: Process noise covariance matrix 
+- `x0::V`: Initial state vector (length `latent_dim`).
+- `P0::M`: Initial state covariance matrix (size `latent_dim×latent_dim
 """
 mutable struct GaussianStateModel{T<:Real, M<:AbstractMatrix{T}, V<:AbstractVector{T}} <: AbstractStateModel{T}
     A::M
@@ -24,13 +24,13 @@ end
 """
     GaussianStateModel(; A, Q, x0, P0, latent_dim)
 
-Construct a GaussianStateModel with the given parameters or random initializations.
+Construct a GaussianStateModel with the given parameters.
 
 # Arguments
-- `A::Matrix{T}=Matrix{T}(undef, 0, 0)`: Transition matrix
-- `Q::Matrix{T}=Matrix{T}(undef, 0, 0)`: Process noise covariance
-- `x0::Vector{T}=Vector{T}(undef, 0)`: Initial state
-- `P0::Matrix{T}=Matrix{T}(undef, 0, 0)`: Initial state covariance
+- `A::AbstractMatrix{T}=Matrix{T}(undef, 0, 0)`: Transition matrix
+- `Q::AbstractMatrix{T}=Matrix{T}(undef, 0, 0)`: Process noise covariance
+- `x0::AbstractVector{T}=Vector{T}(undef, 0)`: Initial state
+- `P0::AbstractMatrix{T}=Matrix{T}(undef, 0, 0)`: Initial state covariance
 - `latent_dim::Int`: Dimension of the latent state (required if any matrix is not provided.)
 """
 function GaussianStateModel(;
@@ -49,18 +49,18 @@ function GaussianStateModel(;
     x0 = isempty(x0) ? randn(T, latent_dim) : x0
     P0 = isempty(P0) ? Matrix{T}(I, latent_dim, latent_dim) : P0
 
-    return GaussianStateModel{T, typeof(A), typeof(x0)}(A, Q, x0, P0)
+    return GaussianStateModel(A, Q, x0, P0)
 end
 
 
 """
-    GaussianObservationModel{T<:Real} <: AbstractObservationModel
+    GaussianObservationModel{T<:Real, M<:AbstractMatrix{T}} 
 
 Represents the observation model of a Linear Dynamical System with Gaussian noise.
 
 # Fields
-- `C::AbstractMatrix{T}`: Observation matrix
-- `R::AbstractMatrix{T}`: Observation noise covariance
+- `C::M`: Observation matrix of size `(obs_dim × latent_dim)`. Maps latent states into observation space. 
+- `R::M`: Observation noise covariance of size `(obs_dim × obs_dim)`. 
 """
 mutable struct GaussianObservationModel{T<:Real, M<:AbstractMatrix{T}} <: AbstractObservationModel{T}
     C::M
@@ -102,13 +102,13 @@ function GaussianObservationModel(;
 end
 
 """
-    PoissonObservationModel{T<:Real} <: AbstractObservationModel
[…]
