+++ src/Utilities.jl
@@ -4,6 +4,11 @@ export row_matrix, stabilize_covariance_matrix, valid_Σ, make_posdef!, gaussian
 export random_rotation_matrix
 
 # Type checking utilities
+"""
+    check_same_type(args...)
+
+Utility function to check if n arguments share the same types. 
+"""
 function check_same_type(args...)
     if length(args) ≤ 1
         return true  # trivial case
@@ -19,12 +24,6 @@ end
     row_matrix(x::AbstractVector)
 
 Convert a vector to a row matrix.
-
-# Arguments
-- `x::AbstractVector`: The input vector.
-
-# Returns
-- A row matrix (1 × n) containing the elements of `x`.
 """
 function row_matrix(x::AbstractVector)
     return reshape(x, 1, length(x))
@@ -35,22 +34,12 @@ end
 
 Compute the inverse of a block tridiagonal matrix.
 
-# Arguments
-- `A`: Lower diagonal blocks.
-- `B`: Main diagonal blocks.
-- `C`: Upper diagonal blocks.
-
-# Returns
-- `λii`: Diagonal blocks of the inverse.
-- `λij`: Off-diagonal blocks of the inverse.
-
 # Notes: This implementation is from the paper:
-
 "An Accelerated Lambda Iteration Method for Multilevel Radiative Transfer” Rybicki, G.B., and Hummer, D.G., Astronomy and Astrophysics, 245, 171–181 (1991), Appendix B.
 """
-function block_tridiagonal_inverse(A::Vector{Matrix{T}},
-                                   B::Vector{Matrix{T}},
-                                   C::Vector{Matrix{T}}) where {T<:Real}
+function block_tridiagonal_inverse(A::Vector{AbstractMatrix{T}},
+                                   B::Vector{AbstractMatrix{T}},
+                                   C::Vector{AbstractMatrix{T}}) where {T<:Real}
     n = length(B)
     block_size = size(B[1], 1)
 
@@ -111,7 +100,9 @@ end
 Compute the inverse of a block tridiagonal matrix using static matrices. See `block_tridiagonal_inverse` for details.
 """
 function block_tridiagonal_inverse_static(
-    A::Vector{Matrix{T}}, B::Vector{Matrix{T}}, C::Vector{Matrix{T}}
+    A::Vector{AbstractMatrix{T}}, 
+    B::Vector{AbstractMatrix{T}},
+    C::Vector{AbstractMatrix{T}}
 ) where {T<:Real}
     n = length(B)
     N = size(B[1], 1)
@@ -174,14 +165,6 @@ end
     block_tridgm(main_diag::Vector{Matrix{T}}, upper_diag::Vector{Matrix{T}}, lower_diag::Vector{Matrix{T}}) where {T<:Real}
 
 Construct a block tridiagonal matrix from three vectors of matrices.
-
-# Arguments
-- `main_diag::Vector{Matrix{T}}`: Vector of matrices for the main diagonal.
-- `upper_diag::Vector{Matrix{T}}`: Vector of matrices for the upper diagonal.
-- `lower_diag::Vector{Matrix{T}}`: Vector of matrices for the lower diagonal.
-
-# Returns
-- A sparse matrix representing the block tridiagonal matrix.
 
 # Throws
 - `ErrorException` if the lengths of `upper_diag` and `lower_diag` are not one less than the length of `main_diag`.
@@ -262,36 +245,23 @@ function block_tridgm(
     return sparse(I, J, V, N, N, +)
 end
 
+
 # Initialization utilities
 """
     euclidean_distance(a::AbstractVector{Float64}, b::AbstractVector{Float64})
 
 Calculate the Euclidean distance between two points.
-
-# Arguments
[…]

[…]
