### 2025-06-11 3261160
Docs and type genericity changes
