+++ src/EmissionModels.jl
@@ -27,16 +27,18 @@ mutable struct GaussianEmission{T<:Real, V<:AbstractVector{T}, M<:AbstractMatrix
 end
 
 """
-    function GaussianEmission(; output_dim::Int, μ::Vector{<:Real}=zeros(output_dim), Σ::Matrix{<:Real}=Matrix{Float64}(I, output_dim, output_dim))
+    function GaussianEmission(; output_dim::Int, μ::AbstractVector, Σ::AbstractMatrix)
 
-Functon to create a GaussianEmission with given output dimension, mean, and covariance.
+Create a GaussianEmission model.
 
 # Arguments
 - `output_dim::Int`: The output dimension of the emission
-- `μ::Vector{<:Real}=zeros(output_dim)`: The mean of the Gaussian
-- `Σ::Matrix{<:Real}=Matrix{Float64}(I, output_dim, output_dim))`: The covariance matrix of the Gaussian
+- `μ::AbstractVector`: The mean of the Gaussian
+- `Σ::AbstractMatrix`: The covariance matrix of the Gaussian
 
 # Returns
+- `GaussianEmission<:EmissionModel`: The Gaussian emission model
+```
 """
 function GaussianEmission(; 
     output_dim::Int,
@@ -51,16 +53,9 @@ end
 
 
 """
-    loglikelihood(model::GaussianEmission, Y::AbstractMatrix{<:Real})
+    loglikelihood(model::GaussianEmission, Y::AbstractMatrix{T}) where {T<:Real}
 
 Calculate the log likelihood of the data `Y` given the Gaussian emission model.
-
-# Arguments
-- `model::GaussianEmission`: The Gaussian emission model for which to calculate the log likelihood.
-- `Y::AbstractMatrix{<:Real}`: The data matrix, where each row represents an observation.
-
-# Returns
-- `Vector{Float64}`: A vector of log likelihoods, one for each observation in the data.
 """
 function loglikelihood(model::GaussianEmission, Y::AbstractMatrix{T}) where {T<:Real}
     # Create MvNormal distribution with the model parameters
@@ -71,28 +66,38 @@ function loglikelihood(model::GaussianEmission, Y::AbstractMatrix{T}) where {T<:
 end
 
 """
-    sample(model::Gaussian; n::Int=1)
+     Random.rand(rng::AbstractRNG, model::GaussianEmission; n::Int=1)
 
-Generate `n` samples from a Gaussian model. Returns a matrix of size `(n, output_dim)`.
+Generate `n` samples from a Gaussian emission model.
+
+# Arguments
+    - `rng::AbstractRNG`: Seed
+    -`model::GaussianEmission`: The GaussianEmission model
+    -`n::Int=1`: The number of samples to generate
+
+# Returns
+    -`samples::Matrix{<:Real}`: Matrix of samples (n, output_dim)
 """
 function Random.rand(rng::AbstractRNG, model::GaussianEmission; n::Int=1)
     raw_samples = rand(rng, MvNormal(model.μ, model.Σ), n)
     return Matrix(raw_samples')
 end
 
+"""
+    Random.rand(model::GaussianEmission; kwargs...)
+
+Generate random samples from  a Gaussian emission model.
+"""
 function Random.rand(model::GaussianEmission; kwargs...)
     return rand(Random.default_rng(), model; kwargs...)
 end
 
 """
-    fit!(model::GaussianEmission, Y::AbstractMatrix{<:Real}, w::AbstractVector{Float64}=ones(size(Y, 1)))
+    function fit!(model::GaussianEmission, 
+            Y::AbstractMatrix{T}, 
[…]
