+++ src/HiddenMarkovModels.jl
@@ -6,14 +6,13 @@ export estep, class_probabilities
 """
     HiddenMarkovModel
 
-A Hidden Markov Model (HMM) with custom emissions.
+Store a Hidden Markov Model (HMM) with custom emissions.
 
 # Fields
+- `A::AbstractMatrix{<:Real}`: Transition matrix.
+- `B::AbstractVector{<:EmissionModel}`: State-dependent emission models.
+- `πₖ::AbstractVector{<:Real}`: Initial state distribution.
 - `K::Int`: Number of states.
-- `B::Vector=Vector()`: Vector of emission models.
-- `emission=nothing`: If B is missing emissions, clones of this model will be used to fill in the rest.
-- `A::Matrix{<:Real}`: Transition matrix.
-- `πₖ::Vector{Float64}`: Initial state distribution.
 """
 mutable struct HiddenMarkovModel{T<:Real, V<:AbstractVector{T}, M<:AbstractMatrix{T}, VE<:AbstractVector{<:EmissionModel}} <: AbstractHMM
     A::M # transition matrix
@@ -22,17 +21,33 @@ mutable struct HiddenMarkovModel{T<:Real, V<:AbstractVector{T}, M<:AbstractMatri
     K::Int # number of states
 end
 
-function initialize_forward_backward(model::AbstractHMM, num_obs::Int)
+"""
+    initialize_forward_backward(model::AbstractHMM, num_obs::Int)
+
+Initialize the forward backward storage struct.
+"""
+function initialize_forward_backward(model::AbstractHMM, num_obs::Int, ::Type{T}) where {T<:Real}
     num_states = model.K
     ForwardBackward(
-        zeros(num_states, num_obs),
-        zeros(num_states, num_obs),
-        zeros(num_states, num_obs),
-        zeros(num_states, num_obs),
-        zeros(Float64, num_states, num_states, num_obs - 1)
+        zeros(T, num_states, num_obs),
+        zeros(T, num_states, num_obs),
+        zeros(T, num_states, num_obs),
+        zeros(T, num_states, num_obs),
+        zeros(T, num_states, num_states, num_obs - 1)
     )
 end
 
+
+"""
+    ForwardBackward(
+        loglikelihoods::Matrix{T},
+        α::Matrix{T},
+        β::Matrix{T},
+        γ::Matrix{T},
+        ξ::Array{T,3}) where {T<:Real}
+
+Initialize the forward backward storage struct.
+"""
 function ForwardBackward(
     loglikelihoods::Matrix{T},
     α::Matrix{T},
@@ -43,6 +58,11 @@ function ForwardBackward(
     ForwardBackward{T, Vector{T}, Matrix{T}, Array{T,3}}(loglikelihoods, α, β, γ, ξ)
 end
 
+"""
+    aggregate_forward_backward!(aggregated_FB::ForwardBackward, FB_storages::Vector{<:ForwardBackward})
+
+Aggregate single trial ForwardBackward structs to one session-wide struct.
+"""
 function aggregate_forward_backward!(
     aggregated_FB::ForwardBackward, 
     FB_storages::Vector{<:ForwardBackward}
@@ -56,6 +76,11 @@ function aggregate_forward_backward!(
     
 end
 
+"""
+    initialize_transition_matrix(K::Int)
+
+Initialize the HMM transition matrix.
+"""
 function initialize_transition_matrix(K::Int)
     # initialize a transition matrix
     A = zeros(Float64, K, K)
@@ -65,38 +90,42 @@ function initialize_transition_matrix(K::Int)
     return A
 end
 
+"""
+    initialize_state_distribution(K::Int)
+
+Initialize the HMM initial state distribution
+"""
 function initialize_state_distribution(K::Int)
     # initialize a state distribution
[…]
