### 2025-06-11 cf02a0d
Type genericity changes
