### 2025-06-11 03a3cbc
Added float32 tests
