### 2025-06-11 4effea4
Added float32 tests -- seems like they cause numerical instabilities and lead to decreasing LL
