### 2025-06-11 2248110
float32 test for gaussianhmm added
