+++ src/Preprocessing.jl
@@ -1,138 +1,61 @@
+# Public API
 export ProbabilisticPCA, loglikelihood, fit!
-
-"""
-    mutable struct ProbabilisticPCA{T<:Real, M<:AbstractMatrix{T}, V<:AbstractVector{T}}
-
-Probabilistic PCA model from Bishop's Pattern Recognition and Machine Learning.
-
-# Fields:
-    W: Weight matrix that maps from latent space to data space.
-    σ²: Noise variance
-    μ: Mean of the data 
-    k: Number of latent dimensions
-    D: Dimension of the data 
-    z: Latent variables
-"""
 mutable struct ProbabilisticPCA{T<:Real, M<:AbstractMatrix{T}, V<:AbstractVector{T}}
-    W::M 
-    σ²::T 
+    W::M
+    σ²::T
     μ::V
     k::Int
-    D::Int 
-    z::M 
+    D::Int
+    z::M
+
+    function ProbabilisticPCA(W::AbstractMatrix{T}, σ²::T, μ::AbstractVector{T}) where {T<:Real}
+        D, k = size(W)
+        z = Matrix{T}(undef, k, 0)  # placeholder, filled after E-step
+        new{T, typeof(W), typeof(μ)}(W, σ², μ, k, D, z)
+    end
 end
 
-"""
-    E_Step(ppca::ProbabilisticPCA, X::Matrix{<:AbstractFloat})
-
-Expectation step of the EM algorithm for PPCA. See Bishop's Pattern Recognition and Machine Learning for more details.
-"""
 function estep(ppca::ProbabilisticPCA, X::Matrix{T}) where {T<:Real}
-    # get dims
-    N, D = size(X)
-    @assert D == ppca.D "Data dimension mismatch: expected $(ppca.D), got $D"
-    
-    # preallocate E_z and E_zz
-    E_z = zeros(T, N, ppca.k)
-    E_zz = zeros(T, N, ppca.k, ppca.k)
-    
-    # calculate M
-    M = ppca.W' * ppca.W + (ppca.σ² * I(ppca.k))
-    M_inv = cholesky(M).U \ (cholesky(M).L \ I(ppca.k))
-    # calculate E_z and E_zz
+    D, N = size(X)
+    E_z = zeros(T, ppca.k, N)
+    E_zz = zeros(T, ppca.k, ppca.k, N)
+    M = ppca.W' * ppca.W + ppca.σ² * I(ppca.k)
+    M_inv = inv(M)
     @views for i in 1:N
-        E_z[i, :] = M_inv * ppca.W' * (X[i, :] .- ppca.μ)
-        E_zz[i, :, :] = (ppca.σ² * M_inv) + (E_z[i, :]  * E_z[i, :]')
+        E_z[:, i] .= M_inv * ppca.W' * (X[:, i] - ppca.μ)
+        E_zz[:, :, i] .= ppca.σ² * M_inv + E_z[:, i] * E_z[:, i]'
     end
-    
     return E_z, E_zz
 end
 
-
-"""
-    M_Step!(model::ProbabilisticPCA, X::Matrix{<:AbstractFloat}, E_z::Matrix{<:AbstractFloat}, E_zz::Array{<:AbstractFloat, 3}
-
-Maximization step of the EM algorithm for PPCA. See Bishop's Pattern Recognition and Machine Learning for more details.
-"""
-function mstep!(
-    ppca::ProbabilisticPCA, X::Matrix{T}, E_z::AbstractArray{T}, E_zz::AbstractArray{T}
-) where {T<:Real}
-    # get dims
-    N, D = size(X)
-    
-    # Calculate the sum of E[z_i z_i^T] across all samples
-    sum_E_zz = sum(E_zz, dims=1)[1, :, :]  # Shape: (k, k)
[…]
