### 2025-06-12 cfa6aa1
New homepage example. Starts addign Literate.jl functionality
