### 2025-06-12 10d048b
Update Switching regression test LL tolerances
