### 2025-06-12 e8d93a9
oops, had some tests commented out. fixed now
