+++ docs/make.jl
@@ -6,6 +6,15 @@ using Random
 # Set up the documentation environment
 DocMeta.setdocmeta!(StateSpaceDynamics, :DocTestSetup, :(using StateSpaceDynamics, Random); recursive=true)
 
+# Run Literate.jl to convert tutorial scripts to markdown in docs/src/tutorials/
+Literate.markdown(
+    joinpath(@__DIR__, "examples", "GaussianLDS.jl"), # input path
+    joinpath(@__DIR__, "src", "tutorials");                        # output directory
+    name = "latent_dynamics_example",                              # output .md filename
+    documenter = true                                              # format for Documenter
+)
+
+# Generate the documentation site
 makedocs(;
     modules=[StateSpaceDynamics],
     authors="Ryan Senne",
@@ -23,10 +32,14 @@ makedocs(;
             "EmissionModels" => "EmissionModels.md",
             "Mixture Models" => "MixtureModels.md",
         ],
+        "Tutorials" => [
+            "Gaussian LDS Example" => "tutorials/latent_dynamics_example.md"
+        ],
         "Miscellaneous" => "Misc.md",
     ],
-    checkdocs = :exports,  # Only check exported functions
-    warnonly = true # Do not fail on warnings
+    checkdocs = :exports,
+    warnonly = true
 )
 
+# Deploy the documentation
 deploydocs(; repo="github.com/depasquale-lab/StateSpaceDynamics.jl", devbranch="docs_dev_")
