+++ docs/src/tutorials/latent_dynamics_example.md
@@ -1,18 +1,25 @@
-# ## Simulating and Fitting a Linear Dynamical System
+```@meta
+EditURL = "../../examples/GaussianLDS.jl"
+```
 
-# This tutorial demonstrates how to use `StateSpaceDynamics.jl` to simulate a latent
-# linear dynamical system and fit it using the EM algorithm.
+Simulating and Fitting a Linear Dynamical System
 
-# ## Load Packages
+This tutorial demonstrates how to use `StateSpaceDynamics.jl` to simulate a latent
+linear dynamical system and fit it using the EM algorithm.
 
+## Load Packages
+
+````@example latent_dynamics_example
 using StateSpaceDynamics
 using LinearAlgebra
 using Random
 using Plots
 using LaTeXStrings
+````
 
-# ## Create a State-Space Model
+## Create a State-Space Model
 
+````@example latent_dynamics_example
 obs_dim = 10
 latent_dim = 2
 
@@ -34,14 +41,18 @@ true_lds = LinearDynamicalSystem(;
     obs_dim=obs_dim,
     fit_bool=fill(true, 6)
 )
+````
 
-# ## Simulate Latent and Observed Data
+## Simulate Latent and Observed Data
 
+````@example latent_dynamics_example
 tSteps = 500
 latents, observations = rand(true_lds; tsteps=tSteps, ntrials=1)
+````
 
-# ## Plot Vector Field of Latent Dynamics
+## Plot Vector Field of Latent Dynamics
 
+````@example latent_dynamics_example
 x = y = -3:0.5:3
 X = repeat(x', length(y), 1)
 Y = repeat(y, 1, length(x))
@@ -65,9 +76,11 @@ p = quiver(X, Y, quiver=(U_norm, V_norm), color=:blue, alpha=0.3,
            linewidth=1, arrow=arrow(:closed, :head, 0.1, 0.1))
 plot!(latents[1, :, 1], latents[2, :, 1], xlabel="x₁", ylabel="x₂",
       color=:black, linewidth=1.5, title="Latent Dynamics", legend=false)
+````
 
-# ## Plot Latent States and Observations
+## Plot Latent States and Observations
 
+````@example latent_dynamics_example
 states = latents[:, :, 1]
 emissions = observations[:, :, 1]
 
@@ -94,9 +107,11 @@ plot!(subplot=2, yticks=(-lim_emissions .* (obs_dim-1:-1:0), [L"y_{%$n}" for n i
       yformatter=y->"", tickfontsize=12)
 
 plot!(link=:x, size=(800, 600), left_margin=10Plots.mm)
+````
 
-# ## Initialize a Model and Perform Smoothing
+## Initialize a Model and Perform Smoothing
 
+````@example latent_dynamics_example
 A_init = random_rotation_matrix(2)
 Q_init = Matrix(0.1 * I(2))
 C_init = randn(obs_dim, latent_dim)
@@ -125,9 +140,11 @@ end
 plot!(subplot=1, yticks=(lim_states .* (0:latent_dim-1), [L"x_%$d" for d in 1:latent_dim]),
       xticks=[], xlims=(0, tSteps), yformatter=y->"", tickfontsize=12,
       title="True vs. Predicted Latent States (Pre-EM)")
+````
 
-# ## Fit Model Using EM Algorithm
+## Fit Model Using EM Algorithm
 
+````@example latent_dynamics_example
 elbo, _ = fit!(naive_ssm, observations; max_iter=100, tol=1e-6)
 
 x_smooth, _, _ = StateSpaceDynamics.smooth(naive_ssm, observations)
[…]
