### 2025-06-12 ca91eb5
Add PoissonLDS example!
