+++ docs/src/tutorials/poisson_latent_dynamics_example.md
@@ -1,26 +1,41 @@
-# # Simulating and Fitting a Poisson Linear Dynamical System
+```@meta
+EditURL = "../../examples/PoissonLDS.jl"
+```
 
-# This tutorial demonstrates how to use `StateSpaceDynamics.jl` to simulate and fit a
-# Linear Dynamical System (LDS) with Poisson observations using the Laplace-EM algorithm.
+# Simulating and Fitting a Poisson Linear Dynamical System
 
-# ## Load Packages
+This tutorial demonstrates how to use `StateSpaceDynamics.jl` to simulate and fit a
+Linear Dynamical System (LDS) with Poisson observations using the Laplace-EM algorithm.
 
+## Load Packages
+
+````@example poisson_latent_dynamics_example
 using StateSpaceDynamics
 using LinearAlgebra
 using Random
 using Plots
 using LaTeXStrings
+````
 
-# ## Create a Poisson Linear Dynamical System
+## Create a Poisson Linear Dynamical System
 
+````@example poisson_latent_dynamics_example
 obs_dim = 10
 latent_dim = 2
+````
 
+State model parameters
+
+````@example poisson_latent_dynamics_example
 A = 0.95 * [cos(0.25) -sin(0.25); sin(0.25) cos(0.25)]
 Q = Matrix(0.1 * I(latent_dim))
 x0 = zeros(latent_dim)
 P0 = Matrix(0.1 * I(latent_dim))
+````
 
+Observation model parameters
+
+````@example poisson_latent_dynamics_example
 log_d = log.(fill(0.1, obs_dim))
 C = permutedims([abs.(randn(obs_dim))'; abs.(randn(obs_dim))'])
 
@@ -34,14 +49,18 @@ true_plds = LinearDynamicalSystem(;
     obs_dim=obs_dim,
     fit_bool=fill(true, 6)
 )
+````
 
-# ## Simulate Latent States and Observations
+## Simulate Latent States and Observations
 
+````@example poisson_latent_dynamics_example
 tSteps = 500
 latents, observations = rand(true_plds; tsteps=tSteps, ntrials=1)
+````
 
-# ## Plot Vector Field of Latent Dynamics
+## Plot Vector Field of Latent Dynamics
 
+````@example poisson_latent_dynamics_example
 x = y = -3:0.5:3
 X = repeat(x', length(y), 1)
 Y = repeat(y, 1, length(x))
@@ -62,15 +81,21 @@ p = quiver(X, Y, quiver=(U_norm, V_norm), color=:blue, alpha=0.3,
            linewidth=1, arrow=arrow(:closed, :head, 0.1, 0.1))
 plot!(latents[1, :, 1], latents[2, :, 1], xlabel="x₁", ylabel="x₂",
       color=:black, linewidth=1.5, title="Latent Dynamics", legend=false)
+````
 
-# ## Plot Latent States and Observations
+## Plot Latent States and Observations
 
+````@example poisson_latent_dynamics_example
 states = latents[:, :, 1]
 emissions = observations[:, :, 1]
 time_bins = size(states, 2)
 
 plot(size=(800, 600), layout=@layout[a{0.3h}; b])
+````
 
+Latent states
+
+````@example poisson_latent_dynamics_example
 lim_states = maximum(abs.(states))
 for d in 1:latent_dim
     plot!(1:time_bins, states[d, :] .+ lim_states * (d-1), color=:black,
@@ -80,7 +105,11 @@ end
 plot!(subplot=1, yticks=(lim_states .* (0:latent_dim-1), [L"x_%$d" for d in 1:latent_dim]),
       xticks=[], xlims=(0, time_bins), title="Simulated Latent States",
[…]
