+++ docs/examples/GaussianLDS.jl
@@ -10,6 +10,10 @@ using LinearAlgebra
 using Random
 using Plots
 using LaTeXStrings
+using StableRNGs
+
+#
+rng = StableRNG(1234);
 
 # ## Create a State-Space Model
 
@@ -38,7 +42,7 @@ true_lds = LinearDynamicalSystem(;
 # ## Simulate Latent and Observed Data
 
 tSteps = 500
-latents, observations = rand(true_lds; tsteps=tSteps, ntrials=1)
+latents, observations = rand(rng, true_lds; tsteps=tSteps, ntrials=1)
 
 # ## Plot Vector Field of Latent Dynamics
 
@@ -97,7 +101,7 @@ plot!(link=:x, size=(800, 600), left_margin=10Plots.mm)
 
 # ## Initialize a Model and Perform Smoothing
 
-A_init = random_rotation_matrix(2)
+A_init = random_rotation_matrix(2, rng)
 Q_init = Matrix(0.1 * I(2))
 C_init = randn(obs_dim, latent_dim)
 R_init = Matrix(0.5 * I(obs_dim))
