+++ docs/make.jl
@@ -60,7 +60,7 @@ makedocs(;
             "Gaussian LDS Example" => "tutorials/gaussian_latent_dynamics_example.md",
             "Poisson LDS Example" => "tutorials/poisson_latent_dynamics_example.md",
             "Hidden Markov Model Example" => "tutorials/hidden_markov_model_example.md",
-            "Gaussian GLM-GMM Example" => "tutorials/gaussian_glm_hmm_example"
+            "Gaussian GLM-GMM Example" => "tutorials/gaussian_glm_hmm_example.md"
 
         ],
         "Miscellaneous" => "Misc.md",
