### 2025-06-12 4b94ef6
Wrote Mixture Model Example docs, fixed Docs for MixtureModels.jl
