+++ Project.toml
@@ -14,6 +14,7 @@ ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
+StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
@@ -30,6 +31,7 @@ ProgressMeter = "1.10.2"
 Random = "1"
 SparseArrays = "1.9.0"
 SpecialFunctions = "2.4.0"
+StableRNGs = "1.0.3"
 StaticArrays = "1.9.8"
 Statistics = "1"
 StatsBase = "0.34.3"

[…]
