### 2025-06-13 c5f3b71
Reduce allocations in Poisson Hessian
