### 2025-06-13 61f9247
More performance uopdates in the Hessians
