### 2025-06-13 d35739d
render plots in Mixture Models
