### 2025-06-13 91525d1
resolve plot rendering
