### 2025-06-13 c479a41
PPCA example docs, updated Make.jl
