### 2025-06-13 eadcc0b
Start Emissions documentation. Fix a few errors
