+++ docs/make.jl
@@ -54,7 +54,7 @@ Literate.markdown(
     documenter = true
 )
 
-# Convert PPCA  example
+# Convert PPCA example
 Literate.markdown(
     joinpath(@__DIR__, "examples", "ProbabilisticPCA.jl"),
     joinpath(@__DIR__, "src", "tutorials");
@@ -78,8 +78,7 @@ makedocs(;
             "Hidden Markov Models" => "HiddenMarkovModels.md",
             "Switching Linear Dynamical Systems" => "SLDS.md",
             "EmissionModels" => "EmissionModels.md",
-            "Mixture Models" => "MixtureModels.md",
-            "Probabilistic PCA" => "ProbabilisticPCA.md"
+            "Mixture Models" => "MixtureModels.md"
         ],
         "Tutorials" => [
             "Gaussian LDS Example" => "tutorials/gaussian_latent_dynamics_example.md",
@@ -88,7 +87,7 @@ makedocs(;
             "Gaussian GLM-GMM Example" => "tutorials/gaussian_glm_hmm_example.md",
             "Gaussian Mixture Model Example" => "tutorials/gaussian_mixture_model_example.md",
             "Poisson Mixture Model Example" => "tutorials/poisson_mixture_model_example.md",
-            "Probabilistic PCA Example" => "tutorials/ProbabilisticPCA.md",
+            "Probabilistic PCA Example" => "tutorials/Probabilistic_PCA_example.md",
         ],
         "Miscellaneous" => "Misc.md",
     ],
