### 2025-06-13 5cf799b
defined w_fit
