### 2025-06-14 1009934
Fix docrefs in Emissions. Also fix PPCA example
