+++ docs/src/tutorials/Probabilistic_PCA_example.md
@@ -1,10 +1,15 @@
-# ## Simulating and Fitting a Gaussian Mixture Model 
+```@meta
+EditURL = "../../examples/ProbabilisticPCA.jl"
+```
 
-# This tutorial demonstrates how to use `StateSpaceDynamics.jl` to create a 
-# Probabilistic PCA model and fit it using the EM algorithm.
+## Simulating and Fitting a Gaussian Mixture Model
 
-# ## Load Packages
+This tutorial demonstrates how to use `StateSpaceDynamics.jl` to create a
+Probabilistic PCA model and fit it using the EM algorithm.
 
+## Load Packages
+
+````@example Probabilistic_PCA_example
 using StateSpaceDynamics
 using LinearAlgebra
 using Random
@@ -12,16 +17,23 @@ using Plots
 using StatsPlots
 using StableRNGs
 using Distributions
+````
 
-#
+````@example Probabilistic_PCA_example
 rng = StableRNG(1234);
+nothing #hide
+````
 
-# ## Create a State-Space Model
+## Create a State-Space Model
 
+````@example Probabilistic_PCA_example
 D = 2
-k = 2 
+k = 2
+````
 
-# define true parameters for the model to sample from 
+define true parameters for the model to sample from
+
+````@example Probabilistic_PCA_example
 W_true = [
    -1.64   0.2;
     0.9  -2.8
@@ -31,13 +43,18 @@ W_true = [
 μ_true = [1.65, -1.3]
 
 ppca = ProbabilisticPCA(W_true, σ²_true, μ_true)
-   
-# Sample data from the model 
-num_obs = 500 
+````
+
+Sample data from the model
+
+````@example Probabilistic_PCA_example
+num_obs = 500
 X, z = rand(rng, ppca, num_obs)
+````
 
-# ## Plot sampled data from the model 
+## Plot sampled data from the model
 
+````@example Probabilistic_PCA_example
 x1 = X[1, :]
 x2 = X[2, :]
 labels = map(i -> (abs(z[1,i]) > abs(z[2,i]) ? 1 : 2), 1:size(z,2))
@@ -56,22 +73,30 @@ scatter!(
 )
 
 display(p)
+````
 
-# ## Paramter recovery: Initialize a new model with default parameters and fit to the data using EM.  
+## Paramter recovery: Initialize a new model with default parameters and fit to the data using EM.
 
-#define default parameters 
+````@example Probabilistic_PCA_example
+#define default parameters
 D = 2
-k = 2 
+k = 2
 W = randn(D, k)
 σ² = 0.5
 μ_vector = randn(2)
 
 fit_ppca = ProbabilisticPCA(W, σ², μ_vector)
+````
 
-# ## Fit model using EM Algorithm 
+## Fit model using EM Algorithm
+
+````@example Probabilistic_PCA_example
 lls = fit!(fit_ppca, X)
+````
 
-# ## Confirm model convergence using log likelihoods 
+## Confirm model convergence using log likelihoods
+
+````@example Probabilistic_PCA_example
 ll_plot = plot(
     lls;
     xlabel="Iteration",
@@ -83,14 +108,16 @@ ll_plot = plot(
 )
 
 display(ll_plot)
+````
 
-# ## Plot the learned lower dimension latent space over the Data 
+## Plot the learned lower dimension latent space over the Data
 
+````@example Probabilistic_PCA_example
 x1, x2 = X[1, :], X[2, :]
 μ1, μ2  = fit_ppca.μ
 W_fit = fit_ppca.W
-w1      = W_fit[:, 1]  
-w2      = W_fit[:, 2] 
[…]

[…]
