+++ docs/examples/GaussianMixtureModel.jl
@@ -56,6 +56,8 @@ scatter!(
   alpha=0.8,
 )
 
+p
+
 # ## Paramter recovery: Initialize a new model with default parameters and fit to the data 
 
 k = 3
@@ -112,3 +114,5 @@ for i in 1:fit_gmm.k
       label     = "Comp $i",
     )
 end
+
+p
