+++ docs/src/tutorials/poisson_mixture_model_example.md
@@ -64,6 +64,8 @@ histogram!(
   alpha     = 0.7,
   legend    = :topright,
 )
+
+p
 ````
 
 ## Paramter recovery: Initialize a new model with default parameters and fit to the data using EM.
@@ -137,6 +139,8 @@ plot!(
   lw    = 3, ls=:dash, c=:black,
   label = "Mixture",
 )
+
+p
 ````
 
 ---
