+++ docs/src/tutorials/Probabilistic_PCA_example.md
@@ -72,7 +72,7 @@ scatter!(
     markersize = 5,
 )
 
-display(p)
+p
 ````
 
 ## Paramter recovery: Initialize a new model with default parameters and fit to the data using EM.
@@ -107,7 +107,7 @@ ll_plot = plot(
     reuse=false,
 )
 
-display(ll_plot)
+ll_plot
 ````
 
 ## Plot the learned lower dimension latent space over the Data
@@ -132,8 +132,8 @@ scatter!(
 )
 
 k = size(W_fit, 2)
-bases_x = repeat([μ_fit[1]], k)
-bases_y = repeat([μ_fit[2]], k)
+bases_x = repeat([μ1], k)
+bases_y = repeat([μ2], k)
 ````
 
 Add the component arrows in both directions
@@ -169,7 +169,7 @@ quiver!(
   label       = ""
 )
 
-display(P)
+P
 ````
 
 ---
