### 2025-06-14 4b430d1
Edits to MixtureModels, kmeans, and examples.
Does the following:
1.) Unifies API
    a.) E_Step->estep
    b.) M_Step! -> mstep!
    c.) log_likelihood -> loglikelihood
2.) Row major to column major in MixtureModels and kmeans
3.) Fix tests related to above
4.) fixes examples related to above
