+++ docs/examples/GaussianMixtureModel.jl
@@ -1,9 +1,7 @@
-# ## Simulating and Fitting a Gaussian Mixture Model 
+# ## Simulating and Fitting a Gaussian Mixture Model
 
 # This tutorial demonstrates how to use `StateSpaceDynamics.jl` to 
 # create a Gaussian Mixture Model and fit it using the EM algorithm.
-
-# ## Load Packages
 
 using StateSpaceDynamics
 using LinearAlgebra
@@ -13,106 +11,87 @@ using StableRNGs
 using Distributions
 using StatsPlots
 
-#
 rng = StableRNG(1234);
 
-# ## Create a State-Space Model
+# ## Create a true GaussianMixtureModel to simulate from
 
 k = 3
-data_dim = 2
+D = 2  # data dimension
 
-# define true parameters for the model to sample from 
 true_μs = [
-    -1.0  -1.0;
-     1.0  -1.5;
-     0.0   2.0
-]
-true_Σs = [0.3 * Matrix{Float64}(I, data_dim, data_dim) for _ in 1:k]
+    -1.0  1.0  0.0;
+    -1.0 -1.5  2.0
+]  # shape (D, K)
+
+true_Σs = [Matrix{Float64}(0.3 * I(2)) for _ in 1:k]
 true_πs = [0.5, 0.2, 0.3]
 
 true_gmm = GaussianMixtureModel(k, true_μs, true_Σs, true_πs)
 
-# Generate sample data
+# ## Sample data from the true GMM
+
 n = 500
-X = rand(true_gmm, n)
+X = rand(rng, true_gmm, n)  # shape (D, N)
 
-# ## Plot sampled data from the model 
-labels = rand(rng, Distributions.Categorical(true_πs), n)
-X2 = Array{Float64,2}(undef, n, data_dim)
-for i in 1:n
-    comp = labels[i]
-    X2[i, :] = rand(rng, MvNormal(true_μs[comp, :], true_Σs[comp]))'
-end
+# Also generate component labels (for plotting)
+labels = rand(rng, Categorical(true_πs), n)
 
-p = plot()
+p1 = scatter(
+    X[1, :], X[2, :];
+    group=labels,
+    title="GMM Samples Colored by Component",
+    xlabel="x₁", ylabel="x₂",
+    markersize=4,
+    alpha=0.8,
+    legend=false,
+)
+p1
 
-scatter!(
-  p, 
-  X2[:,1], X2[:,2];
-  group=labels,
-  title="GMM Samples Coloured by Component",
-  xlabel="x₁", ylabel="x₂",
-  markersize=4,
-  alpha=0.8,
+# ## Fit a new GaussianMixtureModel to the data
+
+fit_gmm = GaussianMixtureModel(k, D)
+
+class_probabilities, lls = fit!(fit_gmm, X;
+    maxiter=100, tol=1e-6, initialize_kmeans=true)
+
+# ## Plot log-likelihoods to visualize EM convergence
+
+p2 = plot(
+    lls;
+    xlabel="Iteration",
+    ylabel="Log-Likelihood",
+    title="EM Convergence",
+    label="log_likelihood",
+    marker=:circle,
+)
+p2
+
+# ## Visualize model contours over the data
+
+xs = range(minimum(X[1, :]) - 1, stop=maximum(X[1, :]) + 1, length=150)
+ys = range(minimum(X[2, :]) - 1, stop=maximum(X[2, :]) + 1, length=150)
+
+p3 = scatter(
+    X[1, :], X[2, :];
+    markersize=3, alpha=0.5,
+    xlabel="x₁", ylabel="x₂",
+    title="Data & Fitted GMM Contours by Component",
[…]
