+++ src/MixtureModels.jl
@@ -1,22 +1,16 @@
-export GaussianMixtureModel,
-    PoissonMixtureModel, fit!, log_likelihood,  E_Step, M_Step!, MixtureModel
+# Public API
+export GaussianMixtureModel, PoissonMixtureModel, fit!, loglikelihood
 
 """
     GaussianMixtureModel
 
 A Gaussian Mixture Model for clustering and density estimation.
-
-# Fields
-- `k::Int`: Number of clusters.
-- `μₖ::Matrix{<:Real}`: Means of each cluster (dimensions: data_dim x k).
-- `Σₖ::Array{Matrix{<:Real}, 1}`: Covariance matrices of each cluster.
-- `πₖ::Vector{Float64}`: Mixing coefficients for each cluster.
 """
 mutable struct GaussianMixtureModel{T<:Real,M<:AbstractMatrix{T},V<:AbstractVector{T}} <: MixtureModel
-    k::Int # Number of clusters
-    μₖ::M # Means of each cluster
-    Σₖ::AbstractArray{M,1} # Covariance matrices of each cluster
-    πₖ::V # Mixing coefficients
+    k::Int
+    μₖ::M           # Means of each cluster (D, K)
+    Σₖ::Vector{M}     # Covariance matrices of each cluster (D, D)
+    πₖ::V             # Mixing coefficients (K)
 end
 
 """
@@ -25,10 +19,10 @@ end
 Constructor for GaussianMixtureModel. Initializes Σₖ's covariance matrices to the 
 identity, πₖ to a uniform distribution, and μₖ's means to zeros.
 """
-function GaussianMixtureModel(k::Int, data_dim::Int)
-    Σ = [Matrix{Float64}(I, data_dim, data_dim) for _ in 1:k]
+function GaussianMixtureModel(k::Int, D::Int)
+    Σ = [Matrix{Float64}(I, D, D) for _ in 1:k]
     πs = ones(k) ./ k
-    μ = zeros(Float64, k, data_dim)  # Mean of each cluster initialized to zero matrix
+    μ = randn(D, k)
     return GaussianMixtureModel(k, μ, Σ, πs)
 end
 
@@ -38,99 +32,72 @@ end
 Draw 'n' samples from gmm. Returns a Matrix{Float64}, where each row is a data point.
 """
 function Random.rand(rng::AbstractRNG, gmm::GaussianMixtureModel, n::Int)
-    component_samples = rand(rng, Multinomial(n, gmm.πₖ), 1)
-    samples = Matrix{Float64}(undef, n, size(gmm.μₖ, 2))
-    start_idx = 1
-
-    for i in 1:gmm.k
-        num_samples = component_samples[i]
-        if num_samples > 0
-            dist = MvNormal(gmm.μₖ[i, :], gmm.Σₖ[i])
-            samples[start_idx:(start_idx + num_samples - 1), :] .= transpose(
-                rand(rng, dist, num_samples)
-            )
-            start_idx += num_samples
+    counts = rand(rng, Multinomial(n, gmm.πₖ), 1)
+    D = size(gmm.μₖ, 1)
+    samples = Matrix{Float64}(undef, D, n)
+    idx = 1
+    for k in 1:gmm.k
+        nk = counts[k]
+        if nk > 0
+            dist = MvNormal(gmm.μₖ[:, k], gmm.Σₖ[k])
+            samples[:, idx:idx + nk - 1] .= rand(rng, dist, nk)
+            idx += nk
         end
     end
     return samples
 end
 
-function Random.rand(gmm::GaussianMixtureModel, n::Int)
-    return rand(Random.default_rng(), gmm, n)
-end
[…]

[…]
