+++ docs/examples/GaussianMixtureModel.jl
@@ -68,8 +68,8 @@ p2
 
 # ## Visualize model contours over the data
 
-xs = range(minimum(X[1, :]) - 1, stop=maximum(X[1, :]) + 1, length=150)
-ys = range(minimum(X[2, :]) - 1, stop=maximum(X[2, :]) + 1, length=150)
+xs = collect(range(minimum(X[1, :]) - 1, stop=maximum(X[1, :]) + 1, length=150))
+ys = collect(range(minimum(X[2, :]) - 1, stop=maximum(X[2, :]) + 1, length=150))
 
 p3 = scatter(
     X[1, :], X[2, :];
