### 2025-06-14 8c94c66
Update GaussianMixtureModel.jl
