### 2025-06-14 33a8db0
Fix to GMM example
