+++ docs/src/HiddenMarkovModels.md
@@ -45,10 +45,6 @@ Where:
 - ``\mu_k`` is the mean of the emission distribution for state $k$
 - ``\Sigma_k`` is the covariance of the emission distribution for state $k$
 
-```@docs
-GaussianEmission
-```
-
 # What is a Generalized Linear Model - Hidden Markov Model
 
 A **Hidden Markov Model - Generalized Linear Model (GLM-HMM)** - also known as **Switching Regression Model** - is an extension to classic HMMs where the emission models are state-dependent GLMs that link an observed input to an observed output. This formulation allows each hidden state to define its own regression relationship between inputs and outputs, enabling the model to capture complex, state-dependent dynamics in the data. Currently, StateSpaceDynamics.jl support Gaussian, Bernoulli, Poisson, and Autoregressive GLMs as emission models.
@@ -96,24 +92,11 @@ Where:
 - ``\sigma(\cdot)`` is the logistic sigmoid function for binary outputs
 - ``\mu_k`` is the state-dependent bias
 
-```@docs
-GaussianRegressionEmission
-BernoulliRegressionEmission
-PoissonRegressionEmission
-AutoRegressionEmission
-```
-
 # Sampling from an HMM
-You can generate synthetic data from an HMM or GLM-HMM:
+You can generate synthetic data from an HMM:
 
 ```@docs
-function Random.rand(
-    rng::AbstractRNG,
-    model::HiddenMarkovModel,
-    X::Union{Matrix{<:Real}, Nothing}=nothing;
-    n::Int,
-    autoregressive::Bool=false,
-)
+Random.rand(rng::AbstractRNG,model::HiddenMarkovModel,X::Union{Matrix{<:Real}, Nothing}=nothing;n::Int,autoregressive::Bool=false,)
 ```
 
 # Learning in an HMM
