+++ docs/examples/SLDS.jl
@@ -1,0 +1,159 @@
+# ## Simulating and Fitting a Switching Linear Dynamical System
+
+# ## Load Packages
+using StateSpaceDynamics
+using LinearAlgebra
+using Random
+using Plots
+using LaTeXStrings
+using Statistics
+using StableRNGs
+
+rng = StableRNG(123);
+
+# ## Simulate Data from an SLDS model
+state_dim = 2
+obs_dim = 10
+K = 2 ## two states
+
+## Create the HMM parameters
+A_hmm = [0.92 0.08; 0.06 0.94]
+π₀ = [1.0, 0.0]
+
+## Create the state models
+A₁ = 0.95 * [cos(0.05) -sin(0.05); sin(0.05) cos(0.05)] ## slower oscillator
+A₂ = 0.95 * [cos(0.55) -sin(0.55); sin(0.55) cos(0.55)] ## faster oscillator
+
+Q₁ = [0.001 0.0; 0.0 0.001]
+Q₂ = [0.1 0.0; 0.0 0.1]
+
+## Assume same initial distribution for ease
+x0 = [0.0, 0.0]
+P0 = [0.1 0.0; 0.0 0.1]
+
+## create the observation models
+C₁ = randn(rng, obs_dim, state_dim)
+C₂ = randn(rng, obs_dim, state_dim)
+
+R = Matrix(0.1 * I(obs_dim)) ## Assume same noise covariance for both states
+
+## Put it all together for an SLDS model
+model = SwitchingLinearDynamicalSystem(
+    A_hmm,
+    [LinearDynamicalSystem(GaussianStateModel(A₁, Q₁, x0, P0), GaussianObservationModel(C₁, R), state_dim, obs_dim, fill(true, 6)), 
+     LinearDynamicalSystem(GaussianStateModel(A₂, Q₂, x0, P0), GaussianObservationModel(C₂, R), state_dim, obs_dim, fill(true, 6))],
+     π₀,
+    K)
+
+## Simulate data
+T = 1000
+x, y, z = rand(rng, model, 1000)
+
+# ## Plot the true dynamics
+p1 = plot(1:T, x[1, :], label="x₁",  linewidth=1.5)
+plot!(1:T, x[2, :], label="x₂", linewidth=1.5)
+
+## Create a background shading based on the state (z)
+## Find state transition points
+transition_points = [1; findall(diff(z) .!= 0) .+ 1; T + 1]
+
+for i in 1:(length(transition_points) - 1)
+    start_idx = transition_points[i]
+    end_idx = transition_points[i + 1] - 1
+    state_value = z[start_idx]
+    
+    ## Choose color based on state value
+    bg_color = state_value == 1 ? :lightblue : :lightyellow
+    
+    ## Add a background shading for this state region
+    vspan!([start_idx, end_idx], fillalpha=0.5, color=bg_color, 
+           label=(i == 1 ? "State $state_value" : ""))
+end
+
+## Adjust the plot appearance
+title!("Latent Dynamics with State")
+xlabel!("Time")
+ylabel!("State Value")
+ylims!(-3, 3)
+
+## Adjust the plot appearance
+title!("Latent Dynamics with State")
+xlabel!("Time")
+ylabel!("State Value")
+ylims!(-3, 3)
+
+p1
+
+# ## Create a new SLDS model with different parameters and fit to the data
+
+## Create a model to start with for EM, using reasonable guesses
+A = [0.9 0.1; 0.1 0.9]
+A ./= sum(A, dims=2) ## Normalize rows to sum to 1
+
+πₖ = rand(K)
+πₖ ./= sum(πₖ) ## Normalize to sum to 1
+
[…]
