### 2025-06-15 depasquale-lab/StateSpaceDynamics.jl#52 @rsenne
All comments we addressed we did on the dev documentation--if you want us to port over to stable let me know. I will say that the overhaul of the docstrings within the actual code was quite extensive. I think as you rightfully pointed out, we documented every single object and chose to display it, and now we only include functions that an actual user would interact with.

> * Please make sure that all examples included in the documentation are included as [doctests](https://documenter.juliadocs.org/stable/man/doctests/).

We have removed the extensive examples we had in the docstrings in favor of generating examples using Literate.jl. I couldn’t find examples of doctesting these examples, but I did poke around [your package](https://github.com/gdalle/HiddenMarkovModels.jl/blob/main/examples/autodiff.jl) and saw you used the regular testing interface? We could implement something like this.

> * It seems you document every single object and put it on the corresponding page. There needs to be a clearer separation between public and private API. By default, in Julia, what is documented is public unless stated otherwise. Starting with Julia 1.11, the `public` keyword enables a clearer distinction.

We have switched our approach to including only user-facing methods now.

> * The first example on the home page fails with SSD v0.1.1 and the latest Julia 1.11.

This was related to not using stable number generators it seems. I have now removed this example in favor of one that only shows the general creation of such models. There is now a full GaussianLDS example included in the tutorials that uses a StableRNG to make sure it works.

> * The link to the repo is outdated.

We have updated the link!

> * Contribution guidelines should be expanded a little.

We have expanded the contribution guidelines.

> * For citation, you could add a [Zenodo](https://zenodo.org/) DOI.

Great suggestion! We have added a Zenodo doi.

> * It might help to intersperse the docstrings with the text? You can force them to be collapsed by default to avoid taking too much space.

We have switched to this approach. Thanks for the suggestion!

> * The math doesn't display properly. For the list you might need the backtick syntax (instead of dollar signs) to avoid these alignment issues.

We have changed all instances of the dollar sign to use backticks.

> * Why does it make sense to construct the default GaussianStateModel with matrices of size 0?

Originally, our API assumed that a user would use `GaussianLDS` to create the model. Ergo, this logic was used to simply check that a user had not passed matrices and would create random starts for them. This logic has since been removed as I mention in #51. So, this is no longer relevant.
