### 2025-06-15 #8077 @rsenne
Hi @espottesmith. We have pretty much finished the first round of reviews. There is a few outstanding issues, that I mentioned in response to the comments that I left unfinished, but I feel like we are in a relatively good spot to move onto the next round of revision. I'm happy to let @gdalle to be the judge of that/if he wants further changes before this.
