+++ src/MixtureModels.jl
@@ -50,7 +50,7 @@ end
 Random.rand(gmm::GaussianMixtureModel, n::Int) = rand(Random.default_rng(), gmm, n)
 
 """
-    estep(gmm::GaussianMixtureModel, data::Matrix{<:Real})
+    estep(gmm::GaussianMixtureModel, data::AbstractMatrix{<:Real})
 
 Performs the E-step of the Expectation-Maximization Algorithm for a Gaussian Mixture Model. 
 """
@@ -73,11 +73,11 @@ end
 estep(gmm::GaussianMixtureModel, data::AbstractVector{<:Real}) = estep(gmm, reshape(data, :, 1))
 
 """
-    mstep!(gmm::GaussianMixtureModel, data::Matrix{T}, class_probabilities::Matrix{T})
+    mstep!(gmm::GaussianMixtureModel, data::AbstractMatrix{<:Real}, class_probabilities::AbstractMatrix{<:Real})
 
 Performs the M-step of the Expectation-Maximization Algorithm 
 """
-function mstep!(gmm::GaussianMixtureModel, data::Matrix{<:Real}, γ::Matrix{<:Real})
+function mstep!(gmm::GaussianMixtureModel, data::AbstractMatrix{<:Real}, γ::AbstractMatrix{<:Real})
     D, N = size(data)
     K = gmm.k
     N_k = sum(γ, dims=2)
@@ -93,22 +93,22 @@ function mstep!(gmm::GaussianMixtureModel, data::Matrix{<:Real}, γ::Matrix{<:Re
     gmm.πₖ .= vec(N_k) / N
 end
 
-mstep!(gmm::GaussianMixtureModel, data::AbstractVector{<:Real}, γ::Matrix{<:Real}) =
+mstep!(gmm::GaussianMixtureModel, data::AbstractVector{<:Real}, γ::AbstractMatrix{<:Real}) =
     mstep!(gmm, reshape(data, :, 1), γ)
 
 """
-    loglikelihood(gmm::GaussianMixtureModel, data::Matrix{<:Real})
+    loglikelihood(gmm::GaussianMixtureModel, data::AbstractMatrix{<:Real})
 
 Compute the log-likelihood of the data given the Gaussian Mixture Model (GMM). The data matrix should be of shape (# observations, # features).
 
 # Arguments 
 - `gmm::GaussianMixtureModel`: The Gaussian Mixture Model instance 
-- `data::Matrix{<:Real}`: data matrix to calculate the Log-Likelihood 
+- `data::AbstractMatrix{<:Real}`: data matrix to calculate the Log-Likelihood 
 
 # Returns
 - `Float64`: The log-likelihood of the data given the model.
 """
-function loglikelihood(gmm::GaussianMixtureModel, data::Matrix{<:Real})
+function loglikelihood(gmm::GaussianMixtureModel, data::AbstractMatrix{<:Real})
     D, N = size(data)
     K = gmm.k
     ll = 0.0
@@ -119,16 +119,16 @@ function loglikelihood(gmm::GaussianMixtureModel, data::Matrix{<:Real})
     return ll
 end
 
-loglikelihood(gmm::GaussianMixtureModel, data::Vector{<:Real}) = loglikelihood(gmm, reshape(data, :, 1))
+loglikelihood(gmm::GaussianMixtureModel, data::AbstractVector{<:Real}) = loglikelihood(gmm, reshape(data, :, 1))
 
 """
-    fit!(gmm::GaussianMixtureModel, data::Matrix{<:Real}; <keyword arguments>)
+    fit!(gmm::GaussianMixtureModel, data::AbstractMatrix{<:Real}; <keyword arguments>)
 
 Fits a Gaussian Mixture Model (GMM) to the given data using the Expectation-Maximization (EM) algorithm.
 
 # Arguments
[…]
