### 2025-06-16 #8077 @rsenne
Yes we can do that! I was originally planning on holding off on the benchmarking for now as we will likely make more code changes, but after thinking about it, its probably a good idea just to see how the changes affected performance. I will work on this today!
