### 2025-06-16 5a73af0
Initial benchmarking commit
