+++ benchmarking/Instances.jl
@@ -1,0 +1,10 @@
+export Instance
+
+# Define struct for storing glmhmm benchmarking params
+@kwdef struct Instance
+    num_states::Int
+    num_trials::Int
+    seq_length::Int
+    input_dim::Int
+    output_dim::Int
+end
