### 2025-06-16 0641a00
Built bencharmik structure
