### 2025-06-16 f01bf62
controleld Bernoulli HMM implementation in HMM.jl
