### 2025-06-16 ccfb1d3
HMM instances and params structs and data functions
