### 2025-06-16 d791b5f
Implemented HMM.jl benchmarking functions
