+++ benchmarking/SSD_Benchmark.jl
@@ -14,6 +14,7 @@ using PythonCall
 using StableRNGs
 using StatsAPI
 using Random
+using Base.Threads: @threads
 
 # Define type for organizing our implementations
 abstract type Implementation end
@@ -21,6 +22,7 @@ abstract type Implementation end
 # Include files
 include("Instances.jl")
 include("Params.jl")
+include("Controlled_BernoulliHMM.jl")
 include("glmhmm_benchmarking_tools.jl")
 
 end
