+++ benchmarking/Instances.jl
@@ -8,3 +8,10 @@ export HMMInstance
     input_dim::Int
     output_dim::Int
 end
+
+@kwdef struct LDSInstance
+    latent_dim::Int
+    obs_dim::Int
+    num_trials::Int
+    seq_length::Int
+end
