### 2025-06-17 70bf6c2
updated buil and benchmarks for hmm.jl
