### 2025-06-17 bb834e9
Expand LDS benchmark
