### 2025-06-17 1f3ca50
Add LDS benchmarking stuff
