+++ benchmarking/run_benchmark.jl
@@ -4,6 +4,7 @@ using BenchmarkTools
 using StableRNGs
 using Printf
 using CSV
+using DataFrames
 
 struct BenchConfig
     latent_dims::Vector{Int}
@@ -35,28 +36,31 @@ implementations = [
 # Results will be stored here
 all_results = []
 
-# Loop over configurations
+# Loop over configrations
 for latent_dim in config.latent_dims
     for obs_dim in config.obs_dims
         obs_dim < latent_dim && continue
         for seq_len in config.seq_lengths
             println("\n→ Benchmarking LDS with latent_dim=$latent_dim, obs_dim=$obs_dim, seq_len=$seq_len")
 
-            # Construct instance & params
             instance = LDSInstance(; latent_dim=latent_dim, obs_dim=obs_dim, num_trials=1, seq_length=seq_len)
             rng = StableRNG(1234)
             params = init_params(rng, instance)
 
-            results_row = Dict("config" => (latent_dim=latent_dim, obs_dim=obs_dim, seq_len=seq_len))
+            # Step 1: Simulate data using SSD only
+            ssd_model = build_model(SSD_LDSImplem(), instance, params)
+            _, y = build_data(rng, ssd_model, instance)
 
+            # Step 2: Initialize result row
+            results_row = Dict{String, Any}()
+            results_row["config"] = (latent_dim=latent_dim, obs_dim=obs_dim, seq_len=seq_len)
+
+            # Step 3: Benchmark all implementations on same data
             for impl in implementations
                 print("  Running $(string(impl))... ")
-
                 try
                     model = build_model(impl, instance, params)
-                    _, y = build_data(rng, model, instance)
-
-                    result = run_benchmark(impl, model, nothing, y)
+                    result = run_benchmark(impl, model, y)
                     results_row[string(impl)] = result
 
                     if result.success
@@ -70,6 +74,8 @@ for latent_dim in config.latent_dims
                 end
             end
 
+push!(all_results, results_row)
+
             push!(all_results, results_row)
             println("-"^50)
         end
@@ -77,21 +83,23 @@ for latent_dim in config.latent_dims
 end
 
 # write to CSV or show sumary
-using DataFrames
+
 df = DataFrame()
 for row in all_results
     config = row["config"]
     for (name, result) in row
-        name == "config" && continue
+        if name == "config"
+            continue
+        end
         push!(df, (
             implementation = name,
             latent_dim = config.latent_dim,
             obs_dim = config.obs_dim,
             seq_len = config.seq_len,
-            time_sec = result.time / 1e9,
-            memory = result.memory,
-            allocs = result.allocs,
-            success = result.success
+            time_sec = result[:time] / 1e9,
+            memory = result[:memory],
+            allocs = result[:allocs],
+            success = result[:success]
         ))
     end
 end
