### 2025-06-17 f44b218
dynamax implem cleaned. pushing to move to desktop
