### 2025-06-18 1409aac
Fixes LDS speed issue. Re-introduces error for F32 and BigFloat
