### 2025-06-18 995c82f
bernoulli GLMHMM benchmarking results
