### 2025-06-18 3c836f0
Removes BigFloat as part of checks
