### 2025-06-18 117a9bc
Start plotting
