+++ benchmarking/Project.toml
@@ -1,6 +1,7 @@
 [deps]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
 CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
@@ -12,3 +13,4 @@ PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 StateSpaceDynamics = "19cf8126-1d00-4ad8-bf17-4e76d7bc0f6d"
 StatsAPI = "82ae8749-77ed-4fe6-ae5f-f523153014b0"
+StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
