### 2025-06-18 ab42ac5
Update paper stuff
