### 2025-06-18 27f8d78
glmhmm benchmarking plot
