### 2025-06-18 abfb3a3
fixed shapes and organized legend
