+++ paper/paper.md
@@ -64,7 +64,7 @@ To thoroughly assess performance across different scales, we tested three sequen
 For the subsequent benchmarking experiments, we compared `StateSpaceDynamics.jl`, `HiddenMarkovModels.jl`, and `Dynamax` in their abilities to fit a Bernoulli GLM-HMM. We completed the benchmarking on the same combinations of sequence lengths, state dimensions, and observation dimensions as in the Gaussian LDS benchmarking. Synthetic datasets were generated for each of these combinations by sampling from a randomly generated Bernoulli GLM-HMM using `StateSpaceDynamics.jl`. All packages were initialized to the same random parameters and the EM algorithm was run for 200 iterations. We disabled the convergence criteria to ensure each package completed all 200 iterations of EM. These benchmarks were conducted using `PythonCall.jl` and `BenchmarkTools.jl` to ensure timing accuracy.
 
 \begin{figure}
-\includegraphics[width=\textwidth]{SwitchingBernoulli_Benchmark_Plot.pdf}
+\includegraphics[width=\textwidth]{glmhmm_bench.png}
 \end{figure}
 
 In our benchmarking, we find that for the LDS, both variants of `StateSpaceDynamics.jl` are faster than `pykalman` across all sequence lengths and dimension configurations. More generally, `StateSpaceDynamics.jl` outperforms `Dynamax` at shorter sequence lengths across various dimensional settings. However, `Dynamax` exhibits superior scaling. In our current implementation, the Hessian is represented as a sparse matrix with block-tridiagonal structure, resulting in O($Tn^2$) memory scaling — which is optimal. However, we do not yet exploit this structure fully during inference. In particular, our solver does not leverage specialized routines for block-banded systems, which can result in unnecessary fill-in and degraded performance at large T. Future versions will use banded or block-tridiagonal solvers to achieve truly linear-time inference.
