### 2025-06-19 #8077 @gdalle
Thanks! This seems like a good time for the other reviewers to start digging in
