### 2025-06-19 ee82652
Starts to implement the use of FilterSmooth into the LDS.
