+++ src/LinearDynamicalSystems.jl
@@ -266,28 +266,50 @@ function Gradient(
 
     grad = zeros(T, latent_dim, tsteps)
 
+    # Pre-allocate dxt, dxt_next, dyt, and two temps for efficiency
+    dxt       = zeros(T, latent_dim)
+    dxt_next  = zeros(T, latent_dim)
+    dyt       = zeros(T, obs_dim)
+    tmp1      = zeros(T, latent_dim)  # for w[t] * C_inv_R * dyt
+    tmp2      = zeros(T, latent_dim)  # for A_inv_Q * dxt_next
+
+
     # First time step
-    dx1 = x[:, 1] - x0
-    dx2 = x[:, 2] - A * x[:, 1]
-    dy1 = y[:, 1] - C * x[:, 1]
+    dxt .= x[:, 1] - x0
+    dxt_next .= x[:, 2] - A * x[:, 1]
+    dyt .= y[:, 1] - C * x[:, 1]
 
-    grad[:, 1] .= A_inv_Q * dx2 + w[1] * C_inv_R * dy1 - (P0_chol \ dx1)
+    grad[:, 1] .= A_inv_Q * dxt_next + w[1] * C_inv_R * dyt - (P0_chol \ dxt)
 
-    # Pre-allocate dxt, dxt_next, and dyt for efficiency
-    dxt = zeros(T, latent_dim)
-    dxt_next = zeros(T, latent_dim)
-    dyt = zeros(T, obs_dim)
-
-    # Middle time steps
     @views for t in 2:(tsteps - 1)
-        grad[:, t] .= w[t] * C_inv_R * (y[:, t] .- C * x[:, t]) - (Q_chol \ (x[:, t] .- A * x[:, t - 1])) + (A_inv_Q * (x[:, t + 1] .- A * x[:, t]))
+        # dxt = x[:, t] - A * x[:, t-1]
+        mul!(dxt, A, x[:, t - 1])
+        dxt .= x[:, t] .- dxt
+
+        # dxt_next = x[:, t+1] - A * x[:, t]
+        mul!(dxt_next, A, x[:, t])
+        dxt_next .= x[:, t + 1] .- dxt_next
+
+        # dyt = y[:, t] - C * x[:, t]
+        mul!(dyt, C, x[:, t])
+        dyt .= y[:, t] .- dyt
+
+        # tmp1 = w[t] * C_inv_R * dyt
+        mul!(tmp1, C_inv_R, dyt)
+        tmp1 .*= w[t]
+
+        # tmp2 = A_inv_Q * dxt_next
+        mul!(tmp2, A_inv_Q, dxt_next)
+
+        # grad[:, t] = tmp1 - (Q_chol \ dxt) + tmp2
+        grad[:, t] .= tmp1 .- (Q_chol \ dxt) .+ tmp2
     end
 
     # Last time step
-    dxT = x[:, tsteps] - A * x[:, tsteps - 1]
-    dyT = y[:, tsteps] - C * x[:, tsteps]
+    dxt .= x[:, tsteps] - A * x[:, tsteps - 1]
+    dyt .= y[:, tsteps] - C * x[:, tsteps]
 
-    grad[:, tsteps] .= w[tsteps] * (C_inv_R * dyT) - (Q_chol \ dxT)
+    grad[:, tsteps] .= w[tsteps] * (C_inv_R * dyt) - (Q_chol \ dxt)
 
     return grad
 end
@@ -374,8 +396,14 @@ This function performs direct smoothing for a linear dynamical system (LDS) give
 - `inverse_offdiag::Array{T, 3}`: The inverse off-diagonal matrix.
 - `Q_val::T`: The Q-function value.
 """
-function smooth(
[…]
