### 2025-06-20 f53531c
Speedup to PoissonLDS
