+++ test/LinearDynamicalSystems/PoissonLDS.jl
@@ -377,6 +377,7 @@ function test_parameter_gradient()
 
     grad = ForwardDiff.gradient(f, params)
 
+
     @test isapprox(grad, grad_analytical, rtol=1e-5, atol=1e-5)
 end
 
