### 2025-06-20 #8077 @espottesmith
It looks like you haven't defined _joss as your paper branch, which is probably why the editorialbot is pulling from main (the default branch). So, yes, if you want the new paper to appear here, you'll need to push to main.
