### 2025-06-23 41b9d70
Reduce allocations in PLDS gradient etc.
