+++ src/LinearDynamicalSystems.jl
@@ -101,18 +101,34 @@ end
    
 Initialize a `FilterSmooth` object for a given linear dynamical system model and number of observations.
 """
+"""
+    initialize_FilterSmooth(model, num_obs) 
+   
+Initialize a `FilterSmooth` object for a given linear dynamical system model and number of observations.
+"""
 function initialize_FilterSmooth(
     model::LinearDynamicalSystem{T,S,O}, 
     num_obs::Int
 ) where {T<:Real, S<:GaussianStateModel{T}, O<:GaussianObservationModel{T}}
     num_states = model.latent_dim
-    FilterSmooth(
-        zeros(T, num_states, num_obs),
-        zeros(T, num_states, num_states, num_obs),
-        zeros(T, num_states, num_obs, 1),
-        zeros(T, num_states, num_states, num_obs, 1),
-    zeros(T, num_states, num_states, num_obs, 1)
+    FilterSmooth{T}(
+        zeros(T, num_states, num_obs),                    # x_smooth
+        zeros(T, num_states, num_states, num_obs),        # p_smooth  
+        zeros(T, num_states, num_states, num_obs),        # p_smooth_tt1
+        zeros(T, num_states, num_obs),                    # E_z
+        zeros(T, num_states, num_states, num_obs),        # E_zz
+        zeros(T, num_states, num_states, num_obs),        # E_zz_prev
+        zero(T)                                           # entropy
     )
+end
+
+function initialize_FilterSmooth(
+    model::LinearDynamicalSystem{T,S,O},
+    tsteps::Int,
+    ntrials::Int
+) where {T<:Real, S<:GaussianStateModel{T}, O<:GaussianObservationModel{T}}
+    filter_smooths = [initialize_FilterSmooth(model, tsteps) for _ in 1:ntrials]
+    return TrialFilterSmooth(filter_smooths)
 end
 
 function Random.rand(rng::AbstractRNG, lds::LinearDynamicalSystem{T,S,O}; tsteps::Int, ntrials::Int) where {T<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
@@ -416,8 +432,8 @@ function smooth!(
 
     tsteps, D = size(y, 2), lds.latent_dim
 
-    # set initial "solution" and preallocate x_reshape
-    X₀ = zeros(T, D * tsteps)
+    # use old fs if it exists, by default is zeros if no iteration of EM has occurred
+    X₀ = fs.E_z 
 
     function nll(vec_x::Vector{T})
         x = reshape(vec_x, D, tsteps)
@@ -456,33 +472,34 @@ function smooth!(
     res = optimize(td, X₀, Newton(;linesearch=LineSearches.BackTracking()), opts)
 
     # Profit
-    x = reshape(res.minimizer, D, tsteps)
+    fs.x_smooth .= reshape(res.minimizer, D, tsteps)
 
-    H, main, super, sub = Hessian(lds, y, x, w)
+    H, main, super, sub = Hessian(lds, y, fs.x_smooth, w)
 
     # Get the second moments of the latent state path, use static matrices if the latent dimension is small
     if lds.latent_dim > 10
-        p_smooth, inverse_offdiag = block_tridiagonal_inverse(-sub, -main, -super)
[…]
