### 2025-06-25 c97c95f
Change API of Analytical Updates in M-Step
