+++ src/LinearDynamicalSystems.jl
@@ -484,7 +484,7 @@ function smooth!(
     end
 
     # Calculate the entropy, see Utilities.jl for the function
-    fs.gauss_entropy .= gaussian_entropy(Symmetric(H))
+    fs.entropy .= gaussian_entropy(Symmetric(H))
 
     # Symmetrize the covariance matrices
     @views for i in 1:tsteps
@@ -523,35 +523,13 @@ function smooth!(
     tfs::TrialFilterSmooth{T},
     y::AbstractArray{T,3}
 ) where {T<:Real,S<:GaussianStateModel{T},O<:AbstractObservationModel{T}}
-    obs_dim, tsteps, ntrials = size(y)
-    latent_dim = lds.latent_dim
-
-    # Fast path for single trial case
-    if ntrials == 1
-        x_sm, p_sm, p_prev, ent = smooth!(lds, tfs[1], y[:, :, 1])
-        # Return directly in the required shape without additional copying
-        return reshape(x_sm, latent_dim, tsteps, 1),
-               reshape(p_sm, latent_dim, latent_dim, tsteps, 1),
-               reshape(p_prev, latent_dim, latent_dim, tsteps, 1),
-               ent
-    end
-
-    # Pre-allocate output arrays
-    x_smooth = Array{T,3}(undef, latent_dim, tsteps, ntrials)
-    p_smooth = Array{T,4}(undef, latent_dim, latent_dim, tsteps, ntrials)
-    inverse_offdiag = Array{T,4}(undef, latent_dim, latent_dim, tsteps, ntrials)
-    total_entropy = 0.0
-
+    ntrials = size(y, 3)
+    
     @views @threads for trial in 1:ntrials
-        x_sm, p_sm, p_prev, ent = smooth!(lds, tfs[i], y[:, :, trial])
-        total_entropy += ent
-        x_smooth[:, :, trial] .= x_sm
-        p_smooth[:, :, :, trial] .= p_sm
-        inverse_offdiag[:, :, :, trial] .= p_prev
+        smooth!(lds, tfs[trial], y[:, :, trial])
     end
     
-
-    return x_smooth, p_smooth, inverse_offdiag, total_entropy
+    return tfs
 end
 
 """
@@ -745,6 +723,12 @@ function sufficient_statistics!(fs::FilterSmooth{T}) where {T<:Real}
     end
 end
 
+function sufficient_statistics!(tfs::TrialFilterSmooth{T}) where {T<:Real}
+    @threads for i in 1:length(tfs.FilterSmooths)
+        sufficient_statistics!(tfs[i])
+    end
+end
+
 
 """
     estep(lds::LinearDynamicalSystem{T,S,O}, y::AbstractArray{T,3}) 
@@ -755,21 +739,11 @@ Perform the E-step of the EM algorithm for a Linear Dynamical System, treating a
 - This function first smooths the data using the `smooth` function, then computes sufficient statistics.
 - It treats all input as multi-trial, with single-trial being a special case where ntrials = 1.
 """
-function estep!(
-    lds::LinearDynamicalSystem{T,S,O},
-    tfs::TrialFilterSmooth,
-    y::AbstractArray{T,3}
-) where {T<:Real,S<:GaussianStateModel{T},O<:AbstractObservationModel{T}}
-    # smooth
-    smooth!(lds, y, tfs)
-
-    # calculate sufficient statistics
+function estep!(lds, tfs, y)
+    smooth!(lds, tfs, y)  
     sufficient_statistics!(tfs)
-
-    # calculate elbo
-    ml_total = calculate_elbo(lds, tfs.E_z, tfs.E_zz, tfs.E_zz_prev, tfs.p_smooth, y, tfs.total_entropy)
-
[…]
