### 2025-06-25 d0dbb01
Fix variable assignments and add single trial threading bypass
