### 2025-06-25 e9176ff
Massively reduces allocations.
Switches to a temp buffer strategy for update_R/C/Q functions.

Makes inverssion algorithm not preallocate a bunch of static matrices.

Reduces allocations in loglikelihood and hessian.
