+++ src/LinearDynamicalSystems.jl
@@ -226,16 +226,27 @@ function loglikelihood(
     # Allocate temporary arrays with correct type
     temp_dx = zeros(eltype(x), size(x, 1))
     temp_dy = zeros(eltype(x), size(y, 1))
+    
+    # Pre-allocate solve result vectors (reuse these!)
+    temp_solve_Q = zeros(eltype(x), size(x, 1))
+    temp_solve_R = zeros(eltype(x), size(y, 1))
 
     for t in 1:tsteps
         if t > 1
             mul!(temp_dx, A, view(x, :, t-1), -one(eltype(x)), false)
             temp_dx .+= view(x, :, t)
-            ll += sum(abs2, Q_chol \ temp_dx)
+            
+            # In-place solve: temp_solve_Q = Q_chol \ temp_dx
+            ldiv!(temp_solve_Q, Q_chol, temp_dx)
+            ll += sum(abs2, temp_solve_Q)
         end
+        
         mul!(temp_dy, C, view(x, :, t), -one(eltype(x)), false)
         temp_dy .+= view(y, :, t)
-        ll += w[t] * sum(abs2, R_chol \ temp_dy)
+        
+        # In-place solve: temp_solve_R = R_chol \ temp_dy  
+        ldiv!(temp_solve_R, R_chol, temp_dy)
+        ll += w[t] * sum(abs2, temp_solve_R)
     end
 
     return -eltype(x)(0.5) * ll
@@ -378,14 +389,21 @@ function Hessian(
     xt1_given_xt = -A' * inv_Q * A
     x_t = -inv_P0
 
+    # Pre-compute constant terms to avoid repeated allocation
+    middle_base = xt_given_xt_1 + xt1_given_xt  # This is constant for middle entries
+    last_base = xt_given_xt_1                   # This is the term for the last entry
+
     H_diag = [Matrix{T}(undef, size(A, 1), size(A, 1)) for _ in 1:tsteps]
 
-    # Build main diagonal
-    H_diag[1] .= w[1] * yt_given_xt + xt1_given_xt + x_t
+    # Build main diagonal efficiently
+    @. H_diag[1] = w[1] * yt_given_xt + xt1_given_xt + x_t
+
+    # avoid temporaries in the loop
     for i in 2:(tsteps - 1)
-        H_diag[i] .= w[i] * yt_given_xt + xt_given_xt_1 + xt1_given_xt
+        @. H_diag[i] = w[i] * yt_given_xt + middle_base
     end
-    H_diag[tsteps] .= w[tsteps] * (yt_given_xt) + xt_given_xt_1
+
+    @. H_diag[tsteps] = w[tsteps] * yt_given_xt + last_base
 
     H = StateSpaceDynamics.block_tridgm(H_diag, H_super, H_sub)
 
@@ -601,31 +619,34 @@ end
 
 Calculate the a single time step observation component of the Q-function for the EM algorithm in a Linear Dynamical System before the R^-1 is accounted for.
 """
-function Q_obs(
+function Q_obs!(
+    result::AbstractMatrix{T},
     H::AbstractMatrix{T},
     E_z::AbstractVector{T},
     E_zz::AbstractMatrix{T},
     y::AbstractVector{T},
+    buffers
 ) where {T<:Real}
-
-    obs_dim = size(H, 1)
-
-    # Pre-allocate statistics
-    sum_yy = zeros(T, obs_dim, obs_dim)
-    sum_yz = zeros(T, obs_dim, size(E_z, 1))
     
-    mul!(sum_yy, y, y', 1.0, 1.0)
-    mul!(sum_yz, y, E_z', 1.0, 1.0)
-
[…]
