### 2025-06-27 #8077 @espottesmith
@slinderman, @THargreaves have y'all had a chance to dig into this submission?
