### 2025-06-28 cc35754
Huge speedups to PoissonLDS models
