+++ src/LinearDynamicalSystems.jl
@@ -105,7 +105,7 @@ Initialize a `FilterSmooth` object for a given linear dynamical system model and
 function initialize_FilterSmooth(
     model::LinearDynamicalSystem{T,S,O}, 
     num_obs::Int
-) where {T<:Real, S<:GaussianStateModel{T}, O<:GaussianObservationModel{T}}
+) where {T<:Real, S<:GaussianStateModel{T}, O<:AbstractObservationModel{T}}
     num_states = model.latent_dim
     FilterSmooth{T}(
         zeros(T, num_states, num_obs),                    # x_smooth
@@ -122,7 +122,7 @@ function initialize_FilterSmooth(
     model::LinearDynamicalSystem{T,S,O},
     tsteps::Int,
     ntrials::Int
-) where {T<:Real, S<:GaussianStateModel{T}, O<:GaussianObservationModel{T}}
+) where {T<:Real, S<:GaussianStateModel{T}, O<:AbstractObservationModel{T}}
     filter_smooths = [initialize_FilterSmooth(model, tsteps) for _ in 1:ntrials]
     return TrialFilterSmooth(filter_smooths)
 end
@@ -762,7 +762,7 @@ function sufficient_statistics!(tfs::TrialFilterSmooth{T}) where {T<:Real}
     if ntrials == 1
         sufficient_statistics!(tfs[1])
     else 
-        @threads for i in ntrials
+        @threads for i in 1:ntrials
             sufficient_statistics!(tfs[i])
         end
     end
@@ -1221,28 +1221,51 @@ function loglikelihood(
     inv_p0 = inv(plds.state_model.P0)
     inv_Q = inv(plds.state_model.Q)
 
-    # Calculate p(yₜ|xₜ)
+    # Get dimensions
     C = plds.obs_model.C
+    A = plds.state_model.A
+    x0 = plds.state_model.x0
     obs_dim, latent_dim = size(C)
-    temp  = Vector{eltype(x)}(undef, obs_dim) # Temporary vector for calculations
 
+    # Pre-allocate ALL temporary vectors
+    Cx_temp = Vector{eltype(x)}(undef, obs_dim)      # For C * x[:, t]
+    obs_temp = Vector{eltype(x)}(undef, obs_dim)     # For C * x[:, t] + d
+    Ax_temp = Vector{eltype(x)}(undef, latent_dim)   # For A * x[:, t-1]
+    state_diff = Vector{eltype(x)}(undef, latent_dim) # For x[:, t] - A * x[:, t-1]
+    quad_temp = Vector{eltype(x)}(undef, latent_dim)  # For inv_Q * state_diff
+
+    # Calculate p(yₜ|xₜ)
     pygivenx_sum = zero(T)  
     @views for t in 1:tsteps
-        temp .= C * x[:, t] .+ d
-        pygivenx_sum += dot(y[:, t], temp) - sum(exp, temp)
+        # Replace: temp .= C * x[:, t] .+ d
+        mul!(Cx_temp, C, x[:, t])           # Cx_temp = C * x[:, t]
+        obs_temp .= Cx_temp .+ d            # obs_temp = C * x[:, t] + d
+        
+        # Compute dot(y[:, t], obs_temp) - sum(exp, obs_temp)
+        dot_product = dot(y[:, t], obs_temp)
+        exp_sum = zero(eltype(x))
+        for i in 1:obs_dim
+            exp_sum += exp(obs_temp[i])
+        end
+        pygivenx_sum += dot_product - exp_sum
     end
 
-    # Calculate p(x₁)
-    dx1 = @view(x[:, 1]) .- plds.state_model.x0
-    px1 = -U(0.5) * dot(dx1, inv_p0 * dx1)
+    # Calculate p(x₁) - avoid creating dx1 temporary
[…]
