### 2025-06-29 depasquale-lab/StateSpaceDynamics.jl#50 @gdalle
Thanks for addressing these remarks. Here are my remaining concerns:

**Benchmarks**

- L86: Unlike Julia, JAX does not automatically get JIT-compiled with a warm-up iteration. Instead, you have to explicitly call `jax.jit` or annotate your function with the macro `@jit` for it to work (see <https://docs.jax.dev/en/latest/jit-compilation.html>).
- L99 and L108: It doesn't make a lot of sense to represent benchmarks of various sizes on the same scale. What does the bar height mean in this case? The average is not a meaningful quantity. In addition, the profusion of symbols makes the plots hard to read. I think the most principled visualization might be a table of 4 $\times$ 4 plots, each comparing the packages on instances of a given state dimension $n$ and observation dimension $m$, while varying $T$ from 100 to 1000
- L121: In your updated benchmarks, HiddenMarkovModels.jl consistently performs _worst_ and not _best_ (although I suspect that might be a benchmarking artifact, will look into it).

**Typos**

- L33: "neural activity followS"
- L56: "are block-tridiagonal"
- L99 and L108: I assume the plot titles should be $T = 100$ and not $n = 100$?

**Reproducibility**

- Code in the `benchmarking` folder should come with a `README.md`. Ideally, the experiments should be available in runnable files instead of notebooks.
