### 2025-06-29 depasquale-lab/StateSpaceDynamics.jl#51 @gdalle
Thanks for these changes. My main remaining gripe is that I still get the SingularException on my computer when running the example in the README. I think it would be best to provide reproducible RNGs with StableRNGs.jl for all of your documentation examples (and run the README example as part of your test suite somewhere).
