### 2025-06-29 depasquale-lab/StateSpaceDynamics.jl#50 @rsenne
> Reproducibility

Similar to my comment in #51, the code you used was not up to date. This was our old benchmarking suite. I think you may not have seen my updated comment:

> Our benchmarks are still inside the benchmarking folder, but the overall structure has changed. You'll need to dev the SSD package into the benchmarking project and install the dependencies. From here you should be able to run run_benchmark.jl and run_benchmark_glmhmm.jl

This however is my fault for not having the proper branch set in the JOSS review issue. So my apologies for this! I will address the other comments and also include a properly formatted README.md for reproducibility. Thank you for all of the suggestions!
