### 2025-06-30 depasquale-lab/StateSpaceDynamics.jl#51 @gdalle
And the code finally runs!
