### 2025-06-30 depasquale-lab/StateSpaceDynamics.jl#50 @gdalle
Okay I profiled your use of HiddenMarkovModels.jl with `@profview` and most of the time is spent in external functions that are written suboptimally. Here are some tips to speed things up (which are also rather general):

https://github.com/depasquale-lab/StateSpaceDynamics.jl/blob/ccb0cf193875078a34c7931a7619d51dad1ca352/benchmarking/Controlled_BernoulliHMM.jl#L10

You often compute `p = logistic(z)` only to use `log(p)` afterwards. There is a faster implementation for that:

```julia
using LogExpFunctions
log_logistic(z) = -log1pexp(-z)
```

A similar shortcut exists for computing `log(1 - p)` based directly on `z`.

https://github.com/depasquale-lab/StateSpaceDynamics.jl/blob/ccb0cf193875078a34c7931a7619d51dad1ca352/benchmarking/Controlled_BernoulliHMM.jl#L39

This is much slower due to splatting a large number of arguments (with `control...`) than either of the following equivalents:

```julia
X = reduce(hcat, control)'
X = stack(control)'
```

https://github.com/depasquale-lab/StateSpaceDynamics.jl/blob/ccb0cf193875078a34c7931a7619d51dad1ca352/benchmarking/Controlled_BernoulliHMM.jl#L42

This needlessly materializes a large vector before summing it: an explicit for loop would be faster:

```julia
val = zero(Base.promote_eltype(weights, obs_seq, logits))
for i in eachindex(weights, obs_seq, logits)
    p_i = logistic(logits[i])  # to optimize, see above
    o_i = obs_seq[i]
    val -= weights[i] * (o_i * log(p_i) + (1 - o_i) * log(1 - p_i))
end
```

https://github.com/depasquale-lab/StateSpaceDynamics.jl/blob/ccb0cf193875078a34c7931a7619d51dad1ca352/benchmarking/Controlled_BernoulliHMM.jl#L58

Since the goal is to perform computations in-place, the other out-of-place computations could be rerouted to buffers instead.
For instance, you could pre-allocate `logits` before the optimization hot loop, and then in each iteration do
```julia
mul!(logits_preallocated, X, dists_coeffs)
```

https://github.com/depasquale-lab/StateSpaceDynamics.jl/blob/ccb0cf193875078a34c7931a7619d51dad1ca352/benchmarking/Controlled_BernoulliHMM.jl#L93-L94

There are a lot of common intermediate quantities between the objective and its gradients. To leverage that property, it would make sense to use the single-function interface of Optim.jl, which is [documented in NLSolversBase.jl](https://github.com/JuliaNLSolvers/NLSolversBase.jl?tab=readme-ov-file#special-single-function-interface).
