### 2025-06-30 depasquale-lab/StateSpaceDynamics.jl#50 @ZachLoschin
Thank you for the useful performance tips @gdalle! I will look into this, but I am leaning towards option 2. To @gdalle's point, spending lots of time in emission model optimization introduces lots of places for discrepancies and unfair comparisons. Benchmarking a standard HMM with Gaussian emissions limits this and shifts focus back onto the fundamental HMM algorithms instead.
