+++ src/SLDS.jl
@@ -1,4 +1,5 @@
 export SwitchingLinearDynamicalSystem, fit!, initialize_slds, variational_expectation!
+
 """
 Switching Linear Dynamical System
 
@@ -16,7 +17,6 @@ mutable struct SwitchingLinearDynamicalSystem{T<:Real,M<:AbstractMatrix{T}, V<:A
     πₖ::V  
     K::Int 
 end
-
 
 """
     Random.rand(rng, slds, T)
@@ -133,6 +133,68 @@ function initialize_slds(;K::Int=2, d::Int=2, p::Int=10, self_bias::Float64=5.0,
 end
 
 """
+    initialize_FilterSmooth_slds(slds::SwitchingLinearDynamicalSystem{T}, T_step::Int) where {T<:Real}
+
+Initialize FilterSmooth objects for each regime in the SLDS.
+"""
+function initialize_FilterSmooth_slds(slds::SwitchingLinearDynamicalSystem{T}, T_step::Int) where {T<:Real}
+    K = slds.K
+    FS = Vector{FilterSmooth{T}}(undef, K)
+    
+    for k in 1:K
+        num_states = slds.B[k].latent_dim
+        FS[k] = FilterSmooth(
+            zeros(T, num_states, T_step),           # x_smooth
+            zeros(T, num_states, num_states, T_step), # p_smooth
+            zeros(T, num_states, T_step, 1),         # E_z
+            zeros(T, num_states, num_states, T_step, 1), # E_zz
+            zeros(T, num_states, num_states, T_step, 1)  # E_zz_prev
+        )
+    end
+    
+    return FS
+end
+
+"""
+    smooth_with_weights!(fs::FilterSmooth{T}, lds::LinearDynamicalSystem{T,S,O}, 
+                        y::AbstractMatrix{T}, weights::AbstractVector{T}) where {T<:Real,S,O}
+
+Perform smoothing with weights and update the FilterSmooth object in-place.
+"""
+function smooth_with_weights!(
+    fs::FilterSmooth{T}, 
+    lds::LinearDynamicalSystem{T,S,O}, 
+    y::AbstractMatrix{T}, 
+    weights::AbstractVector{T}
+) where {T<:Real,S,O}
+    # Perform smoothing with weights
+    x_smooth, p_smooth, inverse_offdiag, total_entropy = smooth(lds, y, weights)
+    
+    # Update the FilterSmooth object
+    fs.x_smooth .= x_smooth
+    fs.p_smooth .= p_smooth
+    
+    # Compute sufficient statistics and store them in FilterSmooth
+    fs.E_z[:, :, 1] .= x_smooth
+    
+    # Compute E_zz and E_zz_prev
+    tsteps = size(x_smooth, 2)
+    @views for t in 1:tsteps
+        xt = x_smooth[:, t]
+        pt = p_smooth[:, :, t]
+        fs.E_zz[:, :, t, 1] .= pt .+ xt * xt'
+        if t > 1
+            xtm1 = x_smooth[:, t-1]
+            pt1 = inverse_offdiag[:, :, t]
+            fs.E_zz_prev[:, :, t, 1] .= pt1 .+ xt * xtm1'
+        end
+    end
+    fs.E_zz_prev[:, :, 1, 1] .= 0
+    
+    return total_entropy
+end
+
+"""
     fit!(slds::SwitchingLinearDynamicalSystem, y::Matrix{T}; 
          max_iter::Int=1000, 
          tol::Real=1e-12, 
@@ -152,11 +214,11 @@ Fit a Switching Linear Dynamical System using the variational Expectation-Maximi
 - `mls::Vector{T}`: Vector of log-likelihood values for each iteration.
[…]
