+++ src/SLDS.jl
@@ -133,29 +133,6 @@ function initialize_slds(;K::Int=2, d::Int=2, p::Int=10, self_bias::Float64=5.0,
 end
 
 """
-    initialize_FilterSmooth_slds(slds::SwitchingLinearDynamicalSystem{T}, T_step::Int) where {T<:Real}
-
-Initialize FilterSmooth objects for each regime in the SLDS.
-"""
-function initialize_FilterSmooth_slds(slds::SwitchingLinearDynamicalSystem{T}, T_step::Int) where {T<:Real}
-    K = slds.K
-    FS = Vector{FilterSmooth{T}}(undef, K)
-    
-    for k in 1:K
-        num_states = slds.B[k].latent_dim
-        FS[k] = FilterSmooth(
-            zeros(T, num_states, T_step),           # x_smooth
-            zeros(T, num_states, num_states, T_step), # p_smooth
-            zeros(T, num_states, T_step, 1),         # E_z
-            zeros(T, num_states, num_states, T_step, 1), # E_zz
-            zeros(T, num_states, num_states, T_step, 1)  # E_zz_prev
-        )
-    end
-    
-    return FS
-end
-
-"""
     smooth_with_weights!(fs::FilterSmooth{T}, lds::LinearDynamicalSystem{T,S,O}, 
                         y::AbstractMatrix{T}, weights::AbstractVector{T}) where {T<:Real,S,O}
 
@@ -167,31 +144,13 @@ function smooth_with_weights!(
     y::AbstractMatrix{T}, 
     weights::AbstractVector{T}
 ) where {T<:Real,S,O}
-    # Perform smoothing with weights
-    x_smooth, p_smooth, inverse_offdiag, total_entropy = smooth(lds, y, weights)
+    # Use the smooth! function from LinearDynamicalSystems.jl
+    smooth!(lds, fs, y, weights)
     
-    # Update the FilterSmooth object
-    fs.x_smooth .= x_smooth
-    fs.p_smooth .= p_smooth
+    # Compute sufficient statistics using the new interface
+    sufficient_statistics!(fs)
     
-    # Compute sufficient statistics and store them in FilterSmooth
-    fs.E_z[:, :, 1] .= x_smooth
-    
-    # Compute E_zz and E_zz_prev
-    tsteps = size(x_smooth, 2)
-    @views for t in 1:tsteps
-        xt = x_smooth[:, t]
-        pt = p_smooth[:, :, t]
-        fs.E_zz[:, :, t, 1] .= pt .+ xt * xt'
-        if t > 1
-            xtm1 = x_smooth[:, t-1]
-            pt1 = inverse_offdiag[:, :, t]
-            fs.E_zz_prev[:, :, t, 1] .= pt1 .+ xt * xtm1'
-        end
-    end
-    fs.E_zz_prev[:, :, 1, 1] .= 0
-    
-    return total_entropy
+    return fs.entropy
 end
 
 """
@@ -238,7 +197,7 @@ function fit!(
     
     # Initialize the FilterSmooth and ForwardBackward structs
     FB = initialize_forward_backward(slds, T_step, T)
-    FS = initialize_FilterSmooth_slds(slds, T_step)
+    FS = [initialize_FilterSmooth(slds.B[k], T_step) for k in 1:K]
 
     # Initialize the parameters by running the Kalman Smoother for each model
     FB.γ = log.(ones(size(y)) * 0.5)
@@ -323,16 +282,11 @@ function variational_expectation!(
             total_entropy = smooth_with_weights!(FS[k], model.B[k], y, weights)
 
             # Calculate the ELBO contribution for the current SSM
-            # Convert single-trial format to multi-trial format for compatibility
-            y_3d = reshape(y, (size(y)..., 1))
[…]
