### 2025-07-02 depasquale-lab/StateSpaceDynamics.jl#50 @gdalle
I understand that I'm giving you a lot of headaches, perhaps more than the average JOSS review, because I'm close to the topic and I want to make sure that every package is given a fair shot. Honestly, I'd be fine with both options. For option 1, using the same optimization algorithm with the same parameters for every package (either Adam or BFGS) would be enough, provided that the implementation of the emission model is not crazy slow. Alternately, getting rid of controls is a much easier solution, and it might end up being fairer. Your call.
