### 2025-07-08 15002e4
Fix more tests. Add compat_helper.yml
