### 2025-07-08 cac4adf
reduced allocs for RegressionEmissions
