+++ src/EmissionModels.jl
@@ -301,25 +301,27 @@ function loglikelihood(
 ) where {T<:Real}
 
     if w === nothing
-        w = ones(eltype(Y), size(Y, 1))
+        w = ones(T, size(Y, 1))
     elseif eltype(w) !== eltype(Y)
         error("weights must be Vector{$(eltype(Y))}; Got Vector{$(eltype(w))}")
     end
 
-    # Add intercept if specified
-    Φ = model.include_intercept ? [ones(size(Φ, 1)) Φ] : Φ
-    
-    # residuals
-    residuals = Y - Φ * model.β
+    Φ = model.include_intercept ? hcat(ones(T, size(Φ, 1), 1), Φ) : Φ
+    n, _  = size(Φ)
+    _, d = size(Y)
+    β = model.β
+    lls = Vector{T}(undef, n)
 
-    # Calculate weighted least squares
-    weighted_residuals = residuals .^ 2 .* w
-
-    if size(weighted_residuals,2) > 1
-        weighted_residuals = vec(sum(weighted_residuals, dims=2))
+     @inbounds for i in 1:n
+        acc = zero(T)
+        for j in 1:d
+            residual = Y[i, j] - dot(Φ[i, :], β[:, j])
+            acc += residual^2
+        end
+        lls[i] = -0.5 * acc * w[i]
     end
 
-    return -0.5 .* weighted_residuals
+    return lls   
 end
 
 """
@@ -500,18 +502,25 @@ end
 Define the objective function for Gaussian/AR regression emission models.
 """
 function objective(
-    opt::Union{RegressionOptimization{<:GaussianRegressionEmission}, RegressionOptimization{<:AutoRegressionEmission}}, β_vec::AbstractVector{T}
+    opt::Union{RegressionOptimization{<:GaussianRegressionEmission}, RegressionOptimization{<:AutoRegressionEmission}}, 
+    β_vec::AbstractVector{T}
 ) where {T<:Real}
     β_mat = vec_to_matrix(β_vec, opt.β_shape)
-    residuals = opt.y - opt.X * β_mat
-    w_reshaped = reshape(opt.w, :, 1)
+    n, _ = size(opt.X)
+    total = zero(T)
+
+    # calculate residuals and w inside loop to avoid allocs 
+    @inbounds for i in 1:n
+        prediction = dot(opt.X[i, :], β_mat[:, 1])
+        residual = opt.y[i, 1] - prediction
+        total += opt.w[i] * residual^2
+    end
 
     # calculate regularization
     regularization = calc_regularization(β_mat, opt.model.λ, opt.model.include_intercept)
 
     # calculate pseudo log-likelihood
-    pseudo_ll = 0.5 * sum(w_reshaped .* residuals .^ 2) + regularization
-    return pseudo_ll
+    return 0.5 * total + regularization
 end
 
 """
@@ -528,16 +537,23 @@ function objective_gradient!(
     β_vec::AbstractVector{T},
 ) where {T<:Real}
     β_mat = vec_to_matrix(β_vec, opt.β_shape)
-    residuals = opt.y - opt.X * β_mat
+    n, p = size(opt.X)
+    _, d = size(opt.y)
+    weighted_residuals = similar(opt.y)
+
+    @inbounds for i in 1:n, j in 1:d
+        prediction = dot(opt.X[i, :], β_mat[:, j])
+        weighted_residuals[i, j] = (opt.y[i, j] - prediction) * opt.w[i]
+    end 
+
+    gradient = -(opt.X' * weighted_residuals) 
 
     # calc gradient of penalty
[…]
