### 2025-07-11 0a672ea
Fixes to LDS tests
