### 2025-07-11 6f39071
Refactor PoissonLDS tests to use FilterSmooth objects
Updated PoissonLDS test functions to use StateSpaceDynamics.initialize_FilterSmooth and pass FilterSmooth objects to smooth!, estep!, and mstep! methods. Adjusted test assertions and parameter update logic to work with the new structure. Commented out GaussianLDS test set in runtests.jl for focus on PoissonLDS tests.
