+++ test/LinearDynamicalSystems/PoissonLDS.jl
@@ -331,18 +331,24 @@ end
 function test_smooth()
     plds, x, y = toy_PoissonLDS()
 
+    tfs = StateSpaceDynamics.initialize_FilterSmooth(plds, size(y, 2), size(y, 3))
+
     # smooth data
-    x_smooth, p_smooth, inverseoffdiag = StateSpaceDynamics.smooth(plds, y)
+    ml_total = StateSpaceDynamics.smooth!(plds, tfs, y)
 
     nTrials = size(y, 3)
     nTsteps = size(y, 2)
 
-    @test size(x_smooth) == size(x)
-    @test size(p_smooth) == (plds.latent_dim, plds.latent_dim, nTsteps, nTrials)
-    @test size(inverseoffdiag) == (plds.latent_dim, plds.latent_dim, nTsteps, nTrials)
+    x_smooth = tfs[1].x_smooth
+    p_smooth = tfs[1].p_smooth
+    p_smooth_tt1 = tfs[1].p_smooth_tt1
+
+    @test size(x_smooth) == (plds.latent_dim, nTsteps)
+    @test size(p_smooth) == (plds.latent_dim, plds.latent_dim, nTsteps)
+    @test size(p_smooth_tt1) == (plds.latent_dim, plds.latent_dim, nTsteps)
 
     # test gradient is zero
-    for i in axes(y, 3)
+    for i in 1:nTrials
         # may as well test the gradient here too 
         f = latents -> StateSpaceDynamics.loglikelihood(latents, plds, y[:, :, i])
         grad_numerical = ForwardDiff.gradient(f, x_smooth[:, :, i])
@@ -355,24 +361,28 @@ end
 function test_parameter_gradient()
     plds, x, y = toy_PoissonLDS()
 
+    tfs = StateSpaceDynamics.initialize_FilterSmooth(plds, size(y, 2), size(y, 3))
+
     # run estep
-    E_z, E_zz, E_zz_prev, x_smooth, P_smooth, elbo = StateSpaceDynamics.estep(plds, y)
+    ml_total = StateSpaceDynamics.estep!(plds, tfs, y)
 
     # params
     C, log_d = plds.obs_model.C, plds.obs_model.log_d
     params = vcat(vec(C), log_d)
-
+    
     # get analytical gradient
     grad_analytical = StateSpaceDynamics.gradient_observation_model!(
-        zeros(length(params)), C, log_d, E_z, P_smooth, y
+        zeros(length(params)), C, log_d, tfs, y
     )
 
     # get numerical gradient
+    E_z = tfs[1].E_z
+    P_smooth = tfs[1].p_smooth
     function f(params::AbstractVector{<:Real})
         C_size = plds.obs_dim * plds.latent_dim
         log_d = params[(end - plds.obs_dim + 1):end]
         C = reshape(params[1:C_size], plds.obs_dim, plds.latent_dim)
-        return -StateSpaceDynamics.Q_observation_model(C, log_d, E_z, P_smooth, y)
+        return -StateSpaceDynamics.Q_observation_model(C, log_d, reshape(E_z, size(E_z)..., 1), reshape(P_smooth, size(P_smooth)..., 1), y)
     end
 
     grad = ForwardDiff.gradient(f, params)
@@ -383,17 +393,20 @@ end
 function test_initial_observation_parameter_updates(ntrials::Int=1)
     plds, x, y = toy_PoissonLDS(ntrials, [true, true, false, false, false, false])
 
-    # run the E_Step
-    E_z, E_zz, E_zz_prev, x_smooth, p_smooth, ml_total = StateSpaceDynamics.estep(plds, y)
[…]
