+++ test/runtests.jl
@@ -52,35 +52,35 @@ Tests for LDS.jl
 
 include("LinearDynamicalSystems//GaussianLDS.jl")
 
-@testset "GaussianLDS Tests" begin
-    @testset "Constructor Tests" begin
-        test_lds_with_params()
-        test_gaussian_obs_constructor_type_preservation()
-        test_gaussian_lds_constructor_type_preservation()
-        test_gaussian_sample_type_preservation()
-        test_gaussian_fit_type_preservation()
-        test_gaussian_loglikelihood_type_preservation()
-    end
-    @testset "Smoother tests" begin
-        test_Gradient()
-        test_Hessian()
-        test_smooth()
-    end
-    @testset "EM tests" begin
-        test_estep()
-        # test when ntrials=1
-        test_initial_observation_parameter_updates()
-        test_state_model_parameter_updates()
-        test_obs_model_params_updates()
-        # test when ntrials>1
-        test_initial_observation_parameter_updates(3)
-        test_state_model_parameter_updates(3)
-        test_obs_model_params_updates(3)
-        # test fit method using n=1 and n=3
-        test_EM()
-        test_EM(3)
-    end
-end
+# @testset "GaussianLDS Tests" begin
+#     @testset "Constructor Tests" begin
+#         test_lds_with_params()
+#         test_gaussian_obs_constructor_type_preservation()
+#         test_gaussian_lds_constructor_type_preservation()
+#         test_gaussian_sample_type_preservation()
+#         test_gaussian_fit_type_preservation()
+#         test_gaussian_loglikelihood_type_preservation()
+#     end
+#     @testset "Smoother tests" begin
+#         test_Gradient()
+#         test_Hessian()
+#         test_smooth()
+#     end
+#     @testset "EM tests" begin
+#         test_estep()
+#         # test when ntrials=1
+#         test_initial_observation_parameter_updates()
+#         test_state_model_parameter_updates()
+#         test_obs_model_params_updates()
+#         # test when ntrials>1
+#         test_initial_observation_parameter_updates(3)
+#         test_state_model_parameter_updates(3)
+#         test_obs_model_params_updates(3)
+#         # test fit method using n=1 and n=3
+#         test_EM()
+#         test_EM(3)
+#     end
+# end
 
 """
 Tests for PoissonLDS.jl
