### 2025-07-12 a475a10
Fix smoother variable usage and enable tests
Replaces E_zz with p_smooth in Q_observation_model and related calls to ensure correct smoother variable is used. Also uncomments and enables package-wide and GaussianLDS tests in runtests.jl for improved test coverage.
