### 2025-07-16 05fe644
adjusted global types, deleted HMM file
