+++ src/HiddenMarkovModels.jl
@@ -1,751 +1,0 @@
-# Public API
-export HiddenMarkovModel, fit!, rand, loglikelihood, viterbi, class_probabilities
-export kmeans_init!
-
-"""
-    HiddenMarkovModel
-
-Store a Hidden Markov Model (HMM) with custom emissions.
-
-# Fields
-- `A::AbstractMatrix{<:Real}`: Transition matrix.
-- `B::AbstractVector{<:EmissionModel}`: State-dependent emission models.
-- `πₖ::AbstractVector{<:Real}`: Initial state distribution.
-- `K::Int`: Number of states.
-"""
-mutable struct HiddenMarkovModel{T<:Real, V<:AbstractVector{T}, M<:AbstractMatrix{T}, VE<:AbstractVector{<:EmissionModel}} <: AbstractHMM
-    A::M # transition matrix
-    B::VE # Vector of emission Models
-    πₖ::V # initial state distribution
-    K::Int # number of states
-end
-
-"""
-    initialize_forward_backward(model::AbstractHMM, num_obs::Int)
-
-Initialize the forward backward storage struct.
-"""
-function initialize_forward_backward(model::AbstractHMM, num_obs::Int, ::Type{T}) where {T<:Real}
-    num_states = model.K
-    ForwardBackward(
-        zeros(T, num_states, num_obs),
-        zeros(T, num_states, num_obs),
-        zeros(T, num_states, num_obs),
-        zeros(T, num_states, num_obs),
-        zeros(T, num_states, num_states, num_obs - 1)
-    )
-end
-
-
-"""
-    ForwardBackward(
-        loglikelihoods::Matrix{T},
-        α::Matrix{T},
-        β::Matrix{T},
-        γ::Matrix{T},
-        ξ::Array{T,3}) where {T<:Real}
-
-Initialize the forward backward storage struct.
-"""
-function ForwardBackward(
-    loglikelihoods::Matrix{T},
-    α::Matrix{T},
-    β::Matrix{T},
-    γ::Matrix{T},
-    ξ::Array{T,3}
-) where {T<:Real}
-    ForwardBackward{T, Vector{T}, Matrix{T}, Array{T,3}}(loglikelihoods, α, β, γ, ξ)
-end
-
-"""
-    aggregate_forward_backward!(aggregated_FB::ForwardBackward, FB_storages::Vector{<:ForwardBackward})
-
-Aggregate single trial ForwardBackward structs to one session-wide struct.
-"""
-function aggregate_forward_backward!(
-    aggregated_FB::ForwardBackward, 
-    FB_storages::Vector{<:ForwardBackward}
-)
-    # Concatenate each field into the respective field in the aggregated struct
-    aggregated_FB.loglikelihoods .= hcat([fb.loglikelihoods for fb in FB_storages]...)
-    aggregated_FB.α .= hcat([fb.α for fb in FB_storages]...)
-    aggregated_FB.β .= hcat([fb.β for fb in FB_storages]...)
-    aggregated_FB.γ .= hcat([fb.γ for fb in FB_storages]...)
-    aggregated_FB.ξ = cat([fb.ξ for fb in FB_storages]..., dims=3)
-    
-end
-
-"""
-    initialize_transition_matrix(K::Int)
-
-Initialize the HMM transition matrix.
-"""
-function initialize_transition_matrix(K::Int)
-    # initialize a transition matrix
-    A = zeros(Float64, K, K)
-    @threads for i in 1:K
-        A[i, :] = rand(Dirichlet(ones(K)))
-    end
-    return A
-end
-
-"""
-    initialize_state_distribution(K::Int)
-
-Initialize the HMM initial state distribution
-"""
[…]
