+++ src/StateSpaceDynamics.jl
@@ -6,8 +6,11 @@ using ForwardDiff
 using LinearAlgebra
 using LineSearches
 using Optim
+using HiddenMarkovModels
 using ProgressMeter
+using DensityInterface
 using Random
+using FillArrays
 using SparseArrays
 using StaticArrays
 using Statistics
@@ -16,12 +19,12 @@ using StatsFuns
 using SpecialFunctions
 using Base.Threads: @threads
 
+import HiddenMarkovModels: initialization, transition_matrix, obs_distributions, logdensityof
+
 include("GlobalTypes.jl")
 include("Utilities.jl")
 include("LinearDynamicalSystems.jl")
 include("EmissionModels.jl")
-include("HiddenMarkovModels.jl")
-include("HMMConstructors.jl")
 include("SLDS.jl")
 include("MixtureModels.jl")
 include("Preprocessing.jl")
