### 2025-07-18 779649a
Emissions allocs refactoring, column based, rewriting tests
