+++ test/RegressionModels/PoissonRegression.jl
@@ -1,8 +1,8 @@
 # PoissonRegression.jl
 function PoissonRegression_simulation(; include_intercept::Bool=true)
     # Generate synthetic data
-    n = 1000
-    X = randn(n, 2)  # Remove intercept from data generation
+    n = 1000  
+    X = randn(2, n)  # Remove intercept from data generation
     true_β = [-1.2, 2.3]
     if include_intercept
         true_β = vcat(0.5, true_β)
@@ -10,9 +10,9 @@ function PoissonRegression_simulation(; include_intercept::Bool=true)
     true_β = reshape(true_β, :, 1)
 
     # Generate y with or without intercept
-    X_with_intercept = include_intercept ? hcat(ones(n), X) : X
-    λ = exp.(X_with_intercept * true_β)
-    y = reshape(rand.(Poisson.(λ)), :, 1)
+    X_with_intercept = include_intercept ? vcat(ones(1, n), X) : X
+    λ = exp.(true_β' * X_with_intercept)
+    y = reshape(rand.(Poisson.(vec(λ))), 1, :)
 
     return X, y, true_β, n
 end
@@ -59,7 +59,7 @@ function test_PoissonRegression_loglikelihood()
     @test all(isfinite.(ll))
 
     # Test single observation
-    single_ll = StateSpaceDynamics.loglikelihood(model, X[1:1, :], y[1:1, :])
+    single_ll = StateSpaceDynamics.loglikelihood(model, X[:, 1:1], y[:, 1:1])
     @test length(single_ll) == 1
     @test isfinite(single_ll[1])
 
@@ -108,8 +108,8 @@ end
 
 function test_PoissonRegression_sklearn()
     # create a set of dummy data we will fit equivalently in sklearn
-    X = reshape([1.0, 2.0, 3.0, 4.1], 4, 1)
-    y = reshape([0.0, 0.0, 2.0, 1.0], 4, 1)
+    X = reshape([1.0, 2.0, 3.0, 4.1], 1, :) 
+    y = reshape([0.0, 0.0, 2.0, 1.0], 1, :)
     w = [1.0, 1.0, 1.0, 0.1]
 
     # test the regression with no weights and no regularization
