+++ src/EmissionModels.jl
@@ -9,7 +9,7 @@ This module implements various emission models for state space modeling, includi
 # Exports
 export EmissionModel, RegressionEmission
 export GaussianEmission, GaussianRegressionEmission, BernoulliRegressionEmission, PoissonRegressionEmission, AutoRegressionEmission
-export loglikelihood, fit!
+export loglikelihood, fit!, create_optimization
 
 #=
 Gaussian Emission Models
@@ -144,10 +144,10 @@ function create_optimization(
     model::RegressionEmission,
     X::AbstractMatrix{<:Real},
     y::AbstractMatrix{<:Real},
-    w::AbstractVector{<:Real}=ones(size(y, 1)),
+    w::AbstractVector{<:Real}=ones(size(y, 2)),
 )
     if model.include_intercept
-        X = hcat(ones(size(X, 1)), X)
+        X = vcat(ones(eltype(X), 1, size(X, 2)), X)
     end
 
     β_shape = size(model.β)
@@ -267,12 +267,15 @@ Generate samples from a Gaussian regression emission model.
     -`samples::Matrix{<:Real}`: Matrix of samples (n, output_dim)
 """
 function Random.rand(rng::AbstractRNG, model::GaussianRegressionEmission, Φ::Union{Matrix{<:Real},Vector{<:Real}})
-    Φ = size(Φ, 2) == 1 ? reshape(Φ, 1, :) : Φ
-    if model.include_intercept
-        Φ = hcat(ones(size(Φ, 1)), Φ)
+    Φ = Φ isa Vector ? reshape(Φ, :, 1) : Φ 
+    n = size(Φ, 2)
+
+    if model.include_intercept 
+        Φ = vcat(ones(eltype(Φ), 1, n), Φ)
     end
-    noise = rand(rng, MvNormal(zeros(model.output_dim), model.Σ), size(Φ, 1))'
-    return Φ * model.β + noise
+   
+    noise = rand(rng, MvNormal(zeros(model.output_dim), model.Σ), n)
+    return model.β' * Φ + noise 
 end
 
 """
@@ -295,33 +298,39 @@ Calculate the log likelihood of the data `Y` given the Gaussian regression emiss
 """
 function loglikelihood(
     model::GaussianRegressionEmission,
-    Φ::AbstractMatrix{T},
-    Y::AbstractMatrix{T},
+    Φ::AbstractMatrix{T}, # f x n 
+    Y::AbstractMatrix{T}, # d x n
     w::Union{Nothing,AbstractVector{T}} = nothing,
 ) where {T<:Real}
 
+    n = size(Φ, 2)
+    d = size(Y, 1)
+    β = model.β
+
     if w === nothing
-        w = ones(T, size(Y, 1))
+        w = ones(T, size(Y, 2))
     elseif eltype(w) !== eltype(Y)
         error("weights must be Vector{$(eltype(Y))}; Got Vector{$(eltype(w))}")
     end
 
-    Φ = model.include_intercept ? hcat(ones(T, size(Φ, 1), 1), Φ) : Φ
-    n, _  = size(Φ)
-    _, d = size(Y)
-    β = model.β
-    lls = Vector{T}(undef, n)
-
-     @inbounds for i in 1:n
-        acc = zero(T)
-        for j in 1:d
-            residual = Y[i, j] - dot(Φ[i, :], β[:, j])
-            acc += residual^2
+    if model.include_intercept
+        if size(Φ, 1) == size(β, 1) - 1
+            Φ = vcat(ones(T, 1, n), Φ)
+        elseif size(Φ, 1) != size(β, 1)
+            error("Φ has incompatible size with model.β: got $(size(Φ)), expected $(size(β,1)) rows.")
[…]

[…]
