### 2025-07-18 3a3951c
more allocs reductions
