+++ v0.2.0/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"tutorials/poisson_mixture_model_example/#Simulating-and-Fitting-a-Poisson-Mixture-Model","page":"Poisson Mixture Model Example","title":"Simulating and Fitting a Poisson Mixture Model","text":"","category":"section"},{"location":"tutorials/poisson_mixture_model_example/","page":"Poisson Mixture Model Example","title":"Poisson Mixture Model Example","text":"This tutorial demonstrates how to use StateSpaceDynamics.jl to create a Poisson Mixture Model and fit it using the EM algorithm.","category":"page"},{"location":"tutorials/poisson_mixture_model_example/","page":"Poisson Mixture Model Example","title":"Poisson Mixture Model Example","text":"using StateSpaceDynamics\nusing LinearAlgebra\nusing Random\nusing Plots\nusing StableRNGs\nusing StatsPlots\nusing Distributions\n\nrng = StableRNG(1234);\nnothing #hide","category":"page"},{"location":"tutorials/poisson_mixture_model_example/#Create-a-true-PoissonMixtureModel-to-simulate-from","page":"Poisson Mixture Model Example","title":"Create a true PoissonMixtureModel to simulate from","text":"","category":"section"},{"location":"tutorials/poisson_mixture_model_example/","page":"Poisson Mixture Model Example","title":"Poisson Mixture Model Example","text":"k = 3\ntrue_λs = [5.0, 10.0, 25.0]  # Poisson means\ntrue_πs = [0.25, 0.45, 0.3]   # Mixing weights\n\ntrue_pmm = PoissonMixtureModel(k, true_λs, true_πs)","category":"page"},{"location":"tutorials/poisson_mixture_model_example/#Generate-data-from-the-true-model","page":"Poisson Mixture Model Example","title":"Generate data from the true model","text":"","category":"section"},{"location":"tutorials/poisson_mixture_model_example/","page":"Poisson Mixture Model Example","title":"Poisson Mixture Model Example","text":"n = 500\nlabels = rand(rng, Categorical(true_πs), n)\ndata = [rand(rng, Poisson(true_λs[labels[i]])) for i in 1:n]  # Vector{Int}","category":"page"},{"location":"tutorials/poisson_mixture_model_example/#Plot-histogram-of-Poisson-samples-by-component","page":"Poisson Mixture Model Example","title":"Plot histogram of Poisson samples by component","text":"","category":"section"},{"location":"tutorials/poisson_mixture_model_example/","page":"Poisson Mixture Model Example","title":"Poisson Mixture Model Example","text":"p1 = histogram(\n    data;\n    group=labels,\n    bins=0:1:maximum(data),\n    bar_position=:dodge,\n    xlabel=\"Count\",\n    ylabel=\"Frequency\",\n    title=\"Poisson Mixture Samples by Component\",\n    alpha=0.7,\n    legend=:topright,\n)\np1","category":"page"},{"location":"tutorials/poisson_mixture_model_example/#Fit-a-new-PoissonMixtureModel-to-the-data","page":"Poisson Mixture Model Example","title":"Fit a new PoissonMixtureModel to the data","text":"","category":"section"},{"location":"tutorials/poisson_mixture_model_example/","page":"Poisson Mixture Model Example","title":"Poisson Mixture Model Example","text":"fit_pmm = PoissonMixtureModel(k)\n_, lls = fit!(fit_pmm, data; maxiter=100, tol=1e-6, initialize_kmeans=true)","category":"page"},{"location":"tutorials/poi
[…]
