### 2025-07-20 c50e1a1
start redoing hmm benchmarking
