+++ benchmarking/SSD_Benchmark.jl
@@ -24,8 +24,7 @@ export Implementation
 # Include files
 include("Instances.jl")
 include("Params.jl")
-include("Controlled_BernoulliHMM.jl")
-include("glmhmm_benchmarking_tools.jl")
+include("hmm_benchmark_tools.jl")
 include("lds_benchmark_tools.jl")
 
 end
