### 2025-07-24 f8f4100
Update README and benchmarking scripts
README.md updated with improved documentation, including inference details and new references. Benchmarking scripts now activate the correct environment and reshape input data for Dynamax HMMs. Benchmark results CSV updated with new timing and memory data.
