### 2025-07-24 ea9605f
trying to get GLM-HMM to work in test files
