+++ test/HiddenMarkovModels/SwitchingGaussianRegression.jl
@@ -1,18 +1,20 @@
 function test_SwitchingGaussianRegression_fit()
     # Create Emission Models
-    emission_1 = GaussianRegressionEmission(input_dim=3, output_dim=1, include_intercept=true, β=reshape([3.0, 2.0, 2.0, 3.0], :, 1), Σ=[1.0;;], λ=0.0)
-    emission_2 = GaussianRegressionEmission(input_dim=3, output_dim=1, include_intercept=true, β=reshape([-4.0, -2.0, 3.0, 2.0], :, 1), Σ=[1.0;;], λ=0.0)
+    emission_1_true = GaussianRegressionEmission(input_dim=3, output_dim=1, include_intercept=true, β=reshape([3.0, 2.0, 2.0, 3.0], :, 1), Σ=[1.0;;], λ=0.0)
+    emission_2_true = GaussianRegressionEmission(input_dim=3, output_dim=1, include_intercept=true, β=reshape([-4.0, -2.0, 3.0, 2.0], :, 1), Σ=[1.0;;], λ=0.0)
 
     # Create Switching Regression Model
-    A = [0.99 0.01; 0.05 0.95]
-    πₖ = [0.8; 0.2]
+    A_true = [0.99 0.01; 0.05 0.95]
+    πₖ_true = [0.8; 0.2]
+    true_model = HMM(πₖ_true, A_true, [emission_1_true, emission_2_true])
 
-    true_model = HiddenMarkovModel(K=2, A=A, πₖ=πₖ, B=[emission_1, emission_2])
+    print(true_model)
 
     # Sample from the model
     n = 20000
-    Φ = randn(3, n)
-    true_labels, data = rand(true_model, Φ; n=n)
+    control_seq = Fill(nothing, n)
+    simulated_data = rand(true_model, control_seq)
+    state_seq, obs_seq = simulated_data.state_seq, simulated_data.obs_seq 
 
     # Fit a new model to the data
     A = [0.8 0.2; 0.1 0.9]
@@ -20,94 +22,20 @@ function test_SwitchingGaussianRegression_fit()
     emission_1 = GaussianRegressionEmission(input_dim=3, output_dim=1, include_intercept=true, β=reshape([2.0, -1.0, 1.0, 2.0], :, 1), Σ=[2.0;;], λ=0.0)
     emission_2 = GaussianRegressionEmission(input_dim=3, output_dim=1, include_intercept=true, β=reshape([-2.5, -1.0, 3.5, 3.0], :, 1), Σ=[0.5;;], λ=0.0)
 
-    test_model = HiddenMarkovModel(K=2, A=A, πₖ=πₖ, B=[emission_1, emission_2])
-    lls = StateSpaceDynamics.fit!(test_model, data, Φ)
+    test_model = HMM(πₖ, A, [emission_1, emission_2])
+    hmm, lls = baum_welch(test_model, obs_seq, control_seq; max_iterations=200)
 
     # Test transition matrix
-    @test isapprox(true_model.A, test_model.A, atol=0.1)
+    @test isapprox(true_model.trans, hmm.trans, atol=0.1)
 
     # Test regression fit
-    @test isapprox(test_model.B[1].β, true_model.B[1].β; atol=0.1) ||
-        isapprox(test_model.B[1].β, true_model.B[2].β; atol=0.1)
-    @test isapprox(test_model.B[2].β, true_model.B[2].β; atol=0.1) ||
-        isapprox(test_model.B[2].β, true_model.B[1].β; atol=0.1)
+    @test isapprox(hmm.dists[1].β, true_model.dists[1].β; atol=0.1) ||
+        isapprox(hmm.dists[1].β, true_model.dists[2].β; atol=0.1)
[…]

[…]
