### 2025-07-25 depasquale-lab/StateSpaceDynamics.jl#67 @slinderman
JOSS Review from Scott Linderman
Dear package authors,

Great work on this exciting project! I'll use this issue to document my comments and suggestions as I go through the paper and code. It will take a few posts, so I'll let you know when the issue is complete so that you can start your response.

Thanks,
Scott
