### 2025-07-25 depasquale-lab/StateSpaceDynamics.jl#67 @slinderman
# Comments on the paper

Overall, it's very well written and comprehensive. There's a bit of friction between the broad utility of the package — lots of disciplines use SSMs — and the specific neuroscience applications that I believe motivated your work. You might add some sentences to clarify that while your intended audience is quite broad, the specific motivating examples come from neuroscience.

One of our goals with Dynamax was to have a package that supports various hardware backends, especially CPUs and GPUs. Is that easily done with Julia and your package? Some discussion of hardware support/requirements would be nice.

Likewise, we were motivated by the automatic differentiation utilities afforded by JAX, and I know there are comparable libraries in Julia. Those could be quite helpful for calculating the gradients and Hessian blocks in your Laplace EM algorithm. From the [code](https://github.com/depasquale-lab/StateSpaceDynamics.jl/blob/main/src/LinearDynamicalSystems.jl), it looks like these quantities are hard-coded for each observation model. Some discussion of this current design and any future plans for automatic differentiation would be a nice addition as well.

On line 115, the maunscript says "our solver does not leverage specialized routines for block-banded systems, which can result in unnecessary fill-in and degraded performance at large T." It looks like you have implemented some [message passing algorithms](https://github.com/depasquale-lab/StateSpaceDynamics.jl/blob/main/src/Utilities.jl#L40) for solving block tridagonal systems. I'm curious what further improvements you have in mind.

Minor:
- Please cite the Dynamax JOSS paper instead of the repo
```@article{Linderman_Dynamax_A_Python_2025,author = {Linderman, Scott W. and Chang, Peter and Harper-Donnelly, Giles and Kara, Aleyna and Li, Xinglong and Duran-Martin, Gerardo and Murphy, Kevin},doi = {10.21105/joss.07069},journal = {Journal of Open Source Software},month = apr,number = {108},pages = {7069},title = {{Dynamax: A Python package for probabilistic state space modeling with JAX}},url = {https://joss.theoj.org/papers/10.21105/joss.07069},volume = {10},year = {2025}}```
- On line 121 you say, "HiddenMarkovModels.jl consistently performs best across a range of sequence lengths and dimensional configurations.", but from fig. 2 it appears Dynamax is generally faster.
