### 2025-07-25 depasquale-lab/StateSpaceDynamics.jl#67 @slinderman
# Comments on the package and docs

## Miscellaneous
- [ ] The repo is about 100MB even though the source code and other assets are very small. You might be able to reduce the size with `git gc`.
- [ ] Thanks for the shoutout to SSM in the README! It would be great to reference Dynamax as well, since that's where our current effort is focused. We're not really maintaining SSM much anymore.

- [ ] It wasn't obvious how to get to the documentation from the github repo. Consider adding a link in the right side bar and/or in the README.
- [ ] The LDS, HMM, and Emission model pages do a nice job of describing the models and algorithms, with appropriate links to code. The SLDS and Mixture Model pages are a bit sparse by comparison.
- [ ] I would love to see more text and explanation in the tutorials. Currently, they pretty much just have the code along with some section headers. You could also link back and forth between the model pages, which give the mathematical details, and the tutorial pages, which show how to use these models/algorithms.
- [ ] In the Gaussian and Poisson LDS tutorials, it could be confusing that the inferred latent states do not match the ground truth, even after EM. I assume that this is due to the non-identifiability of unconstrained Gaussian LDS. If so, that is worth discussing, and you could even show how to find the linear transformation that best recovers the true parameters. You could also show that even though you don't get the ground truth latent states, you still smooth the emissions accurately.
- [ ] As above, it would be good to mention identifiability issues with latent states of HMMs and mixture models.
- [ ] One of the big questions we always get is how to choose the number of states. A tutorial (or section in existing tutorials) on cross validation (or other preferred methods) would be a nice addition.
- [ ] Per the checklist, please describe the automated tests in the docs.
