### 2025-07-25 depasquale-lab/StateSpaceDynamics.jl#67 @slinderman
# Comments on the code

I'm not a Julia user so apologies in advance!  These are mostly just a bunch of miscellaneous thoughts as I went through the code, in no particular order.

- [ ]  Most HMMs have stationary transition matrices, so the expected transition counts could be directly summed in the forward-backward algorithm. That would cut your memory from $TK^2$ to $K^2$. For large models and long time series, that's a big cost saving.
- [ ] For the emission M-steps, it might be helpful to have a prior (or simple regularization if you prefer to think of it that way) to ensure that your covariances are full rank. One issue I've encountered is when the number of time steps assigned to a state drops below the number of states, the M-step yields a low rank covariance that causes log pdfs to nan out. A weak prior would address this issue. (Same goes for smoothing the m-steps of transition matrices and initial distributions.)
- [ ] I don't see any affine terms in the dynamics or emissions of the [Gaussian LDS models](https://github.com/depasquale-lab/StateSpaceDynamics.jl/blob/main/src/LinearDynamicalSystems.jl#L3-L33). Without a bias, how do you handle emissions that are not mean zero?
- [ ] The exponential mean function in the Poisson emissions is mathematically convenient but can be numerically unstable, especially with float32. Have you considered other mean functions like softplus?
- [ ] Looks like the SLDS code is currently for the standard Gaussian SLDS model. How hard would it be to implement the variational Laplace EM algorithm (Zoltowski et al., 2020) for recurrent SLDS with Gaussian and/or GLM emissions? It seems like you have all the building blocks!
- [ ] The code is (very reasonably) geared towards point estimation of the model parameters. Is there a plan to implement functionality for statistical inference (frequentist or Bayesian)?
