### 2025-07-25 #8077 @slinderman
Ok, I've completed my review and posted several comments in https://github.com/depasquale-lab/StateSpaceDynamics.jl/issues/67. I did not attempt to reproduce the performance claims, but I trust the authors.
