### 2025-07-27 32dc0c8
Pretty printing GaussianStateModel
