### 2025-07-27 ab5607b
Pretty print obs models
