### 2025-07-27 a9191bc
Pretty Print LDS
