### 2025-07-27 8c6df3c
Found all objects to pretty print
