### 2025-07-27 e5672e1
Pretty printing for HMM, SLDS
