### 2025-07-27 d10e25a
Pretty printing Mixture Models
