### 2025-07-27 c99ab58
pretty printing for mixture models
