### 2025-07-28 fffa09d
Fixed some doc bugs, set up tests for Pretty Printing
