+++ test/PrettyPrinting/PrettyPrinting.jl
@@ -1,4 +1,114 @@
 function test_pretty_printing()
+    io = IOBuffer()
+
+    # Gaussian Emission
+
+    ge1 = GaussianEmission(5, rand(5), rand(5,5))
+    ge2 = GaussianEmission(2, rand(2), rand(2,2))
+    @test println(io, ge1) === nothing
+    @test println(io, ge2) === nothing
+
+    # Gaussian Regression Emission
+
+    gre = GaussianRegressionEmission(3, 3, rand(3,3), rand(3,3), true, 0.5)
+    @test println(io, gre) === nothing
+
+    # AutoRegression Emission 
+
+    are = AutoRegressionEmission(3, 3, gre)
+    @test println(io, are) === nothing
+
+    # Bernoulli Regression Emission (incomplete)
+
+    bre = BernoulliRegressionEmission(5, 5, rand(5,5), false, 0.5)
+
+    @test println(io, bre) === nothing
+
+    # Poisson regression emission (incomplete)
+
+    pre = PoissonRegressionEmission(5, 5, rand(5,5), false, 0.5)
+    
+    @test println(io, pre) === nothing
+
+    # hidden markov model 
+
+    hmm1 = HiddenMarkovModel(rand(5,5), [gre, are, bre, pre, gre], rand(5), 5)
+    hmm2 = HiddenMarkovModel(rand(2,2), [gre, are], rand(2), 2)
+
+    @test println(io, hmm1) === nothing
+    @test println(io, hmm2) === nothing
+
+    # Gaussian State model 
+
+    gsm1 = GaussianStateModel(rand(5,5), rand(5,5), rand(5), rand(5,5))
+    gsm2 = GaussianStateModel(rand(2,2), rand(2,2), rand(2), rand(2,2))
+
+    @test println(io, gsm1) === nothing
+    @test println(io, gsm2) === nothing
+
+    # gaussian observation model 
+
+    gom1 = GaussianObservationModel(rand(5,5), rand(5,5))
+    gom2 = GaussianObservationModel(rand(3,3), rand(3,3))
+
+    @test println(io, gom1) === nothing
+    @test println(io, gom2) === nothing
+
+    # poisson observation model 
+
+    pom1 = PoissonObservationModel(rand(5,5))
+    pom2 = PoissonObservationModel(rand(2,2))
+
+    @test println(io, pom1) === nothing
+    @test println(io, pom2) === nothing
+
+    # linear dynamical system 
+
+    lds1 = LinearDynamicalSystem(gsm1, gom1, 5, 5, trues(6))
+    lds2 = LinearDynamicalSystem(gsm2, pom2, 2, 2, trues(5))
+
+    @test println(io, lds1) === nothing
+    @test println(io, lds2) === nothing
+
+    # gaussian mixture model 
+
+    gmm1 = GaussianMixtureModel(5, rand(5,5), [rand(5,5) for _ in 1:5], rand(5))
+    gmm2 = GaussianMixtureModel(2, rand(2,2), [rand(2,2) for _ in 1:2], rand(2))
+
+    @test println(io, gmm1) === nothing
+    @test println(io, gmm2) === nothing
+
+    # Poisson Mixture model 
+
+    pmm1 = PoissonMixtureModel(5, rand(5), rand(5))
+    pmm2 = PoissonMixtureModel(2, rand(2), rand(2))
+
+    @test println(io, pmm1) === nothing
+    @test println(io, pmm2) === nothing
+
+    # Probabalistic PCA (incomplete)
+
+    ppca = ProbabilisticPCA()
+
+    @test println(io, ppca) === nothing
 
 
+    # switching linear dynamical system 
+
[…]
