### 2025-07-28 40a4005
Fixed bugs in pretty printing
