### 2025-07-28 f271c0c
Pretty printing for Regression Emission and PPCA
