+++ test/PrettyPrinting/PrettyPrinting.jl
@@ -18,19 +18,19 @@ function test_pretty_printing()
     are = AutoRegressionEmission(3, 3, gre)
     @test println(io, are) === nothing
 
-    # Bernoulli Regression Emission (incomplete)
+    # Bernoulli Regression Emission
 
     bre = BernoulliRegressionEmission(5, 5, rand(5,5), false, 0.5)
 
     @test println(io, bre) === nothing
 
-    # Poisson regression emission (incomplete)
+    # Poisson Regression Emission
 
     pre = PoissonRegressionEmission(5, 5, rand(5,5), false, 0.5)
     
     @test println(io, pre) === nothing
 
-    # hidden markov model 
+    # Hidden Markov Model 
 
     hmm1 = HiddenMarkovModel(rand(5,5), [gre, are, bre, pre, gre], rand(5), 5)
     hmm2 = HiddenMarkovModel(rand(2,2), [gre, are], rand(2), 2)
@@ -38,7 +38,7 @@ function test_pretty_printing()
     @test println(io, hmm1) === nothing
     @test println(io, hmm2) === nothing
 
-    # Gaussian State model 
+    # Gaussian State Model 
 
     gsm1 = GaussianStateModel(rand(5,5), rand(5,5), rand(5), rand(5,5))
     gsm2 = GaussianStateModel(rand(2,2), rand(2,2), rand(2), rand(2,2))
@@ -46,7 +46,7 @@ function test_pretty_printing()
     @test println(io, gsm1) === nothing
     @test println(io, gsm2) === nothing
 
-    # gaussian observation model 
+    # Gaussian Observation Model 
 
     gom1 = GaussianObservationModel(rand(5,5), rand(5,5))
     gom2 = GaussianObservationModel(rand(3,3), rand(3,3))
@@ -54,7 +54,7 @@ function test_pretty_printing()
     @test println(io, gom1) === nothing
     @test println(io, gom2) === nothing
 
-    # poisson observation model 
+    # Poisson Observation Model 
 
     pom1 = PoissonObservationModel(rand(5,5), rand(5))
     pom2 = PoissonObservationModel(rand(2,2), rand(2))
@@ -62,7 +62,7 @@ function test_pretty_printing()
     @test println(io, pom1) === nothing
     @test println(io, pom2) === nothing
 
-    # linear dynamical system 
+    # Linear Dynamical System 
 
     lds1 = LinearDynamicalSystem(gsm1, gom1, 5, 5, [true, true, true, true, true, true])
     lds2 = LinearDynamicalSystem(gsm2, pom2, 2, 2, [true, true, true, true, true])
@@ -70,7 +70,7 @@ function test_pretty_printing()
     @test println(io, lds1) === nothing
     @test println(io, lds2) === nothing
 
-    # gaussian mixture model 
+    # Gaussian Mixture Model 
 
     gmm1 = GaussianMixtureModel(5, rand(5,5), [rand(5,5) for _ in 1:5], rand(5))
     gmm2 = GaussianMixtureModel(2, rand(2,2), [rand(2,2) for _ in 1:2], rand(2))
@@ -78,7 +78,7 @@ function test_pretty_printing()
     @test println(io, gmm1) === nothing
     @test println(io, gmm2) === nothing
 
-    # Poisson Mixture model 
+    # Poisson Mixture Model 
 
     pmm1 = PoissonMixtureModel(5, rand(5), rand(5))
     pmm2 = PoissonMixtureModel(2, rand(2), rand(2))
@@ -86,14 +86,14 @@ function test_pretty_printing()
     @test println(io, pmm1) === nothing
     @test println(io, pmm2) === nothing
 
[…]
