### 2025-07-29 0b81e88
Update test/PrettyPrinting/PrettyPrinting.jl
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
