### 2025-07-29 1bc25a2
Implemented 'print_full' as discussed in PR#68
