### 2025-07-29 76cc508
Bugfix and implemented tests for `print_full`
