+++ test/PrettyPrinting/PrettyPrinting.jl
@@ -1,118 +1,143 @@
 function test_pretty_printing()
     # Set up IO buffer for printing
     
-    io = IOBuffer()
+    io1 = IOBuffer()
+    objs = []
 
     # Gaussian Emission
 
     ge1 = GaussianEmission(5, rand(5), rand(5,5))
     ge2 = GaussianEmission(2, rand(2), rand(2,2))
+    push!(objs, ge1, ge2)
 
-    @test println(io, ge1) === nothing
-    @test println(io, ge2) === nothing
+    @test println(io1, ge1) === nothing
+    @test println(io1, ge2) === nothing
 
     # Gaussian Regression Emission
 
     gre = GaussianRegressionEmission(3, 3, rand(3,3), rand(3,3), true, 0.5)
+    push!(objs, gre)
 
-    @test println(io, gre) === nothing
+    @test println(io1, gre) === nothing
 
     # AutoRegression Emission 
 
     are = AutoRegressionEmission(3, 3, gre)
+    push!(objs, are)
 
-    @test println(io, are) === nothing
+    @test println(io1, are) === nothing
 
     # Bernoulli Regression Emission
 
     bre = BernoulliRegressionEmission(5, 5, rand(5,5), false, 0.5)
+    push!(objs, bre)
 
-    @test println(io, bre) === nothing
+    @test println(io1, bre) === nothing
 
     # Poisson Regression Emission
 
     pre = PoissonRegressionEmission(5, 5, rand(5,5), false, 0.5)
+    push!(objs, pre)
     
-    @test println(io, pre) === nothing
+    @test println(io1, pre) === nothing
 
     # Hidden Markov Model 
 
     hmm1 = HiddenMarkovModel(rand(5,5), [gre, are, bre, pre, gre], rand(5), 5)
     hmm2 = HiddenMarkovModel(rand(2,2), [gre, are], rand(2), 2)
+    push!(objs, hmm1, hmm2)
 
-    @test println(io, hmm1) === nothing
-    @test println(io, hmm2) === nothing
+    @test println(io1, hmm1) === nothing
+    @test println(io1, hmm2) === nothing
 
     # Gaussian State Model 
 
     gsm1 = GaussianStateModel(rand(5,5), rand(5,5), rand(5), rand(5,5))
     gsm2 = GaussianStateModel(rand(2,2), rand(2,2), rand(2), rand(2,2))
+    push!(objs, gsm1, gsm2)
 
-    @test println(io, gsm1) === nothing
-    @test println(io, gsm2) === nothing
+    @test println(io1, gsm1) === nothing
+    @test println(io1, gsm2) === nothing
 
     # Gaussian Observation Model 
 
     gom1 = GaussianObservationModel(rand(5,5), rand(5,5))
     gom2 = GaussianObservationModel(rand(3,3), rand(3,3))
+    push!(objs, gom1, gom2)
 
-    @test println(io, gom1) === nothing
-    @test println(io, gom2) === nothing
+    @test println(io1, gom1) === nothing
+    @test println(io1, gom2) === nothing
 
     # Poisson Observation Model 
 
     pom1 = PoissonObservationModel(rand(5,5), rand(5))
     pom2 = PoissonObservationModel(rand(2,2), rand(2))
+    push!(objs, pom1, pom2)
 
-    @test println(io, pom1) === nothing
-    @test println(io, pom2) === nothing
+    @test println(io1, pom1) === nothing
+    @test println(io1, pom2) === nothing
 
     # Linear Dynamical System 
 
     lds1 = LinearDynamicalSystem(gsm1, gom1, 5, 5, [true, true, true, true, true, true])
[…]
