+++ src/Utilities.jl
@@ -2,6 +2,7 @@ export kmeanspp_initialization,
     kmeans_clustering, fit!, block_tridgm, block_tridiagonal_inverse, block_tridiagonal_inverse_static
 export row_matrix, stabilize_covariance_matrix, valid_Σ, make_posdef!, gaussian_entropy
 export random_rotation_matrix
+export print_full
 
 # Type checking utilities
 """
@@ -516,7 +517,7 @@ end
 
 # Pretty print function that doesn't truncate arrays of model objects
 
-function print_full(io::Union{IOStream, Base.TTY}, obj)
+function print_full(io::Union{IO, Base.TTY}, obj)
     println(IOContext(io, :limit => false), obj)
 end
 
