### 2025-07-29 8553a75
Added docstring for `print_full`
