+++ src/Utilities.jl
@@ -517,6 +517,13 @@ end
 
 # Pretty print function that doesn't truncate arrays of model objects
 
+"""
+    print_full([io::Union{IO, Base.TTY}, ] obj)
+
+Prints full description of object `obj`, overriding both `io`-based limits as
+well as the limits set in the default pretty printing of `StateSpaceDynamics`
+objects.
+"""
 function print_full(io::Union{IO, Base.TTY}, obj)
     println(IOContext(io, :limit => false), obj)
 end
