### 2025-07-30 91d0afc
Missed the RegressionOptimization object!
