### 2025-07-31 df2e5f2
Missed structs in GlobalTypes.jl
