### 2025-08-04 d7cef17
Regression Emission tests, Emissions.jl HMM interface
