+++ src/EmissionModels.jl
@@ -7,10 +7,12 @@ This module implements various emission models for state space modeling, includi
 """
 
 # Exports
-export EmissionModel, RegressionEmission
+export EmissionModel
 export GaussianEmission, GaussianRegressionEmission, BernoulliRegressionEmission, PoissonRegressionEmission, AutoRegressionEmission
 export loglikelihood, fit!, create_optimization, objective, objective_gradient!
 
+import Random: rand, Sampler 
+import StatsAPI: fit!
 #=
 Gaussian Emission Models
 =#
@@ -150,6 +152,9 @@ function create_optimization(
         X = vcat(ones(eltype(X), 1, size(X, 2)), X)
     end
 
+    @assert size(X, 1) == size(model.β, 1) "X rows $(size(X,1)) must match β rows $(size(model.β,1))"
+    @assert size(y, 2) == length(w) "Weights length $(length(w)) must match sample count $(size(y,2))"
+
     β_shape = size(model.β)
     return RegressionOptimization(model, X, y, w, β_shape)
 end
@@ -220,55 +225,6 @@ mutable struct GaussianRegressionEmission{T<:Real, M<:AbstractMatrix{T}} <: Regr
     λ::T # regularization parameter
 end
 
-# abstract type AbstractResult end 
-
-# struct GaussianRegressionObservation{T} <: AbstractResult
-#     Φ::Vector{T}
-#     y::Vector{T}
-# end
-
-"""
-    rand interface for HMM rand function
-"""
-struct RegressionSampler{E}
-    emission::E
-    x::Union{Nothing, AbstractVector}
-    xdist::Union{Nothing, Distribution}
-end
-
-RegressionSampler(emission::E, x::AbstractVector) where {E} =
-    RegressionSampler{E}(emission, x, nothing)
-
-RegressionSampler(emission::E, xdist::Distribution) where {E} =
-    RegressionSampler{E}(emission, nothing, xdist)
-
-Random.Sampler(::Type{<:Random.SamplerTrivial}, sampler::RegressionSampler) =
-    Random.SamplerTrivial(sampler)
-
-function Random.rand(rng::AbstractRNG, sampler::Random.SamplerTrivial{<:RegressionSampler})
-    emission = sampler[].emissionk
-    x = sampler[].x
-    xdist = sampler[].xdist
-
-    if x === nothing
-        @assert xdist !== nothing "Either x or xdist must be provided"
-        x = rand(rng, xdist)
-    end
-
-    Φ = x isa Vector ? reshape(x, :, 1) : x
-    if emission.include_intercept
-        Φ = vcat(ones(eltype(Φ), 1, size(Φ, 2)), Φ)
-    end
-    μ = emission.β' * Φ
-    return vec(rand(rng, MvNormal(vec(μ), emission.Σ)))
-end
-
-function obs_distributions(hmm::AbstractHMM)
-    input_dim = hmm.emissions[1].input_dim
-    xdist = MvNormal(zeros(input_dim), I)
-    return RegressionSampler.(hmm.emissions, Ref(xdist))
-end
-
 """
     GaussianRegressionEmission(input_dim, output_dim, include_intercept, β, Σ, λ)
 
@@ -301,8 +257,6 @@ function GaussianRegressionEmission(;
 
     return GaussianRegressionEmission(input_dim, output_dim, β, Σ, include_intercept, λ)
 end
-
-DensityInterface.DensityKind(::GaussianRegressionEmission) = HasDensity()
 
 """
[…]
