### 2025-08-05 9b0543e
Attempt 1 at implementing Blue formatting style
