+++ src/Preprocessing.jl
@@ -12,7 +12,8 @@ mutable struct ProbabilisticPCA{T<:Real, M<:AbstractMatrix{T}, V<:AbstractVector
     function ProbabilisticPCA(W::AbstractMatrix{T}, σ²::T, μ::AbstractVector{T}) where {T<:Real}
         D, k = size(W)
         z = Matrix{T}(undef, k, 0)  # placeholder, filled after E-step
-        new{T, typeof(W), typeof(μ)}(W, σ², μ, k, D, z)
+
+        return new{T, typeof(W), typeof(μ)}(W, σ², μ, k, D, z)
     end
 end
 
@@ -22,12 +23,14 @@ function Base.show(io::IO, ppca::ProbabilisticPCA; gap = "")
     println(io, gap, " size(W) = ($(size(ppca.W,1)), $(size(ppca.W,2)))")
     println(io, gap, " size(z) = ($(size(ppca.z,1)), $(size(ppca.z,2)))")
     println(io, gap, "      σ² = $(round(ppca.σ², digits=2))")
-    
+
     if length(ppca.μ) > 6
         println(io, gap, " size(μ) = ($(length(ppca.μ)),)")
     else
         println(io, gap, "      μ  = $(round.(ppca.μ, digits=2))")
     end
+
+    return nothing
 end
 
 function estep(ppca::ProbabilisticPCA, X::Matrix{T}) where {T<:Real}
@@ -36,18 +39,26 @@ function estep(ppca::ProbabilisticPCA, X::Matrix{T}) where {T<:Real}
     E_zz = zeros(T, ppca.k, ppca.k, N)
     M = ppca.W' * ppca.W + ppca.σ² * I(ppca.k)
     M_inv = inv(M)
+
     @views for i in 1:N
         E_z[:, i] .= M_inv * ppca.W' * (X[:, i] - ppca.μ)
         E_zz[:, :, i] .= ppca.σ² * M_inv + E_z[:, i] * E_z[:, i]'
     end
+
     return E_z, E_zz
 end
 
-function mstep!(ppca::ProbabilisticPCA, X::Matrix{T}, E_z::Matrix{T}, E_zz::Array{T, 3}) where {T<:Real}
+function mstep!(
+    ppca::ProbabilisticPCA,
+    X::Matrix{T},
+    E_z::Matrix{T},
+    E_zz::Array{T, 3}
+) where {T<:Real}
     D, N = size(X)
     W_new = zeros(T, D, ppca.k)
     σ²_sum = zero(T)
     WW = ppca.W' * ppca.W
+
     for i in 1:N
         x_centered = X[:, i] .- ppca.μ
         ez = @view(E_z[:, i])
@@ -55,9 +66,12 @@ function mstep!(ppca::ProbabilisticPCA, X::Matrix{T}, E_z::Matrix{T}, E_zz::Arra
         W_new .+= x_centered * ez'
         σ²_sum += sum(x_centered .^ 2) - 2 * dot(ez, ppca.W' * x_centered) + tr(ezz * WW)
     end
+
     ppca.z = E_z
     ppca.W = W_new * inv(sum(E_zz, dims=3)[:, :, 1])
     ppca.σ² = σ²_sum / (N * D)
+
+    return nothing
 end
 
 function loglikelihood(ppca::ProbabilisticPCA, X::AbstractMatrix{T}) where {T<:Real}
@@ -66,10 +80,12 @@ function loglikelihood(ppca::ProbabilisticPCA, X::AbstractMatrix{T}) where {T<:R
     C = ppca.W * ppca.W' + ppca.σ² * I(D)
     X_centered = X .- ppca.μ
     S = (X_centered * X_centered') / N
+
     try
         C_chol = cholesky(Symmetric(C))
[…]
