+++ src/LinearDynamicalSystems.jl
@@ -1,27 +1,30 @@
-export LinearDynamicalSystem, GaussianStateModel, GaussianObservationModel, PoissonObservationModel, rand, smooth, fit!
+export LinearDynamicalSystem, GaussianStateModel, GaussianObservationModel,
+    PoissonObservationModel, rand, smooth, fit!
 
 """
     GaussianStateModel{T<:Real. M<:AbstractMatrix{T}, V<:AbstractVector{T}}}
 
 Represents the state model of a Linear Dynamical System with Gaussian noise.
 
-# Fields 
-- `A::M`: Transition matrix (size `latent_dim×latent_dim`). 
-- `Q::M`: Process noise covariance matrix 
+# Fields
+- `A::M`: Transition matrix (size `latent_dim×latent_dim`).
+- `Q::M`: Process noise covariance matrix
 - `x0::V`: Initial state vector (length `latent_dim`).
 - `P0::M`: Initial state covariance matrix (size `latent_dim×latent_dim
 """
-Base.@kwdef mutable struct GaussianStateModel{T<:Real, M<:AbstractMatrix{T}, V<:AbstractVector{T}} <: AbstractStateModel{T}
+Base.@kwdef mutable struct GaussianStateModel{
+    T<:Real, M<:AbstractMatrix{T}, V<:AbstractVector{T}
+} <: AbstractStateModel{T}
     A::M
     Q::M
     x0::V
-    P0::M 
+    P0::M
 end
 
 function Base.show(io::IO, gsm::GaussianStateModel; gap = "")
     println(io, gap, "Gaussian State Model:")
     println(io, gap, "---------------------")
-    
+
     if size(gsm.A, 1) > 4
         println(io, gap, " State Parameters:")
         println(io, gap, "  size(A)  = ($(size(gsm.A,1)), $(size(gsm.A,2)))")
@@ -37,18 +40,23 @@ function Base.show(io::IO, gsm::GaussianStateModel; gap = "")
         println(io, gap, "  x0 = $(round.(gsm.x0, digits=2))")
         println(io, gap, "  P0 = $(round.(gsm.P0, sigdigits=3))")
     end
+
+    return nothing
 end
 
 """
-    GaussianObservationModel{T<:Real, M<:AbstractMatrix{T}} 
+    GaussianObservationModel{T<:Real, M<:AbstractMatrix{T}}
 
 Represents the observation model of a Linear Dynamical System with Gaussian noise.
 
 # Fields
-- `C::M`: Observation matrix of size `(obs_dim × latent_dim)`. Maps latent states into observation space. 
-- `R::M`: Observation noise covariance of size `(obs_dim × obs_dim)`. 
+- `C::M`: Observation matrix of size `(obs_dim × latent_dim)`. Maps latent states into
+    observation space.
+- `R::M`: Observation noise covariance of size `(obs_dim × obs_dim)`.
 """
-Base.@kwdef mutable struct GaussianObservationModel{T<:Real, M<:AbstractMatrix{T}} <: AbstractObservationModel{T}
+Base.@kwdef mutable struct GaussianObservationModel{
+    T<:Real, M<:AbstractMatrix{T}
+} <: AbstractObservationModel{T}
     C::M
     R::M
 end
@@ -64,18 +72,27 @@ function Base.show(io::IO, gom::GaussianObservationModel; gap = "")
         println(io, gap, " C = $(round.(gom.C, digits=2))")
         println(io, gap, " R = $(round.(gom.R, digits=2))")
     end
+
+    return nothing
 end
 
 """
-    PoissonObservationModel{T<:Real, M<:AbstractMatrix{T}, V<:AbstractVector{T}} <: AbstractObservationModel{T}
+    PoissonObservationModel{
+        T<:Real,
+        M<:AbstractMatrix{T},
[…]

[…]
