+++ src/SLDS.jl
@@ -14,7 +14,7 @@ mutable struct SwitchingLinearDynamicalSystem{
     T<:Real,
     M<:AbstractMatrix{T},
     V<:AbstractVector{T},
-    VL<:AbstractVector{<:LinearDynamicalSystem}
+    VL<:AbstractVector{<:LinearDynamicalSystem},
 } <: AbstractHMM
     A::M
     B::VL
@@ -22,7 +22,7 @@ mutable struct SwitchingLinearDynamicalSystem{
     K::Int
 end
 
-function Base.show(io::IO, slds::SwitchingLinearDynamicalSystem; gap = "")
+function Base.show(io::IO, slds::SwitchingLinearDynamicalSystem; gap="")
     println(io, gap, "Switching Linear Dynamical System:")
     println(io, gap, "----------------------------------")
 
@@ -41,13 +41,13 @@ function Base.show(io::IO, slds::SwitchingLinearDynamicalSystem; gap = "")
 
     if slds.K > 4 && !show_all
         for lds in slds.B[1:3]
-            Base.show(io, lds, gap = gap * "  ")
+            Base.show(io, lds; gap=gap * "  ")
             println(io, gap, "  --------------")
         end
         println(io, gap, "  $(slds.K-3) more ..., see `print_full()`")
     else
         for (i, lds) in enumerate(slds.B)
-            Base.show(io, lds; gap = gap * "  ")
+            Base.show(io, lds; gap=gap * "  ")
             if i < slds.K
                 println(io, gap, "  --------------")
             end
@@ -82,17 +82,19 @@ function Random.rand(rng::AbstractRNG, slds::SwitchingLinearDynamicalSystem, T::
     # Sample initial mode
     z[1] = rand(rng, Categorical(slds.πₖ / sum(slds.πₖ)))
     x[:, 1] = rand(rng, MvNormal(zeros(state_dim), slds.B[z[1]].state_model.Q))
-    y[:, 1] = rand(rng, MvNormal(slds.B[z[1]].obs_model.C * x[:, 1], slds.B[z[1]].obs_model.R))
+    y[:, 1] = rand(
+        rng, MvNormal(slds.B[z[1]].obs_model.C * x[:, 1], slds.B[z[1]].obs_model.R)
+    )
 
     @views for t in 2:T
         # Sample mode based on transition probabilities
-        z[t] = rand(Categorical(slds.A[z[t-1], :] ./ sum(slds.A[z[t-1], :])))
+        z[t] = rand(Categorical(slds.A[z[t - 1], :] ./ sum(slds.A[z[t - 1], :])))
         # Update latent state and observation
         x[:, t] = rand(
-          MvNormal(slds.B[z[t]].state_model.A * x[:, t-1], slds.B[z[t]].state_model.Q)
+            MvNormal(slds.B[z[t]].state_model.A * x[:, t - 1], slds.B[z[t]].state_model.Q)
         )
         y[:, t] = rand(
-          MvNormal(slds.B[z[t]].obs_model.C * x[:, t], slds.B[z[t]].obs_model.R)
+            MvNormal(slds.B[z[t]].obs_model.C * x[:, t], slds.B[z[t]].obs_model.R)
         )
     end
 
@@ -108,7 +110,9 @@ end
 
 Initialize a Switching Linear Dynamical System with random parameters.
 """
-function initialize_slds(;K::Int=2, d::Int=2, p::Int=10, self_bias::Float64=5.0, seed::Int=42)
+function initialize_slds(;
+    K::Int=2, d::Int=2, p::Int=10, self_bias::Float64=5.0, seed::Int=42
+)
     Random.seed!(seed)
 
     # Transition matrix using Dirichlet with self-bias
[…]
