+++ test/Utilities/Utilities.jl
@@ -77,11 +77,13 @@ end
 function test_autoregressive_setters_and_getters()
 
     # Define AR emission
-    AR = AutoRegressionEmission(output_dim=2, order=1, include_intercept=false, β=rand(4,4), Σ=rand(2,2), λ=0.0)
+    AR = AutoRegressionEmission(;
+        output_dim=2, order=1, include_intercept=false, β=rand(4, 4), Σ=rand(2, 2), λ=0.0
+    )
 
     # Define parameters
-    β = rand(4,4)
-    Σ = rand(4,4)
+    β = rand(4, 4)
+    Σ = rand(4, 4)
     λ = 1.0
 
     # Set parameters of inner gaussian regression of AR emission using defined parameters
@@ -98,8 +100,14 @@ function test_autoregressive_setters_and_getters()
     @test AR.innerGaussianRegression.λ == AR.λ
 
     # Test setting innerGaussianRegression directly
-    GR = GaussianRegressionEmission(input_dim=1, output_dim=2, include_intercept=false, β=2*rand(1, 2), Σ=rand(2,2), λ=0.0)
+    GR = GaussianRegressionEmission(;
+        input_dim=1,
+        output_dim=2,
+        include_intercept=false,
+        β=2*rand(1, 2),
+        Σ=rand(2, 2),
+        λ=0.0,
+    )
     AR.innerGaussianRegression = GR
     @test AR.innerGaussianRegression == GR
-
 end
