### 2025-08-06 a0d2615
docstring wraparounds
