### 2025-08-06 2550dc3
debugging Dynamax benchmarking
