### 2025-08-06 e98fb3a
revised Switching_gaussian_regression tests and tried to fix PoissonRegression tests
