### 2025-08-11 50847e1
Switched to format check in `]test`
