### 2025-08-11 cdb3e5e
My one disagreement with the formatter
