+++ Project.toml
@@ -7,7 +7,6 @@ version = "0.2.0"
 ArrayLayouts = "4c555306-a7a7-4459-81d9-ec55ddd5c99a"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-JuliaFormatter = "98e50ef6-434e-11e9-1051-2b60c6c9e899"
 LineSearches = "d3d80556-e9d4-5f37-9878-2ab0fcc64255"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Optim = "429524aa-4258-5aef-a3af-852621145aeb"
@@ -24,7 +23,6 @@ StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 ArrayLayouts = "v1.10.4"
 Distributions = "0.25.109"
 ForwardDiff = "0.10.36"
-JuliaFormatter = "2.1.6"
 LineSearches = "7.3.0"
 LinearAlgebra = "1"
 Optim = "1.10.0"
