### 2025-08-18 de256f8
dynamax jit compiltation working now. Errors in other two implementations
