### 2025-08-19 99dbfb0
Update README.md
add contributor guidelines
