### 2025-08-19 ab75fd6
lds and hmm benchmark results uploaded
