+++ benchmarking/run_benchmark.jl
@@ -20,98 +20,90 @@ end
 
 BenchmarkTools.DEFAULT_PARAMETERS.seconds = 1.0
 
-# # ----------------------
-# # LDS Benchmarks
-# # ----------------------
+# ----------------------
+# LDS Benchmarks
+# ----------------------
 
-# lds_config = BenchConfig(
-#     [2, 4, 6, 8],  # latent_dims
-#     [2, 4, 6, 8],  # obs_dims
-#     [100, 500, 1000],  # seq_lengths
-#     100,  # n_iters
-#     5     # n_repeats
-# )
+lds_config = BenchConfig(
+    [2, 4, 6, 8],  # latent_dims
+    [2, 4, 6, 8],  # obs_dims
+    [100, 500, 1000],  # seq_lengths
+    100,  # n_iters
+    5     # n_repeats
+)
 
-# lds_implementations = [
-#     SSD_LDSImplem(),
-#     pykalman_LDSImplem(),
-#     Dynamax_LDSImplem()
-# ]
+lds_implementations = [
+    SSD_LDSImplem(),
+    pykalman_LDSImplem(),
+    Dynamax_LDSImplem()
+]
 
-# lds_results = []
+lds_results = []
 
-# for latent_dim in lds_config.latent_dims
-#     for obs_dim in lds_config.obs_dims
-#         obs_dim < latent_dim && continue
-#         for seq_len in lds_config.seq_lengths
-#             println("\n→ Benchmarking LDS with latent_dim=$latent_dim, obs_dim=$obs_dim, seq_len=$seq_len")
+for latent_dim in lds_config.latent_dims
+    for obs_dim in lds_config.obs_dims
+        obs_dim < latent_dim && continue
+        for seq_len in lds_config.seq_lengths
+            println("\n→ Benchmarking LDS with latent_dim=$latent_dim, obs_dim=$obs_dim, seq_len=$seq_len")
 
-#             instance = LDSInstance(; latent_dim=latent_dim, obs_dim=obs_dim, num_trials=1, seq_length=seq_len)
-#             rng = StableRNG(1234)
-#             params = init_params(rng, instance)
+            instance = LDSInstance(; latent_dim=latent_dim, obs_dim=obs_dim, num_trials=1, seq_length=seq_len)
+            rng = StableRNG(1234)
+            params = init_params(rng, instance)
 
-#             ssd_model = build_model(SSD_LDSImplem(), instance, params)
-#             _, y = build_data(rng, ssd_model, instance)
+            ssd_model = build_model(SSD_LDSImplem(), instance, params)
+            _, y = build_data(rng, ssd_model, instance)
 
-#             results_row = Dict{String, Any}()
-#             results_row["config"] = (latent_dim=latent_dim, obs_dim=obs_dim, seq_len=seq_len)
+            results_row = Dict{String, Any}()
+            results_row["config"] = (latent_dim=latent_dim, obs_dim=obs_dim, seq_len=seq_len)
 
-#             for impl in lds_implementations
-#                 print("  Running $(string(impl))... ")
-#                 try
-#                     model = build_model(impl, instance, params)
-#                     result = run_benchmark(impl, model, y)
-#                     results_row[string(impl)] = result
+            for impl in lds_implementations
+                print("  Running $(string(impl))... ")
+                try
+                    model = build_model(impl, instance, params)
+                    result = run_benchmark(impl, model, y)
+                    results_row[string(impl)] = result
 
-#                     if result.success
-#                         @printf("✓ time = %.3f sec\n", result.time / 1e9)
-#                     else
-#                         println("✗ failed")
-#                     end
-#                 catch e
-#                     results_row[string(impl)] = (time=NaN, memory=0, allocs=0, success=false)
[…]
