### 2025-08-21 aa18275
Integration of filter smooth
