### 2025-08-21 f7da0fc
updates to paper and add plotting.jl etc.
