+++ docs/examples/Gaussian_GLM_HMM.jl
@@ -1,9 +1,17 @@
 # ## Simulating and Fitting a Hidden Markov Model
 
 # This tutorial demonstrates how to use `StateSpaceDynamics.jl` to create, sample from, and fit Hidden 
-# Markov Models (HMMs).
+# Markov Models (HMMs). Unlike Linear Dynamical Systems which have continuous latent states, HMMs have
+# discrete latent states that switch between a finite number of modes. This makes them ideal for modeling
+# data with distinct behavioral regimes, switching dynamics, or categorical latent structure.
 
-# ## Load Packages
+# We'll focus on a Gaussian generalized linear model HMM (GLM-HMM), where each hidden state corresponds
+# to a different regression relationship between inputs and outputs. This is particularly useful for
+# modeling data where the input-output relationship changes over time in discrete jumps.
+
+# ## Load Required Packages
+
+# We load the essential packages for HMM modeling, visualization, and reproducible random sampling.
 
 using LinearAlgebra
 using Plots
@@ -11,28 +19,85 @@ using Random
 using StateSpaceDynamics
 using StableRNGs
 
-#
+# Set up reproducible random number generation
 rng = StableRNG(1234);
 
-# ## Create a Gaussian generalized linear model-hidden Markov model (GLM-HMM)
+# ## Create a Gaussian Generalized Linear Model-Hidden Markov Model (GLM-HMM)
 
-emission_1 = GaussianRegressionEmission(input_dim=3, output_dim=1, include_intercept=true, β=reshape([3.0, 2.0, 2.0, 3.0], :, 1), Σ=[1.0;;], λ=0.0)
-emission_2 = GaussianRegressionEmission(input_dim=3, output_dim=1, include_intercept=true, β=reshape([-4.0, -2.0, 3.0, 2.0], :, 1), Σ=[1.0;;], λ=0.0)
+# In a GLM-HMM, each hidden state defines a different regression model. The system switches
+# between these regression models according to Markovian dynamics. This is useful for modeling
+# scenarios where the relationship between predictors and outcomes changes over time.
 
-A = [0.99 0.01; 0.05 0.95]
-πₖ = [0.8; 0.2]
+# Define emission models for each hidden state
+# State 1: Positive relationship between input and output
+emission_1 = GaussianRegressionEmission(
+    input_dim=3,                                    # Number of input features
+    output_dim=1,                                   # Number of output dimensions  
+    include_intercept=true,                         # Include intercept term
+    β=reshape([3.0, 2.0, 2.0, 3.0], :, 1),        # Regression coefficients [intercept, β1, β2, β3]
+    Σ=[1.0;;],                                     # Observation noise variance
+    λ=0.0                                          # Regularization parameter
+)
 
-true_model = HiddenMarkovModel(K=2, A=A, πₖ=πₖ, B=[emission_1, emission_2])
+# State 2: Different relationship (negative intercept, different slopes)
+emission_2 = GaussianRegressionEmission(
+    input_dim=3,
+    output_dim=1,
+    include_intercept=true,
+    β=reshape([-4.0, -2.0, 3.0, 2.0], :, 1),      # Different regression coefficients
+    Σ=[1.0;;],                                     # Same noise level
+    λ=0.0
+)
+
+# Define the state transition matrix A
+# A[i,j] = probability of transitioning from state i to state j
[…]
