+++ docs/examples/PoissonLDS.jl
@@ -2,8 +2,16 @@
 
 # This tutorial demonstrates how to use `StateSpaceDynamics.jl` to simulate and fit a
 # Linear Dynamical System (LDS) with Poisson observations using the Laplace-EM algorithm.
+# Unlike the standard Gaussian LDS, this model is designed for count data (e.g., neural
+# spike counts, customer arrivals, or any discrete event data) where observations are
+# non-negative integers that follow Poisson distributions.
 
-# ## Load Packages
+# ## Load Required Packages
+
+# We begin by loading the necessary packages. The core difference from the Gaussian
+# case is that we'll be working with Poisson observation models, which require more
+# sophisticated inference algorithms (Laplace approximations) due to the non-conjugate
+# nature of Poisson likelihoods with Gaussian latent states.
 
 using StateSpaceDynamics
 using LinearAlgebra
@@ -12,56 +20,90 @@ using Plots
 using LaTeXStrings
 using StableRNGs
 
-#
+# Set up reproducible random number generation
 rng = StableRNG(123);
 
 # ## Create a Poisson Linear Dynamical System
 
-obs_dim = 10
-latent_dim = 2
+# We define a system where continuous latent dynamics generate discrete count observations.
+# This is particularly relevant in neuroscience (modeling neural spike trains) and other
+# domains where we observe discrete events generated by underlying continuous processes.
 
-A = 0.95 * [cos(0.25) -sin(0.25); sin(0.25) cos(0.25)]
-Q = Matrix(0.1 * I(latent_dim))
-x0 = zeros(latent_dim)
-P0 = Matrix(0.1 * I(latent_dim))
+obs_dim = 10      # Number of observed count variables (e.g., neurons)
+latent_dim = 2    # Number of latent state dimensions
 
-log_d = log.(fill(0.1, obs_dim))
+# Define the latent dynamics: same spiral structure as the Gaussian case
+# The latent states evolve smoothly and continuously according to linear dynamics
+A = 0.95 * [cos(0.25) -sin(0.25); sin(0.25) cos(0.25)]  # Rotation with contraction
+Q = Matrix(0.1 * I(latent_dim))     # Process noise covariance
+x0 = zeros(latent_dim)              # Initial state mean
+P0 = Matrix(0.1 * I(latent_dim))    # Initial state covariance
+
+# Poisson observation model parameters:
+# For Poisson observations, the rate parameter λ is typically modeled as:
+# log(λ_i) = C_i^T * x_t + log_d_i
+# where C maps latent states to log-rates and log_d provides baseline log-rates
+
+log_d = log.(fill(0.1, obs_dim))    # Log baseline rates (small positive rates)
+
+# Observation matrix C: maps 2D latent states to log-rates for each observed dimension
+# We use positive values to ensure that latent activity increases firing rates
 C = permutedims([abs.(randn(rng, obs_dim))'; abs.(randn(rng, obs_dim))'])
 
-state_model = GaussianStateModel(; A, Q, x0, P0)
-obs_model = PoissonObservationModel(; C, log_d)
+# Construct the model components
+state_model = GaussianStateModel(; A, Q, x0, P0)          # Gaussian latent dynamics
+obs_model = PoissonObservationModel(; C, log_d)           # Poisson observations
 
+# Create the complete Poisson Linear Dynamical System
 true_plds = LinearDynamicalSystem(;
     state_model=state_model,
     obs_model=obs_model,
     latent_dim=latent_dim,
     obs_dim=obs_dim,
-    fit_bool=fill(true, 6)
[…]
