### 2025-08-22 50de1fe
Add benchmarking README.md
