### 2025-08-26 3d7b615
Implemented more memory efficient FB (TK^2 -> K^2 scaling)
