+++ src/HiddenMarkovModels.jl
@@ -32,7 +32,7 @@ function initialize_forward_backward(model::AbstractHMM, num_obs::Int, ::Type{T}
         zeros(T, num_states, num_obs),
         zeros(T, num_states, num_obs),
         zeros(T, num_states, num_obs),
-        zeros(T, num_states, num_states, num_obs - 1)
+        zeros(T, num_states, num_states)
     )
 end
 
@@ -52,9 +52,9 @@ function ForwardBackward(
     α::Matrix{T},
     β::Matrix{T},
     γ::Matrix{T},
-    ξ::Array{T,3}
+    ξ::Matrix{T}
 ) where {T<:Real}
-    ForwardBackward{T, Vector{T}, Matrix{T}, Array{T,3}}(loglikelihoods, α, β, γ, ξ)
+    ForwardBackward{T, Vector{T}, Matrix{T}, Matrix{T}}(loglikelihoods, α, β, γ, ξ)
 end
 
 """
@@ -71,9 +71,22 @@ function aggregate_forward_backward!(
     aggregated_FB.α .= hcat([fb.α for fb in FB_storages]...)
     aggregated_FB.β .= hcat([fb.β for fb in FB_storages]...)
     aggregated_FB.γ .= hcat([fb.γ for fb in FB_storages]...)
-    aggregated_FB.ξ = cat([fb.ξ for fb in FB_storages]..., dims=3)
+    aggregated_FB.ξ = hcat([fb.ξ for fb in FB_storages]...)
     
 end
+
+# function aggregate_forward_backward!(
+#     aggregated_FB::ForwardBackward, 
+#     FB_storages::Vector{<:ForwardBackward}
+# )
+#     # Concatenate each field into the respective field in the aggregated struct
+#     aggregated_FB.loglikelihoods .= hcat([fb.loglikelihoods for fb in FB_storages]...)
+#     aggregated_FB.α .= hcat([fb.α for fb in FB_storages]...)
+#     aggregated_FB.β .= hcat([fb.β for fb in FB_storages]...)
+#     aggregated_FB.γ .= hcat([fb.γ for fb in FB_storages]...)
+#     aggregated_FB.ξ = cat([fb.ξ for fb in FB_storages]..., dims=3)
+    
+# end
 
 """
     initialize_transition_matrix(K::Int)
@@ -340,6 +353,7 @@ end
 
 Calculate the joint posterior distribution for state transitions.
 """
+
 function calculate_ξ!(
     model::AbstractHMM,
     FB_storage::ForwardBackward
@@ -347,32 +361,73 @@ function calculate_ξ!(
     α = FB_storage.α
     β = FB_storage.β
     loglikelihoods = FB_storage.loglikelihoods
-    ξ = FB_storage.ξ
     A = model.A
-    log_A = log.(A)  # Precompute log transition probabilities
+    log_A = log.(A)
     K = model.K
     time_steps = size(α, 2)
 
-    # Preallocate reusable arrays
+    # Initialize ξ sum in log-space as -Inf
+    FB_storage.ξ .= -Inf
+    log_ξ_sum = FB_storage.ξ
+
+    # Reusable array
     log_ξ_unnormalized = zeros(K, K)
 
     for t in 1:(time_steps - 1)
-        @views begin
-            for i in 1:K
-                α_t = α[i, t]  # scalar, reuse as-is
-                for j in 1:K
-                    log_ξ_unnormalized[i, j] = α_t + log_A[i, j] + loglikelihoods[j, t + 1] + β[j, t + 1]
-                end
+        # Compute unnormalized log-ξ for this time step
+        @inbounds @views for i in 1:K
+            α_t = α[i, t]
+            for j in 1:K
[…]
