+++ benchmarking/README.md
@@ -7,7 +7,7 @@ To set up the benchmarking environment (assuming you have a cloned/forked versio
 1. `cd` into the `benchmarking` directory.
 2. Start the Julia REPL.
 3. Type `]` to open the Pkg manager then type `activate .` to activate the benchmarking environment.
-4. Type `instantiate` to install all necessary packages. 
+4. Type `instantiate` to install all necessary packages.
 5. We also recommend that you type `dev your/local/path/to/StateSpaceDynamics.jl` so that you ensure your local version is being benchmarked if developing the package.
 
 ## Running the code
