### 2025-08-26 c96fd02
docstrings updated, logsumexp ->  logaddexp where applicable due to deprecation of logsumexp
