### 2025-08-26 ad141f8
hmm benchmark plot
