+++ benchmarking/plotting.jl
@@ -4,6 +4,7 @@ using CSV
 using DataFrames
 using Plots
 using Statistics: median, quantile
+using Measures
 
 function plot_lds_benchmark(fp::AbstractString, sp::AbstractString="benchmarking/results/lds_benchmark_results.svg")
     df = CSV.read(fp, DataFrame)
@@ -95,4 +96,83 @@ function plot_lds_benchmark(fp::AbstractString, sp::AbstractString="benchmarking
     return final_plt
 end
 
-plot_lds_benchmark("benchmarking/results/lds_benchmark_results.csv")
+using CSV, DataFrames, Plots, Statistics, Measures
+
+function plot_hmm_benchmark(fp::AbstractString, sp::AbstractString="benchmarking/results/hmm_benchmark_results.png")
+    df = CSV.read(fp, DataFrame)
+    df = filter(row -> !isnan(row.time_sec), df)
+
+    gb = groupby(df, [:implementation, :num_states, :seq_len])
+    agg = combine(gb, :time_sec => median => :median_time)
+
+    state_vals = sort(unique(agg.num_states))
+    impls = sort(unique(agg.implementation))
+
+    x_min, x_max = extrema(agg.seq_len)
+    y_min, y_max = extrema(agg.median_time)
+    x_ticks = sort(unique(agg.seq_len))
+
+    subplots = Plots.Plot[]
+
+    # create per-column panels with NO x-label (we will add a single global xlabel)
+    for (j, ns) in enumerate(state_vals)
+        panel = @view agg[agg.num_states .== ns, :]
+        if isempty(panel)
+            push!(subplots, plot(framestyle=:none, showaxis=false))
+            continue
+        end
+
+        p = plot(
+            xlabel = "",                  # WARNING: keep empty so there is no per-panel xlabel
+            ylabel = (j == 1) ? "Time (s)" : "",   # optional local ylabel only on first column
+            title = "States = $ns",
+            framestyle = :box,
+            grid = :y,
+            minorgrid = true,
+            legend = false,
+            xlims = (x_min * 0.9, x_max * 1.1),
+            ylims = (y_min * 0.9, y_max * 1.1),
+            xticks = x_ticks
+        )
+
+        for impl in impls
+            impl_data = @view panel[panel.implementation .== impl, :]
+            if !isempty(impl_data)
+                plot!(p, impl_data.seq_len, impl_data.median_time, label = impl, marker = :circle, linewidth = 2)
+            end
+        end
+        push!(subplots, p)
+    end
+
+    # Compose the grid of subplots (no global labels here)
+    plt = plot(subplots...; layout = (1, length(state_vals)),
+               size = (350 * length(state_vals), 300),
+               left_margin = 18mm, bottom_margin = 18mm)
+
+    # Legend panel
+    legend_plot = plot(framestyle = :none, showaxis = false, size = (150, 300))
+    for impl in impls
+        plot!(legend_plot, [NaN], [NaN], label = impl, linewidth = 2, marker = :circle)
+    end
+
+    # Combine with legend and *set the global labels on the combined plot*
+    final_plt = plot(plt, legend_plot,
+                     layout = @layout([a{0.85w} b{0.15w}]),
+                     left_margin = 20mm, bottom_margin = 20mm,
+                     xlabel = "Sequence Length (T)",   # ONE global xlabel
+                     ylabel = "Time (s)")              # ONE global ylabel (optional)
+
+    # Tidy fonts, etc.
+    plot!(final_plt; tickfontsize = 11, guidefontsize = 13, legendfontsize = 12)
+    savefig(sp)
+    return final_plt
+end
+
+
+
+
+# Example usage:
[…]
