### 2025-08-27 14dfa73
sometimes hmm.jl fails a to complete a specific benchmark. pushing up to move to laptop
