### 2025-08-27 99056c7
update some docs and hmm stuff
