### 2025-08-27 508f36f
paper edits and README.md edits
