+++ paper/paper.md
@@ -52,7 +52,7 @@ By providing these features, `StateSpaceDynamics.jl` fills a critical gap in the
 
 # Benchmarks
 
-To evaluate the performance of `StateSpaceDynamics.jl`, we conducted two benchmarking studies focusing on fitting a Gaussian LDS and a Bernoulli GLM-HMM. For the Gaussian LDS benchmark, we compared our package against two alternatives: the NumPy-based Kalman filter-smoother package `pykalman` and the more recent JAX-based `Dynamax`. We intentionally excluded `StateSpaceModels.jl` from our comparison as its scope is geared towards structured time-series models. `Dynamax` was properly JIT-compiled by running a warm-up iteration before benchmarking to ensure fair comparison.
+To evaluate the performance of `StateSpaceDynamics.jl`, we conducted two benchmarking studies focusing on fitting a Gaussian LDS and a Gaussian HMM. For the Gaussian LDS benchmark, we compared our package against two alternatives: the NumPy-based Kalman filter-smoother package `pykalman` and the more recent JAX-based `Dynamax`. We intentionally excluded `StateSpaceModels.jl` from our comparison as its scope is geared towards structured time-series models. `Dynamax` was properly JIT-compiled using the `jax.jit` function prior to benchmarking to ensure fair comparison.
 
 For our Gaussian LDS experiments, we constructed a synthetic dataset as follows. The state transition matrix $A$ was generated as a random $n$-dimensional rotation matrix, while the observation matrix $C$ was created as a random $m \times n$ matrix. Both the state noise covariance $Q$ and observation noise covariance $R$ were set to identity matrices. To ensure a fair comparison, all packages were initialized using identical random parameters, after which we executed the EM algorithm for 100 iterations. We conducted these benchmarks using `PythonCall.jl` [@pythoncall] and `BenchmarkTools.jl` [@bench], with the assumption that Julia-to-Python overhead is negligible for these computationally intensive operations.
 
@@ -67,9 +67,9 @@ For the subsequent benchmarking experiments, we compared `StateSpaceDynamics.jl`
 \includegraphics[width=\textwidth]{glmhmm_bench.png}
 \end{figure}
 
-In our benchmarking, we find that for the LDS, both variants of `StateSpaceDynamics.jl` are faster than `pykalman` across all sequence lengths and dimension configurations. More generally, `StateSpaceDynamics.jl` outperforms `Dynamax` at shorter sequence lengths across various dimensional settings. However, `Dynamax` exhibits superior scaling. In our current implementation, the Hessian is represented as a sparse matrix with block tridiagonal structure, resulting in O($Tn^2$) memory scaling — which is optimal. However, we do not yet exploit this structure fully during inference. In particular, our solver does not leverage specialized routines for block-banded systems, which can result in unnecessary fill-in and degraded performance at large T. Future versions will use banded or block tridiagonal solvers to achieve truly linear-time inference.
[…]
