### 2025-08-27 aae9ef4
Update plotting for consistent labeling and consistent coloring
