+++ paper/paper.md
@@ -61,15 +61,14 @@ To thoroughly assess performance across different scales, we tested three sequen
 \includegraphics[width=\textwidth]{lds_bench.png}
 \end{figure}
 
-For the subsequent benchmarking experiments, we compared `StateSpaceDynamics.jl`, `HiddenMarkovModels.jl`, and `Dynamax` in their abilities to fit a Bernoulli GLM-HMM. We completed the benchmarking on the same combinations of sequence lengths, state dimensions, and observation dimensions as in the Gaussian LDS benchmarking. Synthetic datasets were generated for each of these combinations by sampling from a randomly generated Bernoulli GLM-HMM using `StateSpaceDynamics.jl`. All packages were initialized to the same random parameters and the EM algorithm was run for 200 iterations. We disabled the convergence criteria to ensure each package completed all 200 iterations of EM. These benchmarks were conducted using `PythonCall.jl` and `BenchmarkTools.jl` to ensure timing accuracy.
-
+For the second benchmarking study, we compared `StateSpaceDynamics.jl`, `HiddenMarkovModels.jl`, and `Dynamax` in their ability to fit a Gaussian HMM. Once again, we ensured that `Dynamax` was JIT-compiled for a fair comparison. To construct synthetic datasets, we sampled from a Gaussian HMM with randomly selected emission models, transition matrices, and initial state distributions. Each package was initialized using identical random parameters to maintain consistency. EM was run for 100 iterations.
 \begin{figure}
-\includegraphics[width=\textwidth]{glmhmm_bench.png}
+\includegraphics[width=\textwidth]{hmm_bench.png}
 \end{figure}
 
-In our benchmarking, we find that for the LDS, both `StateSpaceDynamics.jl` and `Dynamax` are faster than `pykalman` across all sequence lengths and dimension configurations. More generally, `StateSpaceDynamics.jl` and `Dynamax` exhibit similar performance at lower sequence lengths (with `Dynamax` slighlty outperforming `StateSpaceDynamics.jl`). However, `Dynamax` exhibits superior scaling in both the dimensions of the state and observation matrices as well as the temporal sequnce length. In our current implementation, the Hessian is represented as a sparse matrix with block tridiagonal structure, resulting in O($Tn^2$) memory scaling — which is optimal. However, we do not yet exploit this structure fully during inference. In particular, our solver does not leverage specialized routines for block-banded systems (e.g., the block Thomas algorithm), which can result in unnecessary fill-in and degraded performance at large T. Future versions will use banded or block tridiagonal solvers to achieve truly linear-time inference.
+In our benchmarking, we find that for the LDS, both `StateSpaceDynamics.jl` and `Dynamax` are faster than `pykalman` across all sequence lengths and dimension configurations. More generally, `StateSpaceDynamics.jl` and `Dynamax` exhibit similar performance at lower sequence lengths (with `Dynamax` slightly outperforming `StateSpaceDynamics.jl`). However, `Dynamax` exhibits superior scaling in both the dimensions of the state and observation matrices as well as the temporal sequnce length. In our current implementation, the Hessian is represented as a sparse matrix with block tridiagonal structure, resulting in O($Tn^2$) memory scaling — which is optimal. However, we do not yet exploit this structure fully during inference. In particular, our solver does not leverage specialized routines for block-banded systems (e.g., the block Thomas algorithm), which can result in unnecessary fill-in and degraded performance at large T. Future versions will use banded or block tridiagonal solvers to achieve truly linear-time inference.
 
[…]
