### 2025-08-27 0b7b799
Add Gaussian bias terms
