+++ src/LinearDynamicalSystems.jl
@@ -10,14 +10,16 @@ Represents the state model of a Linear Dynamical System with Gaussian noise.
 # Fields
 - `A::M`: Transition matrix (size `latent_dim×latent_dim`).
 - `Q::M`: Process noise covariance matrix
+- `b::V`: Bias vector (length `latent_dim`).
 - `x0::V`: Initial state vector (length `latent_dim`).
-- `P0::M`: Initial state covariance matrix (size `latent_dim×latent_dim
+- `P0::M`: Initial state covariance matrix (size `latent_dim×latent_dim`).
 """
 Base.@kwdef mutable struct GaussianStateModel{
     T<:Real,M<:AbstractMatrix{T},V<:AbstractVector{T}
 } <: AbstractStateModel{T}
     A::M
     Q::M
+    b::V
     x0::V
     P0::M
 end
@@ -54,11 +56,13 @@ Represents the observation model of a Linear Dynamical System with Gaussian nois
 - `C::M`: Observation matrix of size `(obs_dim × latent_dim)`. Maps latent states into
     observation space.
 - `R::M`: Observation noise covariance of size `(obs_dim × obs_dim)`.
+- `d::V`: Bias vector of length `(obs_dim)`.
 """
-Base.@kwdef mutable struct GaussianObservationModel{T<:Real,M<:AbstractMatrix{T}} <:
+Base.@kwdef mutable struct GaussianObservationModel{T<:Real,M<:AbstractMatrix{T},V<:AbstractVector{T}} <:
                            AbstractObservationModel{T}
     C::M
     R::M
+    d::V
 end
 
 function Base.show(io::IO, gom::GaussianObservationModel; gap="")
@@ -169,13 +173,13 @@ Extract the state parameters from a Linear Dynamical System.
 """
 function stateparams(
     lds::LinearDynamicalSystem{T,S,O}
-) where {T<:Real,S<:AbstractStateModel,O<:AbstractObservationModel}
+) where {T<:Real,S<:AbstractStateModel{T},O<:AbstractObservationModel{T}}
     if isa(lds.state_model, GaussianStateModel)
         return [
-            lds.state_model.A, lds.state_model.Q, lds.state_model.x0, lds.state_model.P0
+            lds.state_model.A, lds.state_model.Q, lds.state_model.b,
+            lds.state_model.x0, lds.state_model.P0
         ]
     end
-
     return nothing
 end
 
@@ -186,13 +190,12 @@ Extract the observation parameters from a Linear Dynamical System.
 """
 function obsparams(
     lds::LinearDynamicalSystem{T,S,O}
-) where {T<:Real,S<:AbstractStateModel,O<:AbstractObservationModel}
+) where {T<:Real,S<:AbstractStateModel{T},O<:AbstractObservationModel{T}}
     if isa(lds.obs_model, GaussianObservationModel)
-        return [lds.obs_model.C, lds.obs_model.R]
+        return [lds.obs_model.C, lds.obs_model.R, lds.obs_model.d]
     elseif isa(lds.obs_model, PoissonObservationModel)
         return [lds.obs_model.C, lds.obs_model.log_d]
     end
-
     return nothing
 end
 
@@ -219,20 +222,20 @@ function Random.rand(
     rng::AbstractRNG, lds::LinearDynamicalSystem{T,S,O}; tsteps::Int, ntrials::Int
[…]
