### 2025-08-27 fa4722d
paper edits
