+++ paper/paper.bib
@@ -244,3 +244,34 @@ state space models},
   date     = {2017},
   howpublished = {\url{https://github.com/JuliaCI/BenchmarkTools.jl}}
 }
+
+@article{ForwardDiff,
+    title = {Forward-Mode Automatic Differentiation in {J}ulia},
+   author = {{Revels}, J. and {Lubin}, M. and {Papamarkou}, T.},
+  journal = {arXiv:1607.07892 [cs.MS]},
+     year = {2016},
+      url = {https://arxiv.org/abs/1607.07892}
+}
+
+@article{Zygote,
+  author       = {Michael Innes},
+  title        = {Don't Unroll Adjoint: Differentiating SSA-Form Programs},
+  journal      = {CoRR},
+  volume       = {abs/1810.07951},
+  year         = {2018},
+  url          = {http://arxiv.org/abs/1810.07951},
+  eprinttype    = {arXiv},
+  eprint       = {1810.07951},
+  timestamp    = {Tue, 30 Oct 2018 20:39:56 +0100},
+  biburl       = {https://dblp.org/rec/journals/corr/abs-1810-07951.bib},
+  bibsource    = {dblp computer science bibliography, https://dblp.org}
+}
+
+@software{DiffInterface,
+      author={Dalle, Guillaume and Hill, Adrian},
+      title={Differentiation{I}nterface.jl},
+      year={2024},
+      publisher={Zenodo},
+      doi={10.5281/zenodo.11092033},
+      url={https://doi.org/10.5281/zenodo.11092033},
+}
