### 2025-09-08 #8077 @gdalle
@rsenne is the "dev docs" available at https://depasquale-lab.github.io/StateSpaceDynamics.jl/dev/ the most up-to-date version to review, or do we need to manually build it from the `_joss` branch of the repo?
