+++ paper/paper.md
@@ -36,7 +36,7 @@ State-space models (SSMs) are powerful tools for modeling time series data that 
 
 Advances in neuroscience have enabled the collection of massive, multivariate, and complex time-series datasets, where simultaneous observations from hundreds to thousands of neurons are increasingly common. Interpreting these high-dimensional datasets presents significant challenges. Recent modeling approaches suggest that neural activity can be characterized by a set of latent factors evolving within a low-dimensional manifold. Consequently, there is a growing need for models that combine dimensionality reduction with temporal dynamics, for which state-space models provide a natural framework.
 
-While state-space model implementations exist in Python, such as the SSM package [@PySSM2022] and Dynamax [@Linderman_Dynamax_A_Python_2025], the Julia programming language lacks an equivalent that meets the needs of modern neuroscience. Existing Julia offerings, like `StateSpaceModels.jl` [@SSMjl2024], can accommodate continuous-state SSMs (e.g., LDS) but are limited to Gaussian observation models and rely on analytical calculation of the marginal log-likelihood. This latter limitation precludes model inference and parameter learning for non-conjugate observations which are common in neuroscience, where neural activity follow Poisson or other discrete distributions. Packages for performing inference and learning using sampling-based methods exist in Julia (such as `Turing.jl` [@Turing2018]) but are computationally inefficient compared to tailored approaches based on Expectation-Maximization (EM). For discrete SSMs, an existing Julia offering, `HiddenMarkovModels.jl` [@HMMjl2024], is efficient and scalable but not intentionally designed with the functionality for mixing models that contain both discrete and continuous latent variables, such as the switching linear dynamical system model (SLDS) [@Linderman2016-xe; @slds] increasingly used in neuroscience. Although our primary motivation arises from challenges in modeling high-dimensional neural population activity, the package is not specific to neuroscience. The algorithms and abstractions apply equally well to time-series problems in engineering, econometrics, and other fields where latent variable models and structured inference are required.
+While state-space model implementations exist in Python, such as the SSM package [@PySSM2022] and Dynamax [@Linderman_Dynamax_A_Python_2025], the Julia programming language lacks an equivalent that meets the needs of modern neuroscience. Existing Julia offerings, like `StateSpaceModels.jl` [@SaavedraBodinSouto2019], can accommodate continuous-state SSMs (e.g., LDS) but are limited to Gaussian observation models and rely on analytical calculation of the marginal log-likelihood. This latter limitation precludes model inference and parameter learning for non-conjugate observations which are common in neuroscience, where neural activity follow Poisson or other discrete distributions. Packages for performing inference and learning using sampling-based methods exist in Julia (such as `Turing.jl` [@10.1145/3711897,@pmlr-v84-ge18b]) but are computationally inefficient compared to tailored approaches based on Expectation-Maximization (EM). For discrete SSMs, an existing Julia offering, `HiddenMarkovModels.jl` [@Dalle2024], is efficient and scalable but not intentionally designed with the functionality for mixing models that contain both discrete and continuous latent variables, such as the switching linear dynamical system model (SLDS) [@Linderman2016-xe; @slds] increasingly used in neuroscience. Although our primary motivation arises from challenges in modeling high-dimensional neural population activity, the package is not specific to neuroscience. The algorithms and abstractions apply equally well to time-series problems in engineering, econometrics, and other fields where latent variable models and structured inference are required.
 
 # Package design
 
[…]
