### 2025-09-09 29f54fd
new figures/fixes pdf
