### 2025-09-09 612e7b6
Small paper edits
