+++ docs/src/MixtureModels.md
@@ -13,6 +13,7 @@ f_{mix}(x; \Theta, \pi) = \sum_{k=1}^K \pi_k f_k(x; \theta_k)
 ```
 
 Where:
+
 - $f_k(x; \theta_k)$ is the $k$-th **component distribution** with parameters $\theta_k$
 - $\pi_k$ is the **mixing coefficient** (or mixing weight) for component $k$
 - $\Theta = \{\theta_1, \ldots, \theta_K\}$ are the component parameters
@@ -30,6 +31,7 @@ The generative process for a mixture model can be described as follows:
 ```
 
 Where:
+
 - $z_i$ is a **latent assignment variable** indicating which component generated observation $x_i$
 - $z_i = k$ means observation $x_i$ was generated by component $k$
 - $\text{Cat}(\pi)$ is the categorical distribution with probabilities $\pi$
@@ -44,7 +46,7 @@ This two-step process first selects a component according to the mixing weights,
 
 **Latent Structure:** The latent variables $z_i$ provide a natural clustering interpretation, where observations are probabilistically assigned to components.
 
-# Gaussian Mixture Model
+## Gaussian Mixture Model
 
 A **Gaussian Mixture Model (GMM)** is a mixture model where each component is a multivariate Gaussian distribution. GMMs are among the most widely used mixture models due to their mathematical tractability and flexibility in modeling continuous data.
 
@@ -68,18 +70,20 @@ Where each component is a multivariate Gaussian:
 ```
 
 The parameters for each component $k$ are:
+
 - $\mu_k \in \mathbb{R}^D$: the mean vector
 - $\Sigma_k \in \mathbb{R}^{D \times D}$: the covariance matrix (positive definite)
 
 ## Applications
 
 GMMs are particularly effective for:
+
 - **Density estimation** for continuous multimodal data
 - **Soft clustering** where observations can belong to multiple clusters with different probabilities
 - **Dimensionality reduction** when combined with factor analysis
 - **Anomaly detection** by identifying low-probability regions
 
-# Poisson Mixture Model
+## Poisson Mixture Model
 
 A **Poisson Mixture Model (PMM)** is designed for modeling count data that exhibits overdispersion or multimodality. Each component follows a Poisson distribution, making it suitable for discrete, non-negative integer observations.
 
@@ -88,7 +92,7 @@ PoissonMixtureModel
 PoissonMixtureModel(k::Int)
 ```
 
-## Mathematical Formulation
+### Mathematical Formulation
 
 For a PMM with $K$ components, the mixture density is:
 
@@ -103,17 +107,19 @@ Where each component is a Poisson distribution:
 ```
 
 The parameter for each component $k$ is:
+
 - $\lambda_k > 0$: the rate parameter (both mean and variance of the Poisson distribution)
 
-## Applications
+### Applications
 
 PMMs are commonly used for:
+
 - **Count data modeling** with heterogeneous populations
 - **Modeling overdispersed count data** where variance exceeds the mean
 - **Analyzing arrival processes** with multiple underlying rates
 - **Biological applications** such as modeling gene expression counts
 
-# Learning in Mixture Models
+## Learning in Mixture Models
 
 `StateSpaceDynamics.jl` implements the **Expectation-Maximization (EM) algorithm** for parameter estimation in mixture models. EM is an iterative algorithm that finds maximum likelihood estimates in the presence of latent variables.
 
[…]
