+++ docs/src/SLDS.md
@@ -7,6 +7,7 @@ CollapsedDocStrings = true
 A **Switching Linear Dynamical System (SLDS)** is a powerful probabilistic model that combines the temporal structure of linear dynamical systems with the discrete switching behavior of Hidden Markov Models. SLDS can model complex time series data that exhibits multiple dynamical regimes, where the system can switch between different linear dynamics over time.
 
 An SLDS extends the standard Linear Dynamical System (LDS) by introducing a discrete latent state that determines which linear dynamics are active at each time step. This makes SLDS particularly suitable for modeling systems with:
+
 - **Multiple operational modes** (e.g., different flight phases of an aircraft)
 - **Regime changes** (e.g., economic cycles, behavioral states)
 - **Non-stationary dynamics** where linear dynamics change over time
@@ -31,6 +32,7 @@ An SLDS with $K$ discrete states is defined by the following generative model:
 ```
 
 Where:
+
 - $s_t \in \{1, 2, \ldots, K\}$ is the **discrete switching state** at time $t$
 - $z_t \in \mathbb{R}^D$ is the **continuous latent state** at time $t$  
 - $y_t \in \mathbb{R}^P$ is the **observed data** at time $t$
@@ -55,10 +57,11 @@ end
 ```
 
 Each mode in the `B` vector contains its own `LinearDynamicalSystem` with:
+
 - **State model**: Defines the continuous latent dynamics $F_k$, $Q_k$
 - **Observation model**: Defines the emission process $C_k$, $R_k$
 
-# Sampling from SLDS
+## Sampling from SLDS
 
 You can generate synthetic data from an SLDS to test algorithms or create simulated datasets:
 
@@ -68,13 +71,14 @@ Random.rand(slds::SwitchingLinearDynamicalSystem, T::Int)
 ```
 
 The sampling process follows the generative model:
+
 1. **Initialize**: Sample initial discrete state from $\pi_k$ and initial continuous state
 2. **For each time step**:
    - Sample next discrete state based on current state and transition matrix $A$
    - Sample continuous state using the dynamics of the current discrete state
    - Generate observation using the observation model of the current discrete state
 
-# Learning in SLDS
+## Learning in SLDS
 
 `StateSpaceDynamics.jl` implements a **Variational Expectation-Maximization (EM)** algorithm for parameter estimation in SLDS. This approach handles the interaction between discrete and continuous latent variables efficiently.
 
@@ -88,22 +92,18 @@ The variational EM algorithm maximizes the **Evidence Lower Bound (ELBO)** inste
 
 ### Variational Expectation Step
 
-The E-step iteratively updates the variational distributions until convergence:
-
-```@docs
-variational_expectation!(model::SwitchingLinearDynamicalSystem, y, FB, FS)
-```
-
-This involves two coupled updates:
+The E-step iteratively updates the variational distributions until convergence. This involves two coupled updates:
 
 **1. Update continuous state posteriors ($q(z_{1:T})$):**
 For each discrete state $k$, run weighted Kalman smoothing:
+
 ```math
 q(z_{1:T} \mid s_{1:T} = k) = \prod_{t=1}^T \mathcal{N}(z_t; \hat{z}_{t|T}^{(k)}, P_{t|T}^{(k)})
 ```
 
 **2. Update discrete state posteriors ($q(s_{1:T})$):**
[…]
