+++ docs/examples/Gaussian_GLM_HMM.jl
@@ -38,7 +38,7 @@ emission_1 = GaussianRegressionEmission(
     β=reshape([3.0, 2.0, 2.0, 3.0], :, 1),        # Regression coefficients [intercept, β1, β2, β3]
     Σ=[1.0;;],                                     # Observation noise variance
     λ=0.0                                          # Regularization parameter
-)
+);
 
 # State 2: Different relationship (negative intercept, different slopes)
 emission_2 = GaussianRegressionEmission(
@@ -48,16 +48,16 @@ emission_2 = GaussianRegressionEmission(
     β=reshape([-4.0, -2.0, 3.0, 2.0], :, 1),      # Different regression coefficients
     Σ=[1.0;;],                                     # Same noise level
     λ=0.0
-)
+);
 
 # Define the state transition matrix A
 # A[i,j] = probability of transitioning from state i to state j
 # Diagonal elements are high (states are persistent), off-diagonal elements are low
 A = [0.99 0.01;    # From state 1: 99% stay, 1% switch to state 2
-     0.05 0.95]    # From state 2: 5% switch to state 1, 95% stay
+     0.05 0.95];    # From state 2: 5% switch to state 1, 95% stay
 
 # Initial state distribution: probability of starting in each state
-πₖ = [0.8; 0.2]    # 80% chance of starting in state 1, 20% in state 2
+πₖ = [0.8; 0.2];    # 80% chance of starting in state 1, 20% in state 2
 
 # Construct the complete HMM
 true_model = HiddenMarkovModel(
@@ -65,7 +65,7 @@ true_model = HiddenMarkovModel(
     A=A,                        # Transition matrix
     πₖ=πₖ,                     # Initial state distribution
     B=[emission_1, emission_2]  # Emission models for each state
-)
+);
 
 println("Created GLM-HMM with 2 states and 3 input features")
 println("State 1 regression: y = 3.0 + 2.0*x₁ + 2.0*x₂ + 3.0*x₃ + ε")
@@ -81,10 +81,10 @@ n = 20000  # Number of time points
 println("Generating $n samples from the GLM-HMM...")
 
 # Generate random input features (predictors)
-Φ = randn(rng, 3, n)  # 3 features × n time points
+Φ = randn(rng, 3, n);  # 3 features × n time points
 
 # Sample from the HMM: returns both hidden states and observations
-true_labels, data = rand(rng, true_model, Φ, n=n)
+true_labels, data = rand(rng, true_model, Φ, n=n);
 
 println("Generated data summary:")
 println("  - Input features shape: $(size(Φ))")
@@ -98,7 +98,7 @@ println("  - State 2 proportion: $(round(mean(true_labels .== 2), digits=3))")
 # Create a scatter plot showing how the input-output relationship differs between
 # the two hidden states. Points are colored by their true hidden state.
 
-colors = [:dodgerblue, :crimson]  # Blue for state 1, red for state 2
+colors = [:dodgerblue, :crimson];  # Blue for state 1, red for state 2
 
 scatter(Φ[1, :], vec(data);
     color = colors[true_labels],
@@ -108,12 +108,12 @@ scatter(Φ[1, :], vec(data);
     ylabel = "Output",
     title = "GLM-HMM Sampled Data (colored by true state)",
     alpha = 0.6
-)
+);
 
 # Overlay the true regression lines for each state
 # We'll plot the relationship between feature 1 and output, holding other features at 0
 
[…]
