+++ docs/examples/GaussianLDS.jl
@@ -29,32 +29,32 @@ rng = StableRNG(123);
 # generates 10-dimensional observations.
 
 obs_dim = 10      # Number of observed variables at each time step
-latent_dim = 2    # Number of latent state variables
+latent_dim = 2;   # Number of latent state variables
 
 # Define the state transition matrix A. This matrix governs how the latent state
 # evolves from one time step to the next: x_{t+1} = A * x_t + noise.
 # We create a rotation matrix scaled by 0.95, which creates a stable spiral
 # dynamic that slowly contracts toward the origin.
-A = 0.95 * [cos(0.25) -sin(0.25); sin(0.25) cos(0.25)]
+A = 0.95 * [cos(0.25) -sin(0.25); sin(0.25) cos(0.25)];
 
 # Process noise covariance Q controls how much random variation we add to the
 # latent state transitions. A smaller Q means more predictable dynamics.
-Q = Matrix(0.1 * I(2))
+Q = Matrix(0.1 * I(2));
 
 # Initial state parameters: where the latent trajectory starts and how uncertain
 # we are about this initial position.
 x0 = [0.0; 0.0]           # Mean of initial state
-P0 = Matrix(0.1 * I(2))   # Covariance of initial state
+P0 = Matrix(0.1 * I(2));  # Covariance of initial state
 
 # Observation parameters: how the latent states map to observed data.
 # C is the observation matrix (latent-to-observed mapping), and R is the
 # observation noise covariance.
 C = randn(rng, obs_dim, latent_dim)  # Random linear mapping from 2D latent to 10D observed
-R = Matrix(0.5 * I(obs_dim))         # Independent noise on each observation dimension
+R = Matrix(0.5 * I(obs_dim));         # Independent noise on each observation dimension
 
 # Construct the state and observation model components
 true_gaussian_sm = GaussianStateModel(;A=A, Q=Q, x0=x0, P0=P0)
-true_gaussian_om = GaussianObservationModel(;C=C, R=R)
+true_gaussian_om = GaussianObservationModel(;C=C, R=R);
 
 # Combine them into a complete Linear Dynamical System
 # The fit_bool parameter indicates which parameters should be learned during fitting
@@ -64,7 +64,7 @@ true_lds = LinearDynamicalSystem(;
     latent_dim=latent_dim,
     obs_dim=obs_dim,
     fit_bool=fill(true, 6)  # Fit all 6 parameter matrices: A, Q, C, R, x0, P0
-)
+);
 
 # ## Simulate Latent and Observed Data
 
@@ -72,10 +72,10 @@ true_lds = LinearDynamicalSystem(;
 # latent states (which we'll later try to recover) and the observations (which
 # is all a real algorithm would see).
 
-tSteps = 500  # Number of time points to simulate
+tSteps = 500;  # Number of time points to simulate
 
 # The rand function generates both latent trajectories and corresponding observations
-latents, observations = rand(rng, true_lds; tsteps=tSteps, ntrials=1)
+latents, observations = rand(rng, true_lds; tsteps=tSteps, ntrials=1);
 
 # ## Plot Vector Field of Latent Dynamics
 
@@ -86,12 +86,12 @@ latents, observations = rand(rng, true_lds; tsteps=tSteps, ntrials=1)
 # Create a grid of starting points
 x = y = -3:0.5:3
 X = repeat(x', length(y), 1)
-Y = repeat(y, 1, length(x))
+Y = repeat(y, 1, length(x));
 
[…]

[…]
