+++ docs/examples/PoissonLDS.jl
@@ -30,29 +30,29 @@ rng = StableRNG(123);
 # domains where we observe discrete events generated by underlying continuous processes.
 
 obs_dim = 10      # Number of observed count variables (e.g., neurons)
-latent_dim = 2    # Number of latent state dimensions
+latent_dim = 2;    # Number of latent state dimensions
 
 # Define the latent dynamics: same spiral structure as the Gaussian case
 # The latent states evolve smoothly and continuously according to linear dynamics
 A = 0.95 * [cos(0.25) -sin(0.25); sin(0.25) cos(0.25)]  # Rotation with contraction
 Q = Matrix(0.1 * I(latent_dim))     # Process noise covariance
 x0 = zeros(latent_dim)              # Initial state mean
-P0 = Matrix(0.1 * I(latent_dim))    # Initial state covariance
+P0 = Matrix(0.1 * I(latent_dim));    # Initial state covariance
 
 # Poisson observation model parameters:
 # For Poisson observations, the rate parameter λ is typically modeled as:
 # log(λ_i) = C_i^T * x_t + log_d_i
 # where C maps latent states to log-rates and log_d provides baseline log-rates
 
-log_d = log.(fill(0.1, obs_dim))    # Log baseline rates (small positive rates)
+log_d = log.(fill(0.1, obs_dim));    # Log baseline rates (small positive rates)
 
 # Observation matrix C: maps 2D latent states to log-rates for each observed dimension
 # We use positive values to ensure that latent activity increases firing rates
-C = permutedims([abs.(randn(rng, obs_dim))'; abs.(randn(rng, obs_dim))'])
+C = permutedims([abs.(randn(rng, obs_dim))'; abs.(randn(rng, obs_dim))']);
 
 # Construct the model components
 state_model = GaussianStateModel(; A, Q, x0, P0)          # Gaussian latent dynamics
-obs_model = PoissonObservationModel(; C, log_d)           # Poisson observations
+obs_model = PoissonObservationModel(; C, log_d);           # Poisson observations
 
 # Create the complete Poisson Linear Dynamical System
 true_plds = LinearDynamicalSystem(;
@@ -61,7 +61,7 @@ true_plds = LinearDynamicalSystem(;
     latent_dim=latent_dim,
     obs_dim=obs_dim,
     fit_bool=fill(true, 6)  # Learn all parameters: A, Q, C, log_d, x0, P0
-)
+);
 
 # ## Simulate Latent States and Observations
 
@@ -69,11 +69,11 @@ true_plds = LinearDynamicalSystem(;
 # to the linear dynamics, while observations are drawn from Poisson distributions
 # whose rates depend on the current latent state.
 
-tSteps = 500
+tSteps = 500;
 println("Simulating $tSteps time steps of Poisson LDS data...")
 
 # Generate both latent trajectories and count observations
-latents, observations = rand(rng, true_plds; tsteps=tSteps, ntrials=1)
+latents, observations = rand(rng, true_plds; tsteps=tSteps, ntrials=1);
 
 println("Generated latent states with range: [$(minimum(latents)), $(maximum(latents))]")
 println("Generated count observations with range: [$(minimum(observations)), $(maximum(observations))]")
@@ -88,7 +88,7 @@ x = y = -3:0.5:3
 X = repeat(x', length(y), 1)
 Y = repeat(y, 1, length(x))
 U = zeros(size(X))  # Flow in x-direction
-V = zeros(size(Y))  # Flow in y-direction
+V = zeros(size(Y));  # Flow in y-direction
 
[…]
