+++ docs/examples/HMM.jl
@@ -1,17 +1,15 @@
-# ## Simulating and Fitting a Hidden Markov Model with Gaussian Emissions
+# # Simulating and Fitting a Hidden Markov Model with Gaussian Emissions
 
 # This tutorial demonstrates how to use `StateSpaceDynamics.jl` to create, sample from, and fit Hidden 
 # Markov Models (HMMs) with Gaussian emission distributions. This is the classical HMM formulation where
 # each hidden state generates observations from a different multivariate Gaussian distribution.
-# 
-# Unlike the GLM-HMM in the previous tutorial, this model doesn't have input features - each state
-# simply emits observations from its own characteristic Gaussian distribution. This makes it ideal
-# for clustering time series data, identifying behavioral regimes, or modeling switching dynamics
-# in systems where each state has a distinct statistical signature.
+
+# Unlike GLM-HMMs, this model doesn't use input features - each state simply emits observations 
+# from its own characteristic Gaussian distribution. This makes it ideal for clustering time series data, 
+# identifying behavioral regimes, or modeling switching dynamics where each state has a distinct 
+# statistical signature.
 
 # ## Load Required Packages
-
-# We load the essential packages for HMM modeling, visualization, and reproducible analysis.
 
 using LinearAlgebra
 using Plots
@@ -19,6 +17,7 @@ using Random
 using StateSpaceDynamics
 using StableRNGs
 using Statistics: mean, std
+using LaTeXStrings
 
 # Set up reproducible random number generation
 rng = StableRNG(1234);
@@ -26,28 +25,28 @@ rng = StableRNG(1234);
 # ## Create a Gaussian Emission HMM
 
 # We'll create an HMM with two hidden states, each emitting 2D Gaussian observations.
-# This creates a simple but illustrative model where the hidden states correspond
+# This creates a simple but illustrative model where hidden states correspond
 # to different regions in the observation space.
 
-output_dim = 2;  # Each observation is a 2D vector
+output_dim = 2  # Each observation is a 2D vector
 
-# Define the state transition dynamics
+# Define state transition dynamics: $A_{ij} = P(\text{state}_t = j \mid \text{state}_{t-1} = i)$
 # High diagonal values mean states are "sticky" (tend to persist)
-A = [0.99 0.01;    # From state 1: 99% stay in state 1, 1% switch to state 2
-     0.05 0.95];   # From state 2: 5% switch to state 1, 95% stay in state 2
+A = [0.99 0.01;    # From state 1: 99% stay, 1% switch to state 2
+     0.05 0.95];   # From state 2: 5% switch to state 1, 95% stay
 
-# Initial state probabilities (equal probability of starting in either state)
+# Initial state probabilities: $\pi_k = P(\text{state}_1 = k)$
 πₖ = [0.5; 0.5];
 
 # Define emission distributions for each hidden state
 # State 1: Centered at (-1, -1) with small variance (tight cluster)
-μ_1 = [-1.0, -1.0]                                          # Mean vector
-Σ_1 = 0.1 * Matrix{Float64}(I, output_dim, output_dim)      # Covariance matrix (0.1 * Identity)
+μ_1 = [-1.0, -1.0]
+Σ_1 = 0.1 * Matrix{Float64}(I, output_dim, output_dim)
 emission_1 = GaussianEmission(output_dim=output_dim, μ=μ_1, Σ=Σ_1);
 
[…]
