+++ docs/examples/PoissonMixtureModel.jl
@@ -1,35 +1,34 @@
 # # Simulating and Fitting a Poisson Mixture Model
-# 
-# This tutorial shows how to build and fit a **Poisson Mixture Model (PMM)**
-# with `StateSpaceDynamics.jl` using the Expectation–Maximization (EM) algorithm.
-# It starts from simulation, fits the model, and then walks through diagnostics,
-# interpretation, and common extensions.
-#
+
+# This tutorial demonstrates how to build and fit a **Poisson Mixture Model (PMM)**
+# with `StateSpaceDynamics.jl` using the Expectation-Maximization (EM) algorithm.
+# We'll cover simulation, fitting, diagnostics, interpretation, and practical considerations.
+
 # ## What is a Poisson Mixture Model?
-# A PMM assumes each observation \(x_i\in \{0,1,2,\dots\}\) is drawn from one of
-# \(k\) Poisson distributions, with means (rates) \(\lambda_1,\dots,\lambda_k\).
-# The component is a latent categorical variable \(z_i\in\{1,\dots,k\}\) with
-# mixing weights \(\pi_1,\dots,\pi_k\) (\(\sum_j \pi_j = 1\)). The generative process:
-#
-# 1. Draw \(z_i \sim \text{Categorical}(\pi)\)
-# 2. Given \(z_i=j\), draw \(x_i \sim \text{Poisson}(\lambda_j)\)
-#
-# PMMs are handy for **count data** that come from a few heterogeneous sub-populations
-# (e.g., spike counts from cells with different firing rates, click counts from multiple
-# user segments, etc.).
-#
-# ## EM in one paragraph
-# EM maximizes the marginal log-likelihood \(\log p(x\,|\,\pi,\lambda)\) by iterating:
-# - **E-step:** compute responsibilities \(\gamma_{ij} = p(z_i=j\,|\,x_i,\theta)\)
-# - **M-step:** update \(\pi_j\) and \(\lambda_j\) to maximize the expected complete-data log-likelihood.
-# For Poisson mixtures, the closed-form M-step is:
-# \[
-# \pi_j \leftarrow \frac{1}{n} \sum_i \gamma_{ij}, \qquad
-# \lambda_j \leftarrow \frac{\sum_i \gamma_{ij} x_i}{\sum_i \gamma_{ij}}.
-# \]
-# `fit!` in `StateSpaceDynamics.jl` performs these steps under the hood.
-#
-# ---
+
+# A PMM assumes each observation $x_i \in \{0,1,2,\ldots\}$ is drawn from one of
+# $k$ Poisson distributions with rates $\lambda_1,\ldots,\lambda_k$.
+# The component assignment is a latent categorical variable $z_i \in \{1,\ldots,k\}$ 
+# with mixing weights $\pi_1,\ldots,\pi_k$ where $\sum_j \pi_j = 1$.
+
+# **Generative process:**
+# 1. Draw $z_i \sim \text{Categorical}(\boldsymbol{\pi})$
+# 2. Given $z_i = j$, draw $x_i \sim \text{Poisson}(\lambda_j)$
+
+# PMMs are useful for **count data** from heterogeneous sub-populations
+# (e.g., spike counts from different neuron types, customer transaction counts 
+# from different segments, or event frequencies across different regimes).
+
+# ## EM Algorithm Overview
+
+# EM maximizes the marginal log-likelihood $\log p(\mathbf{x} | \boldsymbol{\pi}, \boldsymbol{\lambda})$ by iterating:
+# - **E-step:** Compute responsibilities $\gamma_{ij} = P(z_i = j | x_i, \boldsymbol{\theta})$
+# - **M-step:** Update parameters to maximize expected complete-data log-likelihood
+
+# For Poisson mixtures, the M-step has closed-form updates:
[…]

[…]
