+++ docs/examples/SLDS.jl
@@ -92,7 +92,6 @@ print("Number of mode switches: $(sum(diff(vec(z)) .!= 0))\n");
 p1 = plot(1:T, x[1, :], label=L"x_1", linewidth=1.5, color=:black)
 plot!(1:T, x[2, :], label=L"x_2", linewidth=1.5, color=:blue)
 
-# Add mode-specific background shading
 transition_points = [1; findall(diff(vec(z)) .!= 0) .+ 1; T + 1]
 for i in 1:(length(transition_points) - 1)
     start_idx = transition_points[i]
@@ -119,22 +118,14 @@ A_init ./= sum(A_init, dims=2)  # Ensure row-stochastic
 πₖ_init = rand(rng, K)
 πₖ_init ./= sum(πₖ_init)
 
-# Initialize dynamics with different frequencies
-Q_init = Matrix(0.001 * I(state_dim))
-P0_init = Matrix(0.001 * I(state_dim))
-C_init = randn(rng, obs_dim, state_dim)
-R_init = Matrix(0.1 * I(obs_dim))
-
-# Create mode-specific dynamics with different rotation frequencies
-dynamics_init = [
-    LinearDynamicalSystem(
-        GaussianStateModel(0.95 * [cos(f) -sin(f); sin(f) cos(f)], Q_init, x0, P0_init),
-        GaussianObservationModel(C_init, R_init),
-        state_dim, obs_dim, fill(true, 6)
-    ) for f in (0.7, 0.1)  # Different initial frequencies
-]
-
-learned_model = SwitchingLinearDynamicalSystem(A_init, dynamics_init, πₖ_init, K)
+# Use the initialize_slds function for proper initialization
+learned_model = initialize_slds(;
+    K=K, 
+    d=state_dim, 
+    p=obs_dim, 
+    self_bias=5.0, 
+    seed=456  # Different seed from true model
+)
 
 print("Running variational EM algorithm...")
 
@@ -176,19 +167,16 @@ end
 p3 = plot(size=(900, 400))
 offset = 2.5
 
-# True states
 plot!(1:T, x[1, :] .+ offset, label=L"x_1 \text{ (true)}", 
       linewidth=2, color=:black, alpha=0.8)
 plot!(1:T, x[2, :] .- offset, label=L"x_2 \text{ (true)}", 
       linewidth=2, color=:black, alpha=0.8)
 
-# Learned states
 plot!(1:T, latents_learned[1, :] .+ offset, label=L"x_1 \text{ (learned)}", 
       linewidth=1.5, color=:red, alpha=0.8)
 plot!(1:T, latents_learned[2, :] .- offset, label=L"x_2 \text{ (learned)}", 
       linewidth=1.5, color=:blue, alpha=0.8)
 
-# Reference lines
 hline!([offset, -offset], color=:gray, alpha=0.3, linestyle=:dash, label="")
 
 plot!(title="True vs. Learned Latent States",
@@ -250,31 +238,6 @@ end
 print("True vs. Learned HMM Transitions:\n")
 print("True A_hmm:\n$(round.(A_hmm, digits=3))\n")
 print("Learned A_hmm:\n$(round.(learned_model.A, digits=3))\n");
-
-# ## Practical Considerations
-
-print("\n=== Practical Guidelines ===\n")
-print("• Initialization: Use different frequencies/directions for mode-specific dynamics\n")
-print("• Stickiness: Increase diagonal elements in A_hmm for more persistent modes\n")
-print("• Identifiability: Consider sharing parameters (e.g., R) across modes to avoid degeneracies\n") 
-print("• Diagnostics: Monitor ELBO convergence and visualize mode responsibilities\n")
-print("• Model selection: Use information criteria or cross-validation to choose K\n")
[…]
