+++ docs/examples/PoissonLDS.jl
@@ -20,7 +20,7 @@ using LaTeXStrings
 using StableRNGs
 
 # Set up reproducible random number generation
-rng = StableRNG(1234);
+rng = StableRNG(54321);
 
 # ## Create a Poisson Linear Dynamical System
 
@@ -83,7 +83,7 @@ x = y = -3:0.5:3
 X = repeat(x', length(y), 1)
 Y = repeat(y, 1, length(x))
 U = zeros(size(X))  # Flow in x-direction
-V = zeros(size(Y))  # Flow in y-direction
+V = zeros(size(Y));  # Flow in y-direction
 
 for i in 1:size(X, 1), j in 1:size(X, 2)
     v = A * [X[i,j], Y[i,j]]
@@ -93,7 +93,7 @@ end
 
 magnitude = @. sqrt(U^2 + V^2)  # Normalize arrow lengths for cleaner visualization
 U_norm = U ./ magnitude
-V_norm = V ./ magnitude
+V_norm = V ./ magnitude;
 
 # Plot vector field with simulated trajectory
 p1 = quiver(X, Y, quiver=(U_norm, V_norm), color=:blue, alpha=0.3,
@@ -124,8 +124,6 @@ plot!(subplot=1, yticks=(lim_states .* (0:latent_dim-1), [L"x_%$d" for d in 1:la
       xticks=[], xlims=(0, tSteps), title="Simulated Latent States",
       yformatter=y->"", tickfontsize=12)
 
-# Plot discrete observations as spike rasters
-# Each row represents one observed dimension, spikes shown as vertical lines
 colors = palette(:default, obs_dim)
 for f in 1:obs_dim
     spike_times = findall(x -> x > 0, emissions[f, :])
@@ -150,7 +148,7 @@ Q_init = Matrix(0.1 * I(latent_dim))              # Process noise guess
 C_init = randn(rng, obs_dim, latent_dim)          # Random observation mapping
 log_d_init = log.(fill(0.1, obs_dim))             # Baseline log-rate guess
 x0_init = zeros(latent_dim)                       # Start from origin
-P0_init = Matrix(0.1 * I(latent_dim))             # Initial uncertainty
+P0_init = Matrix(0.1 * I(latent_dim));             # Initial uncertainty
 
 # Construct naive model
 sm_init = GaussianStateModel(; A=A_init, Q=Q_init, x0=x0_init, P0=P0_init)
@@ -163,9 +161,6 @@ naive_plds = LinearDynamicalSystem(;
     obs_dim=obs_dim,
     fit_bool=fill(true, 6)
 );
-
-# Perform initial smoothing with random parameters
-print("Initial smoothing with random parameters...")
 
 # For Poisson observations, this requires Laplace approximations since
 # the posterior is no longer Gaussian (unlike linear-Gaussian case)
@@ -185,15 +180,6 @@ plot!(yticks=(lim_states .* (0:latent_dim-1), [L"x_%$d" for d in 1:latent_dim]),
       yformatter=y->"", tickfontsize=12, legend=:topright)
 
 # ## Fit Using Laplace-EM Algorithm
-
-# Use Laplace-EM to learn parameters. This is more complex than standard EM because:
-# 1. E-step requires Laplace approximations for non-Gaussian posteriors
-# 2. M-step updates account for Poisson likelihood structure  
-# 3. Convergence can be slower due to non-conjugate nature
-
-print("Starting Laplace-EM algorithm...")
-print("Note: Poisson LDS fitting is computationally intensive\n")
-
 # Fit the model - using fewer iterations due to computational cost
 elbo, _ = fit!(naive_plds, observations; max_iter=25, tol=1e-6);
 
@@ -225,42 +211,13 @@ p5 = plot(elbo, xlabel="Iteration", ylabel="ELBO",
[…]
