### 2025-09-10 4a80be6
Fix plds elbo bug
