### 2025-09-10 436c6d7
Clean up HMM and GLM-HMM examples for docs page
