+++ docs/examples/Gaussian_GLM_HMM.jl
@@ -25,9 +25,11 @@ rng = StableRNG(1234);
 
 # In a GLM-HMM, each hidden state defines a different regression model. The system switches
 # between these regression models according to Markovian dynamics. This is useful for modeling
-# scenarios where the relationship between predictors and outcomes changes over time.
+# scenarios where the relationship between predictors and outcomes changes over time. In this example,
+# we will demonstrate how to use `StateSpaceDynamics.jl` to create a GLM-HMM, generate synthetic data,
+# fit the model using the EM algorithm, and perform state inference with the Viterbi algorithm.
 
-# Define emission models for each hidden state
+# We will start by defining a simple GLM-HMM with two hidden states.
 # State 1: Positive relationship between input and output
 emission_1 = GaussianRegressionEmission(
     input_dim=3,                                    # Number of input features
@@ -50,14 +52,15 @@ emission_2 = GaussianRegressionEmission(
 
 # Define the state transition matrix $\mathbf{A}$:
 # $A_{ij} = P(\text{state}_t = j \mid \text{state}_{t-1} = i)$
-# Diagonal elements are high (states are persistent), off-diagonal elements are low
+
+# Diagonal elements are high (states are persistent)
 A = [0.99 0.01;    # From state 1: 99% stay, 1% switch to state 2
      0.05 0.95];    # From state 2: 5% switch to state 1, 95% stay
 
 # Initial state distribution: $\pi_k = P(\text{state}_1 = k)$
 πₖ = [0.8; 0.2];    # 80% chance of starting in state 1, 20% in state 2
 
-# Construct the complete HMM
+# Now that we have defined our emission models, we can construct the complete GLM-HMM.
 true_model = HiddenMarkovModel(
     K=2,                        # Number of hidden states
     A=A,                        # Transition matrix
@@ -71,23 +74,23 @@ print("State 2: y = -4.0 - 2.0x₁ + 3.0x₂ + 2.0x₃ + ε\n");
 
 # ## Sample from the GLM-HMM
 
-# Generate synthetic data from our true model. This will give us both the observed
-# data (inputs and outputs) and the true hidden state sequence.
+# Generate synthetic data from our true model. To do this, we must define the inputs to the model.
+# Then, sampling will yield the outputs nad the true hidden state sequence.
 
 n = 20000  # Number of time points
 
-# Generate random input features (predictors)
-Φ = randn(rng, 3, n);  # 3 features × n time points
+Φ = randn(rng, 3, n);  # Generate random input features (predictors)
 
-# Sample from the HMM: returns both hidden states and observations
-true_labels, data = rand(rng, true_model, Φ, n=n);
+true_labels, data = rand(rng, true_model, Φ, n=n);  # Sample from the HMM: returns both hidden states and observations
 
 print("Generated $(n) samples: State 1 ($(round(mean(true_labels .== 1)*100, digits=1))%), State 2 ($(round(mean(true_labels .== 2)*100, digits=1))%)\n");
 
 # ## Visualize the Sampled Dataset
 
-# Create a scatter plot showing how the input-output relationship differs between
-# the two hidden states. We'll plot feature 1 vs output, with points colored by true state.
+# Here, we create a scatter plot to show how the input-output relationship differs between
+# the two hidden states. We plot feature 1 vs output, with points colored by true state. State 1 (blue)
[…]
