+++ docs/examples/HMM.jl
@@ -64,17 +64,17 @@ print("State 2: μ = $μ_2, σ² = $(Σ_2[1,1]) (looser cluster)\n");
 # ## Sample from the HMM
 
 # Generate synthetic data from our true model. Each state generates observations 
-# from its own Gaussian distribution without requiring input features.
+# from its own Gaussian distribution without requiring input features. The rand function
+# samples both the hidden state sequence and the corresponding observations.
 
 num_samples = 10000;
-
-# Sample both hidden state sequence and corresponding observations
 true_labels, data = rand(rng, model, n=num_samples);
 
 # ## Visualize the Sampled Dataset
 
 # Create a 2D scatter plot showing observations colored by their true hidden state.
 # This illustrates how each state generates observations from a distinct region of space.
+# We will also plot a trajectory line to show the temporal evolution for the first 1000 timepoints.
 
 x_vals = data[1, 1:num_samples]
 y_vals = data[2, 1:num_samples]
@@ -93,7 +93,6 @@ for state in 1:2
         alpha=0.6)
 end
 
-# Add trajectory line showing temporal evolution (first 1000 points)
 plot!(x_vals[1:1000], y_vals[1:1000];
     color=:gray, lw=1, alpha=0.3, label="Trajectory")
 
@@ -104,14 +103,15 @@ scatter!([x_vals[end]], [y_vals[end]]; marker=:diamond, markersize=6,
 
 plot!(xlabel=L"x_1", ylabel=L"x_2", 
       title="HMM Emissions by True Hidden State",
-      legend=:topright)
+      legend=:topleft)
 
 # ## Initialize and Fit HMM with EM
 
-# Now simulate the realistic scenario: observe only data, not hidden states.
-# Initialize HMM with incorrect parameters and use EM to learn true parameters.
+# In reality, we only observe the data, not the hidden states. The goal of fitting
+# is to learn the latent state sequence and the model parameters that best explain the data.
+# We will initialize a new HMM with incorrect parameters and use the Expectation-Maximization (EM)
+# algorithm to iteratively refine the parameters and infer the hidden states.
 
-# Initialize with biased/incorrect parameters
 μ_1_init = [-0.25, -0.25]  # Closer to center than true
 Σ_1_init = 0.3 * Matrix{Float64}(I, output_dim, output_dim)  # Larger variance
 emission_1_init = GaussianEmission(output_dim=output_dim, μ=μ_1_init, Σ=Σ_1_init);
@@ -120,11 +120,9 @@ emission_1_init = GaussianEmission(output_dim=output_dim, μ=μ_1_init, Σ=Σ_1_
 Σ_2_init = 0.5 * Matrix{Float64}(I, output_dim, output_dim)  # Much larger variance
 emission_2_init = GaussianEmission(output_dim=output_dim, μ=μ_2_init, Σ=Σ_2_init);
 
-# Different transition matrix and initial distribution
 A_init = [0.8 0.2; 0.05 0.95]  # Less persistent than true model
 πₖ_init = [0.6, 0.4];           # Biased toward state 1
 
-# Create test model with naive initialization
 test_model = HiddenMarkovModel(K=2, B=[emission_1_init, emission_2_init], 
                               A=A_init, πₖ=πₖ_init);
 
@@ -145,15 +143,18 @@ print("State 2: μ = $(round.(test_model.B[2].μ, digits=3)), σ² = $(round(tes
 
 # ## Hidden State Decoding with Viterbi
 
-# Use Viterbi algorithm to find the most likely sequence of hidden states
-# given the observed data and learned parameters.
[…]
