+++ docs/src/tutorials/gaussian_latent_dynamics_example.md
@@ -28,11 +28,17 @@ nothing #hide
 ````
 
 ## Create a State-Space Model
+## Mathematical Foundation of Linear Dynamical Systems
 
-We start by defining the dimensions of our system. A linear dynamical system (LDS)
-models how a low-dimensional latent state evolves over time and generates high-dimensional
-observations. Here we use a 2D latent space (which we can visualize easily) that
-generates 10-dimensional observations.
+A Linear Dynamical System describes how a hidden state evolves over time and
+generates observations through two key equations:
+
+**State Evolution**: x_{t+1} = A * x_t + ε_t,  where ε_t ~ N(0, Q)
+**Observation**: y_t = C * x_t + η_t,  where η_t ~ N(0, R)
+
+The beauty of this formulation is that it separates the underlying dynamics
+(governed by A) from what we can actually measure (governed by C). The noise
+terms ε and η represent our uncertainty about the process and measurements.
 
 ````@example gaussian_latent_dynamics_example
 obs_dim = 10      # Number of observed variables at each time step
@@ -42,8 +48,10 @@ nothing #hide
 
 Define the state transition matrix $\mathbf{A}$. This matrix governs how the latent state
 evolves from one time step to the next: $\mathbf{x}_{t+1} = \mathbf{A} \mathbf{x}_t + \boldsymbol{\epsilon}$.
-We create a rotation matrix scaled by 0.95, which creates a stable spiral
-dynamic that slowly contracts toward the origin.
+The rotation angle of 0.25 radians (≈14.3°) creates a gentle spiral, while
+the 0.95 scaling ensures the system is stable (eigenvalues < 1). Without the
+scaling factor, trajectories would spiral outward indefinitely. This particular
+combination creates visually appealing dynamics that are easy to interpret.
 
 ````@example gaussian_latent_dynamics_example
 A = 0.95 * [cos(0.25) -sin(0.25); sin(0.25) cos(0.25)];
@@ -197,11 +205,16 @@ plot!(subplot=2, yticks=(-lim_emissions .* (obs_dim-1:-1:0), [L"y_{%$n}" for n i
       yformatter=y->"", tickfontsize=12, left_margin=10Plots.mm)
 ````
 
-## Initialize a Model and Perform Smoothing
+## The Learning Problem
 
-In a real scenario, we would only observe the emissions, not the latent states.
-Our goal is to learn the parameters $\mathbf{A}$, $\mathbf{Q}$, $\mathbf{C}$, $\mathbf{R}$ from the observations alone.
-We start by creating a "naive" model with random initial parameters.
+In real applications, we only observe y_t (the emissions) - the latent states x_t
+are hidden from us. Our challenge is to recover both:
+1. The system parameters (A, Q, C, R) that generated the data
+2. The most likely latent state sequence given our observations
+
+This is a classic "chicken and egg" problem: if we knew the parameters, we could
+infer the states; if we knew the states, we could estimate the parameters.
+The EM algorithm elegantly solves this by alternating between these two problems.
 
 Initialize with random parameters (this simulates not knowing the true system)
 
@@ -215,7 +228,10 @@ P0_init = Matrix(0.1 * I(latent_dim));      # Same initial uncertainty
 nothing #hide
 ````
 
-Create the naive model components
+Our "naive" initialization uses random parameters, simulating a real scenario
+where we don't know the true system. The quality of initialization can affect
+convergence speed and which local optimum we find, but EM is generally robust
+to reasonable starting points.
 
 ````@example gaussian_latent_dynamics_example
[…]
