### 2025-09-10 5de4924
Add cross validation tutorial
