+++ docs/src/tutorials/hmm_model_selection_example.md
@@ -1,0 +1,397 @@
+```@meta
+EditURL = "../../examples/HMM_ModelSelection.jl"
+```
+
+# Choosing "K" for a hidden Markov model.
+In principle, one can fit an HMM with any number of states, but how do we choose?
+One generally has no ground truth, except for the most rare cases. So it begs the question:
+How do we select the number of hidden states K? In this tutorial we will demonstrate a few
+typical approaches for model selection.
+
+## Load Required Packages
+
+````@example hmm_model_selection_example
+using StateSpaceDynamics
+using LinearAlgebra
+using Random
+using Plots
+using Statistics
+using StableRNGs
+````
+
+Fix RNG for reproducible simulation and k-means seeding
+
+````@example hmm_model_selection_example
+rng = StableRNG(1234);
+nothing #hide
+````
+
+## Create a True HMM
+For our example we will create a simple HMM with K=3 states and 2D Gaussian emissions.
+We will then generate ground truth data from this model for use in the tutorial.
+
+````@example hmm_model_selection_example
+K = 3  # Number of hidden states
+D = 2  # Dimensionality of observations
+T = 500  # Number of time steps to simulate
+````
+
+Define the true transition matrix
+
+````@example hmm_model_selection_example
+true_A = [0.90 0.05 0.05;
+          0.10 0.80 0.10;
+          0.15 0.15 0.70];
+nothing #hide
+````
+
+Define the true emission parameters (means and covariances)
+
+````@example hmm_model_selection_example
+true_μs = [-1.0 0.0 1.0;
+            0.6 -1.0 0.0]
+
+true_Σs =[[0.1 -0.3; -0.3 1.0],
+          [0.5 0.2; 0.2 0.3],
+          [0.3 0.0; 0.0 0.3]];
+nothing #hide
+````
+
+Define the initial state distribution
+
+````@example hmm_model_selection_example
+true_π = [0.5, 0.3, 0.2];
+nothing #hide
+````
+
+Create the true HMM
+
+````@example hmm_model_selection_example
+true_hmm = HiddenMarkovModel(true_A,
+                            [GaussianEmission(2, true_μs[:, k],
+                            true_Σs[k]) for k in 1:K],
+                            true_π,
+                            K)
+
+states, observations = rand(rng, true_hmm; n=T)
+````
+
+Now plot the data colored by the true states
+
+````@example hmm_model_selection_example
+p1 = scatter(observations[1, :], observations[2, :],
+    group=states,
+    xlabel="x1", ylabel="x2",
+    title="HMM Observations Colored by True State",
+    legend=:topright, alpha=0.7)
+````
+
+## Model Selection Approaches
+
+With real data, we don't know the true number of states K. We'll explore several
+methods to estimate the optimal K by fitting HMMs with different numbers of states
+and comparing their performance using various criteria.
+
+Define range of K values to test
+
+````@example hmm_model_selection_example
+K_range = 1:8
+````
+
+We'll compare models using multiple criteria:
+1. Log-likelihood (higher is better, but tends to overfit)
+2. AIC (Akaike Information Criterion - penalizes complexity)
+3. BIC (Bayesian Information Criterion - stronger complexity penalty)
+4. Cross-validation likelihood
+
+````@example hmm_model_selection_example
+results = Dict(
+    "K" => Int[],
+    "log_likelihood" => Float64[],
+    "AIC" => Float64[],
+    "BIC" => Float64[],
+    "n_params" => Int[]
+)
+````
+
[…]
