### 2025-09-12 9ef7aed
doc example changes
