+++ docs/examples/HMM_ModelSelection.jl
@@ -2,7 +2,7 @@
 # In principle, one can fit an HMM with any number of states, but how do we choose?
 # One generally has no ground truth, except for the most rare cases. So it begs the question:
 # How do we select the number of hidden states K? In this tutorial we will demonstrate a few 
-# typical approaches for model selection.
+# typical approaches for model selection with enhanced cross-validation integration.
 
 # ## Load Required Packages
 using StateSpaceDynamics
@@ -23,6 +23,7 @@ rng = StableRNG(1234);
 K = 3  # Number of hidden states
 D = 2  # Dimensionality of observations
 T = 500  # Number of time steps to simulate
+
 # Define the true transition matrix
 true_A = [0.90 0.05 0.05;
           0.10 0.80 0.10;
@@ -55,7 +56,7 @@ p1 = scatter(observations[1, :], observations[2, :],
     title="HMM Observations Colored by True State", 
     legend=:topright, alpha=0.7)
 
-    # ## Model Selection Approaches
+# ## Model Selection Approaches
 #
 # With real data, we don't know the true number of states K. We'll explore several
 # methods to estimate the optimal K by fitting HMMs with different numbers of states
@@ -75,8 +76,9 @@ results = Dict(
     "log_likelihood" => Float64[],
     "AIC" => Float64[],
     "BIC" => Float64[],
-    "n_params" => Int[]
-)
+    "n_params" => Int[],
+    "CV_score" => Float64[]
+);
 
 # ## Helper Functions for Model Selection
 
@@ -118,65 +120,6 @@ function count_parameters(hmm)
     return transition_params + initial_params + emission_params
 end
 
-println("Fitting HMMs with different numbers of states...")
-
-for k in K_range
-    println("Fitting HMM with K=$k states...")
-    
-    hmm_k = initialize_hmm_kmeans(observations, k, rng)
-    
-    fit!(hmm_k, observations; max_iters=100, tol=1e-6)
-    
-    ll = loglikelihood(hmm_k, observations)
-    n_params = count_parameters(hmm_k)
-    aic_val = -2*ll + 2*n_params
-    bic_val = -2*ll + log(T)*n_params
-    
-    push!(results["K"], k)
-    push!(results["log_likelihood"], ll)
-    push!(results["AIC"], aic_val)
-    push!(results["BIC"], bic_val)
-    push!(results["n_params"], n_params)
-end
-
-# ## Visualize Model Selection Results
-
-# Create a comprehensive plot showing all criteria
-p2 = plot(layout=(2, 2), size=(800, 600))
-
-plot!(results["K"], results["log_likelihood"], 
-      marker=:circle, linewidth=2, label="Log-likelihood",
-      xlabel="Number of States (K)", ylabel="Log-likelihood",
-      title="Model Log-likelihood", subplot=1)
-vline!([K], linestyle=:dash, color=:red, label="True K=$K", subplot=1)
-
-plot!(results["K"], results["AIC"], 
-      marker=:circle, linewidth=2, label="AIC", color=:orange,
-      xlabel="Number of States (K)", ylabel="AIC",
-      title="Akaike Information Criterion", subplot=2)
-aic_min_idx = argmin(results["AIC"])
-vline!([results["K"][aic_min_idx]], linestyle=:dash, color=:orange, 
-       label="AIC min (K=$(results["K"][aic_min_idx]))", subplot=2)
-vline!([K], linestyle=:dash, color=:red, label="True K=$K", subplot=2)
-
-plot!(results["K"], results["BIC"], 
-      marker=:circle, linewidth=2, label="BIC", color=:green,
[…]
