+++ docs/src/tutorials/gaussian_glm_hmm_example.md
@@ -35,9 +35,11 @@ nothing #hide
 
 In a GLM-HMM, each hidden state defines a different regression model. The system switches
 between these regression models according to Markovian dynamics. This is useful for modeling
-scenarios where the relationship between predictors and outcomes changes over time.
+scenarios where the relationship between predictors and outcomes changes over time. In this example,
+we will demonstrate how to use `StateSpaceDynamics.jl` to create a GLM-HMM, generate synthetic data,
+fit the model using the EM algorithm, and perform state inference with the Viterbi algorithm.
 
-Define emission models for each hidden state
+We will start by defining a simple GLM-HMM with two hidden states.
 State 1: Positive relationship between input and output
 
 ````@example gaussian_glm_hmm_example
@@ -68,7 +70,8 @@ nothing #hide
 
 Define the state transition matrix $\mathbf{A}$:
 $A_{ij} = P(\text{state}_t = j \mid \text{state}_{t-1} = i)$
-Diagonal elements are high (states are persistent), off-diagonal elements are low
+
+Diagonal elements are high (states are persistent)
 
 ````@example gaussian_glm_hmm_example
 A = [0.99 0.01;    # From state 1: 99% stay, 1% switch to state 2
@@ -83,7 +86,7 @@ Initial state distribution: $\pi_k = P(\text{state}_1 = k)$
 nothing #hide
 ````
 
-Construct the complete HMM
+Now that we have defined our emission models, we can construct the complete GLM-HMM.
 
 ````@example gaussian_glm_hmm_example
 true_model = HiddenMarkovModel(
@@ -101,24 +104,15 @@ nothing #hide
 
 ## Sample from the GLM-HMM
 
-Generate synthetic data from our true model. This will give us both the observed
-data (inputs and outputs) and the true hidden state sequence.
+Generate synthetic data from our true model. To do this, we must define the inputs to the model.
+Then, sampling will yield the outputs nad the true hidden state sequence.
 
 ````@example gaussian_glm_hmm_example
 n = 20000  # Number of time points
-````
 
-Generate random input features (predictors)
+Φ = randn(rng, 3, n);  # Generate random input features (predictors)
 
-````@example gaussian_glm_hmm_example
-Φ = randn(rng, 3, n);  # 3 features × n time points
-nothing #hide
-````
-
-Sample from the HMM: returns both hidden states and observations
-
-````@example gaussian_glm_hmm_example
-true_labels, data = rand(rng, true_model, Φ, n=n);
+true_labels, data = rand(rng, true_model, Φ, n=n);  # Sample from the HMM: returns both hidden states and observations
 
 print("Generated $(n) samples: State 1 ($(round(mean(true_labels .== 1)*100, digits=1))%), State 2 ($(round(mean(true_labels .== 2)*100, digits=1))%)\n");
 nothing #hide
@@ -126,8 +120,10 @@ nothing #hide
 
 ## Visualize the Sampled Dataset
 
-Create a scatter plot showing how the input-output relationship differs between
-the two hidden states. We'll plot feature 1 vs output, with points colored by true state.
+Here, we create a scatter plot to show how the input-output relationship differs between
+the two hidden states. We plot feature 1 vs output, with points colored by true state. State 1 (blue)
+has a positive slope, while State 2 (red) has a negative slopes. In addition to the scatter plot, we overlay
+the true regression lines for each state.
 
[…]

[…]
