+++ test/LinearDynamicalSystems/GaussianLDS.jl
@@ -4,17 +4,19 @@ l = 1.0 # length of pendulum
 dt = 0.01 # time step
 
 # Discrete-time dynamics
-A = [1.0 dt; -g / l*dt 1.0]
-Q = Matrix{Float64}(0.00001 * I(2))  # Process noise covariance
+A = [1.0 dt; -g / l * dt 1.0]
+Q = Matrix{Float64}(0.00001 * I(2))               # Process noise covariance
 
 # Initial state/ covariance
 x0 = [0.0; 1.0]
-P0 = Matrix{Float64}(0.1 * I(2))  # Initial state covariance
+P0 = Matrix{Float64}(0.1 * I(2))                  # Initial state covariance
 
-# Observation params 
-C = Matrix{Float64}(I(2))  # Observation matrix (assuming direct observation)
+# Observation params
+C = Matrix{Float64}(I(2))                         # Direct observation
 observation_noise_std = 0.5
-R = Matrix{Float64}((observation_noise_std^2) * I(2))  # Observation noise covariance
+R = Matrix{Float64}((observation_noise_std^2) * I(2))
+b = zeros(Float64, 2)                              # state bias
+d = zeros(Float64, 2)                              # observation bias
 
 function toy_lds(
     ntrials::Int=1, fit_bool::Vector{Bool}=[true, true, true, true, true, true]
@@ -28,27 +30,25 @@ function toy_lds(
         obs_dim=2,
         fit_bool=fill(true, 6)
     )
-
     # sample data
     T = 100
-    x, y = StateSpaceDynamics.rand(lds; tsteps=T, ntrials=ntrials) # 100 timepoints, 1 trials
-
+    x, y = StateSpaceDynamics.rand(lds; tsteps=T, ntrials=ntrials)
     return lds, x, y
 end
 
 function test_lds_properties(lds)
-    # check state and observation model are of correct type
     @test isa(lds.state_model, StateSpaceDynamics.GaussianStateModel)
     @test isa(lds.obs_model, StateSpaceDynamics.GaussianObservationModel)
     @test isa(lds, StateSpaceDynamics.LinearDynamicalSystem)
 
-    # check model param dimensions
     @test size(lds.state_model.A) == (lds.latent_dim, lds.latent_dim)
     @test size(lds.obs_model.C) == (lds.obs_dim, lds.latent_dim)
     @test size(lds.state_model.Q) == (lds.latent_dim, lds.latent_dim)
     @test size(lds.obs_model.R) == (lds.obs_dim, lds.obs_dim)
     @test size(lds.state_model.x0) == (lds.latent_dim,)
     @test size(lds.state_model.P0) == (lds.latent_dim, lds.latent_dim)
+    @test size(lds.state_model.b) == (lds.latent_dim,)
+    @test size(lds.obs_model.d) == (lds.obs_dim,)
 end
 
 function test_gaussian_obs_constructor_type_preservation()
@@ -69,6 +69,7 @@ function test_gaussian_obs_constructor_type_preservation()
     @test eltype(gsm_int.Q) === Int
     @test eltype(gsm_int.x0) === Int
     @test eltype(gsm_int.P0) === Int
+    @test eltype(gsm_int.b)  === Int
 
     # Float32
     A_f32 = Float32[1 2; 3 4]
@@ -88,7 +89,7 @@ function test_gaussian_obs_constructor_type_preservation()
     @test eltype(gsm_f32.x0)  === Float32
     @test eltype(gsm_f32.P0)  === Float32
 
-    # BigFloat
[…]
