+++ docs/examples/HMM_ModelSelection.jl
@@ -120,98 +120,6 @@ function count_parameters(hmm)
     return transition_params + initial_params + emission_params
 end
 
-println("Fitting HMMs with different numbers of states...")
-
-for k in K_range
-    println("Fitting HMM with K=$k states...")
-    
-    hmm_k = initialize_hmm_kmeans(observations, k, rng)
-    
-    fit!(hmm_k, observations; max_iters=100, tol=1e-6)
-    
-    ll = loglikelihood(hmm_k, observations)
-    n_params = count_parameters(hmm_k)
-    aic_val = -2*ll + 2*n_params
-    bic_val = -2*ll + log(T)*n_params
-    
-    push!(results["K"], k)
-    push!(results["log_likelihood"], ll)
-    push!(results["AIC"], aic_val)
-    push!(results["BIC"], bic_val)
-    push!(results["n_params"], n_params)
-end
-
-# ## Visualize Model Selection Results
-
-# Create a comprehensive plot showing all criteria
-p2 = plot(layout=(2, 2), size=(800, 600))
-
-plot!(results["K"], results["log_likelihood"], 
-      marker=:circle, linewidth=2, label="Log-likelihood",
-      xlabel="Number of States (K)", ylabel="Log-likelihood",
-      title="Model Log-likelihood", subplot=1)
-vline!([K], linestyle=:dash, color=:red, label="True K=$K", subplot=1)
-
-plot!(results["K"], results["AIC"], 
-      marker=:circle, linewidth=2, label="AIC", color=:orange,
-      xlabel="Number of States (K)", ylabel="AIC",
-      title="Akaike Information Criterion", subplot=2)
-aic_min_idx = argmin(results["AIC"])
-vline!([results["K"][aic_min_idx]], linestyle=:dash, color=:orange, 
-       label="AIC min (K=$(results["K"][aic_min_idx]))", subplot=2)
-vline!([K], linestyle=:dash, color=:red, label="True K=$K", subplot=2)
-
-plot!(results["K"], results["BIC"], 
-      marker=:circle, linewidth=2, label="BIC", color=:green,
-      xlabel="Number of States (K)", ylabel="BIC",
-      title="Bayesian Information Criterion", subplot=3)
-bic_min_idx = argmin(results["BIC"])
-vline!([results["K"][bic_min_idx]], linestyle=:dash, color=:green, 
-       label="BIC min (K=$(results["K"][bic_min_idx]))", subplot=3)
-vline!([K], linestyle=:dash, color=:red, label="True K=$K", subplot=3)
-
-plot!(results["K"], results["n_params"], 
-      marker=:circle, linewidth=2, label="# Parameters", color=:purple,
-      xlabel="Number of States (K)", ylabel="Number of Parameters",
-      title="Model Complexity", subplot=4)
-
-display(p2)
-
- ## Interpreting Information Criteria
-#
-# **AIC (Akaike Information Criterion)**:
-# - Estimates relative model quality for prediction
-# - Asymptotically equivalent to leave-one-out cross-validation
-# - Tends to select more complex models (higher K)
-# - Better for prediction tasks
-#
-# **BIC (Bayesian Information Criterion)**:
-# - Estimates probability that model is true among candidates
-# - Stronger complexity penalty, especially for large datasets
-# - Tends to select simpler models (lower K)
-# - Better for identifying "true" model structure
-#
-# **Key insight**: Lower values are better for both AIC and BIC
-# (they measure "badness" - deviance plus penalty)
-
-# ## Cross-Validation: The Gold Standard
-#
-# Cross-validation provides the most honest estimate of generalization performance:
-# - Trains on subset of data, tests on held-out portion
-# - Directly measures what we care about: performance on unseen data
[…]
