### 2025-09-16 e8b870c
lds identifiability example
