+++ docs/make.jl
@@ -11,6 +11,7 @@ tutorials = [
     ("GaussianLDS.jl", "gaussian_latent_dynamics_example"),
     ("PoissonLDS.jl", "poisson_latent_dynamics_example"),
     ("LDSModelSelection.jl", "lds_model_selection_example"),
+    ("LDSIdentifiability.jl", "lds_identifiability_example"),
     ("SLDS.jl", "switching_linear_dynamical_system_example"),
     ("HMM.jl", "hidden_markov_model_example"),
     ("HMM_ModelSelection.jl", "hmm_model_selection_example"),
@@ -55,6 +56,7 @@ makedocs(;
             "Gaussian LDS Example" => "tutorials/gaussian_latent_dynamics_example.md",
             "Poisson LDS Example" => "tutorials/poisson_latent_dynamics_example.md",
             "LDS Model Selection Example" => "tutorials/lds_model_selection_example.md",
+            "Non-Identifiability in LDS Models" => "tutorials/lds_identifiability_example.md",
             "Hidden Markov Model Example" => "tutorials/hidden_markov_model_example.md",
             "HMM Model Selection" => "tutorials/hmm_model_selection_example.md",
             "Gaussian GLM-HMM Example" => "tutorials/gaussian_glm_hmm_example.md",
