### 2025-09-18 2251ba9
adds hmm tutorial etc.
