+++ docs/make.jl
@@ -15,6 +15,7 @@ tutorials = [
     ("SLDS.jl", "switching_linear_dynamical_system_example"),
     ("HMM.jl", "hidden_markov_model_example"),
     ("HMM_ModelSelection.jl", "hmm_model_selection_example"),
+    ("HMMIdentifiability.jl", "hmm_identifiability_example"),
     ("Gaussian_GLM_HMM.jl", "gaussian_glm_hmm_example"),
     ("GaussianMixtureModel.jl", "gaussian_mixture_model_example"),
     ("PoissonMixtureModel.jl", "poisson_mixture_model_example"),
@@ -60,6 +61,7 @@ makedocs(;
             "Hidden Markov Model Example" => "tutorials/hidden_markov_model_example.md",
             "HMM Model Selection" => "tutorials/hmm_model_selection_example.md",
             "Gaussian GLM-HMM Example" => "tutorials/gaussian_glm_hmm_example.md",
+            "HMM Identifiability" => "tutorials/hmm_identifiability_example.md",
             "Gaussian Mixture Model Example" => "tutorials/gaussian_mixture_model_example.md",
             "Poisson Mixture Model Example" => "tutorials/poisson_mixture_model_example.md",
             "Probabilistic PCA Example" => "tutorials/Probabilistic_PCA_example.md",
