+++ docs/src/index.md
@@ -111,7 +111,7 @@ lds = LinearDynamicalSystem(state_model, obs_model, latent_dim, obs_dim, fill(tr
 
 ## Contributing
 
-If you encounter a bug or would like to contribute to the package, please [open an issue](https://github.com/depasquale-lab/StateSpaceDynamics.jl/issues) on our GitHub repository. Once the suggested change has received positive feedback feel free to submit a PR adhering to the [blue](https://github.com/JuliaDiff/BlueStyle) style guide.
+If you encounter a bug or would like to contribute to the package, please [open an issue](https://github.com/depasquale-lab/StateSpaceDynamics.jl/issues) on our GitHub repository. Once the suggested change has received positive feedback feel free to submit a PR adhering to the [blue](https://github.com/JuliaDiff/BlueStyle) style guide. 
 
 ## Citing StateSpaceDynamics.jl
 
