+++ docs/src/tutorials/hmm_identifiability_example.md
@@ -1,0 +1,231 @@
+```@meta
+EditURL = "../../examples/HMMIdentifiability.jl"
+```
+
+# Understanding Non-Identifiability in Hidden Markov Models
+
+This hands-on literate tutorial mirrors the LDS/SSM label-indeterminacy example but for HMMs. We will:
+
+1. **Build a reference HMM and synthesize data.**
+2. **Show permutation invariance of the likelihood** (log-likelihood is unchanged under state relabelings).
+3. **Fit from multiple random initializations** to surface **label switching**.
+4. **Align states post‑hoc** to a reference ordering and outline a simple canonicalization.
+5. **Summarize diagnostics** and clarify what **is** and **is not** identifiable.
+
+> **Takeaway.** In mixture models and HMMs, state labels are arbitrary. Without constraints or post‑hoc alignment, EM fits from different starts can represent the *same* model up to a permutation of labels.
+
+## Load Required Packages
+We’ll use `StateSpaceDynamics.jl` for HMM construction and EM fitting. `StableRNGs` gives bit‑reproducibility across runs.
+
+````@example hmm_identifiability_example
+using StateSpaceDynamics
+using Random
+using LinearAlgebra
+using Statistics
+using StableRNGs
+using Printf
+using Distributions
+
+rng = StableRNG(12345);
+nothing #hide
+````
+
+## Create a Reference ("True") Gaussian-Emission HMM
+We define a simple 3‑state, 1D Gaussian‑emission HMM with well‑separated means. The transition matrix is near‑symmetric
+with strong self‑transition probability, making states persist for stretches—this helps visualize label switching cleanly.
+
+````@example hmm_identifiability_example
+K = 3
+D = 1
+T = 400
+
+π_true = fill(1/3, K)  # uniform prior
+
+ρ = 0.88                # near-symmetric transitions with strong self-prob
+A_true = fill((1-ρ)/(K-1), K, K)
+for i in 1:K
+    A_true[i,i] = ρ
+end
+
+μ_true  = [-2.5, 0.0, 2.5]
+σ2_true = fill(0.1, K)
+Σ_true  = [fill(σ2_true[k], D, D) for k in 1:K]
+
+emissions = [GaussianEmission(1, [μ_true[k]], Σ_true[k]) for k in 1:K]
+true_hmm = HiddenMarkovModel(A_true, emissions, π_true, K)
+````
+
+Generate data
+
+````@example hmm_identifiability_example
+s_true, x_true = rand(rng, true_hmm; n=T)
+````
+
+## Permutation (Label) Invariance: Likelihood is Unchanged
+The HMM likelihood is **invariant** to permutation of state labels. If we apply a permutation matrix `P` to the transition
+matrix and reorder emissions accordingly, the joint distribution (and hence log-likelihood on data) is identical.
+We implement a helper `permute_states!` and verify that the log-likelihood difference is numerically ~0.
+
+````@example hmm_identifiability_example
+function permute_states!(m::HiddenMarkovModel, perm::Vector{Int})
+    K = length(perm)
+    P = zeros(eltype(m.A), K, K)
+    for (i,j) in enumerate(perm)   # new i  <- old j
+        P[i,j] = 1
+    end
+    m.A  .= P * m.A * P'
+    m.πₖ .= P * m.πₖ
+    m.B   = m.B[perm]
+    return m
+end
+
+mtest = deepcopy(true_hmm)
+ll0   = StateSpaceDynamics.loglikelihood(mtest, x_true)
+perm  = [2,3,1]  # arbitrary relabeling
+permute_states!(mtest, perm)
+ll1   = StateSpaceDynamics.loglikelihood(mtest, x_true)
+
[…]
