### 2025-09-19 9d69837
Fix some typos in the docs
