+++ docs/examples/PoissonLDS.jl
@@ -82,12 +82,12 @@ C = permutedims([abs.(randn(rng, obs_dim))'; abs.(randn(rng, obs_dim))']);
 # Instead of linear observations y = Cx + noise, we use:
 #
 # $λᵢ(t) = exp(Cᵢᵀx_t + dᵢ)$
-# $yᵢ(t) ~ Poisson(λᵢ(t))$
+# $yᵢ(t) \sim Poisson(λᵢ(t))$
 #
 # The exponential ensures rates are always positive (required for Poisson), and
 # the log-linear relationship means latent states multiplicatively affect firing rates.
 #
-# The baseline parameter d_i sets the minimum firing rate when latent states are zero.
+# The baseline parameter $d_i$ sets the minimum firing rate when latent states are zero.
 
 # Construct the model components
 state_model = GaussianStateModel(; A, Q, x0, P0)          # Gaussian latent dynamics
