+++ docs/make.jl
@@ -50,7 +50,7 @@ makedocs(;
             "Linear Dynamical Systems" => "LinearDynamicalSystems.md",
             "Hidden Markov Models" => "HiddenMarkovModels.md",
             "Switching Linear Dynamical Systems" => "SLDS.md",
-            "EmissionModels" => "EmissionModels.md",
+            "Emission Models" => "EmissionModels.md",
             "Mixture Models" => "MixtureModels.md"
         ],
         "Tutorials" => [
