+++ docs/examples/GaussianMixtureModel.jl
@@ -95,6 +95,8 @@ for i in 1:k
         label="") 
 end
 
+p1
+
 # ## Fit GMM Using EM Algorithm
 
 # Now we simulate the realistic scenario: observe only data points $\mathbf{X}$,
@@ -144,6 +146,8 @@ p3 = scatter(X[1, :], X[2, :];
     legend=:topright,
     label="Data points"
 )
+
+p3
 
 colors = [:red, :green, :blue] # Plot probability density contours for each learned component
 for i in 1:fit_gmm.k
