### 2025-09-19 eb58c51
Update ProbabilisticPCA.jl
