### 2025-09-19 877f21c
fixes mixture model rendering
