### 2025-09-19 7e2a577
Fix a few math display errors
