### 2025-09-19 03df42d
Changes to LDS files
