+++ docs/examples/GaussianLDS.jl
@@ -27,8 +27,8 @@ rng = StableRNG(123);
 # **Observation**: $y_t = \mathbf{C}  x_t + η_t$,  where $η_t \sim N(0, \mathbf{R})$
 #
 # The beauty of this formulation is that it separates the underlying dynamics 
-# (governed by A) from what we can actually measure (governed by C). The noise
-# terms ε and η represent our uncertainty about the process and measurements.
+# (governed by $\mathbf{A}$) from what we can actually measure (governed by $\mathbf{C}$). The noise
+# terms $\boldsymbol{\epsilon}$ and $\boldsymbol{\eta}$ represent our uncertainty about the process and measurements.
 
 obs_dim = 10      # Number of observed variables at each time step
 latent_dim = 2;   # Number of latent state variables
@@ -155,7 +155,7 @@ plot!(subplot=2, yticks=(-lim_emissions .* (obs_dim-1:-1:0), [L"y_{%$n}" for n i
 #
 # In real applications, we only observe $y_t$ (the emissions) - the latent states $x_t$
 # are hidden from us. Our challenge is to recover both:
-# 1. The system parameters (A, Q, C, R) that generated the data
+# 1. The system parameters ($\mathbf{A}$, $\mathbf{Q}$, $\mathbf{C}$, $\mathbf{R}$) that generated the data
 # 2. The most likely latent state sequence given our observations
 #
 # This is a classic "chicken and egg" problem: if we knew the parameters, we could
@@ -210,11 +210,11 @@ plot!(yticks=(lim_states .* (0:latent_dim-1), [L"x_%$d" for d in 1:latent_dim]),
 #
 # **E-step (Expectation)**: Given current parameter estimates, compute the posterior
 # distribution over latent states using the Kalman smoother. This gives us
-# `p(x_{1:T} | y_1:T, θ_current)`.
+# $p(x_{1:T} | y_1:T, θ_{current})$.
 #
 # **M-step (Maximization)**: Given the state estimates from the E-step, update
 # the parameters to maximize the expected log-likelihood. This involves solving
-# closed-form equations for A, Q, C, and R.
+# closed-form equations for $\mathbf{A}$, $\mathbf{Q}$, $\mathbf{C}$, and $\mathbf{R}$.
 #
 # The Evidence Lower BOund (ELBO) measures how well our model explains the data.
 # It's guaranteed to increase (or stay constant) at each iteration, ensuring
