### 2025-09-24 ab55833
Add bias terms to Gaussian LDS state and observation models
This commit introduces explicit bias vectors to the Gaussian state (b) and observation (d) models throughout the codebase. All relevant functions, Q-function calculations, EM steps, and tests are updated to support and verify the correct handling of these bias terms. Test coverage is expanded to ensure type preservation and correct parameter updates for the new bias parameters.
