+++ test/LinearDynamicalSystems/GaussianLDS.jl
@@ -17,12 +17,14 @@ P0 = Matrix{Float64}(0.1 * I(2))  # Initial state covariance
 C = Matrix{Float64}(I(2))  # Observation matrix (assuming direct observation)
 observation_noise_std = 0.5
 R = Matrix{Float64}((observation_noise_std^2) * I(2))  # Observation noise covariance
+b = zeros(2) # no bias
+d = zeros(2) # no bias
 
 function toy_lds(
     ntrials::Int=1, fit_bool::Vector{Bool}=[true, true, true, true, true, true]
 )
-    gaussian_sm = GaussianStateModel(; A=A, Q=Q, x0=x0, P0=P0)
-    gaussian_om = GaussianObservationModel(; C=C, R=R)
+    gaussian_sm = GaussianStateModel(; A=A, Q=Q, b=b, x0=x0, P0=P0)
+    gaussian_om = GaussianObservationModel(; C=C, d=d, R=R)
     lds = LinearDynamicalSystem(;
         state_model=gaussian_sm,
         obs_model=gaussian_om,
@@ -59,8 +61,9 @@ function test_gaussian_obs_constructor_type_preservation()
     Q_int = [1 0; 0 1]
     x0_int = [5; 6]
     P0_int = [2 0; 0 2]
+    b_int = [0, 0]
 
-    gsm_int = GaussianStateModel(; A=A_int, Q=Q_int, x0=x0_int, P0=P0_int)
+    gsm_int = GaussianStateModel(; A=A_int, Q=Q_int, b=b_int, x0=x0_int, P0=P0_int)
 
     @test eltype(gsm_int.A) === Int
     @test eltype(gsm_int.Q) === Int
@@ -72,8 +75,9 @@ function test_gaussian_obs_constructor_type_preservation()
     Q_f32 = Float32[1 0; 0 1]
     x0_f32 = Float32[0.5; 0.6]
     P0_f32 = Float32[0.2 0; 0 0.2]
+    b_f32 = Float32[0.0, 0.0]
 
-    gsm_f32 = GaussianStateModel(; A=A_f32, Q=Q_f32, x0=x0_f32, P0=P0_f32)
+    gsm_f32 = GaussianStateModel(; A=A_f32, Q=Q_f32, b=b_f32, x0=x0_f32, P0=P0_f32)
 
     @test eltype(gsm_f32.A) === Float32
     @test eltype(gsm_f32.Q) === Float32
@@ -85,8 +89,9 @@ function test_gaussian_obs_constructor_type_preservation()
     Q_bf = BigFloat[1 0; 0 1]
     x0_bf = BigFloat[0.1; 0.2]
     P0_bf = BigFloat[0.3 0; 0 0.3]
+    b_bf = BigFloat[0.0, 0.0]
 
-    gsm_bf = GaussianStateModel(; A=A_bf, Q=Q_bf, x0=x0_bf, P0=P0_bf)
+    gsm_bf = GaussianStateModel(; A=A_bf, Q=Q_bf, b=b_bf, x0=x0_bf, P0=P0_bf)
 
     @test eltype(gsm_bf.A) === BigFloat
     @test eltype(gsm_bf.Q) === BigFloat
@@ -102,9 +107,11 @@ function test_gaussian_lds_constructor_type_preservation()
     R_int = [3 0; 0 3]
     x0_int = [5, 6]
     P0_int = [4 0; 0 4]
+    b_int = [0, 0]
+    d_int = [0, 0]
 
-    gsm_int = GaussianStateModel(; A=A_int, Q=Q_int, x0=x0_int, P0=P0_int)
-    gom_int = GaussianObservationModel(; C=C_int, R=R_int)
[…]
