+++ test/runtests.jl
@@ -18,38 +18,39 @@ using Test
 Package Wide Tests
 """
 
-@testset "Package Wide Tests" begin
-    Aqua.test_all(StateSpaceDynamics; ambiguities=false)
-    @test isempty(Test.detect_ambiguities(StateSpaceDynamics))
-end
+# @testset "Package Wide Tests" begin
+#     Aqua.test_all(StateSpaceDynamics; ambiguities=false)
+#     @test isempty(Test.detect_ambiguities(StateSpaceDynamics))
+# end
 
-@testset "Blue Formatting" begin
-    @test JuliaFormatter.format(StateSpaceDynamics; verbose=false, overwrite=false)
-end
+# @testset "Blue Formatting" begin
+#     @test JuliaFormatter.format(StateSpaceDynamics; verbose=false, overwrite=false)
+# end
 
-@testset "Code linting using JET " begin
-    if VERSION >= v"1.11"
-        JET.test_package(StateSpaceDynamics; target_defined_modules=true)
-    end
-end
+# @testset "Code linting using JET " begin
+#     if VERSION >= v"1.11"
+#         JET.test_package(StateSpaceDynamics; target_defined_modules=true)
+#     end
+# end
 
 include("helper_functions.jl")
+
 """
 Tests for SLDS.jl
 """
 
-include("LinearDynamicalSystems//SLDS.jl")
+# include("LinearDynamicalSystems//SLDS.jl")
 
-@testset "SLDS Tests" begin
-    @testset "Constructor Tests" begin
-        test_init()
-        test_sample()
-    end
+# @testset "SLDS Tests" begin
+#     @testset "Constructor Tests" begin
+#         test_init()
+#         test_sample()
+#     end
 
-    @testset "vEM Tests" begin
-        test_vEstep()
-    end
-end
+#     @testset "vEM Tests" begin
+#         test_vEstep()
+#     end
+# end
 
 """
 Tests for LDS.jl
@@ -76,11 +77,11 @@ include("LinearDynamicalSystems//GaussianLDS.jl")
         # test when ntrials=1
         test_initial_observation_parameter_updates()
         test_state_model_parameter_updates()
-        test_obs_model_params_updates()
+        test_obs_model_parameter_updates()
         # test when ntrials>1
         test_initial_observation_parameter_updates(3)
         test_state_model_parameter_updates(3)
-        test_obs_model_params_updates(3)
+        test_obs_model_parameter_updates(3)
         # test fit method using n=1 and n=3
         test_EM()
         test_EM(3)

[…]
