### 2025-09-24 262eac9
Improve state model parameter update tests in PoissonLDS
Refactors the test for state model parameter updates to jointly optimize [A b] and Q, and checks both A, b, and Q for correctness. Updates test_EM and test_EM_matlab to explicitly set the state offset b, ensuring consistency with the new state model parameterization.
