+++ src/LinearDynamicalSystems.jl
@@ -58,8 +58,9 @@ Represents the observation model of a Linear Dynamical System with Gaussian nois
 - `R::M`: Observation noise covariance of size `(obs_dim × obs_dim)`.
 - `d::V`: Bias vector of length `(obs_dim)`.
 """
-Base.@kwdef mutable struct GaussianObservationModel{T<:Real,M<:AbstractMatrix{T},V<:AbstractVector{T}} <:
-                           AbstractObservationModel{T}
+Base.@kwdef mutable struct GaussianObservationModel{
+    T<:Real,M<:AbstractMatrix{T},V<:AbstractVector{T}
+} <: AbstractObservationModel{T}
     C::M
     R::M
     d::V
@@ -176,8 +177,11 @@ function stateparams(
 ) where {T<:Real,S<:AbstractStateModel{T},O<:AbstractObservationModel{T}}
     if isa(lds.state_model, GaussianStateModel)
         return [
-            lds.state_model.A, lds.state_model.Q, lds.state_model.b,
-            lds.state_model.x0, lds.state_model.P0
+            lds.state_model.A,
+            lds.state_model.Q,
+            lds.state_model.b,
+            lds.state_model.x0,
+            lds.state_model.P0,
         ]
     end
     return nothing
@@ -222,8 +226,8 @@ function Random.rand(
     rng::AbstractRNG, lds::LinearDynamicalSystem{T,S,O}; tsteps::Int, ntrials::Int
 ) where {T<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
     A, Q, x0, P0 = lds.state_model.A,
-                   lds.state_model.Q, lds.state_model.x0,
-                   lds.state_model.P0
+    lds.state_model.Q, lds.state_model.x0,
+    lds.state_model.P0
     C, R, b, d = lds.obs_model.C, lds.obs_model.R, lds.state_model.b, lds.obs_model.d
 
     x = Array{T,3}(undef, lds.latent_dim, tsteps, ntrials)
@@ -316,8 +320,8 @@ function loglikelihood(
 
     tsteps = size(y, 2)
     A, Q, x0, P0 = lds.state_model.A,
-                   lds.state_model.Q, lds.state_model.x0,
-                   lds.state_model.P0
+    lds.state_model.Q, lds.state_model.x0,
+    lds.state_model.P0
     C, R, b, d = lds.obs_model.C, lds.obs_model.R, lds.state_model.b, lds.obs_model.d
 
     R_chol = cholesky(Symmetric(R)).U
@@ -364,8 +368,8 @@ function Gradient(
     end
     latent_dim, tsteps = size(x)
     A, Q, x0, P0 = lds.state_model.A,
-                   lds.state_model.Q, lds.state_model.x0,
-                   lds.state_model.P0
+    lds.state_model.Q, lds.state_model.x0,
+    lds.state_model.P0
     C, R, b, d = lds.obs_model.C, lds.obs_model.R, lds.state_model.b, lds.obs_model.d
 
     R_chol = cholesky(Symmetric(R))
@@ -606,22 +610,21 @@ function smooth(
         )
     end
 
[…]
