### 2025-09-24 d828906
Improve model show methods and add Gaussian entropy test
Enhanced the Base.show methods for GaussianStateModel, GaussianObservationModel, and PoissonObservationModel to display additional parameter information, including vector sizes and values for b and d. Updated LinearDynamicalSystem show output to clarify parameter names. Added a new test for Gaussian entropy in Utilities.jl and enabled package-wide, formatting, and linting tests in runtests.jl.
