### 2025-09-25 c6dad47
Add SLDS validation and refactor SLDS structure
Introduces a new Valid.jl module with comprehensive validation functions for LinearDynamicalSystem and SLDS structures. Refactors SLDS.jl to define a new SLDS struct and removes legacy code. Updates tests to cover SLDS validation, adds dependency on SparseArrays for tests, and includes the new validation logic in the main module.
