+++ src/Valid.jl
@@ -1,0 +1,229 @@
+export isvalid_SLDS, isvalid_LDS, isvalid_probvec
+
+"""
+    isvalid_LDS(lds::LinearDynamicalSystem{T,S,O}) where {T,S,O}
+
+Validate that all parameters in a LinearDynamicalSystem are dimensionally consistent
+and mathematically valid.
+
+# Checks performed:
+- Matrix dimensions are consistent
+- Covariance matrices are positive definite
+- fit_bool has correct length for the observation model type
+- All parameters have consistent data types
+"""
+function isvalid_LDS(lds::LinearDynamicalSystem{T,S,O}) where {T,S,O}
+    try
+        # Check state model dimensions and properties
+        if !_isvalid_state_model(lds.state_model, lds.latent_dim)
+            return false
+        end
+        
+        # Check observation model dimensions and properties
+        if !_isvalid_obs_model(lds.obs_model, lds.obs_dim, lds.latent_dim)
+            return false
+        end
+        
+        # Check fit_bool length
+        expected_fit_length = lds.obs_model isa PoissonObservationModel ? 5 : 6
+        if length(lds.fit_bool) != expected_fit_length
+            @warn "fit_bool has length $(length(lds.fit_bool)), expected $expected_fit_length"
+            return false
+        end
+        
+        # Check consistency between inferred and stored dimensions
+        inferred_latent = size(lds.state_model.A, 1)
+        inferred_obs = size(lds.obs_model.C, 1)
+        
+        if lds.latent_dim != inferred_latent
+            @warn "Stored latent_dim ($(lds.latent_dim)) ≠ inferred from A matrix ($inferred_latent)"
+            return false
+        end
+        
+        if lds.obs_dim != inferred_obs
+            @warn "Stored obs_dim ($(lds.obs_dim)) ≠ inferred from C matrix ($inferred_obs)"
+            return false
+        end
+        
+        return true
+        
+    catch e
+        @warn "Error during LDS validation: $e"
+        return false
+    end
+end
+
+"""
+    _isvalid_state_model(state_model::GaussianStateModel{T}, latent_dim::Int) where T
+
+Validate GaussianStateModel parameters.
+"""
+function _isvalid_state_model(state_model::GaussianStateModel{T}, latent_dim::Int) where T
+    # Check A matrix
+    if size(state_model.A) != (latent_dim, latent_dim)
+        @warn "A matrix size $(size(state_model.A)) ≠ expected ($latent_dim, $latent_dim)"
+        return false
+    end
+    
+    # Check Q matrix (process noise covariance)
+    if size(state_model.Q) != (latent_dim, latent_dim)
+        @warn "Q matrix size $(size(state_model.Q)) ≠ expected ($latent_dim, $latent_dim)"
+        return false
+    end
+    
+    if !issymmetric(state_model.Q)
+        @warn "Q matrix is not symmetric"
+        return false
+    end
+    
+    if !isposdef(state_model.Q)
+        @warn "Q matrix is not positive definite"
+        return false
+    end
+    
+    # Check bias vector b
+    if length(state_model.b) != latent_dim
+        @warn "Bias vector b length $(length(state_model.b)) ≠ expected $latent_dim"
+        return false
+    end
+    
+    # Check initial state x0
+    if length(state_model.x0) != latent_dim
+        @warn "Initial state x0 length $(length(state_model.x0)) ≠ expected $latent_dim"
+        return false
+    end
+    
+    # Check P0 matrix (initial covariance)
[…]
