+++ src/SLDS.jl
@@ -1,495 +1,10 @@
-export SwitchingLinearDynamicalSystem, fit!, initialize_slds, variational_expectation!
-"""
-Switching Linear Dynamical System
-
-Struct to Encode a Hidden Markov model that switches among K distinct LinearDyanmicalSystems
-
-# Fields
-- `A::V`: Transition matrix for mode switching.
-- `B::VL`: Vector of Linear Dynamical System models.
-- `πₖ::V`: Initial state distribution.
-- `K::Int`: Number of modes.
-"""
-mutable struct SwitchingLinearDynamicalSystem{
-    T<:Real,
-    M<:AbstractMatrix{T},
-    V<:AbstractVector{T},
-    VL<:AbstractVector{<:LinearDynamicalSystem},
-} <: AbstractHMM
-    A::M
-    B::VL
-    πₖ::V
-    K::Int
-end
-
-function Base.show(io::IO, slds::SwitchingLinearDynamicalSystem; gap="")
-    println(io, gap, "Switching Linear Dynamical System:")
-    println(io, gap, "----------------------------------")
-
-    if slds.K > 3
-        println(io, gap, " size(A)  = ($(size(slds.A,1)), $(size(slds.A,2)))")
-        println(io, gap, " size(πₖ) = ($(size(slds.πₖ,1)),)")
-    else
-        println(io, gap, " A  = $(round.(slds.A, sigdigits=3))")
-        println(io, gap, " πₖ = $(round.(slds.πₖ, sigdigits=3))")
-    end
-
-    println(io, gap, " Switching Models:")
-    println(io, gap, " -----------------")
-
-    show_all = get(io, :limit, true) == false
-
-    if slds.K > 4 && !show_all
-        for lds in slds.B[1:3]
-            Base.show(io, lds; gap=gap * "  ")
-            println(io, gap, "  --------------")
-        end
-        println(io, gap, "  $(slds.K-3) more ..., see `print_full()`")
-    else
-        for (i, lds) in enumerate(slds.B)
-            Base.show(io, lds; gap=gap * "  ")
-            if i < slds.K
-                println(io, gap, "  --------------")
-            end
-        end
-    end
-
-    return nothing
-end
-
-"""
-    Random.rand(rng, slds, T)
-
-Generate synthetic data with switching LDS models
-
-#Arguments
-- `rng:AbstractRNG`: Random number generator
-- `slds::SwitchingLinearDynamicalSystem`: The switching LDS model
-- `T::Int`: Number of time steps to sample
-
-# Returns
-- `Tuple{Array,Array, Array}`: Latent states (x), observations (y), and mode sequences (z).
-"""
-function Random.rand(rng::AbstractRNG, slds::SwitchingLinearDynamicalSystem, T::Int)
-    state_dim = slds.B[1].latent_dim
-    obs_dim = slds.B[1].obs_dim
-    K = slds.K
-
-    x = zeros(state_dim, T)  # Latent states
-    y = zeros(obs_dim, T)    # Observations
-    z = zeros(Int, T)        # Mode sequence
-
-    # Sample initial mode
-    z[1] = rand(rng, Categorical(slds.πₖ / sum(slds.πₖ)))
-    x[:, 1] = rand(rng, MvNormal(zeros(state_dim), slds.B[z[1]].state_model.Q))
-    y[:, 1] = rand(
-        rng, MvNormal(slds.B[z[1]].obs_model.C * x[:, 1], slds.B[z[1]].obs_model.R)
-    )
-
-    @views for t in 2:T
-        # Sample mode based on transition probabilities
-        z[t] = rand(Categorical(slds.A[z[t - 1], :] ./ sum(slds.A[z[t - 1], :])))
-        # Update latent state and observation
-        x[:, t] = rand(
[…]
