### 2025-09-29 be050cc
Refactor SLDS sampling and validation, add tests
Refactored SLDS sampling to support both Gaussian and Poisson observation models with a unified interface. Replaced `valid_SLDS` with `isvalid_SLDS` and improved validation logic to check only for consistent latent and observation dimensions. Added comprehensive tests for SLDS sampling, edge cases, and reproducibility. Removed obsolete parameter extraction functions and replaced them with a systematic approach. Enabled parallel sampling for large numbers of trials in LDS sampling.
