+++ test/LinearDynamicalSystems/SLDS.jl
@@ -52,7 +52,7 @@ function test_valid_SLDS_happy_path()
     lds = _make_gaussian_lds(2, 4)
     s = SLDS(A=_rowstochastic(K), Z₀=_probvec(K), LDSs=fill(lds, K))
     @test begin
-        valid_SLDS(s)    # should not throw
+        isvalid_SLDS(s)    # should not throw
         true
     end
 end
@@ -63,11 +63,11 @@ function test_valid_SLDS_dimension_mismatches()
 
     # size(A,1)=K, but length(Z₀) ≠ K
     s_badZ0 = SLDS(A=_rowstochastic(K), Z₀=_probvec(K+1), LDSs=fill(lds, K))
-    @test_throws AssertionError valid_SLDS(s_badZ0)
+    @test_throws AssertionError isvalid_SLDS(s_badZ0)
 
     # size(A,1)=K, but number of LDSs ≠ K
     s_badLDSs = SLDS(A=_rowstochastic(K), Z₀=_probvec(K), LDSs=fill(lds, K+1))
-    @test_throws AssertionError valid_SLDS(s_badLDSs)
+    @test_throws AssertionError isvalid_SLDS(s_badLDSs)
 end
 
 function test_valid_SLDS_nonstochastic_rows_and_invalid_Z0()
@@ -78,12 +78,12 @@ function test_valid_SLDS_nonstochastic_rows_and_invalid_Z0()
     A_bad = _rowstochastic(K)
     A_bad[2, :] .= (-0.1, 0.5, 0.6)  # sums to 1 but has a negative entry
     s_badA = SLDS(A=A_bad, Z₀=_probvec(K), LDSs=fill(lds, K))
-    @test_throws AssertionError valid_SLDS(s_badA)
+    @test_throws AssertionError isvalid_SLDS(s_badA)
 
     # Z0 does not sum to 1
     Z0_bad = _probvec(K); Z0_bad[1] += 0.1
     s_badZ0 = SLDS(A=_rowstochastic(K), Z₀=Z0_bad, LDSs=fill(lds, K))
-    @test_throws AssertionError valid_SLDS(s_badZ0)
+    @test_throws AssertionError isvalid_SLDS(s_badZ0)
 end
 
 function test_valid_SLDS_mixed_observation_model_types()
@@ -91,8 +91,8 @@ function test_valid_SLDS_mixed_observation_model_types()
     K = 3
     lds_g = _make_gaussian_lds(2, 2)
     lds_p = _make_poisson_lds(2, 2)  # different obs model type
-    s = SLDS(A=_rowstochastic(K), Z₀=_probvec(K), LDSs=[lds_g, lds_g, lds_p])
-    @test_throws AssertionError valid_SLDS(s)
+    @test_throws MethodError SLDS(A=_rowstochastic(K), Z₀=_probvec(K), LDSs=[lds_g, lds_g, lds_p])
+    
 end
 
 function test_valid_SLDS_inconsistent_latent_or_obs_dims()
@@ -102,8 +102,129 @@ function test_valid_SLDS_inconsistent_latent_or_obs_dims()
     lds_b_obs   = _make_gaussian_lds(2, 4) # different obs_dim
 
     s_bad_state = SLDS(A=_rowstochastic(K), Z₀=_probvec(K), LDSs=[lds_a, lds_b_state])
-    @test_throws AssertionError valid_SLDS(s_bad_state)
+    @test_throws AssertionError isvalid_SLDS(s_bad_state)
 
     s_bad_obs = SLDS(A=_rowstochastic(K), Z₀=_probvec(K), LDSs=[lds_a, lds_b_obs])
[…]

[…]
