### 2025-09-30 e656e27
Add Grads and Hessians for SLDS
