+++ src/LinearDynamicalSystems.jl
@@ -334,7 +334,6 @@ observed data.
 - `x::AbstractMatrix{T}`: The state sequence of the LDS.
 - `lds::LinearDynamicalSystem{S,O}`: The Linear Dynamical System.
 - `y::AbstractMatrix{T}`: The observed data.
-- `w::AbstractVector{T}`: coeffcients to weight the data.
 
 # Returns
 - `ll::T`: The complete-data log-likelihood of the LDS.
@@ -343,14 +342,7 @@ function loglikelihood(
     x::AbstractMatrix{U},
     lds::LinearDynamicalSystem{T,S,O},
     y::AbstractMatrix{T},
-    w::Union{Nothing,AbstractVector{T}}=nothing,
 ) where {U<:Real,T<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
-    if w === nothing
-        w = ones(eltype(y), size(y, 2))
-    elseif eltype(w) !== eltype(y)
-        error("weights must be Vector{$(eltype(y))}; Got Vector{$(eltype(w))}")
-    end
-
     tsteps = size(y, 2)
     A, Q, x0, P0 = lds.state_model.A,
     lds.state_model.Q, lds.state_model.x0,
@@ -369,16 +361,14 @@ function loglikelihood(
 
     for t in 1:tsteps
         if t > 1
-            # residual: x_t - (A x_{t-1} + b)
             mul!(temp_dx, A, view(x, :, t-1), -one(eltype(x)), false)
             temp_dx .+= view(x, :, t) .- b
             ll += sum(abs2, Q_chol \ temp_dx)
         end
 
-        # residual: y_t - (C x_t + d)
         mul!(temp_dy, C, view(x, :, t), -one(eltype(x)), false)
         temp_dy .+= view(y, :, t) .- d
-        ll += w[t] * sum(abs2, R_chol \ temp_dy)
+        ll += sum(abs2, R_chol \ temp_dy)
     end
 
     return -eltype(x)(0.5) * ll
@@ -394,11 +384,7 @@ function Gradient(
     lds::LinearDynamicalSystem{T,S,O},
     y::AbstractMatrix{T},
     x::AbstractMatrix{T},
-    w::Union{Nothing,AbstractVector{T}}=nothing,
 ) where {T<:Real,S<:GaussianStateModel{T},O<:GaussianObservationModel{T}}
-    if w === nothing
-        w = ones(T, size(y, 2))
-    end
     latent_dim, tsteps = size(x)
     A, Q, x0, P0 = lds.state_model.A,
     lds.state_model.Q, lds.state_model.x0,
@@ -418,12 +404,12 @@ function Gradient(
     dx1 = x[:, 1] .- x0
     dx2 = x[:, 2] .- (A * x[:, 1] .+ b)
     dy1 = y[:, 1] .- (C * x[:, 1] .+ d)
-    grad[:, 1] .= A_inv_Q * dx2 + w[1] * C_inv_R * dy1 - (P0_chol \ dx1)
+    grad[:, 1] .= A_inv_Q * dx2 + C_inv_R * dy1 - (P0_chol \ dx1)
 
     # Middle steps
     @views for t in 2:(tsteps - 1)
         grad[:, t] .=
-            w[t] * C_inv_R * (y[:, t] .- (C * x[:, t] .+ d)) -
+            C_inv_R * (y[:, t] .- (C * x[:, t] .+ d)) -
             (Q_chol \ (x[:, t] .- (A * x[:, t - 1] .+ b))) +
             (A_inv_Q * (x[:, t + 1] .- (A * x[:, t] .+ b)))
     end
@@ -431,7 +417,7 @@ function Gradient(
     # Last time step
     dxT = x[:, tsteps] .- (A * x[:, tsteps - 1] .+ b)
[…]
