+++ test/runtests.jl
@@ -58,6 +58,14 @@ include("LinearDynamicalSystems/SLDS.jl")
         test_SLDS_minimal_dimensions()
         test_valid_SLDS_probability_helper_functions()
     end
+    
+    @testset "SLDS Gradient and Hessian Tests" begin
+        test_SLDS_gradient_numerical()
+        test_SLDS_hessian_numerical()
+        test_SLDS_gradient_reduces_to_single_LDS()
+        test_SLDS_hessian_block_structure()
+        test_SLDS_gradient_weight_normalization()
+    end
 end
 
 """
