### 2025-10-01 9f8c9a0
Refactor loglikelihood to return per-timestep vector
The loglikelihood functions for Gaussian and Poisson LDS now return a vector of per-timestep log-likelihoods instead of a scalar. All usages and tests have been updated to sum the vector when a scalar is needed. Added a loglikelihood method for SLDS that computes weighted per-timestep log-likelihoods. This change improves flexibility for downstream algorithms that require per-timestep likelihoods.
