### 2025-10-02 a206e0f
Optimize temporary array usage in loglikelihood and Gradient
Pre-allocate and reuse temporary arrays in loglikelihood and Gradient functions for improved efficiency and reduced memory allocations. Replaces repeated allocations and direct solves with ldiv! and mul! to optimize performance.
