### 2025-10-03 fa143d5
Add smoothing
