+++ test/runtests.jl
@@ -17,26 +17,26 @@ using Test
 
 const CHECKED_TYPES = [Float32, Float64] #, BigFloat] UMFPACK does not support BigFloat for Sparse Arrays see: https://github.com/JuliaSparse/SparseArrays.jl/blob/main/src/solvers/umfpack.jl
 
-"""
-Package Wide Tests
-"""
+# """
+# Package Wide Tests
+# """
 
-@testset "Package Wide Tests" begin
-    Aqua.test_all(StateSpaceDynamics; ambiguities=false)
-    @test isempty(Test.detect_ambiguities(StateSpaceDynamics))
-end
+# @testset "Package Wide Tests" begin
+#     Aqua.test_all(StateSpaceDynamics; ambiguities=false)
+#     @test isempty(Test.detect_ambiguities(StateSpaceDynamics))
+# end
 
-@testset "Blue Formatting" begin
-    @test JuliaFormatter.format(StateSpaceDynamics; verbose=false, overwrite=false)
-end
+# @testset "Blue Formatting" begin
+#     @test JuliaFormatter.format(StateSpaceDynamics; verbose=false, overwrite=false)
+# end
 
-@testset "Code linting using JET " begin
-    if VERSION >= v"1.11"
-        JET.test_package(StateSpaceDynamics; target_defined_modules=true)
-    end
-end
+# @testset "Code linting using JET " begin
+#     if VERSION >= v"1.11"
+#         JET.test_package(StateSpaceDynamics; target_defined_modules=true)
+#     end
+# end
 
-include("helper_functions.jl")
+# include("helper_functions.jl")
 
 """
 Tests for SLDS.jl
@@ -67,6 +67,16 @@ include("LinearDynamicalSystems/SLDS.jl")
         test_SLDS_gradient_reduces_to_single_LDS()
         test_SLDS_hessian_block_structure()
         test_SLDS_gradient_weight_normalization()
+    end
+
+    @testset "SLDS Smooth Test" begin
+        test_SLDS_smooth_basic()
+        test_SLDS_smooth_reduces_to_single_LDS()
+        test_SLDS_smooth_with_realistic_weights()
+        test_SLDS_smooth_consistency_with_gradients()
+        test_SLDS_smooth_entropy_calculation()
+        test_SLDS_smooth_covariance_symmetry()
+        test_SLDS_smooth_different_weight_patterns()
     end
 end
 
