### 2025-10-04 68000b6
add smoothing and begin mstep
