+++ src/LinearDynamicalSystems.jl
@@ -1018,23 +1018,23 @@ Update the initial state mean of the Linear Dynamical System using the average a
 trials.
 """
 function update_initial_state_mean!(
-    lds::LinearDynamicalSystem{T,S,O}, 
+    lds::LinearDynamicalSystem{T,S,O},
     tfs::TrialFilterSmooth{T},
-    w::Union{Nothing,AbstractVector{<:AbstractVector{T}}} = nothing
+    w::Union{Nothing,AbstractVector{<:AbstractVector{T}}}=nothing,
 ) where {T<:Real,S<:GaussianStateModel{T},O<:AbstractObservationModel{T}}
     if lds.fit_bool[1]
         ntrials = length(tfs.FilterSmooths)
         x0_new = zeros(T, lds.latent_dim)
         total_weight = zero(T)
-        
+
         for trial in 1:ntrials
             fs = tfs[trial]
             weight = isnothing(w) ? one(T) : w[trial][1]  # Weight at t=1
-            
+
             x0_new .+= weight .* fs.E_z[:, 1]
             total_weight += weight
         end
-        
+
         lds.state_model.x0 .= x0_new ./ total_weight
     end
 end
@@ -1050,27 +1050,28 @@ Update the initial state covariance of the Linear Dynamical System using the ave
 all trials.
 """
 function update_initial_state_covariance!(
-    lds::LinearDynamicalSystem{T,S,O}, 
+    lds::LinearDynamicalSystem{T,S,O},
     tfs::TrialFilterSmooth{T},
-    w::Union{Nothing,AbstractVector{<:AbstractVector{T}}} = nothing
+    w::Union{Nothing,AbstractVector{<:AbstractVector{T}}}=nothing,
 ) where {T<:Real,S<:GaussianStateModel{T},O<:AbstractObservationModel{T}}
     if lds.fit_bool[2]
         ntrials = length(tfs.FilterSmooths)
         state_dim = lds.latent_dim
         p0_new = zeros(T, state_dim, state_dim)
         total_weight = zero(T)
-        
+
         for trial in 1:ntrials
             fs = tfs[trial]
             weight = isnothing(w) ? one(T) : w[trial][1]  # Weight at t=1
-            
-            p0_new .+= weight .* (fs.E_zz[:, :, 1] - (lds.state_model.x0 * lds.state_model.x0'))
+
+            p0_new .+=
+                weight .* (fs.E_zz[:, :, 1] - (lds.state_model.x0 * lds.state_model.x0'))
             total_weight += weight
         end
-        
+
         p0_new ./= total_weight
         p0_new .= 0.5 * (p0_new + p0_new')
-        
+
         # Set the new P0 matrix
         lds.state_model.P0 = p0_new
     end
@@ -1088,30 +1089,30 @@ Update the transition matrix A of the Linear Dynamical System.
 
 """
 function update_A_b!(
-    lds::LinearDynamicalSystem{T,S,O}, 
+    lds::LinearDynamicalSystem{T,S,O},
     tfs::TrialFilterSmooth{T},
-    w::Union{Nothing,AbstractVector{<:AbstractVector{T}}} = nothing
+    w::Union{Nothing,AbstractVector{<:AbstractVector{T}}}=nothing,
 ) where {T<:Real,S<:GaussianStateModel{T},O<:AbstractObservationModel{T}}
     lds.fit_bool[3] || return nothing
     D = lds.latent_dim
     ntrials = length(tfs)
-    
+
     # Accumulate statistics for [A b] jointly
     Sxz = zeros(T, D, D + 1)
     Szz = zeros(T, D + 1, D + 1)
-    
+
[…]
