+++ test/LinearDynamicalSystems/SLDS.jl
@@ -599,7 +599,7 @@ function test_SLDS_smooth_entropy_calculation()
 
     w = ones(Float64, K, tsteps) ./ K
 
-    # Call smooth! directly to access FilterSmooth
+    # Call smooth! directly to access StateSpaceDynamics.FilterSmooth
     fs = StateSpaceDynamics.initialize_FilterSmooth(slds.LDSs[1], tsteps)
     StateSpaceDynamics.smooth!(slds, fs, y[:, :, 1], w)
 
@@ -894,4 +894,232 @@ function test_SLDS_estep_elbo_components()
 
     # Entropy should be positive
     # @test tfs[1].entropy > 0
+end
+
+function test_weighted_update_initial_state_mean()
+    """Test weighted update of initial state mean"""
+    K = 2
+    latent_dim = 2
+    obs_dim = 3
+    lds = _make_gaussian_lds(latent_dim, obs_dim)
+    slds = SLDS(; A=_rowstochastic(K), πₖ=_probvec(K), LDSs=fill(lds, K))
+
+    ntrials = 4
+    tsteps = 10
+    z, x, y = rand(slds; tsteps=tsteps, ntrials=ntrials)
+
+    # Create per-trial, per-timestep weights
+    w = [rand(K, tsteps) for _ in 1:ntrials]
+    for trial in 1:ntrials
+        w[trial] ./= sum(w[trial]; dims=1)
+    end
+
+    # Mock FilterSmooth objects with ALL required fields
+    tfs_array = Vector{StateSpaceDynamics.FilterSmooth{Float64}}(undef, ntrials)
+    for k in 1:ntrials
+        x_smooth = x[:, :, k]
+        p_smooth = zeros(Float64, latent_dim, latent_dim, tsteps)
+        p_smooth_tt1 = zeros(Float64, latent_dim, latent_dim, tsteps)
+        E_z = x[:, :, k]
+        E_zz = zeros(Float64, latent_dim, latent_dim, tsteps)
+        E_zz_prev = zeros(Float64, latent_dim, latent_dim, tsteps)
+
+        for t in 1:tsteps
+            p_smooth[:, :, t] .= 0.1 * I(latent_dim)
+            p_smooth_tt1[:, :, t] .= 0.1 * I(latent_dim)
+            E_zz[:, :, t] .= E_z[:, t] * E_z[:, t]' .+ 0.1 * I(latent_dim)
+            if t > 1
+                E_zz_prev[:, :, t] .= E_z[:, t] * E_z[:, t - 1]'
+            end
+        end
+
+        entropy = 0.0
+        tfs_array[k] = StateSpaceDynamics.FilterSmooth(
+            x_smooth, p_smooth, p_smooth_tt1, E_z, E_zz, E_zz_prev, entropy
+        )
+    end
+    tfs = StateSpaceDynamics.TrialFilterSmooth(tfs_array)
+
+    for active_k in 1:K
+        lds_k = slds.LDSs[active_k]
+        lds_k.fit_bool[1] = true
+
+        # Create weights for this LDS
+        w_k = [w[trial][active_k, :] for trial in 1:ntrials]
+
+        StateSpaceDynamics.update_initial_state_mean!(lds_k, tfs, w_k)
+
+        # Check result is finite
+        @test all(isfinite.(lds_k.state_model.x0))
+    end
+end
+
+function test_weighted_update_A_b()
+    """Test weighted update of A and b matrices"""
+    K = 2
+    latent_dim = 2
+    obs_dim = 3
+    lds = _make_gaussian_lds(latent_dim, obs_dim)
[…]
