### 2025-10-07 ec21a97
add more SLDS stuff
