+++ src/StateSpaceDynamics.jl
@@ -23,7 +23,6 @@ include("Utilities.jl")
 include("LinearDynamicalSystems.jl")
 include("EmissionModels.jl")
 include("HiddenMarkovModels.jl")
-include("HMMConstructors.jl")
 include("SLDS.jl")
 include("MixtureModels.jl")
 include("Preprocessing.jl")
