+++ src/SLDS.jl
@@ -29,6 +29,25 @@ y_t | x_t, z_t ~ N(C^{(z_t)} x_t + d^{(z_t)}, R^{(z_t)})
 end
 
 """
+    initialize_forward_backward(model::AbstractHMM, num_obs::Int)
+
+Initialize the forward backward storage struct.
+"""
+function initialize_forward_backward(
+    model::SLDS, num_obs::Int, ::Type{T}
+) where {T<:Real}
+    num_states = size(model.A, 1)
+
+    return ForwardBackward(
+        zeros(T, num_states, num_obs),
+        zeros(T, num_states, num_obs),
+        zeros(T, num_states, num_obs),
+        zeros(T, num_states, num_obs),
+        zeros(T, num_states, num_states),
+    )
+end
+
+"""
     rand(rng::AbstractRNG, slds::SLDS{T,S,O}; tsteps::Int, ntrials::Int=1) where {T<:Real, S<:AbstractStateModel, O<:AbstractObservationModel}
 
 Sample from a Switching Linear Dynamical System (SLDS). Returns a tuple `(z, x, y)` where:
@@ -315,7 +334,6 @@ function smooth!(
 
     fs.p_smooth .= p_smooth_result
     fs.p_smooth_tt1[:, :, 2:end] .= p_smooth_tt1_result
-    fs.entropy = gaussian_entropy(Symmetric(H))
 
     # Symmetrize
     @views for t in 1:tsteps
@@ -332,21 +350,260 @@ function smooth(slds::SLDS, y::AbstractMatrix{T}, w::AbstractMatrix{T}) where {T
     return fs.x_smooth, fs.p_smooth
 end
 
-function mstep!(slds::SLDS, 
-                tfs::TrialFilterSmooth, 
-                fb::ForwardBackward, 
-                y::AbstractMatrix)
+"""
+    sample_posterior(rng::AbstractRNG, fs::FilterSmooth{T}) where {T<:Real}
 
-    # update hmm parameters
-    update_initial_state_distribution!(slds, fb)
-    update_transmision_matrix!(slds, fb)
-    
+Sample a trajectory from the posterior over continuous states and compute its entropy.
 
-    # update LDS paramerers
-    for (k, lds) in enumerate(slds.LDSs)
-        weights = w[k, :]
-        mstep!(lds, tfs, y, weights)
+Returns:
+- x_sample: matrix of size (latent_dim, tsteps) representing one sample from 
+  q(x) = ∏ₜ N(x_t | x_smooth_t, p_smooth_t)
+- entropy: H[q(x)] = ∑ₜ H[N(x_t | x_smooth_t, p_smooth_t)]
+"""
+function sample_posterior(rng::AbstractRNG, fs::FilterSmooth{T}) where {T<:Real}
+    latent_dim, tsteps = size(fs.x_smooth)
+    x_sample = similar(fs.x_smooth)
+    entropy = zero(T)
+
+    for t in 1:tsteps
+        μ = fs.x_smooth[:, t]
+        Σ = Symmetric(fs.p_smooth[:, :, t])
+
+        # Sample
+        x_sample[:, t] = rand(rng, MvNormal(μ, Σ))
+
+        # Accumulate entropy: H[N(μ, Σ)] = 0.5 * (d * (1 + log(2π)) + log|Σ|)
+        logdet_Σ = logdet(Σ)
+        entropy += 0.5 * (latent_dim * (1 + log(2π)) + logdet_Σ)
+    end
+
+    return x_sample, entropy
+end
+
+# Convenience method
+function sample_posterior(fs::FilterSmooth{T}) where {T<:Real}
+    return sample_posterior(Random.GLOBAL_RNG, fs)
+end
+
+"""
[…]
