### 2025-10-08 25f6f21
First pass at vLEM for the SLDS
