+++ test/LinearDynamicalSystems/SLDS.jl
@@ -56,7 +56,7 @@ end
 function test_valid_SLDS_happy_path()
     K = 3
     lds = _make_gaussian_lds(2, 4)
-    s = SLDS(; A=_rowstochastic(K), Z₀=_probvec(K), LDSs=fill(lds, K))
+    s = SLDS(; A=_rowstochastic(K), πₖ=_probvec(K), LDSs=fill(lds, K))
     @test begin
         isvalid_SLDS(s)    # should not throw
         true
@@ -67,12 +67,12 @@ function test_valid_SLDS_dimension_mismatches()
     K = 2
     lds = _make_gaussian_lds(2, 3)
 
-    # size(A,1)=K, but length(Z₀) ≠ K
-    s_badZ0 = SLDS(; A=_rowstochastic(K), Z₀=_probvec(K+1), LDSs=fill(lds, K))
+    # size(A,1)=K, but length(πₖ) ≠ K
+    s_badZ0 = SLDS(; A=_rowstochastic(K), πₖ=_probvec(K+1), LDSs=fill(lds, K))
     @test_throws AssertionError isvalid_SLDS(s_badZ0)
 
     # size(A,1)=K, but number of LDSs ≠ K
-    s_badLDSs = SLDS(; A=_rowstochastic(K), Z₀=_probvec(K), LDSs=fill(lds, K+1))
+    s_badLDSs = SLDS(; A=_rowstochastic(K), πₖ=_probvec(K), LDSs=fill(lds, K+1))
     @test_throws AssertionError isvalid_SLDS(s_badLDSs)
 end
 
@@ -83,13 +83,13 @@ function test_valid_SLDS_nonstochastic_rows_and_invalid_Z0()
     # Non-probability row in A (negative entry)
     A_bad = _rowstochastic(K)
     A_bad[2, :] .= (-0.1, 0.5, 0.6)  # sums to 1 but has a negative entry
-    s_badA = SLDS(; A=A_bad, Z₀=_probvec(K), LDSs=fill(lds, K))
+    s_badA = SLDS(; A=A_bad, πₖ=_probvec(K), LDSs=fill(lds, K))
     @test_throws AssertionError isvalid_SLDS(s_badA)
 
     # Z0 does not sum to 1
     Z0_bad = _probvec(K);
     Z0_bad[1] += 0.1
-    s_badZ0 = SLDS(; A=_rowstochastic(K), Z₀=Z0_bad, LDSs=fill(lds, K))
+    s_badZ0 = SLDS(; A=_rowstochastic(K), πₖ=Z0_bad, LDSs=fill(lds, K))
     @test_throws AssertionError isvalid_SLDS(s_badZ0)
 end
 
@@ -99,7 +99,7 @@ function test_valid_SLDS_mixed_observation_model_types()
     lds_g = _make_gaussian_lds(2, 2)
     lds_p = _make_poisson_lds(2, 2)  # different obs model type
     @test_throws MethodError SLDS(
-        A=_rowstochastic(K), Z₀=_probvec(K), LDSs=[lds_g, lds_g, lds_p]
+        A=_rowstochastic(K), πₖ=_probvec(K), LDSs=[lds_g, lds_g, lds_p]
     )
 end
 
@@ -109,17 +109,17 @@ function test_valid_SLDS_inconsistent_latent_or_obs_dims()
     lds_b_state = _make_gaussian_lds(3, 3) # different latent_dim
     lds_b_obs = _make_gaussian_lds(2, 4) # different obs_dim
 
-    s_bad_state = SLDS(; A=_rowstochastic(K), Z₀=_probvec(K), LDSs=[lds_a, lds_b_state])
[…]

[…]
