### 2025-10-09 9cf6ecb
add more poisson slds tests
