+++ test/LinearDynamicalSystems/SLDS.jl
@@ -1123,3 +1123,332 @@ function test_zero_weights_behavior()
     H, H_diag, H_super, H_sub = StateSpaceDynamics.Hessian(slds, y[:, :, 1], x[:, :, 1], w)
     @test all(isfinite.(H))
 end
+
+# Poisson SLDS Tests
+
+function test_SLDS_sampling_poisson_extended()
+    # Extended Poisson sampling test with more trials and edge cases
+    K = 3
+    lds = _make_poisson_lds(3, 5)
+    s = SLDS(; A=_rowstochastic(K), πₖ=_probvec(K), LDSs=fill(lds, K))
+
+    tsteps, ntrials = 50, 10
+    z, x, y = rand(s; tsteps=tsteps, ntrials=ntrials)
+
+    @test all(y[i, t, n] ≥ 0 for i in 1:5, t in 1:tsteps, n in 1:ntrials)
+    @test all(y[i, t, n] == round(y[i, t, n]) for i in 1:5, t in 1:tsteps, n in 1:ntrials)
+    @test all(isfinite, y)
+end
+
+function test_SLDS_gradient_numerical_poisson()
+    K = 2
+    lds = _make_poisson_lds(2, 3)
+    slds = SLDS(; A=_rowstochastic(K), πₖ=_probvec(K), LDSs=fill(lds, K))
+
+    z, x, y = rand(slds; tsteps=20, ntrials=1)
+
+    tsteps = size(y, 2)
+    w = rand(K, tsteps)
+    w ./= sum(w; dims=1)
+
+    y_trial = y[:, :, 1]
+    x_trial = x[:, :, 1]
+
+    # Analytical gradient
+    grad_analytical = StateSpaceDynamics.Gradient(slds, y_trial, x_trial, w)
+
+    # For numerical gradient, manually compute weighted log-likelihood
+    function weighted_ll(x_flat)
+        x_mat = reshape(x_flat, size(x_trial))
+        ll = 0.0
+
+        for k in 1:K
+            lds_k = slds.LDSs[k]
+
+            # Extract parameters
+            A_k = lds_k.state_model.A
+            Q_k = lds_k.state_model.Q
+            b_k = lds_k.state_model.b
+            x0_k = lds_k.state_model.x0
+            P0_k = lds_k.state_model.P0
+            C_k = lds_k.obs_model.C
+            log_d_k = lds_k.obs_model.log_d
+            d_k = exp.(log_d_k)
+
+            # Precompute inverses
+            inv_P0 = inv(P0_k)
+            inv_Q = inv(Q_k)
+
+            # Initial state (weighted by w[k, 1])
+            dx0 = x_mat[:, 1] - x0_k
+            ll += w[k, 1] * (-0.5 * dot(dx0, inv_P0 * dx0))
+
+            # Dynamics and emissions
+            for t in 1:tsteps
+                # Poisson emission (weighted by w[k, t])
+                obs_mean = C_k * x_mat[:, t] .+ d_k
+                ll += w[k, t] * (dot(y_trial[:, t], obs_mean) - sum(exp, obs_mean))
+
+                # Dynamics (weighted by w[k, t] for t > 1)
+                if t > 1
+                    dx = x_mat[:, t] - (A_k * x_mat[:, t - 1] + b_k)
+                    ll += w[k, t] * (-0.5 * dot(dx, inv_Q * dx))
+                end
+            end
+        end
+
+        return ll
+    end
+
+    grad_numerical = ForwardDiff.gradient(weighted_ll, vec(x_trial))
[…]
