+++ test/runtests.jl
@@ -17,26 +17,26 @@ using Test
 
 const CHECKED_TYPES = [Float32, Float64] #, BigFloat] UMFPACK does not support BigFloat for Sparse Arrays see: https://github.com/JuliaSparse/SparseArrays.jl/blob/main/src/solvers/umfpack.jl
 
-# """
-# Package Wide Tests
-# """
+"""
+Package Wide Tests
+"""
 
-# @testset "Package Wide Tests" begin
-#     Aqua.test_all(StateSpaceDynamics; ambiguities=false)
-#     @test isempty(Test.detect_ambiguities(StateSpaceDynamics))
-# end
+@testset "Package Wide Tests" begin
+    Aqua.test_all(StateSpaceDynamics; ambiguities=false)
+    @test isempty(Test.detect_ambiguities(StateSpaceDynamics))
+end
 
-# @testset "Blue Formatting" begin
-#     @test JuliaFormatter.format(StateSpaceDynamics; verbose=false, overwrite=false)
-# end
+@testset "Blue Formatting" begin
+    @test JuliaFormatter.format(StateSpaceDynamics; verbose=false, overwrite=false)
+end
 
-# @testset "Code linting using JET " begin
-#     if VERSION >= v"1.11"
-#         JET.test_package(StateSpaceDynamics; target_defined_modules=true)
-#     end
-# end
+@testset "Code linting using JET " begin
+    if VERSION >= v"1.11"
+        JET.test_package(StateSpaceDynamics; target_defined_modules=true)
+    end
+end
 
-# include("helper_functions.jl")
+include("helper_functions.jl")
 
 """
 Tests for SLDS.jl
@@ -111,7 +111,6 @@ include("LinearDynamicalSystems/SLDS.jl")
         test_SLDS_poisson_log_d_interpretation()
         test_SLDS_gradient_weight_normalization_poisson()
     end
-
 end
 
 """
@@ -186,203 +185,203 @@ include("LinearDynamicalSystems//PoissonLDS.jl")
     end
 end
 
-# """
-# Tests for Switching Regression Models
-# """
+"""
+Tests for Switching Regression Models
+"""
 
-# include("HiddenMarkovModels/GaussianHMM.jl")
+include("HiddenMarkovModels/GaussianHMM.jl")
 
-# @testset "GaussianHMM Tests" begin
-#     test_SwitchingGaussian_fit()
-#     test_SwitchingGaussian_SingleState_fit()
-#     test_kmeans_init()
-#     test_trialized_GaussianHMM()
-#     test_SwitchingGaussian_fit_float32()
-# end
+@testset "GaussianHMM Tests" begin
+    test_SwitchingGaussian_fit()
+    test_SwitchingGaussian_SingleState_fit()
+    test_kmeans_init()
+    test_trialized_GaussianHMM()
+    test_SwitchingGaussian_fit_float32()
+end
 
-# include("HiddenMarkovModels/SwitchingGaussianRegression.jl")
+include("HiddenMarkovModels/SwitchingGaussianRegression.jl")
 
-# @testset "Switching Gaussian Regression Tests" begin
-#     test_SwitchingGaussianRegression_fit()
-#     test_SwitchingGaussianRegression_SingleState_fit()
-#     test_trialized_SwitchingGaussianRegression()
-#     # test_SwitchingGaussianRegression_fit_float32()
-# end
+@testset "Switching Gaussian Regression Tests" begin
+    test_SwitchingGaussianRegression_fit()
+    test_SwitchingGaussianRegression_SingleState_fit()
+    test_trialized_SwitchingGaussianRegression()
+    # test_SwitchingGaussianRegression_fit_float32()
+end
 
[…]
