+++ src/SLDS.jl
@@ -28,6 +28,37 @@ y_t | x_t, z_t ~ N(C^{(z_t)} x_t + d^{(z_t)}, R^{(z_t)})
     LDSs::Vector{LinearDynamicalSystem{T,S,O}}
 end
 
+function Base.show(io::IO, slds::SLDS; gap="")
+    K = length(slds.LDSs)
+
+    println(io, gap, "Switching Linear Dynamical System (SLDS):")
+    println(io, gap, "-----------------------------------------")
+    println(io, gap, " Number of discrete states: $K")
+
+    if K > 3
+        println(io, gap, " size(A)  = ($(size(slds.A,1)), $(size(slds.A,2)))")
+        println(io, gap, " size(πₖ) = ($(length(slds.πₖ)),)")
+    else
+        println(io, gap, " A  = $(round.(slds.A, sigdigits=3))")
+        println(io, gap, " πₖ = $(round.(slds.πₖ, sigdigits=3))")
+    end
+
+    println(io, gap, " Linear Dynamical Systems:")
+    println(io, gap, " -------------------------")
+
+    # Show details of first LDS
+    if K > 0
+        println(io, gap, "  State 1:")
+        Base.show(io, slds.LDSs[1]; gap=gap * "   ")
+
+        if K > 1
+            println(io, gap, "  ... and $(K-1) more state(s)")
+        end
+    end
+
+    return nothing
+end
+
 """
     initialize_forward_backward(model::AbstractHMM, num_obs::Int)
 
