+++ docs/src/SLDS.md
@@ -23,43 +23,48 @@ An SLDS with $K$ discrete states is defined by the following generative model:
 
 ```math
 \begin{align*}
-    s_1 &\sim \text{Cat}(\pi_k) \\
-    z_1 &\sim \mathcal{N}(\mu_{0}, P_{0}) \\
-    s_t &\mid s_{t-1} \sim \text{Cat}(A_{s_{t-1}, :}) \\
-    z_t &\mid z_{t-1}, s_t \sim \mathcal{N}(F_{s_t} z_{t-1}, Q_{s_t}) \\
-    y_t &\mid z_t, s_t \sim \mathcal{N}(C_{s_t} z_t, R_{s_t})
+    z_1 &\sim \text{Cat}(\pi_k) \\
+    x_1 &\sim \mathcal{N}(\mu_{0}, P_{0}) \\
+    z_t &\mid z_{t-1} \sim \text{Cat}(A_{z_{t-1}, :}) \\
+    x_t &\mid x_{t-1}, z_t \sim \mathcal{N}(F_{z_t} x_{t-1}, Q_{z_t}) \\
+    y_t &\mid x_t, z_t \sim \mathcal{N}(C_{z_t} x_t, R_{z_t})
 \end{align*}
 ```
 
 Where:
 
-- ``s_t ∈ {1, 2, …, K}`` is the **discrete switching state** at time ``t``
-- ``z_t ∈ ℝᴰ`` is the **continuous latent state** at time ``t``
+- ``z_t ∈ {1, 2, …, K}`` is the **discrete switching state** at time ``t``
+- ``x_t ∈ ℝᴰ`` is the **continuous latent state** at time ``t``
 - ``y_t ∈ ℝᴾ`` is the **observed data** at time ``t``
 - ``π_k`` is the **initial discrete state distribution**
 - ``A`` is the **discrete state transition matrix**
-- ``F_{s_t}`` is the **state-dependent dynamics matrix** for discrete state ``s_t``
-- ``Q_{s_t}`` is the **state-dependent process noise covariance** for discrete state ``s_t``
-- ``C_{s_t}`` is the **state-dependent observation matrix** for discrete state ``s_t``
-- ``R_{s_t}`` is the **state-dependent observation noise covariance** for discrete state ``s_t``
+- ``F_{z_t}`` is the **state-dependent dynamics matrix** for discrete state ``z_t``
+- ``Q_{z_t}`` is the **state-dependent process noise covariance** for discrete state ``z_t``
+- ``C_{z_t}`` is the **state-dependent observation matrix** for discrete state ``z_t``
+- ``R_{z_t}`` is the **state-dependent observation noise covariance** for discrete state ``z_t``
 
 ## Implementation Structure
 
 In `StateSpaceDynamics.jl`, an SLDS is represented as:
 
 ```julia
-mutable struct SwitchingLinearDynamicalSystem <: AbstractHMM
-    A::M     # Transition matrix for mode switching (K × K)
-    B::VL    # Vector of LinearDynamicalSystem models (length K)
-    πₖ::V    # Initial state distribution (length K)
-    K::Int   # Number of modes
+mutable struct SLDS{
+    T<:Real,
+    S<:AbstractStateModel,
+    O<:AbstractObservationModel,
+    TM<:AbstractMatrix{T},
+    ISV<:AbstractVector{T},
+} <: AbstractHMM
+    A::TM # Transition matrix
+    πₖ::ISV # Initial state distribution
+    LDSs::Vector{LinearDynamicalSystem{T,S,O}} # Vector of LDS models
 end
 ```
 
-Each mode in the `B` vector contains its own `LinearDynamicalSystem` with:
[…]
