### 2025-10-21 cbf9bdd
fix jet errors
