+++ src/Utilities.jl
@@ -265,9 +265,7 @@ end
 
 Calculate the Euclidean distance between two points.
 """
-function euclidean_distance(
-    a::AbstractVector{T1}, b::AbstractVector{T2}
-) where {T1<:Real,T2<:Real}
+function euclidean_distance(a::AbstractVector, b::AbstractVector)
     return sqrt(sum((a .- b) .^ 2))
 end
 
