### 2025-10-21 974efeb
Fix variable shadowing bug
