### 2025-10-22 d1cbef4
Adds covariance priors
