+++ docs/src/LinearDynamicalSystems.md
@@ -168,3 +168,102 @@ Given the latent structure of state-space models, we must rely on either the Exp
 ```@docs
 fit!(lds::LinearDynamicalSystem{T,S,O}, y::AbstractArray{T,3}; max_iter::Int=1000, tol::Float64=1e-12) where {T<:Real,S<:GaussianStateModel{T},O<:AbstractObservationModel{T}}
 ```
+
+## Inverse-Wishart Priors on Covariances (MAP)
+
+To encourage well-conditioned covariance estimates, StateSpaceDynamics.jl supports **Inverse-Wishart (IW)** priors on the covariance matrices of a Linear Dynamical System.  
+These can be used to impose **shrinkage** toward a target scale, yielding **maximum a posteriori (MAP)** estimates instead of pure MLEs.
+
+You can attach an IW prior to:
+
+- The **process noise covariance** `Q_prior`
+- The **initial state covariance** `P0_prior`
+- The **observation noise covariance** `R_prior` (Gaussian models only)
+
+```@docs
+IWPrior
+```
+
+### Definition
+
+For a covariance matrix ` \Sigma \in \mathbb{R}^{d\times d} `,
+
+```math
+\Sigma \sim \text{IW}(\Psi, \nu),
+\qquad
+p(\Sigma)\propto |\Sigma|^{-(\nu+d+1)/2}\exp\!\Big[-\tfrac12\operatorname{tr}(\Psi\,\Sigma^{-1})\Big].
+```
+
+- ` \Psi ` is the **scale matrix** (positive-definite).  
+- ` \nu ` is the **degrees of freedom**.  
+  A proper mode exists when ` \nu > d+1 `.
+
+Given `n` effective samples and sample statistic `S`,
+the posterior is
+
+```math
+\Sigma\mid\text{data}\sim\text{IW}(\Psi+S,\;\nu+n)
+```
+
+and the **MAP** (mode) update used internally is
+
+```math
+\widehat{\Sigma}_{\text{MAP}}
+=\frac{\Psi+S}{\nu+n+d+1}.
+```
+
+### Example: Adding IW Priors to an LDS
+
+```julia
+using LinearAlgebra, Random, StateSpaceDynamics
+
+rng = MersenneTwister(42)
+D, P = 3, 4
+
+# Proper SPD scale matrices (avoid UniformScaling)
+ΨQ  = diagm(0 => fill(0.01, D))
+ΨP0 = diagm(0 => fill(0.01, D))
+ΨR  = diagm(0 => fill(0.01, P))
+
+Qprior  = IWPrior(Ψ = ΨQ,  ν = D + 3.0)
+P0prior = IWPrior(Ψ = ΨP0, ν = D + 3.0)
+Rprior  = IWPrior(Ψ = ΨR,  ν = P + 3.0)
+
+# Gaussian LDS
+A = 0.9I + 0.05randn(rng, D, D)
+Q = Matrix(I, D, D) .* 0.3
+b = zeros(D); x0 = zeros(D); P0 = Matrix(I, D, D) .* 0.8
+
+C = randn(rng, P, D)
+R = Matrix(I, P, P) .* 0.25
+d = 0.1 .* randn(rng, P)
+
+gsm = GaussianStateModel(A=A, Q=Q, b=b, x0=x0, P0=P0,
+                         Q_prior=Qprior, P0_prior=P0prior)
+gom = GaussianObservationModel(C=C, R=R, d=d, R_prior=Rprior)
+lds = LinearDynamicalSystem(gsm, gom)
+
+X, Y = rand(rng, lds; tsteps=100, ntrials=5)
+fit!(lds, Y; max_iter=20, progress=false)
+```
+
+### Effect on Learning
+
+- When a prior is present, the M-step uses the MAP mode formula above.  
[…]
