+++ test/runtests.jl
@@ -17,105 +17,105 @@ using Test
 
 const CHECKED_TYPES = [Float32, Float64] #, BigFloat] UMFPACK does not support BigFloat for Sparse Arrays see: https://github.com/JuliaSparse/SparseArrays.jl/blob/main/src/solvers/umfpack.jl
 
-# """
-# Package Wide Tests
-# """
+"""
+Package Wide Tests
+"""
 
-# @testset "Package Wide Tests" begin
-#     Aqua.test_all(StateSpaceDynamics; ambiguities=false)
-#     @test isempty(Test.detect_ambiguities(StateSpaceDynamics))
-# end
+@testset "Package Wide Tests" begin
+    Aqua.test_all(StateSpaceDynamics; ambiguities=false)
+    @test isempty(Test.detect_ambiguities(StateSpaceDynamics))
+end
 
-# @testset "Blue Formatting" begin
-#     @test JuliaFormatter.format(StateSpaceDynamics; verbose=false, overwrite=false)
-# end
+@testset "Blue Formatting" begin
+    @test JuliaFormatter.format(StateSpaceDynamics; verbose=false, overwrite=false)
+end
 
-# @testset "Code linting using JET " begin
-#     if VERSION >= v"1.11"
-#         JET.test_package(StateSpaceDynamics; target_defined_modules=true)
-#     end
-# end
+@testset "Code linting using JET " begin
+    if VERSION >= v"1.11"
+        JET.test_package(StateSpaceDynamics; target_defined_modules=true)
+    end
+end
 
-# include("helper_functions.jl")
+include("helper_functions.jl")
 
-# """
-# Tests for SLDS.jl
-# """
+"""
+Tests for SLDS.jl
+"""
 
-# include("LinearDynamicalSystems/SLDS.jl")
+include("LinearDynamicalSystems/SLDS.jl")
 
-# @testset "SLDS Tests" begin
-#     @testset "valid_SLDS Tests" begin
-#         test_valid_SLDS_happy_path()
-#         test_valid_SLDS_dimension_mismatches()
-#         test_valid_SLDS_nonstochastic_rows_and_invalid_Z0()
-#         test_valid_SLDS_mixed_observation_model_types()
-#         test_valid_SLDS_inconsistent_latent_or_obs_dims()
-#         test_SLDS_sampling_gaussian()
-#         test_SLDS_sampling_poisson()
-#         test_SLDS_deterministic_transitions()
-#         test_SLDS_single_trial()
-#         test_SLDS_reproducibility()
-#         test_SLDS_single_state_edge_case()
-#         test_SLDS_minimal_dimensions()
-#         test_valid_SLDS_probability_helper_functions()
-#     end
+@testset "SLDS Tests" begin
+    @testset "valid_SLDS Tests" begin
+        test_valid_SLDS_happy_path()
+        test_valid_SLDS_dimension_mismatches()
+        test_valid_SLDS_nonstochastic_rows_and_invalid_Z0()
+        test_valid_SLDS_mixed_observation_model_types()
+        test_valid_SLDS_inconsistent_latent_or_obs_dims()
+        test_SLDS_sampling_gaussian()
+        test_SLDS_sampling_poisson()
+        test_SLDS_deterministic_transitions()
+        test_SLDS_single_trial()
+        test_SLDS_reproducibility()
+        test_SLDS_single_state_edge_case()
+        test_SLDS_minimal_dimensions()
+        test_valid_SLDS_probability_helper_functions()
+    end
 
-#     @testset "SLDS Gradient and Hessian Tests" begin
-#         test_SLDS_gradient_numerical()
-#         test_SLDS_hessian_numerical()
-#         test_SLDS_gradient_reduces_to_single_LDS()
-#         test_SLDS_hessian_block_structure()
[…]
