### 2025-11-04 depasquale-lab/StateSpaceDynamics.jl#67 @slinderman
Thanks for all of these improvements. It looks great now!
