### 2025-11-04 #8077 @slinderman
Thank you for all your hard work @rsenne! The package is much improved! I've closed my issue (https://github.com/depasquale-lab/StateSpaceDynamics.jl/issues/67).
