### 2025-11-08 7e4370a
Fixes a bug in BernoulliRegression. Removes unnecessary utility, and updates paper to include vLEM
