+++ paper/paper.bib
@@ -289,3 +289,19 @@ state space models},
       primaryClass={cs.MS},
       url={https://arxiv.org/abs/2505.05542},
 }
+
+@InProceedings{pmlr-v119-zoltowski20a,
+  title = 	 {A general recurrent state space framework for modeling neural dynamics during decision-making},
+  author =       {Zoltowski, David and Pillow, Jonathan and Linderman, Scott},
+  booktitle = 	 {Proceedings of the 37th International Conference on Machine Learning},
+  pages = 	 {11680--11691},
+  year = 	 {2020},
+  editor = 	 {III, Hal Daumé and Singh, Aarti},
+  volume = 	 {119},
+  series = 	 {Proceedings of Machine Learning Research},
+  month = 	 {13--18 Jul},
+  publisher =    {PMLR},
+  pdf = 	 {http://proceedings.mlr.press/v119/zoltowski20a/zoltowski20a.pdf},
+  url = 	 {https://proceedings.mlr.press/v119/zoltowski20a.html},
+  abstract = 	 {An open question in systems and computational neuroscience is how neural circuits accumulate evidence towards a decision. Fitting models of decision-making theory to neural activity helps answer this question, but current approaches limit the number of these models that we can fit to neural data. Here we propose a general framework for modeling neural activity during decision-making. The framework includes the canonical drift-diffusion model and enables extensions such as multi-dimensional accumulators, variable and collapsing boundaries, and discrete jumps. Our framework is based on constraining the parameters of recurrent state space models, for which we introduce a scalable variational Laplace EM inference algorithm. We applied the modeling approach to spiking responses recorded from monkey parietal cortex during two decision-making tasks. We found that a two-dimensional accumulator better captured the responses of a set of parietal neurons than a single accumulator model, and we identified a variable lower boundary in the responses of a parietal neuron during a random dot motion task. We expect this framework will be useful for modeling neural dynamics in a variety of decision-making settings.}
+}
