### 2025-11-09 2081250
CompatHelper: bump compat for ForwardDiff to 1, (keep existing compat)
