+++ Project.toml
@@ -1,7 +1,7 @@
 name = "StateSpaceDynamics"
 uuid = "19cf8126-1d00-4ad8-bf17-4e76d7bc0f6d"
-authors = ["Ryan Senne rsenne@bu.edu"]
 version = "0.2.0"
+authors = ["Ryan Senne rsenne@bu.edu"]
 
 [deps]
 ArrayLayouts = "4c555306-a7a7-4459-81d9-ec55ddd5c99a"
@@ -22,7 +22,7 @@ StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 [compat]
 ArrayLayouts = "v1.10.4"
 Distributions = "0.25.109"
-ForwardDiff = "0.10.36"
+ForwardDiff = "0.10.36, 1"
 LineSearches = "7.3.0"
 LinearAlgebra = "1"
 Optim = "1.10.0"
