### 2025-11-12 #8077 @rsenne
@espottesmith got it!

>  Zenodo archive had a v1

I didn't put a version on the original so it defaulted to 1. I have amended the original archive. I also added the v0.3.0 release on github.

> Names don't exactly match between Zenodo and your paper

I have amended the names to match!

> missing affiliations in Zenodo

Fixed!

> Zenodo archive should also be identical to the title of your paper

fixed!

I can also confirm i went through the references and could not find any DOIs for those missing. I think this is everything. Let me know if i missed anything or any issues remain.
