### 2025-11-13 #8077 @espottesmith
I've read through your manuscript, and I have no real concerns or critiques. It's very well written. I'm happy to recommend this manuscript and codebase for acceptance in JOSS.
