+++ paper/paper.bib
@@ -49,7 +49,7 @@ month = feb,
 }
 
 @article{SaavedraBodinSouto2019,
-  title={StateSpaceModels.jl: a Julia Package for Time-Series Analysis in a State-Space Framework},
+  title={StateSpaceModels.jl: a {J}ulia Package for Time-Series Analysis in a State-Space Framework},
   author={Raphael Saavedra and Guilherme Bodin and Mario Souto},
   journal={arXiv preprint arXiv:1908.01757},
   year={2019}
@@ -125,7 +125,7 @@ year = {2025}}
 }
 
 @software{SSDjl2024,
-  title = {StateSpaceDynamics.jl: A Julia package for probabilistic
+  title = {StateSpaceDynamics.jl: A {J}ulia package for probabilistic
 state space models},
   author = {Senne, Ryan and Loschinskey, Zachary and Loughridge, Carson  and DePasquale, Brian},
   date = {2025},
@@ -173,9 +173,10 @@ state space models},
 
 @REPORT{Murphy1998-bk,
   type   = {techreport},
-  title  = {Switching Kalman Filters},
+  title  = {Switching {K}alman Filters},
   author = {Murphy, Kevin P},
   date   = {1998}
+  url    = {https://www.cs.ubc.ca/~murphyk/Papers/skf.pdf}
 }
 
 @MISC{Rybicki1990-ky,
@@ -183,6 +184,7 @@ state space models},
             Tridiagonal Matrix},
   author = {Rybicki, G B and Hummer, D G},
   date   = {1990}
+  url    = {https://numerical.recipes/whp/fast/diagonal.pdf}
 }
 
 @BOOK{Bishop2006-kv,
