### 2025-11-14 b30c442
Add URLs to two bibliography entries
Added URL fields to the Murphy (1998) and Rybicki & Hummer (1990) entries in paper.bib for easier reference to the original documents.
