+++ paper/paper.bib
@@ -175,7 +175,7 @@ state space models},
   type   = {techreport},
   title  = {Switching {K}alman Filters},
   author = {Murphy, Kevin P},
-  date   = {1998}
+  date   = {1998},
   url    = {https://www.cs.ubc.ca/~murphyk/Papers/skf.pdf}
 }
 
@@ -183,7 +183,7 @@ state space models},
   title  = {Fast Solution for the Diagonal Elements of the Inverse of a
             Tridiagonal Matrix},
   author = {Rybicki, G B and Hummer, D G},
-  date   = {1990}
+  date   = {1990},
   url    = {https://numerical.recipes/whp/fast/diagonal.pdf}
 }
 
